Estimating Covariance Matrix using Empirical Bayes

Estimating covariance matrix using Empirical Bayes

covestEB(X, order = -3, verbose = FALSE, force.nonbin = FALSE)

covestEB.cor(X, verbose = FALSE)

Arguments

X
a matrix of size n * p, where n is the number of observations and p is the number of variables

order
the level of binning to use when the number of observations passed to the computation is greater than 5000.

verbose
logical; If TRUE, the intermediate results will be shown.

force.nonbin
logical; If TRUE, no binning is performce by force.

Value

a list. a covariance matrix estimate of size p * p is given in mat, whether correction is done is given in correction, and the method for computing the density of sample correlation coefficients is given in method.

Details

The function covestEB performs covariance matrix estimation using Fisher transformation, while the function covestEB.cor performs covariance estimation directly on sample correlation coefficients using one-parameter normal approximation.

Covariance matrix estimation using Fisher transformation supports estimation sparsity as well as large-scale computation, while estimation on the original scale supports neither and it is for comparison only. It is recommended to perform estimation on Fisher-transformed sample correlation coefficients.

Examples

n = 100; p = 10
X = matrix(rnorm(n * p), nrow = n, ncol = p)
covestEB(X)

#> $mat
#>               [,1]        [,2]          [,3]         [,4]          [,5]
#>  [1,]  1.087845234 -0.01328512  0.0038987447  0.150505766  0.0854740735
#>  [2,] -0.013285125  0.99534837  0.0457961290  0.123665730 -0.0133924617
#>  [3,]  0.003898745  0.04579613  0.8611970716 -0.015824964 -0.0307962178
#>  [4,]  0.150505766  0.12366573 -0.0158249638  1.027849627 -0.0032617257
#>  [5,]  0.085474073 -0.01339246 -0.0307962178 -0.003261726  0.9724777423
#>  [6,] -0.087499381  0.01270599 -0.0226349333 -0.007531817 -0.0181517568
#>  [7,]  0.171873209 -0.02764924 -0.0201519483 -0.024824673 -0.0006337502
#>  [8,] -0.010105035 -0.02200756 -0.0393621149 -0.023003919 -0.0130967371
#>  [9,] -0.025779682  0.01507497 -0.0144647767 -0.056895882 -0.0176065704
#> [10,] -0.025406925 -0.01222215 -0.0004530902 -0.011215158 -0.0202731238
#>               [,6]          [,7]        [,8]        [,9]         [,10]
#>  [1,] -0.087499381  0.1718732095 -0.01010504 -0.02577968 -0.0254069253
#>  [2,]  0.012705989 -0.0276492364 -0.02200756  0.01507497 -0.0122221516
#>  [3,] -0.022634933 -0.0201519483 -0.03936211 -0.01446478 -0.0004530902
#>  [4,] -0.007531817 -0.0248246726 -0.02300392 -0.05689588 -0.0112151581
#>  [5,] -0.018151757 -0.0006337502 -0.01309674 -0.01760657 -0.0202731238
#>  [6,]  0.790847774 -0.0395497059 -0.02267945 -0.01886187  0.0088558492
#>  [7,] -0.039549706  1.2022786131 -0.02033592  0.08651158 -0.0372984058
#>  [8,] -0.022679449 -0.0203359167  1.04946512 -0.01135037 -0.0284846098
#>  [9,] -0.018861871  0.0865115802 -0.01135037  1.08899449 -0.0246072636
#> [10,]  0.008855849 -0.0372984058 -0.02848461 -0.02460726  1.1817901125
#> 
#> $correction
#> [1] FALSE
#> 
#> $correction.Fnorm
#> [1] 0
#> 
#> $mix.dist
#> $iter
#> [1] 13
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -3.956848e-06
#> 
#> $beta
#> [1] 0.1015346
#> 
#> $mix
#> $mix$pt
#> [1] -0.18431740 -0.02050932  0.00000000  0.17705766
#> 
#> $mix$pr
#> [1] 0.03319553 0.84677412 0.00000000 0.12003035
#> 
#> 
#> $ll
#> [1] -31.23088
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"
#> 

covestEB.cor(X)

#> $mat
#>               [,1]        [,2]         [,3]         [,4]         [,5]
#>  [1,]  1.087845234 -0.01409674  0.003704273  0.152725464  0.089064838
#>  [2,] -0.014096740  0.99534837  0.048149181  0.126642130 -0.014152306
#>  [3,]  0.003704273  0.04814918  0.861197072 -0.016471596 -0.031012179
#>  [4,]  0.152725464  0.12664213 -0.016471596  1.027849627 -0.003833440
#>  [5,]  0.089064838 -0.01415231 -0.031012179 -0.003833440  0.972477742
#>  [6,] -0.083101150  0.01309989 -0.022862680 -0.008198597 -0.018573681
#>  [7,]  0.173847252 -0.02798053 -0.020694369 -0.025264506 -0.001119436
#>  [8,] -0.010917858 -0.02245225 -0.039436647 -0.023423700 -0.013883406
#>  [9,] -0.026152667  0.01554803 -0.015189506 -0.056329070 -0.018301309
#> [10,] -0.025851604 -0.01306842 -0.000899408 -0.012064773 -0.020899661
#>               [,6]         [,7]        [,8]        [,9]        [,10]
#>  [1,] -0.083101150  0.173847252 -0.01091786 -0.02615267 -0.025851604
#>  [2,]  0.013099886 -0.027980526 -0.02245225  0.01554803 -0.013068424
#>  [3,] -0.022862680 -0.020694369 -0.03943665 -0.01518951 -0.000899408
#>  [4,] -0.008198597 -0.025264506 -0.02342370 -0.05632907 -0.012064773
#>  [5,] -0.018573681 -0.001119436 -0.01388341 -0.01830131 -0.020899661
#>  [6,]  0.790847774 -0.039651140 -0.02296227 -0.01932837  0.008962791
#>  [7,] -0.039651140  1.202278613 -0.02104114  0.09039407 -0.037601485
#>  [8,] -0.022962270 -0.021041140  1.04946512 -0.01217951 -0.028815980
#>  [9,] -0.019328366  0.090394068 -0.01217951  1.08899449 -0.025091510
#> [10,]  0.008962791 -0.037601485 -0.02881598 -0.02509151  1.181790112
#> 
#> $correction
#> [1] FALSE
#> 
#> $correction.Fnorm
#> [1] 0
#> 
#> $mix.dist
#> $iter
#> [1] 13
#> 
#> $family
#> [1] "npnormc"
#> 
#> $min.gradient
#> [1] -4.14457e-06
#> 
#> $beta
#> [1] 100
#> 
#> $mix
#> $mix$pt
#> [1] -0.18171703 -0.02066678  0.00000000  0.17771444
#> 
#> $mix$pr
#> [1] 0.03195825 0.84930197 0.00000000 0.11873978
#> 
#> 
#> $ll
#> [1] -31.89158
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"
#>