Computing non-parametric mixing distribution with estimated proportion at 0

computing non-parametric mixing distribution with estimated proportion at 0

estpi0(v, method = "npnormll", ...)

estpi0.npnormll(v, beta, val, order, mix = NULL, ...)

estpi0.npnormcvm(v, beta, val, order, mix = NULL, ...)

estpi0.npnormad(v, beta, val, order, mix = NULL, ...)

estpi0.npnormcll(v, beta, val, order, mix = NULL, ...)

estpi0.nptll(v, beta, val, order, mix = NULL, ...)

estpi0.npnormllw(v, beta, val, order, mix = NULL, ...)

estpi0.npnormcvmw(v, beta, val, order, mix = NULL, ...)

estpi0.npnormadw(v, beta, val, order, mix = NULL, ...)

Arguments

v
observations

method
An implemented family; see details

...
parameters above passed to the specific method.

beta
structural parameter

val
thresholding function

order
the parameter for the binned version.

mix
initial mixing distribution.

Details

This is a function for computing non-parametric mixing distribution with estimated proportion at 0. Different families will have different threshold values.

The parameters are listed as follows:

tol: tolerance to stop the code.

verbose: logical; Whether to print the intermediate results.

It is not shown in the parameter section since various method have different default threshold values and this function essentially calls the class method in the object.

The full list of implemented families is in computemixdist.

IMPORTANT: although estpi0 is implemented for nptll and npnormcll, the theory is not yet done, and hence it is recommand to use with precaution.

Examples

data = rnorm(500, c(0, 2))
system.time(r <- estpi0(data, method = "npnormll", val = 2, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.519666 fval: 884.567
#> Iter: 2 lower: 0.519666 upper: 0.759833
#> current val: 0.69535 fval: 903.146
#> Iter: 3 lower: 0.519666 upper: 0.607508
#> current val: 0.530424 fval: 884.909
#> Iter: 4 lower: 0.525045 upper: 0.530424
#> current val: 0.527959 fval: 884.824

#>    user  system elapsed 
#>    0.04    0.00    0.04 

r

#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -1.553542e-05
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] -1.872467  0.000000  2.029319
#> 
#> $mix$pr
#> [1] 0.004690674 0.527961530 0.467347796
#> 
#> 
#> $ll
#> [1] 884.8242
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormllw", val = 2, verbose = TRUE, order = -2))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.519929 fval: 884.598
#> Iter: 2 lower: 0.519929 upper: 0.759964
#> current val: 0.695332 fval: 903.18
#> Iter: 3 lower: 0.519929 upper: 0.607631
#> current val: 0.529775 fval: 884.911
#> Iter: 4 lower: 0.524852 upper: 0.529775
#> current val: 0.527456 fval: 884.832

#>    user  system elapsed 
#>   0.196   0.001   0.197 

r

#> $iter
#> [1] 8
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -3.535435e-06
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] -1.855625  0.000000  2.028591
#> 
#> $mix$pr
#> [1] 0.004814976 0.527458605 0.467726419
#> 
#> 
#> $ll
#> [1] 884.832
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormcvm", val = 0.1, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.511595 fval: 0.107162
#> Iter: 2 lower: 0.255797 upper: 0.511595
#> current val: 0.47515 fval: 0.0598808
#> Iter: 3 lower: 0.493373 upper: 0.511595
#> current val: 0.509688 fval: 0.0598825
#> Iter: 4 lower: 0.510641 upper: 0.511595
#> current val: 0.511495 fval: 0.0600272
#> Iter: 5 lower: 0.511545 upper: 0.511595
#> current val: 0.511589 fval: 0.0600412
#> Iter: 6 lower: 0.511592 upper: 0.511595
#> current val: 0.511594 fval: 0.0600414

#>    user  system elapsed 
#>   0.023   0.000   0.024 

r

#> $iter
#> [1] 2
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] 1.110223e-16
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.041728
#> 
#> $mix$pr
#> [1] 0.5115946 0.4884054
#> 
#> 
#> $ll
#> [1] 0.06004139
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormcvmw", val = 0.1, verbose = TRUE, order = -2))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.510792 fval: 0.106982
#> Iter: 2 lower: 0.255396 upper: 0.510792
#> current val: 0.476423 fval: 0.0590176
#> Iter: 3 lower: 0.493608 upper: 0.510792
#> current val: 0.509067 fval: 0.0590995
#> Iter: 4 lower: 0.50993 upper: 0.510792
#> current val: 0.510705 fval: 0.0593779
#> Iter: 5 lower: 0.510749 upper: 0.510792
#> current val: 0.510788 fval: 0.0593971
#> Iter: 6 lower: 0.51079 upper: 0.510792
#> current val: 0.510792 fval: 0.0593979

#>    user  system elapsed 
#>   0.105   0.001   0.107 

r

#> $iter
#> [1] 2
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] 0
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.046075
#> 
#> $mix$pr
#> [1] 0.5107923 0.4892077
#> 
#> 
#> $ll
#> [1] 0.05939799
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormad", val = 1, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.521417 fval: 38.8876
#> Iter: 2 lower: 0 upper: 0.260709
#> current val: 0.00211049 fval: 61.1937
#> Iter: 3 lower: 0.00105525 upper: 0.00211049
#> current val: 0.00107291 fval: 0.480842
#> Iter: 4 lower: 0.0015917 upper: 0.00211049
#> current val: 0.00160042 fval: 0.482417
#> Iter: 5 lower: 0.00185546 upper: 0.00211049
#> current val: 0.00185972 fval: 0.487528
#> Iter: 6 lower: 0.0019851 upper: 0.00211049
#> current val: 0.00198717 fval: 0.47724
#> Iter: 7 lower: 0.00204883 upper: 0.00211049
#> current val: 0.00204988 fval: 0.475913
#> Iter: 8 lower: 0.00208018 upper: 0.00211049
#> current val: 0.0020807 fval: 0.479419
#> Iter: 9 lower: 0.0020956 upper: 0.00211049
#> current val: 0.00209585 fval: 0.475109
#> Iter: 10 lower: 0.00210317 upper: 0.00211049
#> current val: 0.00210329 fval: 0.475392
#> Iter: 11 lower: 0.00210689 upper: 0.00211049
#> current val: 0.00210695 fval: 0.479343
#> Iter: 12 lower: 0.00210872 upper: 0.00211049
#> current val: 0.00210875 fval: 0.475078

#>    user  system elapsed 
#>   0.056   0.001   0.057 

r

#> $iter
#> [1] 1
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -0.9985344
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 0.000000 2.033374
#> 
#> $mix$pr
#> [1] 0.514002869 0.002109638 0.483887493
#> 
#> 
#> $ll
#> [1] 0.476785
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormadw", val = 1, verbose = TRUE, order = -2))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.521331 fval: 40.0474
#> Iter: 2 lower: 0 upper: 0.260665
#> current val: 0.00206196 fval: 62.5153
#> Iter: 3 lower: 0 upper: 0.00103098
#> current val: 9.09042e-06 fval: 0.464155
#> Iter: 4 lower: 0.000520035 upper: 0.00103098
#> current val: 0.000528701 fval: 0.46424
#> Iter: 5 lower: 0.00077984 upper: 0.00103098
#> current val: 0.000784097 fval: 0.464895
#> Iter: 6 lower: 0.000907538 upper: 0.00103098
#> current val: 0.000909626 fval: 0.466445
#> Iter: 7 lower: 0.000970302 upper: 0.00103098
#> current val: 0.00097133 fval: 0.465788
#> Iter: 8 lower: 0.00100115 upper: 0.00103098
#> current val: 0.00100166 fval: 0.468141
#> Iter: 9 lower: 0.00101632 upper: 0.00103098
#> current val: 0.00101656 fval: 0.471575
#> Iter: 10 lower: 0.00102377 upper: 0.00103098
#> current val: 0.00102389 fval: 0.476092
#> Iter: 11 lower: 0.00102744 upper: 0.00103098
#> current val: 0.00102749 fval: 0.481687
#> Iter: 12 lower: 0.00102924 upper: 0.00103098
#> current val: 0.00102926 fval: 0.488353

#>    user  system elapsed 
#>   0.025   0.000   0.026 

r

#> $iter
#> [1] 1
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -4.330511
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 0.000000 2.034759
#> 
#> $mix$pr
#> [1] 0.519056713 0.001030135 0.479913152
#> 
#> 
#> $ll
#> [1] 0.4960798
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "nptll", val = 2, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.519729 fval: 884.569
#> Iter: 2 lower: 0.519729 upper: 0.759864
#> current val: 0.695367 fval: 903.149
#> Iter: 3 lower: 0.519729 upper: 0.607548
#> current val: 0.530408 fval: 884.908
#> Iter: 4 lower: 0.525068 upper: 0.530408
#> current val: 0.527959 fval: 884.824

#>    user  system elapsed 
#>   0.118   0.001   0.119 

r

#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npt"
#> 
#> $min.gradient
#> [1] -1.532229e-05
#> 
#> $beta
#> [1] Inf
#> 
#> $mix
#> $mix$pt
#> [1] -1.872464  0.000000  2.029319
#> 
#> $mix$pr
#> [1] 0.004690676 0.527961677 0.467347647
#> 
#> 
#> $ll
#> [1] 884.8242
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(tanh(data), method = "npnormcll", val = 2, beta = 4, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.310889 fval: 42.5774
#> Iter: 2 lower: 0.310889 upper: 0.655445
#> current val: 0.5788 fval: 80.6923
#> Iter: 3 lower: 0.310889 upper: 0.444845
#> current val: 0.37015 fval: 44.5307
#> Iter: 4 lower: 0.340519 upper: 0.37015
#> current val: 0.370124 fval: 44.5283
#> Iter: 5 lower: 0.370124 upper: 0.370137
#> current val: 0.370135 fval: 44.529
#> Iter: 6 lower: 0.370129 upper: 0.370135
#> current val: 0.370134 fval: 44.529
#> Iter: 7 lower: 0.370132 upper: 0.370134
#> current val: 0.370134 fval: 44.529
#> Iter: 8 lower: 0.370133 upper: 0.370134
#> current val: 0.370134 fval: 44.529

#>    user  system elapsed 
#>   0.215   0.002   0.219 

r

#> $iter
#> [1] 4
#> 
#> $family
#> [1] "npnormc"
#> 
#> $min.gradient
#> [1] -2.115578e-06
#> 
#> $beta
#> [1] 4
#> 
#> $mix
#> $mix$pt
#>  [1] -0.9982257 -0.9239238 -0.8118828  0.0000000  0.8282864  0.9176979
#>  [7]  0.9722182  0.9858447  0.9904687  0.9976710  0.9990725
#> 
#> $mix$pr
#>  [1] 0.000513808 0.049375818 0.050063871 0.370134105 0.136139747 0.123406161
#>  [7] 0.121037903 0.002807918 0.079962184 0.064608446 0.001950040
#> 
#> 
#> $ll
#> [1] 44.52898
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"