Find the posterior mean

Find the posterior mean given the observations and the mixing distribution based on the family in the result

posteriormean(x, result, fun = function(x) x)

posteriormean.npnorm(x, result, fun = function(x) x)

posteriormean.npt(x, result, fun = function(x) x)

posteriormean.npnormc(x, result, fun = function(x) x)

Arguments

x
a vector of observations

result
an object of class nspmix

fun
the function to transform the mean. It finds the posterior mean of fun(x). The function fun must be vectorised.

Examples

data = rnorm(500, c(0, 2))
r1 = computemixdist(data, pi0 = 0.5)
posteriormean(data, r1)

#>   [1] 0.52334114 1.62947189 1.62598239 2.08145705 0.07356137 2.06713115
#>   [7] 1.15289892 1.24380358 1.44558215 2.08981688 0.55483761 2.01915716
#>  [13] 0.05360459 2.10845214 0.09061258 1.50533571 0.47154803 1.80509202
#>  [19] 1.54052908 1.15659444 0.09785373 1.87264563 0.50809695 2.46058101
#>  [25] 0.03953543 2.04465673 0.78188255 2.04328280 0.04076686 1.90729464
#>  [31] 0.10376423 0.31745909 0.06099066 0.84556592 1.50391493 1.18386448
#>  [37] 0.78981625 1.97067361 0.06960984 1.95541427 0.50290346 0.74265246
#>  [43] 0.63467220 1.74946149 0.15071460 1.00387285 0.16631552 0.80615025
#>  [49] 0.20918777 1.67311314 0.14939769 2.04435990 0.08753974 1.67339101
#>  [55] 0.75779646 2.06256976 0.18495976 1.78663325 0.98811747 1.89280471
#>  [61] 0.36125975 1.58426443 0.25319813 1.75639658 0.51713641 1.66490097
#>  [67] 0.13625361 1.70929149 0.10927319 2.08030087 0.05083247 2.05348991
#>  [73] 0.21359461 0.21029560 0.09689262 2.04838404 0.03981283 1.81872405
#>  [79] 0.04249195 1.79444915 0.91229069 1.62405965 0.84652097 1.99432955
#>  [85] 1.10021522 2.06351028 0.88914134 1.91555229 0.23894701 1.82104288
#>  [91] 0.37881604 0.98851178 0.09053432 1.38587334 0.09971970 1.56802688
#>  [97] 0.19929697 1.09590479 0.01899066 2.12695462 0.06773766 2.03487683
#> [103] 0.08771247 1.94595397 0.17598500 1.96970204 0.02473635 1.86292305
#> [109] 0.17268040 2.06782991 0.13149084 1.48245278 0.02492147 1.83279927
#> [115] 0.54925064 1.60303068 0.09190149 1.90616120 0.04389693 1.89105660
#> [121] 0.90101452 1.65219491 0.12099326 1.51467222 0.45429827 0.63630000
#> [127] 0.06559833 0.46828875 0.14464707 1.73122028 0.24789061 1.64982948
#> [133] 0.02510268 0.63285057 0.13854532 1.87727091 0.29004032 0.84096541
#> [139] 0.91025447 1.77653892 0.12808056 1.87464777 0.74755447 1.96984774
#> [145] 0.11965732 0.71662836 0.75694739 0.84443488 0.08002414 2.04739621
#> [151] 0.09930614 1.58743640 0.44584821 2.06370911 0.05538304 1.99676622
#> [157] 0.20769484 2.07045668 0.85002887 1.72623529 0.07323430 0.61514372
#> [163] 0.05376770 1.43583425 0.04339047 1.66613257 0.09812789 2.12173591
#> [169] 0.98873713 1.94811340 0.09508673 2.12686306 0.14213805 1.73176025
#> [175] 0.02331212 2.04247111 0.19952069 0.87341593 0.21449023 2.04072477
#> [181] 0.24620005 1.88713256 0.94840890 0.97187247 0.14850794 1.98741413
#> [187] 0.03609380 2.05213528 0.04982600 1.87467074 0.28532280 1.29551127
#> [193] 0.30297104 1.63315743 0.78600275 0.57715780 1.18930073 1.58963222
#> [199] 0.07446291 0.82718304 0.96868474 0.43829564 0.26854578 1.29370731
#> [205] 0.07195779 1.65465241 0.23765693 1.49834812 0.07390080 1.99434592
#> [211] 0.58221123 2.08840034 0.05508991 0.88070876 0.06784487 0.74741208
#> [217] 1.74366875 0.10178030 0.03553867 1.13187767 0.15066849 1.66317105
#> [223] 0.90342304 2.05928998 0.05188025 1.85799961 0.02800980 2.01907670
#> [229] 0.06308301 1.93136599 0.05469741 1.76159633 0.21459071 0.81772664
#> [235] 0.38819232 2.78763084 0.22397798 1.45847737 0.10097955 1.50538269
#> [241] 0.28161536 1.01764842 1.40371012 1.96991823 0.81406075 2.06754225
#> [247] 0.04920151 0.95171263 0.35946261 1.40232941 0.47126812 1.71243320
#> [253] 0.61864335 0.39600069 0.61600453 0.27546406 0.21101689 1.93835173
#> [259] 0.06959755 2.07661047 0.30379303 0.39441426 1.43951407 1.88963264
#> [265] 0.14266958 1.99364199 0.04154425 1.75336478 0.33553224 0.05961604
#> [271] 0.11723904 1.20097969 0.37440592 2.04431040 0.27258775 1.08841672
#> [277] 0.19289515 0.87863407 0.05385639 0.10265975 0.14593774 1.96675821
#> [283] 0.32960960 0.84109491 0.10785043 2.04600090 0.64671757 1.95572979
#> [289] 0.13347170 1.72255793 0.04788545 1.77320851 0.40294015 1.78136886
#> [295] 0.30217670 0.11816293 0.33723468 1.53230671 0.04107483 1.55267415
#> [301] 0.29092727 0.93312044 0.22431218 0.71864891 0.04961759 2.10331564
#> [307] 0.67670877 0.41270209 0.15111743 1.81810652 0.13619770 2.04253528
#> [313] 1.87097360 1.31016601 0.74013582 2.04480657 0.40388749 1.93831151
#> [319] 1.11720156 1.81341897 0.02253117 1.93203334 0.04001587 1.27823622
#> [325] 0.15069465 0.17998107 0.16393277 2.01848621 0.08710400 0.47287205
#> [331] 0.13731726 1.31546207 0.34997965 1.26910868 0.07848341 1.90531636
#> [337] 0.46091080 2.04735807 0.95404205 1.23575730 0.79704661 0.59354968
#> [343] 0.27784043 0.22136225 0.40301757 0.81709254 0.19377628 1.87662251
#> [349] 0.03281806 2.01991056 0.98543760 1.87783820 0.04218148 0.47411454
#> [355] 0.03271230 1.43225701 0.18425377 1.90384689 0.69581506 1.36708933
#> [361] 0.99009943 1.49773274 0.03279236 1.99161526 0.50333815 1.81203805
#> [367] 0.45780482 2.07115493 0.13172503 2.03353988 0.09609415 2.06763295
#> [373] 0.16846260 1.60433161 0.55027061 1.97096242 1.32925620 1.93821388
#> [379] 0.89648780 1.98624460 0.10610443 2.05816693 0.05641394 2.03480024
#> [385] 0.06150957 2.02998922 0.18278094 1.77121740 1.71315231 1.13098369
#> [391] 0.93601406 1.21239563 0.22556448 1.67916853 1.71908354 2.05857626
#> [397] 0.24382355 1.81952234 0.24622065 0.55137924 0.02809845 1.63668619
#> [403] 0.15865256 1.16503604 0.06172109 1.72258171 1.61627240 1.51315677
#> [409] 1.01672112 1.59439317 0.34926051 2.15173223 0.34161514 1.24282782
#> [415] 0.02759829 1.37206670 0.88662612 1.41533112 0.71340293 1.59878961
#> [421] 0.03584842 0.25479918 0.05375008 1.75649126 0.10362198 1.10355855
#> [427] 0.24925139 0.79856804 0.85778650 1.29206432 0.16629353 0.98532760
#> [433] 0.49661122 1.91072321 0.11152038 0.22756250 0.27780416 1.92599432
#> [439] 0.45764278 2.03783506 0.06745763 1.15305028 0.04035883 2.01442457
#> [445] 0.10921708 1.09291307 0.20217893 1.63403482 0.08564869 2.12026534
#> [451] 0.07458504 2.04701834 0.53494598 2.10353614 0.25715696 1.75736460
#> [457] 0.08788554 0.69802390 0.13737596 1.29663470 0.10086576 1.91011230
#> [463] 0.11116533 0.03547352 0.15776576 1.84301640 0.19471222 1.84830443
#> [469] 0.11066297 2.02682300 0.13002734 1.42126660 0.81533665 2.05173639
#> [475] 0.11575941 1.67706414 0.07757658 1.81751951 1.58689796 2.01701449
#> [481] 0.56997144 1.93731502 0.15714693 0.36246001 0.65500999 1.82429223
#> [487] 0.10047334 2.06178632 0.74094479 2.05325790 0.65363479 1.12904446
#> [493] 0.14725395 0.24921821 0.15962915 0.30562341 0.25809399 1.82848608
#> [499] 0.08763675 1.73952136

r2 = computemixdist(data, pi0 = 0.5, method = "nptll") # equivalent to normal
posteriormean(data, r2, fun = function(x) x^2)

#>   [1] 1.01265268 3.35413831 3.34665157 4.41131054 0.09973356 4.35145671
#>   [7] 2.33789921 2.53090930 2.96065738 4.45154800 1.07833030 4.21020079
#>  [13] 0.06389839 4.55208330 0.13162723 3.08830033 0.90487057 3.73243731
#>  [19] 3.16357030 2.34573851 0.14540570 3.87924944 0.98089978 6.79441326
#>  [25] 0.04038537 4.27782485 1.55410097 4.27388116 0.04235400 3.95514043
#>  [31] 0.15673393 0.58633517 0.07691027 1.68814593 3.08526312 2.40360496
#>  [37] 1.57078784 4.09625769 0.09248274 4.06184974 0.97008746 1.47164368
#>  [43] 1.24519452 3.61221807 0.24855256 2.02227787 0.27956812 1.60515436
#>  [49] 0.36563010 3.44785090 0.24594364 4.27696886 0.12581770 3.44844809
#>  [55] 1.50346369 4.33478254 0.31686308 3.69249432 1.98896885 3.92333596
#>  [61] 0.67648353 3.25721233 0.45488797 3.62718061 0.99972569 3.43020473
#>  [67] 0.21999320 3.52566293 0.16735073 4.40604294 0.05911439 4.30445511
#>  [73] 0.37453182 0.36786706 0.14357024 4.28877409 0.04082697 3.76197578
#>  [79] 0.04514603 3.70940003 1.82882683 3.34252672 1.69015788 4.15062527
#>  [85] 2.22620695 4.33813460 1.77999324 3.97331859 0.42590376 3.76700403
#>  [91] 0.71271753 1.98980234 0.13147897 2.83329120 0.14897464 3.22243209
#>  [97] 0.34568461 2.21707405 0.01244955 4.66081179 0.08907014 4.25065500
#> [103] 0.12614363 4.04069173 0.29888220 4.09405458 0.01899571 3.85804201
#> [109] 0.29227436 4.35410686 0.21063377 3.03939646 0.01922789 3.79251516
#> [115] 1.06667333 3.29743047 0.13407100 3.95264851 0.04744732 3.91950608
#> [121] 1.80503634 3.40291328 0.19009910 3.10826198 0.86904044 1.24860215
#> [127] 0.08519000 0.89809790 0.23654525 3.57289073 0.44408520 3.39783400
#> [133] 0.01945631 1.24138139 0.22450540 3.88935015 0.53011496 1.67845494
#> [139] 1.82453034 3.67067485 0.20394796 3.88362074 1.48194206 4.09438484
#> [145] 0.18749599 1.41699620 1.50167928 1.68576329 0.11171691 4.28583459
#> [151] 0.14818303 3.26400855 0.85150174 4.33884881 0.06699829 4.15632729
#> [157] 0.36261647 4.36430825 1.69754822 3.56215002 0.09913107 1.20432962
#> [163] 0.06418174 2.93985191 0.04661507 3.43285081 0.14592961 4.62956543
#> [169] 1.99027867 4.04551143 0.14012681 4.66026056 0.23159007 3.57405433
#> [175] 0.01725034 4.27157287 0.34613522 1.74683610 0.37634204 4.26665795
#> [181] 0.44064630 3.91091409 1.90507171 1.95463696 0.24418184 4.13456289
#> [187] 0.03500430 4.30021020 0.05739306 3.88367090 0.52046087 2.64085849
#> [193] 0.55660596 3.36204664 1.56276647 1.12492844 2.41514448 3.26871369
#> [199] 0.10139636 1.64942924 1.94790153 0.83583350 0.48617734 2.63702062
#> [205] 0.09678358 3.40819073 0.42328355 3.07336391 0.10035925 4.15066351
#> [211] 1.13548459 4.44447714 0.06648580 1.76221148 0.08926514 1.48164289
#> [217] 3.59972480 0.15292389 0.03415472 2.29331832 0.24846118 3.42648823
#> [223] 1.81011730 4.32342057 0.06091542 3.84731459 0.02326531 4.21000087
#> [229] 0.08065633 4.00826266 0.06580052 3.63840327 0.37654514 1.62952007
#> [235] 0.73209008 8.90956984 0.39554005 2.98818777 0.15138823 3.08840079
#> [241] 0.51287802 2.05141081 2.87132102 4.09454465 1.62180338 4.35301267
#> [247] 0.05632945 1.91204901 0.67277750 2.86837665 0.90428889 3.53242608
#> [253] 1.21165070 0.74823378 1.20613031 0.50030507 0.36932380 4.02376543
#> [259] 0.09246030 4.38974083 0.55829134 0.74495310 2.94770538 3.91638742
#> [265] 0.23263930 4.14902056 0.04360743 3.62063870 0.62348598 0.07446258
#> [271] 0.18279020 2.43993975 0.70361064 4.27682632 0.49442971 2.20121037
#> [277] 0.33280136 1.75783716 0.06433588 0.15461191 0.23909660 4.08739040
#> [283] 0.61130394 1.67872772 0.16460376 4.28172969 1.27041488 4.06255742
#> [289] 0.21452334 3.55422851 0.05409963 3.66347937 0.76258890 3.68111306
#> [295] 0.55497740 0.18458704 0.62698897 3.14597829 0.04284960 3.18956237
#> [301] 0.53193068 1.87279021 0.39621697 1.42123757 0.05703771 4.52319648
#> [307] 1.33325578 0.78279478 0.24935090 3.76063689 0.21988319 4.27175479
#> [313] 3.87559999 2.67204149 1.46635716 4.27825780 0.76454915 4.02367603
#> [319] 2.26220547 3.75047615 0.01632549 4.00974173 0.04115091 2.60411228
#> [325] 0.24851303 0.30688214 0.27481885 4.20853552 0.12499586 0.90762221
#> [331] 0.22208677 2.68331306 0.65323160 2.58470222 0.10884717 3.95079158
#> [337] 0.88277137 4.28572166 1.91696895 2.51381065 1.58599853 1.15917743
#> [343] 0.50516100 0.39024339 0.76274910 1.62818526 0.33457327 3.88793367
#> [349] 0.03007474 4.21207555 1.98330440 3.89058954 0.04464073 0.91020457
#> [355] 0.02991915 2.93221800 0.31544686 3.94756269 1.37332224 2.79325816
#> [361] 1.99315835 3.07204862 0.03003691 4.14430070 0.97099231 3.74748371
#> [367] 0.87632113 4.36708476 0.21109337 4.24708348 0.14204686 4.35335719
#> [373] 0.28385118 3.30021937 1.06880124 4.09691295 2.71267697 4.02345904
#> [379] 1.79548767 4.13186228 0.16123744 4.31964117 0.06880537 4.25044959
#> [385] 0.07783707 4.23773376 0.31249330 3.65917828 3.53397424 2.29142284
#> [391] 1.87889919 2.46418241 0.39875397 3.46086636 3.54674560 4.32101239
#> [397] 0.43581377 3.76370669 0.44068821 1.07111423 0.02338546 3.36961952
#> [403] 0.26430959 2.36364792 0.07821529 3.55427975 3.32582342 3.10502158
#> [409] 2.04944945 3.27891640 0.65175002 4.81286148 0.63600502 2.52883562
#> [415] 0.02271047 2.80386446 1.77468903 2.89610643 1.41022619 3.28833930
#> [421] 0.03462809 0.45814854 0.06415113 3.62738494 0.15646049 2.23329133
#> [427] 0.44685400 1.58919959 1.71389410 2.63352534 0.27952427 1.98307189
#> [433] 0.95699145 3.96268286 0.17169623 0.40280299 0.50508687 3.99637111
#> [439] 0.87598465 4.25866808 0.08856103 2.33822028 0.04169937 4.19853646
#> [445] 0.16724232 2.21073572 0.35149129 3.36392947 0.12225466 4.62083045
#> [451] 0.10162185 4.28471757 1.03684052 4.52442114 0.46295174 3.62926963
#> [457] 0.12647028 1.37795585 0.22220233 2.64324864 0.15117011 3.96133842
#> [463] 0.17100911 0.03405537 0.26254671 3.81471249 0.33645588 3.82621160
#> [469] 0.17003725 4.22954523 0.20776293 2.90876822 1.62448906 4.29897236
#> [475] 0.17991508 3.45634283 0.10716177 3.75936425 3.26285485 4.20489662
#> [481] 1.10992059 4.02146183 0.26131690 0.67895903 1.28778362 3.77405198
#> [487] 0.15041807 4.33202301 1.46805647 4.30372347 1.28490288 2.28731127
#> [493] 0.24170001 0.44678648 0.26625168 0.56204491 0.46486112 3.78315201
#> [499] 0.12600075 3.59078261

data = runif(500, min = -0.5, max = 0.5)
r3 = computemixdist(data, method = "npnormcll", beta = 100)
posteriormean(data, r3)

#>   [1] -0.3613611487 -0.3604237292  0.1777899555 -0.2034152216  0.3683695359
#>   [6]  0.3032663682  0.2420179134 -0.0758284624  0.3671256549  0.3692651182
#>  [11] -0.0717481644 -0.0729118804 -0.0775773246  0.2961812027  0.3485953591
#>  [16]  0.3582558765  0.2182582492 -0.3617064946  0.2935154164  0.3252151607
#>  [21] -0.0711287185 -0.3554980208 -0.0633581894 -0.3374863707  0.0960933121
#>  [26]  0.0661425438 -0.0426370074  0.3402833333  0.2911383386 -0.3614911095
#>  [31] -0.0876203986  0.3505755648 -0.3460923502 -0.3606537966  0.3634017329
#>  [36] -0.3583212316 -0.2975420537  0.3445711754  0.1255653765 -0.1739611294
#>  [41] -0.3615615324  0.2901159400 -0.2385816879  0.3304391093  0.3698884344
#>  [46]  0.3465823663 -0.0221391920 -0.0821315712  0.2972591254  0.1676119334
#>  [51]  0.2275953990 -0.0538285306 -0.3609155764  0.2925932553  0.3071313208
#>  [56] -0.0718505457 -0.3542196107  0.3685539359  0.2707854883  0.3660018383
#>  [61] -0.3593538327  0.3500777658 -0.0609727376  0.3543707294  0.2244432616
#>  [66]  0.3696889287 -0.0818940749  0.1562442554 -0.0716379133 -0.1669545912
#>  [71]  0.2348729359 -0.3606624969  0.2727237616  0.2598794396  0.3641932518
#>  [76] -0.0728213530  0.3702042828 -0.0025996751 -0.0697351382  0.1724883714
#>  [81]  0.3589578411 -0.3456685628  0.2654355345  0.2121116932 -0.0323674709
#>  [86] -0.0736516355 -0.0640466329  0.3344858476 -0.3527824387  0.1028753406
#>  [91]  0.3312506335 -0.0804201118  0.3374711584 -0.0357816038 -0.1348579862
#>  [96]  0.3527135852 -0.3592945308 -0.1251299447 -0.0469418738  0.2417532103
#> [101]  0.3459987883 -0.2361304360  0.3672030235  0.3698475739 -0.1710893534
#> [106]  0.3498918171 -0.0702279560 -0.0772714588  0.3669401447  0.3388622885
#> [111] -0.0949149403 -0.3372330561 -0.3536012284  0.2537659197  0.1632247466
#> [116] -0.3616237958  0.3698278327 -0.3555076821 -0.3041534618  0.2095027147
#> [121] -0.1197411784  0.3685727613 -0.0691007350  0.3388309160 -0.0368853397
#> [126]  0.1254372977 -0.0916437207  0.3085156944  0.2112213525 -0.3453087241
#> [131] -0.2477131721  0.2390579648 -0.0710945464 -0.3617062773  0.2610336198
#> [136]  0.2204296957  0.0018537445  0.0009598038  0.2646827864 -0.0607210333
#> [141] -0.3608018707 -0.1023494062 -0.3304107493 -0.0774874127 -0.0721416244
#> [146]  0.2228229010 -0.3612969203 -0.3615443850  0.0463311702 -0.0778211264
#> [151]  0.3197254187  0.2866430304 -0.0709012172 -0.3597895507 -0.3613894258
#> [156] -0.3273322275 -0.0762144120  0.0885342001 -0.3602954856 -0.3350807522
#> [161] -0.0771348599 -0.1135339270 -0.3608436935 -0.0693191469 -0.3616264154
#> [166]  0.0234233185  0.3054526198 -0.3586444164  0.1038384848 -0.0562886788
#> [171] -0.3587124313 -0.0537864539 -0.3608874204 -0.0750181291 -0.3596356049
#> [176]  0.3630072269 -0.1570162459 -0.0713272584  0.3524840425 -0.0688644907
#> [181]  0.1779471294  0.3156006026 -0.3498686803  0.2075655583  0.3292589497
#> [186]  0.3684938502 -0.3608045838 -0.0719133858  0.3665603391 -0.3604557859
#> [191] -0.0763421867 -0.3610639775  0.3622626353  0.1659640511 -0.3539605177
#> [196]  0.3669139356 -0.0607891831 -0.3568175297 -0.1398436197 -0.0756590002
#> [201]  0.1589657782 -0.0745101520 -0.3509338791  0.3534494685 -0.3595947523
#> [206] -0.3550373867 -0.3491787878  0.3686608577  0.3674252194 -0.3560677145
#> [211]  0.0879275807 -0.2221814762  0.3364050420 -0.0976540436 -0.3084513386
#> [216]  0.3274096632 -0.0696399655 -0.0768170885 -0.0681614126 -0.3615377312
#> [221] -0.0777502180  0.3510246948 -0.0713175922 -0.0501257912  0.1308805451
#> [226]  0.2429113487 -0.3195883188  0.3616784082  0.2857864804 -0.0777385560
#> [231]  0.3676625711 -0.3600557202 -0.0649856796 -0.3615790311  0.2218833393
#> [236] -0.0792057228  0.0629959913 -0.3585883327 -0.0618024160  0.1671210553
#> [241]  0.2025811928 -0.3570439638 -0.0646119107 -0.3607942354 -0.3386706184
#> [246]  0.3471111284 -0.0678888370 -0.0677470278  0.3534581997  0.3644568040
#> [251]  0.2341335565 -0.0703918818  0.1845239771 -0.3227688946 -0.0698105394
#> [256] -0.0349262059  0.2299767486  0.2292739479 -0.3616229815  0.0369649097
#> [261] -0.0659196194  0.1187744965 -0.3611450485 -0.3515435988 -0.0955447417
#> [266] -0.3497330628 -0.3608416857 -0.0683547600  0.3562336754 -0.3347862994
#> [271] -0.2699743059  0.2085606802 -0.1275176592 -0.0686262428  0.3545308668
#> [276] -0.0724352346  0.2964915202 -0.0698589747 -0.0838332496 -0.3253090772
#> [281] -0.3616203402  0.3371127992  0.3020999034 -0.0363462107 -0.0885736779
#> [286] -0.0786641725 -0.2804826148  0.3328208892 -0.1586694137  0.0442563106
#> [291] -0.3577731077  0.3631465264 -0.3609126201 -0.3616209024 -0.3537122475
#> [296] -0.3579364270  0.3427067771  0.2106479771 -0.3553352218 -0.1572572087
#> [301] -0.0816594028  0.2544232201  0.2230088351  0.1831162529  0.2019016284
#> [306] -0.2689594616 -0.3615981959  0.3674360346  0.3629936158  0.2995012742
#> [311]  0.0360455086 -0.3032149099  0.3325034308 -0.2941283807 -0.3479717380
#> [316]  0.1486376588 -0.1091191555 -0.3615892264 -0.0773424223 -0.0859526540
#> [321]  0.2548107207 -0.0970003062 -0.3595506197 -0.0777193220 -0.1226799108
#> [326]  0.3700432928  0.2478830329 -0.0911876901  0.3438295905  0.3132465189
#> [331] -0.3442242010 -0.3613441281  0.3617092713 -0.3613041224 -0.0709787929
#> [336]  0.3532727853 -0.0737991350 -0.0554786065  0.2603232562 -0.0790004292
#> [341]  0.3107445026 -0.2370519451 -0.3068204906 -0.0932129034  0.3082696518
#> [346] -0.3280910931 -0.3491104939 -0.1142455737  0.3024586690  0.2113975550
#> [351] -0.1526295349 -0.3239610202 -0.3177356068  0.3541045940  0.3400838826
#> [356] -0.1255900122  0.2090184231 -0.3615508869 -0.3595126403 -0.0702339331
#> [361]  0.3094856833  0.2162021705 -0.0772546400  0.3061405399  0.2337656098
#> [366] -0.3614335710  0.2053073865 -0.0905060112 -0.3509045624  0.3528168130
#> [371] -0.0698768460  0.3082672639 -0.0596033162  0.2677158228 -0.1040319680
#> [376] -0.1284980228 -0.3616129852 -0.2997916939 -0.1382051558 -0.0187246648
#> [381]  0.2143773106  0.2221300450  0.3274229910 -0.0852857619  0.3484069973
#> [386] -0.0709590250  0.2710179907  0.3217958200 -0.0741801618  0.2780884930
#> [391] -0.0842511251 -0.0734848035 -0.1895268953 -0.2793410996  0.3228342294
#> [396]  0.2934336633 -0.2073798204 -0.0549390854  0.0571382875 -0.0752877326
#> [401]  0.3209545046  0.2502076532  0.1684629706 -0.0679942735  0.0432097561
#> [406]  0.3676099282 -0.0995047141 -0.3600484735 -0.3587438364 -0.3606826503
#> [411] -0.0778170073 -0.3604197563 -0.3361817805  0.3679894315  0.0069835145
#> [416] -0.1336906155 -0.0465631802  0.1536175550 -0.0830795748 -0.3604938089
#> [421]  0.2365866019 -0.2589890308  0.3628238438  0.1001482422 -0.0382789598
#> [426]  0.2924022224  0.2270793890  0.3642534239 -0.3544439151 -0.0992392657
#> [431] -0.0688040609 -0.0075972226 -0.0467326862  0.3118491819  0.2056730095
#> [436] -0.3599847638  0.1801133746 -0.0722970115  0.0328548372 -0.0636569673
#> [441]  0.1829668489  0.3698152140 -0.3415230445 -0.0738717928  0.0466475913
#> [446]  0.0994195071 -0.3613735795  0.2870968092  0.3175575114 -0.3201173491
#> [451]  0.2631628648  0.1858395610 -0.0636124933 -0.0414465558  0.3666765306
#> [456] -0.3575106177 -0.1650342891 -0.0761658087  0.3564714177  0.2891431844
#> [461]  0.2676774608  0.2180228668  0.3694647783  0.3090016428  0.3563088971
#> [466] -0.0871105429  0.2835189603 -0.3583191420  0.2372806761  0.2775801532
#> [471]  0.2253560989 -0.3615442397  0.3080376679 -0.3461736022  0.2839985678
#> [476] -0.2479027444  0.3524173450  0.3684968931 -0.0714094935  0.3437633202
#> [481]  0.3674604799 -0.0744792741 -0.0599870836  0.3692279893 -0.3603361433
#> [486] -0.3605391139  0.3566090309 -0.0843314254 -0.3609256127 -0.3616916891
#> [491] -0.0443981243 -0.3492650170 -0.3345631234 -0.0497021417  0.3580699600
#> [496] -0.1139169694  0.2582195345 -0.0304745744 -0.2660958697 -0.3576368771