Computing non-parametric mixing distribution

Computing non-parametric mixing distribution

computemixdist(v, method = "npnormll", ...)

computemixdist.npnormll(v, mu0, pi0, beta, order, mix = NULL, ...)

computemixdist.npnormcvm(v, mu0, pi0, beta, order, mix = NULL, ...)

computemixdist.npnormad(v, mu0, pi0, beta, order, mix = NULL, ...)

computemixdist.npnormcll(v, mu0, pi0, beta, order, mix = NULL, ...)

computemixdist.nptll(v, mu0, pi0, beta, order, mix = NULL, ...)

computemixdist.npnormllw(v, mu0, pi0, beta, order, mix = NULL, ...)

computemixdist.npnormcvmw(v, mu0, pi0, beta, order, mix = NULL, ...)

computemixdist.npnormadw(v, mu0, pi0, beta, order, mix = NULL, ...)

Arguments

v
a vector of observations

method
An implemented family; see details

...
parameters above passed to the specific method

mu0
A vector of support points

pi0
A vector of weights corresponding to the support points

beta
structual parameter.

order
the parameter for the binned version.

mix
initial mixing distribution.

Details

These functions are used to make the object for computing the non-paramtric mixing distribution or estimating the proportion of zero using non-parametric methods.

This is a generic function for making the object for computing the non-parametric mixing distribution or estimating the proportion of zero.

current implemented families are:

npnormll : normal density using maximum likelihood (Chapter 3). The default beta is 1.

npnormllw : Binned version of normal density using maximum likelihood (Chapter 6). The default beta is 1 and the default order is -3.

npnormcvm : normal density using the Cramer-von Mises distance (Chapter 5). The default beta is 1.

npnormcvmw : Binned version of normal density using the Cramer-von Mises distance (Chapter 6). The default beta is 1 and the default order is -3.

npnormad : normal density using the Anderson-Darling distance (Chapter 5). The default beta is 1.

npnormadw : Binned version of normal density using the Anderson-Darling distance (Chapter 6) The default beta is 1 and the default order is -3.

nptll : t-density using maximum likelihood (Chapter 3). The default beta is infinity (normal distribution).

npnormcll : the one-parameter normal distribution used for approximating the sample correlation coefficients using maximum likelihood. This does not have a corresponding estimation of zero due to incompleted theory (Chapter 8). There is no default beta. The structure beta is the number of observations.

The default method used is npnormll.

Examples

set.seed(123)
pi0 = 0
data = rnorm(1000, mean = c(0, 2))
system.time(r <- computemixdist(data, pi0 = pi0, method = "npnormll"))

#>    user  system elapsed 
#>   0.031   0.001   0.033 

r

#> $iter
#> [1] 20
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -0.0001168956
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.0000000 0.1473755 1.9719686 2.9966729
#> 
#> $mix$pr
#> [1] 0.0000000 0.5304402 0.4578319 0.0117279
#> 
#> 
#> $ll
#> [1] 1722.52
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- computemixdist(data, pi0 = pi0, method = "npnormcvm"))

#>    user  system elapsed 
#>   0.014   0.000   0.014 

r

#> $iter
#> [1] 8
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -0.0001803887
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.0000000 0.1565762 2.0131053
#> 
#> $mix$pr
#> [1] 0.0000000 0.5376014 0.4623986
#> 
#> 
#> $ll
#> [1] 0.03069826
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- computemixdist(data, pi0 = pi0, method = "nptll"))

#>    user  system elapsed 
#>    0.07    0.00    0.07 

r

#> $iter
#> [1] 20
#> 
#> $family
#> [1] "npt"
#> 
#> $min.gradient
#> [1] -0.0001168788
#> 
#> $beta
#> [1] Inf
#> 
#> $mix
#> $mix$pt
#> [1] 0.0000000 0.1473755 1.9719686 2.9966730
#> 
#> $mix$pr
#> [1] 0.0000000 0.5304402 0.4578319 0.0117279
#> 
#> 
#> $ll
#> [1] 1722.52
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- computemixdist(data, pi0 = pi0, method = "npnormad"))

#>    user  system elapsed 
#>   0.037   0.000   0.038 

r

#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -0.3125589
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] -0.747729525  0.000000000  0.006250369  0.348921413  1.868616436
#> [6]  2.136038519  2.375932733  5.950000000
#> 
#> $mix$pr
#> [1] 2.169151e-14 0.000000e+00 2.964487e-01 2.337454e-01 3.654475e-01
#> [6] 3.515191e-14 1.040758e-01 2.826597e-04
#> 
#> 
#> $ll
#> [1] 0.1879813
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- computemixdist(tanh(data), pi0 = pi0, method = "npnormcll", beta = 4))

#>    user  system elapsed 
#>   0.070   0.003   0.072 

r

#> $iter
#> [1] 19
#> 
#> $family
#> [1] "npnormc"
#> 
#> $min.gradient
#> [1] -1.227633e-05
#> 
#> $beta
#> [1] 4
#> 
#> $mix
#> $mix$pt
#>  [1] -0.97638733 -0.91582239 -0.78104678 -0.78099638  0.00000000  0.02848914
#>  [7]  0.73499301  0.84816179  0.95214940  0.95219194  0.97872542  0.98392771
#> [13]  0.98725410  0.99613494  0.99818400  0.99921090  0.99958927
#> 
#> $mix$pr
#>  [1] 7.784339e-05 2.710385e-02 3.034553e-02 3.013395e-02 0.000000e+00
#>  [6] 3.285917e-01 1.428765e-01 7.808465e-02 1.187466e-01 1.852541e-02
#> [11] 7.410298e-02 5.148271e-03 7.163621e-02 4.206778e-02 1.704924e-02
#> [16] 1.462971e-02 8.798550e-04
#> 
#> 
#> $ll
#> [1] 130.3718
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- computemixdist(data, pi0 = pi0, method = "npnormllw", order = -2))

#>    user  system elapsed 
#>   0.073   0.001   0.074 

r

#> $iter
#> [1] 19
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -7.743007e-05
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 0.147110 1.972148 3.008075
#> 
#> $mix$pr
#> [1] 0.00000000 0.53027829 0.45831463 0.01140708
#> 
#> 
#> $ll
#> [1] 1722.453
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- computemixdist(data, pi0 = pi0, method = "npnormcvmw", order = -2))

#>    user  system elapsed 
#>   0.034   0.001   0.034 

r

#> $iter
#> [1] 9
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -4.635597e-14
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.0000000 0.1624174 2.0193714
#> 
#> $mix$pr
#> [1] 0.0000000 0.5380375 0.4619625
#> 
#> 
#> $ll
#> [1] 0.03072473
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- computemixdist(data, pi0 = pi0, method = "npnormadw", order = -2))

#>    user  system elapsed 
#>   0.037   0.000   0.037 

r

#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -0.6422916
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] -0.1295192  0.0000000  0.2991013  1.7151648  2.1189265  5.9500000
#> 
#> $mix$pr
#> [1] 0.1494487206 0.0000000000 0.3915918607 0.1268746222 0.3317557910
#> [6] 0.0003290055
#> 
#> 
#> $ll
#> [1] 0.1883913
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- nspmix::cnm(nspmix::npnorm(data)))

#>    user  system elapsed 
#>   0.097   0.028   0.126 

r

#> $family
#> [1] "npnorm"
#> 
#> $num.iterations
#> [1] 14
#> 
#> $max.gradient
#> [1] 9.135504e-07
#> 
#> $convergence
#> [1] 0
#> 
#> $ll
#> [1] -1722.52
#> 
#> $mix
#>             pt         pr
#> [1,] 0.1470637 0.53022704
#> [2,] 1.9708652 0.45769630
#> [3,] 2.9886660 0.01207665
#> 
#> $beta
#> [1] 1
#> 
#> attr(,"class")
#> [1] "nspmix"