Computing non-parametric mixing distribution

Computing non-parametric mixing distribution

computemixdist(v, method = "npnormll", ...)

computemixdist.npnormll(
  v,
  mu0,
  pi0,
  beta,
  order,
  mix = NULL,
  gridpoints = NULL,
  ...
)

computemixdist.npnormcvm(
  v,
  mu0,
  pi0,
  beta,
  order,
  mix = NULL,
  gridpoints = NULL,
  ...
)

computemixdist.npnormad(
  v,
  mu0,
  pi0,
  beta,
  order,
  mix = NULL,
  gridpoints = NULL,
  ...
)

computemixdist.npnormcll(
  v,
  mu0,
  pi0,
  beta,
  order,
  mix = NULL,
  gridpoints = NULL,
  ...
)

computemixdist.nptll(
  v,
  mu0,
  pi0,
  beta,
  order,
  mix = NULL,
  gridpoints = NULL,
  ...
)

computemixdist.npnormllw(
  v,
  mu0,
  pi0,
  beta,
  order = -3,
  mix = NULL,
  gridpoints = NULL,
  ...
)

computemixdist.npnormcvmw(
  v,
  mu0,
  pi0,
  beta,
  order = -3,
  mix = NULL,
  gridpoints = NULL,
  ...
)

computemixdist.npnormadw(
  v,
  mu0,
  pi0,
  beta,
  order = -3,
  mix = NULL,
  gridpoints = NULL,
  ...
)

computemixdist.nppoisll(
  v,
  mu0,
  pi0,
  beta,
  order,
  mix = NULL,
  gridpoints = NULL,
  ...
)

computemixdist.nptllw(
  v,
  mu0,
  pi0,
  beta,
  order = -3,
  mix = NULL,
  gridpoints = NULL,
  ...
)

computemixdist.npnorm2Dll(
  v,
  mu0,
  pi0,
  beta,
  order,
  mix = NULL,
  gridpoints = NULL,
  ...
)

Arguments

v

a vector of observations

method

An implemented family; see details

...

parameters above passed to the specific method

mu0

A vector of support points

pi0

A vector of weights corresponding to the support points

beta

structual parameter.

order

the parameter for the binned version.

mix

initial mixing distribution.

gridpoints

a vector of gridpoints to evaluate new support points

Details

These functions are used to compute the estimates of a mixing distribution non-parametrically using various methods, with possibly fixed components.

current implemented families are:

npnormll : normal density using maximum likelihood (Chapter 3). The default beta is 1.

npnormllw : Binned version of normal density using maximum likelihood (Chapter 6). The default beta is 1 and the default order is -3.

npnormcvm : normal density using the Cramer-von Mises distance (Chapter 5). The default beta is 1.

npnormcvmw : Binned version of normal density using the Cramer-von Mises distance (Chapter 6). The default beta is 1 and the default order is -3.

npnormad : normal density using the Anderson-Darling distance (Chapter 5). The default beta is 1.

npnormadw : Binned version of normal density using the Anderson-Darling distance (Chapter 6) The default beta is 1 and the default order is -3.

nptll : t-density using maximum likelihood (Chapter 3). The default beta is infinity (normal distribution).

nptllw : Binned version of non-centrol t-distribution using maximum likelihood. The default beta is infinity (normal distribution) and the default order is -3.

npnormcll : the one-parameter normal distribution used for approximating the sample correlation coefficients using maximum likelihood. This does not have a corresponding estimation of zero due to incompleted theory (Chapter 8). There is no default beta. The structure beta is the number of observations.

nppoisll : poisson mixture using maximum likelihood. There is no structural parameter, and hence the template will pass beta but this beta will never be referenced.

npnorm2Dll : bi-variate normal mixture using maximum likelihood. The structural parameter is the covariance matrix. Currently this implementation is experimental, and it could be very slow. Finding the new support points uses L-BFGS-B method provided by https://github.com/yixuan/LBFGSpp/.

The default method used is npnormll.

Examples

set.seed(123)
pi0 = 0
data = rnorm(1000, mean = c(0, 2))
system.time(r <- computemixdist(data, pi0 = pi0, method = "npnormll"))
#>    user  system elapsed 
#>   0.036   0.001   0.037 
r
#> $iter
#> [1] 24
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -9.528668e-05
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.0000000 0.1473455 1.9718642 2.9959354
#> 
#> $mix$pr
#> [1] 0.00000000 0.53041976 0.45782007 0.01176017
#> 
#> 
#> $ll
#> [1] 1722.52
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"
system.time(r <- computemixdist(data, pi0 = pi0, method = "npnormcvm"))
#>    user  system elapsed 
#>   0.017   0.000   0.018 
r
#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -2.395306e-14
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.0000000 0.1585935 2.0155947
#> 
#> $mix$pr
#> [1] 0.0000000 0.5386374 0.4613626
#> 
#> 
#> $ll
#> [1] 0.03071343
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"
system.time(r <- computemixdist(data, pi0 = pi0, method = "nptll"))
#>    user  system elapsed 
#>   0.074   0.001   0.075 
r
#> $iter
#> [1] 24
#> 
#> $family
#> [1] "npt"
#> 
#> $min.gradient
#> [1] -9.528091e-05
#> 
#> $beta
#> [1] Inf
#> 
#> $mix
#> $mix$pt
#> [1] 0.0000000 0.1473455 1.9718642 2.9959353
#> 
#> $mix$pr
#> [1] 0.00000000 0.53041975 0.45782007 0.01176018
#> 
#> 
#> $ll
#> [1] 1722.52
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"
system.time(r <- computemixdist(data, pi0 = pi0, method = "npnormad"))
#>    user  system elapsed 
#>   0.027   0.001   0.028 
r
#> $iter
#> [1] 12
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -0.0003599591
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.0000000 0.1483473 1.9544469 2.3759327
#> 
#> $mix$pr
#> [1] 0.00000000 0.53095300 0.42321313 0.04583387
#> 
#> 
#> $ll
#> [1] 0.1816632
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"
system.time(r <- computemixdist(tanh(data), pi0 = pi0, method = "npnormcll", beta = 4))
#>    user  system elapsed 
#>   0.255   0.002   0.259 
r
#> $iter
#> [1] 20
#> 
#> $family
#> [1] "npnormc"
#> 
#> $min.gradient
#> [1] -9.457214e-06
#> 
#> $beta
#> [1] 4
#> 
#> $mix
#> $mix$pt
#>  [1] -0.97638718 -0.91582603 -0.78103107  0.00000000  0.02847452  0.73497632
#>  [7]  0.84822981  0.95217224  0.97797874  0.98652942  0.99610345  0.99818746
#> [13]  0.99921169  0.99958934
#> 
#> $mix$pr
#>  [1] 7.767509e-05 2.710160e-02 6.047962e-02 0.000000e+00 3.285872e-01
#>  [6] 1.428905e-01 7.813434e-02 1.370084e-01 6.495827e-02 8.528828e-02
#> [11] 4.288186e-02 1.711822e-02 1.459931e-02 8.747163e-04
#> 
#> 
#> $ll
#> [1] 130.3605
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"
system.time(r <- computemixdist(data, pi0 = pi0, method = "npnormllw", order = -2))
#>    user  system elapsed 
#>   0.100   0.001   0.101 
r
#> $iter
#> [1] 22
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -0.0001632363
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.0000000 0.1472468 1.9726278 3.0115051
#> 
#> $mix$pr
#> [1] 0.0000000 0.5303727 0.4583648 0.0112625
#> 
#> 
#> $ll
#> [1] 1722.453
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"
system.time(r <- computemixdist(data, pi0 = pi0, method = "npnormcvmw", order = -2))
#>    user  system elapsed 
#>   0.046   0.001   0.047 
r
#> $iter
#> [1] 9
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -6.545629e-14
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.0000000 0.1624174 2.0193714
#> 
#> $mix$pr
#> [1] 0.0000000 0.5380375 0.4619625
#> 
#> 
#> $ll
#> [1] 0.0307244
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"
system.time(r <- computemixdist(data, pi0 = pi0, method = "npnormadw", order = -2))
#>    user  system elapsed 
#>   0.013   0.000   0.014 
r
#> $iter
#> [1] 12
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -5.5992e-09
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.0000000 0.1566451 1.9302599 2.1189265
#> 
#> $mix$pr
#> [1] 0.0000000 0.5333872 0.2785527 0.1880601
#> 
#> 
#> $ll
#> [1] 0.1822308
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"
system.time(r <- computemixdist(data, pi0 = pi0, method = "nptllw", order = -2))
#>    user  system elapsed 
#>   0.210   0.001   0.210 
r
#> $iter
#> [1] 22
#> 
#> $family
#> [1] "npt"
#> 
#> $min.gradient
#> [1] -0.0001632446
#> 
#> $beta
#> [1] Inf
#> 
#> $mix
#> $mix$pt
#> [1] 0.0000000 0.1472463 1.9726271 3.0115030
#> 
#> $mix$pr
#> [1] 0.00000000 0.53037249 0.45836479 0.01126272
#> 
#> 
#> $ll
#> [1] 1722.453
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"
data2 = rpois(1000, c(0, 2))
system.time(r <- computemixdist(data2, pi0 = pi0, method = "nppoisll"))
#>    user  system elapsed 
#>   0.003   0.000   0.003 
r
#> $iter
#> [1] 16
#> 
#> $family
#> [1] "nppois"
#> 
#> $min.gradient
#> [1] -2.52869e-07
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 0.000000 1.694065 2.834590
#> 
#> $mix$pr
#> [1] 0.4889106 0.0000000 0.4004427 0.1106468
#> 
#> 
#> $ll
#> [1] 1341.294
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"
data3 = matrix(rnorm(1000), 500, 2)
system.time(r <- computemixdist(data3, method = "npnorm2Dll"))
#>    user  system elapsed 
#>   2.825   0.047   2.880 
r
#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npnorm2D"
#> 
#> $min.gradient
#> [1] -1.705303e-13
#> 
#> $beta
#>      [,1] [,2]
#> [1,]    1    0
#> [2,]    0    1
#> 
#> $mix
#> $mix$pt
#>             [,1]        [,2]
#> [1,]  0.08330555 -0.04440768
#> [2,]  2.84615385 -0.70588235
#> [3,]  0.14458691 -0.11056646
#> [4,] -0.46153846  0.29411765
#> [5,] -0.37342729  0.28556205
#> [6,]  0.00000000  0.00000000
#> 
#> $mix$pr
#> [1] 0.888877042 0.001978998 0.016918604 0.078323242 0.013902114 0.000000000
#> 
#> 
#> $ll
#> [1] 1400.049
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"