Computing non-parametric mixing distribution with estimated proportion at 0

computing non-parametric mixing distribution with estimated proportion at 0

estpi0(v, method = "npnormll", ...)

estpi0.npnormll(v, beta, val, order, mix = NULL, gridpoints = NULL, ...)

estpi0.npnormcvm(v, beta, val, order, mix = NULL, gridpoints = NULL, ...)

estpi0.npnormad(v, beta, val, order, mix = NULL, gridpoints = NULL, ...)

estpi0.npnormcll(v, beta, val, order, mix = NULL, gridpoints = NULL, ...)

estpi0.nptll(v, beta, val, order, mix = NULL, gridpoints = NULL, ...)

estpi0.npnormllw(v, beta, val, order, mix = NULL, gridpoints = NULL, ...)

estpi0.npnormcvmw(v, beta, val, order, mix = NULL, gridpoints = NULL, ...)

estpi0.npnormadw(v, beta, val, order, mix = NULL, gridpoints = NULL, ...)

Arguments

v
observations

method
An implemented family; see details

...
parameters above passed to the specific method.

beta
structural parameter

val
thresholding function

order
the parameter for the binned version.

mix
initial mixing distribution.

gridpoints
a vector of gridpoints to evaluate new support points

Details

This is a function for computing non-parametric mixing distribution with estimated proportion at 0. Different families will have different threshold values.

The parameters are listed as follows:

tol: tolerance to stop the code.

verbose: logical; Whether to print the intermediate results.

It is not shown in the parameter section since various method have different default threshold values and this function essentially calls the class method in the object.

The full list of implemented families is in computemixdist.

IMPORTANT: although estpi0 is implemented for nptll and npnormcll, the theory is not yet done, and hence it is recommand to use with precaution.

Examples

data = rnorm(500, c(0, 2))
system.time(r <- estpi0(data, method = "npnormll", val = 2, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.519666 fval: 884.567
#> Iter: 2 lower: 0.519666 upper: 0.759833
#> current val: 0.69535 fval: 903.146
#> Iter: 3 lower: 0.519666 upper: 0.607508
#> current val: 0.530424 fval: 884.909
#> Iter: 4 lower: 0.525045 upper: 0.530424
#> current val: 0.527959 fval: 884.824

#>    user  system elapsed 
#>   0.040   0.000   0.039 

r

#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -1.553542e-05
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] -1.872467  0.000000  2.029319
#> 
#> $mix$pr
#> [1] 0.004690674 0.527961530 0.467347796
#> 
#> 
#> $ll
#> [1] 884.8242
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormllw", val = 2, verbose = TRUE, order = -2))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.519929 fval: 884.598
#> Iter: 2 lower: 0.519929 upper: 0.759964
#> current val: 0.695332 fval: 903.18
#> Iter: 3 lower: 0.519929 upper: 0.607631
#> current val: 0.529775 fval: 884.911
#> Iter: 4 lower: 0.524852 upper: 0.529775
#> current val: 0.527456 fval: 884.832

#>    user  system elapsed 
#>   0.196   0.002   0.198 

r

#> $iter
#> [1] 8
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -3.533445e-06
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] -1.855625  0.000000  2.028591
#> 
#> $mix$pr
#> [1] 0.004814977 0.527458605 0.467726418
#> 
#> 
#> $ll
#> [1] 884.832
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormcvm", val = 0.1, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.511595 fval: 0.0547402
#> Iter: 2 lower: 0.511595 upper: 0.755797
#> current val: 0.653472 fval: 1.13416
#> Iter: 3 lower: 0.511595 upper: 0.582534
#> current val: 0.54777 fval: 0.11833
#> Iter: 4 lower: 0.529682 upper: 0.54777
#> current val: 0.541128 fval: 0.0991586

#>    user  system elapsed 
#>   0.042   0.001   0.043 

r

#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] 0
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.00000 2.14288
#> 
#> $mix$pr
#> [1] 0.5414529 0.4585471
#> 
#> 
#> $ll
#> [1] 0.1000002
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormcvmw", val = 0.1, verbose = TRUE, order = -2))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.511263 fval: 0.0555813
#> Iter: 2 lower: 0.511263 upper: 0.755632
#> current val: 0.652105 fval: 1.14135
#> Iter: 3 lower: 0.511263 upper: 0.581684
#> current val: 0.545936 fval: 0.117719
#> Iter: 4 lower: 0.528599 upper: 0.545936
#> current val: 0.539617 fval: 0.0993445
#> Iter: 5 lower: 0.539617 upper: 0.542776
#> current val: 0.539866 fval: 0.100002
#> Iter: 6 lower: 0.539741 upper: 0.539866
#> current val: 0.539865 fval: 0.0999984

#>    user  system elapsed 
#>   0.217   0.002   0.220 

r

#> $iter
#> [1] 2
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] 0
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.144217
#> 
#> $mix$pr
#> [1] 0.539866 0.460134
#> 
#> 
#> $ll
#> [1] 0.0999997
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormad", val = 1, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.521239 fval: 0.50138
#> Iter: 2 lower: 0.521239 upper: 0.76062
#> current val: 0.71334 fval: 12.1407
#> Iter: 3 lower: 0.521239 upper: 0.61729
#> current val: 0.570112 fval: 1.16755
#> Iter: 4 lower: 0.545676 upper: 0.570112
#> current val: 0.563345 fval: 1.00038
#> Iter: 5 lower: 0.55451 upper: 0.563345
#> current val: 0.563328 fval: 0.999999

#>    user  system elapsed 
#>   0.058   0.001   0.059 

r

#> $iter
#> [1] 3
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] 0
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.0000 2.1792
#> 
#> $mix$pr
#> [1] 0.5633282 0.4366718
#> 
#> 
#> $ll
#> [1] 0.9999996
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormadw", val = 1, verbose = TRUE, order = -2))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.520964 fval: 0.435933
#> Iter: 2 lower: 0.520964 upper: 0.760482
#> current val: 0.712784 fval: 12.3067
#> Iter: 3 lower: 0.520964 upper: 0.616874
#> current val: 0.569363 fval: 1.18946
#> Iter: 4 lower: 0.545163 upper: 0.569363
#> current val: 0.561564 fval: 0.993906
#> Iter: 5 lower: 0.561564 upper: 0.565464
#> current val: 0.561829 fval: 1

#>    user  system elapsed 
#>   0.043   0.001   0.044 

r

#> $iter
#> [1] 4
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] 0
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.181053
#> 
#> $mix$pr
#> [1] 0.5618285 0.4381715
#> 
#> 
#> $ll
#> [1] 1.000001
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "nptll", val = 2, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.519729 fval: 884.569
#> Iter: 2 lower: 0.519729 upper: 0.759864
#> current val: 0.695367 fval: 903.149
#> Iter: 3 lower: 0.519729 upper: 0.607548
#> current val: 0.530408 fval: 884.908
#> Iter: 4 lower: 0.525068 upper: 0.530408
#> current val: 0.527959 fval: 884.824

#>    user  system elapsed 
#>   0.119   0.001   0.120 

r

#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npt"
#> 
#> $min.gradient
#> [1] -1.532249e-05
#> 
#> $beta
#> [1] Inf
#> 
#> $mix
#> $mix$pt
#> [1] -1.872464  0.000000  2.029319
#> 
#> $mix$pr
#> [1] 0.004690676 0.527961677 0.467347647
#> 
#> 
#> $ll
#> [1] 884.8242
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(tanh(data), method = "npnormcll", val = 2, beta = 4, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.310872 fval: 42.5775
#> Iter: 2 lower: 0.310872 upper: 0.655436
#> current val: 0.578798 fval: 80.6917
#> Iter: 3 lower: 0.310872 upper: 0.444835
#> current val: 0.37019 fval: 44.5332
#> Iter: 4 lower: 0.340531 upper: 0.37019
#> current val: 0.370162 fval: 44.5301
#> Iter: 5 lower: 0.370176 upper: 0.37019
#> current val: 0.37018 fval: 44.5312
#> Iter: 6 lower: 0.37018 upper: 0.370185
#> current val: 0.370184 fval: 44.5315

#>    user  system elapsed 
#>   0.173   0.001   0.174 

r

#> $iter
#> [1] 4
#> 
#> $family
#> [1] "npnormc"
#> 
#> $min.gradient
#> [1] -2.285373e-06
#> 
#> $beta
#> [1] 4
#> 
#> $mix
#> $mix$pt
#>  [1] -0.9982257 -0.9238673 -0.8121274  0.0000000  0.8293340  0.9188417
#>  [7]  0.9723593  0.9860540  0.9904374  0.9976712  0.9990728
#> 
#> $mix$pr
#>  [1] 0.0005143657 0.0493643684 0.0500462939 0.3701839443 0.1385106786
#>  [6] 0.1219338650 0.1207031300 0.0013148071 0.0808438678 0.0646444371
#> [11] 0.0019402421
#> 
#> 
#> $ll
#> [1] 44.53145
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"