Computing non-parametric mixing distribution with estimated proportion at 0

computing non-parametric mixing distribution with estimated proportion at 0

estpi0(v, method = "npnormll", ...)

estpi0.npnormll(v, beta, val, order, mix = NULL, gridpoints = NULL, ...)

estpi0.npnormcvm(v, beta, val, order, mix = NULL, gridpoints = NULL, ...)

estpi0.npnormad(v, beta, val, order, mix = NULL, gridpoints = NULL, ...)

estpi0.npnormcll(v, beta, val, order, mix = NULL, gridpoints = NULL, ...)

estpi0.nptll(v, beta, val, order, mix = NULL, gridpoints = NULL, ...)

estpi0.npnormllw(v, beta, val, order = -3, mix = NULL, gridpoints = NULL, ...)

estpi0.npnormcvmw(v, beta, val, order = -3, mix = NULL, gridpoints = NULL, ...)

estpi0.npnormadw(v, beta, val, order = -3, mix = NULL, gridpoints = NULL, ...)

estpi0.nppoisll(v, beta, val, order, mix = NULL, gridpoints = NULL, ...)

estpi0.nptllw(v, beta, val, order = -3, mix = NULL, gridpoints = NULL, ...)

Arguments

v

observations

method

An implemented family; see details

...

parameters above passed to the specific method.

beta

structural parameter

val

thresholding function

order

the parameter for the binned version.

mix

initial mixing distribution.

gridpoints

a vector of gridpoints to evaluate new support points

Details

These function are for computing non-parametric mixing distribution with estimated proportion at 0. Different families will have different threshold values.

The parameters are listed as follows:

tol: tolerance to stop the code.

verbose: logical; Whether to print the intermediate results.

It is not shown in the parameter section since various method have different default threshold values and this function essentially calls the class method in the object.

The full list of implemented families is in computemixdist.

IMPORTANT: although estpi0 is implemented for nptll and npnormcll, the theory is not yet done, and hence it is recommand to use with precaution.

Examples

data = rnorm(500, c(0, 2))
system.time(r <- estpi0(data, method = "npnormll", val = 2, verbose = TRUE))
#> Iter: 1 lower: 0 upper: 1
#> current val: 0.587672 fval: 887.52
#> Iter: 2 lower: 0.293836 upper: 0.587672
#> current val: 0.585199 fval: 887.391
#>    user  system elapsed 
#>   0.039   0.000   0.040 
r
#> $iter
#> [1] 6
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -5.684342e-14
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.161092
#> 
#> $mix$pr
#> [1] 0.5871929 0.4128071
#> 
#> 
#> $ll
#> [1] 887.4947
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"
system.time(r <- estpi0(data, method = "npnormllw", val = 2, verbose = TRUE, order = -2))
#> Iter: 1 lower: 0 upper: 1
#> current val: 0.588829 fval: 887.457
#> Iter: 2 lower: 0.294415 upper: 0.588829
#> current val: 0.580664 fval: 887.043
#> Iter: 3 lower: 0.584747 upper: 0.588829
#> current val: 0.587229 fval: 887.371
#>    user  system elapsed 
#>   0.262   0.003   0.267 
r
#> $iter
#> [1] 6
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -2.842171e-14
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.161115
#> 
#> $mix$pr
#> [1] 0.5872292 0.4127708
#> 
#> 
#> $ll
#> [1] 887.3713
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"
system.time(r <- estpi0(data, method = "npnormcvm", val = 0.1, verbose = TRUE))
#> Iter: 1 lower: 0 upper: 1
#> current val: 0.579927 fval: 0.194234
#> Iter: 2 lower: 0.289964 upper: 0.579927
#> current val: 0.461012 fval: 0.0141004
#> Iter: 3 lower: 0.520469 upper: 0.579927
#> current val: 0.554309 fval: 0.0778666
#> Iter: 4 lower: 0.554309 upper: 0.567118
#> current val: 0.560645 fval: 0.1001
#>    user  system elapsed 
#>   0.059   0.001   0.062 
r
#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] 5.551115e-17
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.197569
#> 
#> $mix$pr
#> [1] 0.5606189 0.4393811
#> 
#> 
#> $ll
#> [1] 0.09999996
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"
system.time(r <- estpi0(data, method = "npnormcvmw", val = 0.1, verbose = TRUE, order = -2))
#> Iter: 1 lower: 0 upper: 1
#> current val: 0.580238 fval: 0.209413
#> Iter: 2 lower: 0.290119 upper: 0.580238
#> current val: 0.451065 fval: 0.0146137
#> Iter: 3 lower: 0.515651 upper: 0.580238
#> current val: 0.549983 fval: 0.0719631
#> Iter: 4 lower: 0.549983 upper: 0.565111
#> current val: 0.558342 fval: 0.100156
#>    user  system elapsed 
#>   0.219   0.001   0.220 
r
#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] 0
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.195892
#> 
#> $mix$pr
#> [1] 0.5583012 0.4416988
#> 
#> 
#> $ll
#> [1] 0.09999932
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"
system.time(r <- estpi0(data, method = "npnormad", val = 1, verbose = TRUE))
#> Iter: 1 lower: 0 upper: 1
#> current val: 0.583558 fval: 1.13074
#> Iter: 2 lower: 0.291779 upper: 0.583558
#> current val: 0.558346 fval: 0.490327
#> Iter: 3 lower: 0.570952 upper: 0.583558
#> current val: 0.579373 fval: 0.999686
#>    user  system elapsed 
#>   0.069   0.000   0.070 
r
#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] 1.136868e-13
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.243823
#> 
#> $mix$pr
#> [1] 0.579383 0.420617
#> 
#> 
#> $ll
#> [1] 1
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"
system.time(r <- estpi0(data, method = "npnormadw", val = 1, verbose = TRUE, order = -2))
#> Iter: 1 lower: 0 upper: 1
#> current val: 0.583873 fval: 1.21728
#> Iter: 2 lower: 0.291936 upper: 0.583873
#> current val: 0.544392 fval: 0.314434
#> Iter: 3 lower: 0.564132 upper: 0.583873
#> current val: 0.577024 fval: 0.998811
#>    user  system elapsed 
#>   0.048   0.000   0.048 
r
#> $iter
#> [1] 8
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -1.412204e-13
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.242782
#> 
#> $mix$pr
#> [1] 0.5770636 0.4229364
#> 
#> 
#> $ll
#> [1] 1
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"
system.time(r <- estpi0(data, method = "nptll", val = 2, verbose = TRUE))
#> Iter: 1 lower: 0 upper: 1
#> current val: 0.587672 fval: 887.52
#> Iter: 2 lower: 0.293836 upper: 0.587672
#> current val: 0.585199 fval: 887.391
#>    user  system elapsed 
#>   0.093   0.000   0.093 
r
#> $iter
#> [1] 6
#> 
#> $family
#> [1] "npt"
#> 
#> $min.gradient
#> [1] 1.421085e-13
#> 
#> $beta
#> [1] Inf
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.161092
#> 
#> $mix$pr
#> [1] 0.5871929 0.4128071
#> 
#> 
#> $ll
#> [1] 887.4947
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"
system.time(r <- estpi0(tanh(data), method = "npnormcll", val = 2, beta = 4, verbose = TRUE))
#> Iter: 1 lower: 0 upper: 1
#> current val: 0.302298 fval: 56.845
#> Iter: 2 lower: 0.302298 upper: 0.651149
#> current val: 0.585611 fval: 90.8133
#> Iter: 3 lower: 0.302298 upper: 0.443955
#> current val: 0.378453 fval: 59.1119
#> Iter: 4 lower: 0.340375 upper: 0.378453
#> current val: 0.373701 fval: 58.8379
#>    user  system elapsed 
#>   0.740   0.002   0.746 
r
#> $iter
#> [1] 8
#> 
#> $family
#> [1] "npnormc"
#> 
#> $min.gradient
#> [1] -3.944927e-05
#> 
#> $beta
#> [1] 4
#> 
#> $mix
#> $mix$pt
#>  [1] -0.9933681 -0.9059587 -0.7895701  0.0000000  0.7963924  0.9357594
#>  [7]  0.9757480  0.9884145  0.9964577  0.9994700
#> 
#> $mix$pr
#>  [1] 0.003093495 0.034868789 0.051059911 0.373720488 0.188005423 0.085826481
#>  [7] 0.114814781 0.045314386 0.087308005 0.015988242
#> 
#> 
#> $ll
#> [1] 58.83897
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"
system.time(r <- estpi0(data, method = "nptllw", val = 2, verbose = TRUE))
#> Iter: 1 lower: 0 upper: 1
#> current val: 0.587756 fval: 887.519
#> Iter: 2 lower: 0.293878 upper: 0.587756
#> current val: 0.584811 fval: 887.365
#>    user  system elapsed 
#>   0.623   0.001   0.627 
r
#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npt"
#> 
#> $min.gradient
#> [1] -2.842171e-14
#> 
#> $beta
#> [1] Inf
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.161078
#> 
#> $mix$pr
#> [1] 0.5871847 0.4128153
#> 
#> 
#> $ll
#> [1] 887.4882
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"
data2 = rpois(1000, c(0, 2))
system.time(r <- estpi0(data2, method = "nppoisll", val = 2))
#>    user  system elapsed 
#>   0.006   0.000   0.006 
r
#> $iter
#> [1] 2
#> 
#> $family
#> [1] "nppois"
#> 
#> $min.gradient
#> [1] -5.684342e-14
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.059058
#> 
#> $mix$pr
#> [1] 0.5376284 0.4623716
#> 
#> 
#> $ll
#> [1] 1343.556
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"