Computing non-parametric mixing distribution with estimated proportion at 0

computing non-parametric mixing distribution with estimated proportion at 0

estpi0(v, method = "npnormll", ...)

estpi0.npnormll(v, beta, val, order, mix = NULL, ...)

estpi0.npnormcvm(v, beta, val, order, mix = NULL, ...)

estpi0.npnormad(v, beta, val, order, mix = NULL, ...)

estpi0.npnormcll(v, beta, val, order, mix = NULL, ...)

estpi0.nptll(v, beta, val, order, mix = NULL, ...)

estpi0.npnormllw(v, beta, val, order, mix = NULL, ...)

estpi0.npnormcvmw(v, beta, val, order, mix = NULL, ...)

estpi0.npnormadw(v, beta, val, order, mix = NULL, ...)

Arguments

v
observations

method
An implemented family; see details

...
parameters above passed to the specific method.

beta
structural parameter

val
thresholding function

order
the parameter for the binned version.

mix
initial mixing distribution.

Details

This is a function for computing non-parametric mixing distribution with estimated proportion at 0. Different families will have different threshold values.

The parameters are listed as follows:

tol: tolerance to stop the code.

verbose: logical; Whether to print the intermediate results.

It is not shown in the parameter section since various method have different default threshold values and this function essentially calls the class method in the object.

The full list of implemented families is in computemixdist.

IMPORTANT: although estpi0 is implemented for nptll and npnormcll, the theory is not yet done, and hence it is recommand to use with precaution.

Examples

data = rnorm(500, c(0, 2))
system.time(r <- estpi0(data, method = "npnormll", val = 2, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.519666 fval: 884.567
#> Iter: 2 lower: 0.519666 upper: 0.759833
#> current val: 0.69535 fval: 903.146
#> Iter: 3 lower: 0.519666 upper: 0.607508
#> current val: 0.530424 fval: 884.909
#> Iter: 4 lower: 0.525045 upper: 0.530424
#> current val: 0.527959 fval: 884.824

#>    user  system elapsed 
#>   0.042   0.000   0.042 

r

#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -1.553542e-05
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] -1.872467  0.000000  2.029319
#> 
#> $mix$pr
#> [1] 0.004690674 0.527961530 0.467347796
#> 
#> 
#> $ll
#> [1] 884.8242
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormllw", val = 2, verbose = TRUE, order = -2))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.519929 fval: 884.598
#> Iter: 2 lower: 0.519929 upper: 0.759964
#> current val: 0.695332 fval: 903.18
#> Iter: 3 lower: 0.519929 upper: 0.607631
#> current val: 0.529775 fval: 884.911
#> Iter: 4 lower: 0.524852 upper: 0.529775
#> current val: 0.527456 fval: 884.832

#>    user  system elapsed 
#>   0.203   0.002   0.204 

r

#> $iter
#> [1] 8
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -3.533445e-06
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] -1.855625  0.000000  2.028591
#> 
#> $mix$pr
#> [1] 0.004814977 0.527458605 0.467726418
#> 
#> 
#> $ll
#> [1] 884.832
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormcvm", val = 0.1, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.511595 fval: 0.107162
#> Iter: 2 lower: 0.255797 upper: 0.511595
#> current val: 0.47515 fval: 0.0598808
#> Iter: 3 lower: 0.493373 upper: 0.511595
#> current val: 0.509688 fval: 0.0598825
#> Iter: 4 lower: 0.510641 upper: 0.511595
#> current val: 0.511495 fval: 0.0600272
#> Iter: 5 lower: 0.511545 upper: 0.511595
#> current val: 0.511589 fval: 0.0600413
#> Iter: 6 lower: 0.511592 upper: 0.511595
#> current val: 0.511594 fval: 0.0600414

#>    user  system elapsed 
#>   0.039   0.001   0.040 

r

#> $iter
#> [1] 2
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] 5.551115e-17
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.041729
#> 
#> $mix$pr
#> [1] 0.5115947 0.4884053
#> 
#> 
#> $ll
#> [1] 0.0600414
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormcvmw", val = 0.1, verbose = TRUE, order = -2))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.511263 fval: 0.107317
#> Iter: 2 lower: 0.255632 upper: 0.511263
#> current val: 0.475527 fval: 0.0590176
#> Iter: 3 lower: 0.493395 upper: 0.511263
#> current val: 0.509393 fval: 0.0591394
#> Iter: 4 lower: 0.510328 upper: 0.511263
#> current val: 0.511165 fval: 0.0594917
#> Iter: 5 lower: 0.511214 upper: 0.511263
#> current val: 0.511258 fval: 0.0595166
#> Iter: 6 lower: 0.51126 upper: 0.511263
#> current val: 0.511263 fval: 0.0595177

#>    user  system elapsed 
#>   0.107   0.001   0.108 

r

#> $iter
#> [1] 2
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] 0
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.047599
#> 
#> $mix$pr
#> [1] 0.5112631 0.4887369
#> 
#> 
#> $ll
#> [1] 0.05951781
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormad", val = 1, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.521239 fval: 2.92943
#> Iter: 2 lower: 0 upper: 0.26062
#> current val: 0.19065 fval: 1.35581
#> Iter: 3 lower: 0 upper: 0.095325
#> current val: 0.0203124 fval: 1.35581
#> Iter: 4 lower: 0 upper: 0.0101562
#> current val: 0.00585975 fval: 1.35581
#> Iter: 5 lower: 0 upper: 0.00292987
#> current val: 0.00169043 fval: 1.35581
#> Iter: 6 lower: 0 upper: 0.000845215
#> current val: 0.000487596 fval: 1.35605
#> Iter: 7 lower: 0 upper: 0.000243798
#> current val: 0.000140607 fval: 1.35581
#> Iter: 8 lower: 0 upper: 7.03033e-05
#> current val: 4.05624e-05 fval: 1.35581
#> Iter: 9 lower: 0 upper: 2.02812e-05
#> current val: 1.17015e-05 fval: 1.35581
#> Iter: 10 lower: 0 upper: 5.85076e-06
#> current val: 3.37567e-06 fval: 1.35581
#> Iter: 11 lower: 0 upper: 1.68784e-06
#> current val: 9.73821e-07 fval: 1.35581

#>    user  system elapsed 
#>   0.032   0.000   0.031 

r

#> $iter
#> [1] 1
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -6.683477e-05
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 0.000000 1.820802
#> 
#> $mix$pr
#> [1] 4.638657e-01 2.809296e-07 5.361340e-01
#> 
#> 
#> $ll
#> [1] 1.35581
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormadw", val = 1, verbose = TRUE, order = -2))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.520964 fval: 3.0631
#> Iter: 2 lower: 0 upper: 0.260482
#> current val: 0.191606 fval: 1.36386
#> Iter: 3 lower: 0 upper: 0.0958028
#> current val: 0.0547303 fval: 1.36386
#> Iter: 4 lower: 0 upper: 0.0273652
#> current val: 0.0156332 fval: 1.36386
#> Iter: 5 lower: 0 upper: 0.0078166
#> current val: 0.00441695 fval: 1.3845
#> Iter: 6 lower: 0 upper: 0.00220848
#> current val: 0.00126672 fval: 1.36408
#> Iter: 7 lower: 0 upper: 0.00063336
#> current val: 0.000361842 fval: 1.36386
#> Iter: 8 lower: 0 upper: 0.000180921
#> current val: 0.000103357 fval: 1.36386
#> Iter: 9 lower: 0 upper: 5.16784e-05
#> current val: 2.95229e-05 fval: 1.36386
#> Iter: 10 lower: 0 upper: 1.47615e-05
#> current val: 8.43294e-06 fval: 1.36386
#> Iter: 11 lower: 0 upper: 4.21647e-06
#> current val: 2.40879e-06 fval: 1.36386
#> Iter: 12 lower: 0 upper: 1.2044e-06
#> current val: 6.88048e-07 fval: 1.36386

#>    user  system elapsed 
#>   0.028   0.000   0.029 

r

#> $iter
#> [1] 1
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -2.773959e-11
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 0.000000 1.820354
#> 
#> $mix$pr
#> [1] 4.613331e-01 1.965346e-07 5.386667e-01
#> 
#> 
#> $ll
#> [1] 1.363862
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "nptll", val = 2, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.519729 fval: 884.569
#> Iter: 2 lower: 0.519729 upper: 0.759864
#> current val: 0.695367 fval: 903.149
#> Iter: 3 lower: 0.519729 upper: 0.607548
#> current val: 0.530408 fval: 884.908
#> Iter: 4 lower: 0.525068 upper: 0.530408
#> current val: 0.527959 fval: 884.824

#>    user  system elapsed 
#>   0.122   0.001   0.123 

r

#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npt"
#> 
#> $min.gradient
#> [1] -1.532249e-05
#> 
#> $beta
#> [1] Inf
#> 
#> $mix
#> $mix$pt
#> [1] -1.872464  0.000000  2.029319
#> 
#> $mix$pr
#> [1] 0.004690676 0.527961677 0.467347647
#> 
#> 
#> $ll
#> [1] 884.8242
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(tanh(data), method = "npnormcll", val = 2, beta = 4, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.310872 fval: 42.5775
#> Iter: 2 lower: 0.310872 upper: 0.655436
#> current val: 0.578798 fval: 80.6917
#> Iter: 3 lower: 0.310872 upper: 0.444835
#> current val: 0.37019 fval: 44.5332
#> Iter: 4 lower: 0.340531 upper: 0.37019
#> current val: 0.370162 fval: 44.5301
#> Iter: 5 lower: 0.370176 upper: 0.37019
#> current val: 0.37018 fval: 44.5312
#> Iter: 6 lower: 0.37018 upper: 0.370185
#> current val: 0.370184 fval: 44.5315

#>    user  system elapsed 
#>   0.177   0.001   0.178 

r

#> $iter
#> [1] 4
#> 
#> $family
#> [1] "npnormc"
#> 
#> $min.gradient
#> [1] -2.285373e-06
#> 
#> $beta
#> [1] 4
#> 
#> $mix
#> $mix$pt
#>  [1] -0.9982257 -0.9238673 -0.8121274  0.0000000  0.8293340  0.9188417
#>  [7]  0.9723593  0.9860540  0.9904374  0.9976712  0.9990728
#> 
#> $mix$pr
#>  [1] 0.0005143657 0.0493643684 0.0500462939 0.3701839443 0.1385106786
#>  [6] 0.1219338650 0.1207031300 0.0013148071 0.0808438678 0.0646444371
#> [11] 0.0019402421
#> 
#> 
#> $ll
#> [1] 44.53145
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"