Computing non-parametric mixing distribution with estimated proportion at 0

computing non-parametric mixing distribution with estimated proportion at 0

estpi0(v, method = "npnormll", ...)

estpi0.npnormll(v, beta, val, order, mix = NULL, gridpoints = NULL, ...)

estpi0.npnormcvm(v, beta, val, order, mix = NULL, gridpoints = NULL, ...)

estpi0.npnormad(v, beta, val, order, mix = NULL, gridpoints = NULL, ...)

estpi0.npnormcll(v, beta, val, order, mix = NULL, gridpoints = NULL, ...)

estpi0.nptll(v, beta, val, order, mix = NULL, gridpoints = NULL, ...)

estpi0.npnormllw(v, beta, val, order = -3, mix = NULL, gridpoints = NULL, ...)

estpi0.npnormcvmw(v, beta, val, order = -3, mix = NULL, gridpoints = NULL, ...)

estpi0.npnormadw(v, beta, val, order = -3, mix = NULL, gridpoints = NULL, ...)

estpi0.nppoisll(v, beta, val, order, mix = NULL, gridpoints = NULL, ...)

estpi0.nppoiscvm(v, beta, val, order, mix = NULL, gridpoints = NULL, ...)

Arguments

v
observations

method
An implemented family; see details

...
parameters above passed to the specific method.

beta
structural parameter

val
thresholding function

order
the parameter for the binned version.

mix
initial mixing distribution.

gridpoints
a vector of gridpoints to evaluate new support points

Details

This is a function for computing non-parametric mixing distribution with estimated proportion at 0. Different families will have different threshold values.

The parameters are listed as follows:

tol: tolerance to stop the code.

verbose: logical; Whether to print the intermediate results.

It is not shown in the parameter section since various method have different default threshold values and this function essentially calls the class method in the object.

The full list of implemented families is in computemixdist.

IMPORTANT: although estpi0 is implemented for nptll and npnormcll, the theory is not yet done, and hence it is recommand to use with precaution.

Examples

data = rnorm(500, c(0, 2))
system.time(r <- estpi0(data, method = "npnormll", val = 2, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.584421 fval: 886.323
#> Iter: 2 lower: 0.584421 upper: 0.79221
#> current val: 0.739589 fval: 907.155
#> Iter: 3 lower: 0.584421 upper: 0.662005
#> current val: 0.603503 fval: 887.362
#> Iter: 4 lower: 0.593962 upper: 0.603503
#> current val: 0.601896 fval: 887.26

#>    user  system elapsed 
#>   0.037   0.001   0.038 

r

#> $iter
#> [1] 6
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -8.526513e-14
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.239117
#> 
#> $mix$pr
#> [1] 0.6018971 0.3981029
#> 
#> 
#> $ll
#> [1] 887.2603
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormllw", val = 2, verbose = TRUE, order = -2))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.585008 fval: 886.325
#> Iter: 2 lower: 0.585008 upper: 0.792504
#> current val: 0.739759 fval: 907.172
#> Iter: 3 lower: 0.585008 upper: 0.662384
#> current val: 0.603426 fval: 887.333
#> Iter: 4 lower: 0.594217 upper: 0.603426
#> current val: 0.601855 fval: 887.234

#>    user  system elapsed 
#>   0.164   0.001   0.166 

r

#> $iter
#> [1] 6
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] 8.526513e-14
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.239046
#> 
#> $mix$pr
#> [1] 0.6018564 0.3981436
#> 
#> 
#> $ll
#> [1] 887.2337
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormcvm", val = 0.1, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.561794 fval: 0.0614314
#> Iter: 2 lower: 0.561794 upper: 0.780897
#> current val: 0.663223 fval: 0.926025
#> Iter: 3 lower: 0.561794 upper: 0.612509
#> current val: 0.578257 fval: 0.107011
#> Iter: 4 lower: 0.570025 upper: 0.578257
#> current val: 0.576288 fval: 0.0999762

#>    user  system elapsed 
#>   0.033   0.000   0.034 

r

#> $iter
#> [1] 5
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -5.551115e-17
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.303408
#> 
#> $mix$pr
#> [1] 0.5762947 0.4237053
#> 
#> 
#> $ll
#> [1] 0.1000004
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormcvmw", val = 0.1, verbose = TRUE, order = -2))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.561629 fval: 0.0650354
#> Iter: 2 lower: 0.561629 upper: 0.780815
#> current val: 0.661479 fval: 0.927662
#> Iter: 3 lower: 0.561629 upper: 0.611554
#> current val: 0.576144 fval: 0.106406
#> Iter: 4 lower: 0.568887 upper: 0.576144
#> current val: 0.574338 fval: 0.0999827
#> Iter: 5 lower: 0.574338 upper: 0.575241
#> current val: 0.574343 fval: 0.100001
#> Iter: 6 lower: 0.574341 upper: 0.574343
#> current val: 0.574341 fval: 0.0999931
#> Iter: 7 lower: 0.574342 upper: 0.574343
#> current val: 0.574343 fval: 0.0999983

#>    user  system elapsed 
#>   0.109   0.001   0.110 

r

#> $iter
#> [1] 2
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] 0
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.302959
#> 
#> $mix$pr
#> [1] 0.5743433 0.4256567
#> 
#> 
#> $ll
#> [1] 0.09999965
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormad", val = 1, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.58105 fval: 0.730864
#> Iter: 2 lower: 0.58105 upper: 0.790525
#> current val: 0.738884 fval: 13.6509
#> Iter: 3 lower: 0.58105 upper: 0.659967
#> current val: 0.595263 fval: 1.08828
#> Iter: 4 lower: 0.588157 upper: 0.595263
#> current val: 0.59217 fval: 0.999937

#>    user  system elapsed 
#>   0.049   0.000   0.048 

r

#> $iter
#> [1] 8
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] 0
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.329429
#> 
#> $mix$pr
#> [1] 0.5921727 0.4078273
#> 
#> 
#> $ll
#> [1] 0.9999998
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormadw", val = 1, verbose = TRUE, order = -2))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.581502 fval: 0.784657
#> Iter: 2 lower: 0.581502 upper: 0.790751
#> current val: 0.738407 fval: 13.8668
#> Iter: 3 lower: 0.581502 upper: 0.659954
#> current val: 0.592739 fval: 1.07356
#> Iter: 4 lower: 0.58712 upper: 0.592739
#> current val: 0.590145 fval: 0.999969

#>    user  system elapsed 
#>   0.039   0.000   0.040 

r

#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -2.953193e-14
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.329319
#> 
#> $mix$pr
#> [1] 0.590146 0.409854
#> 
#> 
#> $ll
#> [1] 1.000001
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "nptll", val = 2, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.584421 fval: 886.323
#> Iter: 2 lower: 0.584421 upper: 0.79221
#> current val: 0.739589 fval: 907.155
#> Iter: 3 lower: 0.584421 upper: 0.662005
#> current val: 0.603503 fval: 887.362
#> Iter: 4 lower: 0.593962 upper: 0.603503
#> current val: 0.601896 fval: 887.26

#>    user  system elapsed 
#>   0.183   0.001   0.185 

r

#> $iter
#> [1] 6
#> 
#> $family
#> [1] "npt"
#> 
#> $min.gradient
#> [1] 0
#> 
#> $beta
#> [1] Inf
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.239117
#> 
#> $mix$pr
#> [1] 0.6018971 0.3981029
#> 
#> 
#> $ll
#> [1] 887.2603
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(tanh(data), method = "npnormcll", val = 2, beta = 4, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.310725 fval: 33.87
#> Iter: 2 lower: 0.310725 upper: 0.655363
#> current val: 0.580815 fval: 71.7638
#> Iter: 3 lower: 0.310725 upper: 0.44577
#> current val: 0.366489 fval: 35.6415
#> Iter: 4 lower: 0.338607 upper: 0.366489
#> current val: 0.366042 fval: 35.6147

#>    user  system elapsed 
#>   0.330   0.002   0.334 

r

#> $iter
#> [1] 10
#> 
#> $family
#> [1] "npnormc"
#> 
#> $min.gradient
#> [1] -2.333782e-07
#> 
#> $beta
#> [1] 4
#> 
#> $mix
#> $mix$pt
#> [1] -0.8982123 -0.8141842  0.0000000  0.8384127  0.9692529  0.9894580  0.9963746
#> [8]  0.9981044  0.9999507
#> 
#> $mix$pr
#> [1] 0.019301390 0.086812010 0.366040534 0.211994972 0.147307126 0.072424490
#> [7] 0.033360462 0.061209829 0.001549187
#> 
#> 
#> $ll
#> [1] 35.61458
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

data2 = rpois(1000, c(0, 2))
system.time(r <- estpi0(data2, method = "nppoisll", val = 2))

#>    user  system elapsed 
#>   0.005   0.000   0.006 

r

#> $iter
#> [1] 15
#> 
#> $family
#> [1] "nppois"
#> 
#> $min.gradient
#> [1] -5.684342e-14
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 1.948936
#> 
#> $mix$pr
#> [1] 0.5295802 0.4704198
#> 
#> 
#> $ll
#> [1] 1314.227
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"