Computing non-parametric mixing distribution with estimated proportion at 0

computing non-parametric mixing distribution with estimated proportion at 0

estpi0(v, method = "npnormll", ...)

estpi0.npnormll(v, beta, val, order, mix = NULL, ...)

estpi0.npnormcvm(v, beta, val, order, mix = NULL, ...)

estpi0.npnormad(v, beta, val, order, mix = NULL, ...)

estpi0.npnormcll(v, beta, val, order, mix = NULL, ...)

estpi0.nptll(v, beta, val, order, mix = NULL, ...)

estpi0.npnormllw(v, beta, val, order, mix = NULL, ...)

estpi0.npnormcvmw(v, beta, val, order, mix = NULL, ...)

estpi0.npnormadw(v, beta, val, order, mix = NULL, ...)

Arguments

v
observations

method
An implemented family; see details

...
parameters above passed to the specific method.

beta
structural parameter

val
thresholding function

order
the parameter for the binned version.

mix
initial mixing distribution.

Details

This is a function for computing non-parametric mixing distribution with estimated proportion at 0. Different families will have different threshold values.

The parameters are listed as follows:

tol: tolerance to stop the code.

verbose: logical; Whether to print the intermediate results.

It is not shown in the parameter section since various method have different default threshold values and this function essentially calls the class method in the object.

The full list of implemented families is in computemixdist.

IMPORTANT: although estpi0 is implemented for nptll and npnormcll, the theory is not yet done, and hence it is recommand to use with precaution.

Examples

data = rnorm(500, c(0, 2))
system.time(r <- estpi0(data, method = "npnormll", val = 2, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.519666 fval: 884.567
#> Iter: 2 lower: 0.519666 upper: 0.759833
#> current val: 0.69535 fval: 903.146
#> Iter: 3 lower: 0.519666 upper: 0.607508
#> current val: 0.530424 fval: 884.909
#> Iter: 4 lower: 0.525045 upper: 0.530424
#> current val: 0.527959 fval: 884.824

#>    user  system elapsed 
#>   0.044   0.000   0.045 

r

#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -1.553542e-05
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] -1.872467  0.000000  2.029319
#> 
#> $mix$pr
#> [1] 0.004690674 0.527961530 0.467347796
#> 
#> 
#> $ll
#> [1] 884.8242
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormllw", val = 2, verbose = TRUE, order = -2))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.519865 fval: 884.596
#> Iter: 2 lower: 0.519865 upper: 0.759933
#> current val: 0.695316 fval: 903.177
#> Iter: 3 lower: 0.519865 upper: 0.607591
#> current val: 0.529792 fval: 884.912
#> Iter: 4 lower: 0.524829 upper: 0.529792
#> current val: 0.527456 fval: 884.832

#>    user  system elapsed 
#>   0.216   0.001   0.217 

r

#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -1.452114e-05
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] -1.856877  0.000000  2.028579
#> 
#> $mix$pr
#> [1] 0.004809091 0.527458450 0.467732459
#> 
#> 
#> $ll
#> [1] 884.832
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormcvm", val = 0.1, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.511595 fval: 0.0547401
#> Iter: 2 lower: 0.511595 upper: 0.755797
#> current val: 0.653472 fval: 1.13416
#> Iter: 3 lower: 0.511595 upper: 0.582534
#> current val: 0.54777 fval: 0.11833
#> Iter: 4 lower: 0.529682 upper: 0.54777
#> current val: 0.541128 fval: 0.0991584

#>    user  system elapsed 
#>   0.048   0.001   0.048 

r

#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -1.110223e-16
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.00000 2.14288
#> 
#> $mix$pr
#> [1] 0.5414529 0.4585471
#> 
#> 
#> $ll
#> [1] 0.1000002
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormcvmw", val = 0.1, verbose = TRUE, order = -2))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.510792 fval: 0.055144
#> Iter: 2 lower: 0.510792 upper: 0.755396
#> current val: 0.652139 fval: 1.14195
#> Iter: 3 lower: 0.510792 upper: 0.581466
#> current val: 0.545998 fval: 0.117921
#> Iter: 4 lower: 0.528395 upper: 0.545998
#> current val: 0.539599 fval: 0.0992993
#> Iter: 5 lower: 0.539599 upper: 0.542799
#> current val: 0.539866 fval: 0.100002
#> Iter: 6 lower: 0.539733 upper: 0.539866
#> current val: 0.539865 fval: 0.0999984

#>    user  system elapsed 
#>   0.190   0.001   0.192 

r

#> $iter
#> [1] 2
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] 0
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.144194
#> 
#> $mix$pr
#> [1] 0.539866 0.460134
#> 
#> 
#> $ll
#> [1] 0.09999971
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormad", val = 1, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.521417 fval: 0.433048
#> Iter: 2 lower: 0.521417 upper: 0.760709
#> current val: 0.713417 fval: 12.1516
#> Iter: 3 lower: 0.521417 upper: 0.617417
#> current val: 0.571003 fval: 1.19142
#> Iter: 4 lower: 0.54621 upper: 0.571003
#> current val: 0.56301 fval: 0.992755
#> Iter: 5 lower: 0.56301 upper: 0.567007
#> current val: 0.563328 fval: 1

#>    user  system elapsed 
#>   0.254   0.002   0.257 

r

#> $iter
#> [1] 5
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] 0
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.179164
#> 
#> $mix$pr
#> [1] 0.5633282 0.4366718
#> 
#> 
#> $ll
#> [1] 1
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormadw", val = 1, verbose = TRUE, order = -2))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.520942 fval: 0.435774
#> Iter: 2 lower: 0.520942 upper: 0.760471
#> current val: 0.712778 fval: 12.3058
#> Iter: 3 lower: 0.520942 upper: 0.61686
#> current val: 0.569364 fval: 1.18949
#> Iter: 4 lower: 0.545153 upper: 0.569364
#> current val: 0.561563 fval: 0.993877
#> Iter: 5 lower: 0.561563 upper: 0.565464
#> current val: 0.561829 fval: 1

#>    user  system elapsed 
#>   0.157   0.000   0.159 

r

#> $iter
#> [1] 4
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] 0
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.181052
#> 
#> $mix$pr
#> [1] 0.5618285 0.4381715
#> 
#> 
#> $ll
#> [1] 1.000001
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "nptll", val = 2, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.519666 fval: 884.567
#> Iter: 2 lower: 0.519666 upper: 0.759833
#> current val: 0.69535 fval: 903.146
#> Iter: 3 lower: 0.519666 upper: 0.607508
#> current val: 0.530424 fval: 884.909
#> Iter: 4 lower: 0.525045 upper: 0.530424
#> current val: 0.527959 fval: 884.824

#>    user  system elapsed 
#>   0.124   0.000   0.124 

r

#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npt"
#> 
#> $min.gradient
#> [1] -1.55354e-05
#> 
#> $beta
#> [1] Inf
#> 
#> $mix
#> $mix$pt
#> [1] -1.872467  0.000000  2.029319
#> 
#> $mix$pr
#> [1] 0.004690674 0.527961530 0.467347796
#> 
#> 
#> $ll
#> [1] 884.8242
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(tanh(data), method = "npnormcll", val = 2, beta = 4, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.310865 fval: 42.5755
#> Iter: 2 lower: 0.310865 upper: 0.655433
#> current val: 0.578796 fval: 80.7019
#> Iter: 3 lower: 0.310865 upper: 0.444831
#> current val: 0.370209 fval: 44.5325
#> Iter: 4 lower: 0.340537 upper: 0.370209
#> current val: 0.370151 fval: 44.5287

#>    user  system elapsed 
#>   0.173   0.002   0.175 

r

#> $iter
#> [1] 9
#> 
#> $family
#> [1] "npnormc"
#> 
#> $min.gradient
#> [1] -1.870608e-06
#> 
#> $beta
#> [1] 4
#> 
#> $mix
#> $mix$pt
#>  [1] -0.9982743 -0.9243200 -0.8125827  0.0000000  0.8291025  0.9184382
#>  [7]  0.9722603  0.9858114  0.9905176  0.9976707  0.9990718
#> 
#> $mix$pr
#>  [1] 0.0005131967 0.0490159827 0.0504150032 0.3701508467 0.1378115913
#>  [6] 0.1222547625 0.1202721013 0.0041363243 0.0789202368 0.0645397562
#> [11] 0.0019701984
#> 
#> 
#> $ll
#> [1] 44.52869
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"