Computing non-parametric mixing distribution with estimated proportion at 0

computing non-parametric mixing distribution with estimated proportion at 0

estpi0(v, method = "npnormll", ...)

estpi0.npnormll(v, beta, val, order, mix = NULL, ...)

estpi0.npnormcvm(v, beta, val, order, mix = NULL, ...)

estpi0.npnormad(v, beta, val, order, mix = NULL, ...)

estpi0.npnormcll(v, beta, val, order, mix = NULL, ...)

estpi0.nptll(v, beta, val, order, mix = NULL, ...)

estpi0.npnormllw(v, beta, val, order, mix = NULL, ...)

estpi0.npnormcvmw(v, beta, val, order, mix = NULL, ...)

estpi0.npnormadw(v, beta, val, order, mix = NULL, ...)

Arguments

v
observations

method
An implemented family; see details

...
parameters above passed to the specific method.

beta
structural parameter

val
thresholding function

order
the parameter for the binned version.

mix
initial mixing distribution.

Details

This is a function for computing non-parametric mixing distribution with estimated proportion at 0. Different families will have different threshold values.

The parameters are listed as follows:

tol: tolerance to stop the code.

verbose: logical; Whether to print the intermediate results.

It is not shown in the parameter section since various method have different default threshold values and this function essentially calls the class method in the object.

The full list of implemented families is in computemixdist.

IMPORTANT: although estpi0 is implemented for nptll and npnormcll, the theory is not yet done, and hence it is recommand to use with precaution.

Examples

data = rnorm(500, c(0, 2))
system.time(r <- estpi0(data, method = "npnormll", val = 2, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.519666 fval: 884.567
#> Iter: 2 lower: 0.519666 upper: 0.759833
#> current val: 0.69535 fval: 903.146
#> Iter: 3 lower: 0.519666 upper: 0.607508
#> current val: 0.530424 fval: 884.909
#> Iter: 4 lower: 0.525045 upper: 0.530424
#> current val: 0.527959 fval: 884.824

#>    user  system elapsed 
#>   0.044   0.000   0.045 

r

#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -1.553542e-05
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] -1.872467  0.000000  2.029319
#> 
#> $mix$pr
#> [1] 0.004690674 0.527961530 0.467347796
#> 
#> 
#> $ll
#> [1] 884.8242
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormllw", val = 2, verbose = TRUE, order = -2))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.519865 fval: 884.596
#> Iter: 2 lower: 0.519865 upper: 0.759933
#> current val: 0.695316 fval: 903.177
#> Iter: 3 lower: 0.519865 upper: 0.607591
#> current val: 0.529792 fval: 884.912
#> Iter: 4 lower: 0.524829 upper: 0.529792
#> current val: 0.527456 fval: 884.832

#>    user  system elapsed 
#>   0.213   0.001   0.215 

r

#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -1.45222e-05
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] -1.856877  0.000000  2.028579
#> 
#> $mix$pr
#> [1] 0.004809091 0.527458450 0.467732459
#> 
#> 
#> $ll
#> [1] 884.832
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormcvm", val = 0.1, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.511595 fval: 0.107162
#> Iter: 2 lower: 0.255797 upper: 0.511595
#> current val: 0.47515 fval: 0.0598808
#> Iter: 3 lower: 0.493373 upper: 0.511595
#> current val: 0.509688 fval: 0.0598825
#> Iter: 4 lower: 0.510641 upper: 0.511595
#> current val: 0.511495 fval: 0.0600272
#> Iter: 5 lower: 0.511545 upper: 0.511595
#> current val: 0.511589 fval: 0.0600412
#> Iter: 6 lower: 0.511592 upper: 0.511595
#> current val: 0.511594 fval: 0.0600414

#>    user  system elapsed 
#>   0.041   0.000   0.042 

r

#> $iter
#> [1] 2
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] 1.110223e-16
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.041728
#> 
#> $mix$pr
#> [1] 0.5115946 0.4884054
#> 
#> 
#> $ll
#> [1] 0.06004139
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormcvmw", val = 0.1, verbose = TRUE, order = -2))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.510792 fval: 0.106982
#> Iter: 2 lower: 0.255396 upper: 0.510792
#> current val: 0.476423 fval: 0.0590176
#> Iter: 3 lower: 0.493608 upper: 0.510792
#> current val: 0.509067 fval: 0.0590995
#> Iter: 4 lower: 0.50993 upper: 0.510792
#> current val: 0.510705 fval: 0.0593779
#> Iter: 5 lower: 0.510749 upper: 0.510792
#> current val: 0.510788 fval: 0.0593971
#> Iter: 6 lower: 0.51079 upper: 0.510792
#> current val: 0.510792 fval: 0.0593979

#>    user  system elapsed 
#>   0.104   0.000   0.104 

r

#> $iter
#> [1] 2
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] 0
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 2.046075
#> 
#> $mix$pr
#> [1] 0.5107923 0.4892077
#> 
#> 
#> $ll
#> [1] 0.05939799
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormad", val = 1, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.521417 fval: 2.93358
#> Iter: 2 lower: 0 upper: 0.260709
#> current val: 0.191201 fval: 6.40115
#> Iter: 3 lower: 0 upper: 0.0956007
#> current val: 0.0838447 fval: 2.31905
#> Iter: 4 lower: 0 upper: 0.0419224
#> current val: 0.00686774 fval: 3.61645
#> Iter: 5 lower: 0 upper: 0.00343387
#> current val: 0.000558232 fval: 3.7412
#> Iter: 6 lower: 0 upper: 0.000279116
#> current val: 0.000219458 fval: 1.35363
#> Iter: 7 lower: 0 upper: 0.000109729
#> current val: 6.34518e-05 fval: 1.35363
#> Iter: 8 lower: 0 upper: 3.17259e-05
#> current val: 1.83462e-05 fval: 1.35363
#> Iter: 9 lower: 0 upper: 9.17309e-06
#> current val: 5.30453e-06 fval: 1.35363
#> Iter: 10 lower: 0 upper: 2.65226e-06
#> current val: 1.53373e-06 fval: 1.35363

#>    user  system elapsed 
#>   0.090   0.000   0.092 

r

#> $iter
#> [1] 1
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -0.01664243
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 0.000000 1.821063
#> 
#> $mix$pr
#> [1] 4.639705e-01 4.434540e-07 5.360291e-01
#> 
#> 
#> $ll
#> [1] 1.353628
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "npnormadw", val = 1, verbose = TRUE, order = -2))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.521331 fval: 3.07228
#> Iter: 2 lower: 0 upper: 0.260665
#> current val: 0.192728 fval: 6.533
#> Iter: 3 lower: 0 upper: 0.0963639
#> current val: 0.00409048 fval: 11.8732
#> Iter: 4 lower: 0 upper: 0.00204524
#> current val: 9.62473e-05 fval: 1.35935
#> Iter: 5 lower: 0 upper: 4.81237e-05
#> current val: 2.76188e-05 fval: 1.35935
#> Iter: 6 lower: 0 upper: 1.38094e-05
#> current val: 7.92542e-06 fval: 1.35935
#> Iter: 7 lower: 0 upper: 3.96271e-06
#> current val: 2.27425e-06 fval: 1.35935
#> Iter: 8 lower: 0 upper: 1.13713e-06
#> current val: 6.52613e-07 fval: 1.35935

#>    user  system elapsed 
#>   0.058   0.000   0.058 

r

#> $iter
#> [1] 1
#> 
#> $family
#> [1] "npnorm"
#> 
#> $min.gradient
#> [1] -0.00027761
#> 
#> $beta
#> [1] 1
#> 
#> $mix
#> $mix$pt
#> [1] 0.000000 0.000000 1.820884
#> 
#> $mix$pr
#> [1] 4.614721e-01 1.872718e-07 5.385277e-01
#> 
#> 
#> $ll
#> [1] 1.359351
#> 
#> $flag
#> [1] "d1"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(data, method = "nptll", val = 2, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.519666 fval: 884.567
#> Iter: 2 lower: 0.519666 upper: 0.759833
#> current val: 0.69535 fval: 903.146
#> Iter: 3 lower: 0.519666 upper: 0.607508
#> current val: 0.530424 fval: 884.909
#> Iter: 4 lower: 0.525045 upper: 0.530424
#> current val: 0.527959 fval: 884.824

#>    user  system elapsed 
#>   0.133   0.000   0.134 

r

#> $iter
#> [1] 7
#> 
#> $family
#> [1] "npt"
#> 
#> $min.gradient
#> [1] -1.553531e-05
#> 
#> $beta
#> [1] Inf
#> 
#> $mix
#> $mix$pt
#> [1] -1.872467  0.000000  2.029319
#> 
#> $mix$pr
#> [1] 0.004690674 0.527961530 0.467347796
#> 
#> 
#> $ll
#> [1] 884.8242
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"

system.time(r <- estpi0(tanh(data), method = "npnormcll", val = 2, beta = 4, verbose = TRUE))

#> Iter: 1 lower: 0 upper: 1
#> current val: 0.310868 fval: 42.5754
#> Iter: 2 lower: 0.310868 upper: 0.655434
#> current val: 0.578796 fval: 80.6909
#> Iter: 3 lower: 0.310868 upper: 0.444832
#> current val: 0.370186 fval: 44.5306
#> Iter: 4 lower: 0.340527 upper: 0.370186
#> current val: 0.37015 fval: 44.5283

#>    user  system elapsed 
#>   0.198   0.003   0.200 

r

#> $iter
#> [1] 9
#> 
#> $family
#> [1] "npnormc"
#> 
#> $min.gradient
#> [1] -2.323069e-06
#> 
#> $beta
#> [1] 4
#> 
#> $mix
#> $mix$pt
#>  [1] -0.9982743 -0.9243175 -0.8125801  0.0000000  0.8290951  0.9184491
#>  [7]  0.9723275  0.9856815  0.9904156  0.9976675  0.9990713
#> 
#> $mix$pr
#>  [1] 0.0005132107 0.0490178222 0.0504132856 0.3701497285 0.1378038541
#>  [6] 0.1223190447 0.1210394352 0.0008609066 0.0812226886 0.0646628368
#> [11] 0.0019971868
#> 
#> 
#> $ll
#> [1] 44.52828
#> 
#> $flag
#> [1] "d0"
#> 
#> $convergence
#> [1] 0
#> 
#> attr(,"class")
#> [1] "nspmix"