Find the posterior mean

Find the posterior mean given the observations and the mixing distribution based on the family in the result

posteriormean(x, result, fun = function(x) x)

posteriormean.npnorm(x, result, fun = function(x) x)

posteriormean.npt(x, result, fun = function(x) x)

posteriormean.npnormc(x, result, fun = function(x) x)

Arguments

x
a vector of observations

result
an object of class nspmix

fun
the function to transform the mean. It finds the posterior mean of fun(x). The function fun must be vectorised.

Examples

data = rnorm(500, c(0, 2))
r1 = computemixdist(data, pi0 = 0.5)
posteriormean(data, r1)

#>   [1] 0.52334114 1.62947189 1.62598239 2.08145705 0.07356137 2.06713115
#>   [7] 1.15289892 1.24380358 1.44558215 2.08981688 0.55483761 2.01915716
#>  [13] 0.05360459 2.10845214 0.09061258 1.50533571 0.47154803 1.80509202
#>  [19] 1.54052908 1.15659444 0.09785373 1.87264563 0.50809695 2.46058101
#>  [25] 0.03953543 2.04465673 0.78188255 2.04328280 0.04076686 1.90729464
#>  [31] 0.10376423 0.31745909 0.06099066 0.84556592 1.50391493 1.18386448
#>  [37] 0.78981625 1.97067361 0.06960984 1.95541427 0.50290346 0.74265246
#>  [43] 0.63467220 1.74946149 0.15071460 1.00387285 0.16631552 0.80615025
#>  [49] 0.20918777 1.67311314 0.14939769 2.04435990 0.08753974 1.67339101
#>  [55] 0.75779646 2.06256976 0.18495976 1.78663325 0.98811747 1.89280471
#>  [61] 0.36125975 1.58426443 0.25319813 1.75639658 0.51713641 1.66490097
#>  [67] 0.13625361 1.70929149 0.10927319 2.08030087 0.05083247 2.05348991
#>  [73] 0.21359461 0.21029560 0.09689262 2.04838404 0.03981283 1.81872405
#>  [79] 0.04249195 1.79444915 0.91229069 1.62405965 0.84652097 1.99432955
#>  [85] 1.10021522 2.06351028 0.88914134 1.91555229 0.23894701 1.82104288
#>  [91] 0.37881604 0.98851178 0.09053432 1.38587334 0.09971970 1.56802688
#>  [97] 0.19929697 1.09590479 0.01899066 2.12695462 0.06773766 2.03487683
#> [103] 0.08771247 1.94595397 0.17598500 1.96970204 0.02473635 1.86292305
#> [109] 0.17268040 2.06782991 0.13149084 1.48245278 0.02492147 1.83279927
#> [115] 0.54925064 1.60303068 0.09190149 1.90616120 0.04389693 1.89105660
#> [121] 0.90101452 1.65219491 0.12099326 1.51467222 0.45429827 0.63630000
#> [127] 0.06559833 0.46828875 0.14464707 1.73122028 0.24789061 1.64982948
#> [133] 0.02510268 0.63285057 0.13854532 1.87727091 0.29004032 0.84096541
#> [139] 0.91025447 1.77653892 0.12808056 1.87464777 0.74755447 1.96984774
#> [145] 0.11965732 0.71662836 0.75694739 0.84443488 0.08002414 2.04739621
#> [151] 0.09930614 1.58743640 0.44584821 2.06370911 0.05538304 1.99676622
#> [157] 0.20769484 2.07045668 0.85002887 1.72623529 0.07323430 0.61514372
#> [163] 0.05376770 1.43583425 0.04339047 1.66613257 0.09812789 2.12173591
#> [169] 0.98873713 1.94811340 0.09508673 2.12686306 0.14213805 1.73176025
#> [175] 0.02331212 2.04247111 0.19952069 0.87341593 0.21449023 2.04072477
#> [181] 0.24620005 1.88713256 0.94840890 0.97187247 0.14850794 1.98741413
#> [187] 0.03609380 2.05213528 0.04982600 1.87467074 0.28532280 1.29551127
#> [193] 0.30297104 1.63315743 0.78600275 0.57715780 1.18930073 1.58963222
#> [199] 0.07446291 0.82718304 0.96868474 0.43829564 0.26854578 1.29370731
#> [205] 0.07195779 1.65465241 0.23765693 1.49834812 0.07390080 1.99434592
#> [211] 0.58221123 2.08840034 0.05508991 0.88070876 0.06784487 0.74741208
#> [217] 1.74366875 0.10178030 0.03553867 1.13187767 0.15066849 1.66317105
#> [223] 0.90342304 2.05928998 0.05188025 1.85799961 0.02800980 2.01907670
#> [229] 0.06308301 1.93136599 0.05469741 1.76159633 0.21459071 0.81772664
#> [235] 0.38819232 2.78763084 0.22397798 1.45847737 0.10097955 1.50538269
#> [241] 0.28161536 1.01764842 1.40371012 1.96991823 0.81406075 2.06754225
#> [247] 0.04920151 0.95171263 0.35946261 1.40232941 0.47126812 1.71243320
#> [253] 0.61864335 0.39600069 0.61600453 0.27546406 0.21101689 1.93835173
#> [259] 0.06959755 2.07661047 0.30379303 0.39441426 1.43951407 1.88963264
#> [265] 0.14266958 1.99364199 0.04154425 1.75336478 0.33553224 0.05961604
#> [271] 0.11723904 1.20097969 0.37440592 2.04431040 0.27258775 1.08841672
#> [277] 0.19289515 0.87863407 0.05385639 0.10265975 0.14593774 1.96675821
#> [283] 0.32960960 0.84109491 0.10785043 2.04600090 0.64671757 1.95572979
#> [289] 0.13347170 1.72255793 0.04788545 1.77320851 0.40294015 1.78136886
#> [295] 0.30217670 0.11816293 0.33723468 1.53230671 0.04107483 1.55267415
#> [301] 0.29092727 0.93312044 0.22431218 0.71864891 0.04961759 2.10331564
#> [307] 0.67670877 0.41270209 0.15111743 1.81810652 0.13619770 2.04253528
#> [313] 1.87097360 1.31016601 0.74013582 2.04480657 0.40388749 1.93831151
#> [319] 1.11720156 1.81341897 0.02253117 1.93203334 0.04001587 1.27823622
#> [325] 0.15069465 0.17998107 0.16393277 2.01848621 0.08710400 0.47287205
#> [331] 0.13731726 1.31546207 0.34997965 1.26910868 0.07848341 1.90531636
#> [337] 0.46091080 2.04735807 0.95404205 1.23575730 0.79704661 0.59354968
#> [343] 0.27784043 0.22136225 0.40301757 0.81709254 0.19377628 1.87662251
#> [349] 0.03281806 2.01991056 0.98543760 1.87783820 0.04218148 0.47411454
#> [355] 0.03271230 1.43225701 0.18425377 1.90384689 0.69581506 1.36708933
#> [361] 0.99009943 1.49773274 0.03279236 1.99161526 0.50333815 1.81203805
#> [367] 0.45780482 2.07115493 0.13172503 2.03353988 0.09609415 2.06763295
#> [373] 0.16846260 1.60433161 0.55027061 1.97096242 1.32925620 1.93821388
#> [379] 0.89648780 1.98624460 0.10610443 2.05816693 0.05641394 2.03480024
#> [385] 0.06150957 2.02998922 0.18278094 1.77121740 1.71315231 1.13098369
#> [391] 0.93601406 1.21239563 0.22556448 1.67916853 1.71908354 2.05857626
#> [397] 0.24382355 1.81952234 0.24622065 0.55137924 0.02809845 1.63668619
#> [403] 0.15865256 1.16503604 0.06172109 1.72258171 1.61627240 1.51315677
#> [409] 1.01672112 1.59439317 0.34926051 2.15173223 0.34161514 1.24282782
#> [415] 0.02759829 1.37206670 0.88662612 1.41533112 0.71340293 1.59878961
#> [421] 0.03584842 0.25479918 0.05375008 1.75649126 0.10362198 1.10355855
#> [427] 0.24925139 0.79856804 0.85778650 1.29206432 0.16629353 0.98532760
#> [433] 0.49661122 1.91072321 0.11152038 0.22756250 0.27780416 1.92599432
#> [439] 0.45764278 2.03783506 0.06745763 1.15305028 0.04035883 2.01442457
#> [445] 0.10921708 1.09291307 0.20217893 1.63403482 0.08564869 2.12026534
#> [451] 0.07458504 2.04701834 0.53494598 2.10353614 0.25715696 1.75736460
#> [457] 0.08788554 0.69802390 0.13737596 1.29663470 0.10086576 1.91011230
#> [463] 0.11116533 0.03547352 0.15776576 1.84301640 0.19471222 1.84830443
#> [469] 0.11066297 2.02682300 0.13002734 1.42126660 0.81533665 2.05173639
#> [475] 0.11575941 1.67706414 0.07757658 1.81751951 1.58689796 2.01701449
#> [481] 0.56997144 1.93731502 0.15714693 0.36246001 0.65500999 1.82429223
#> [487] 0.10047334 2.06178632 0.74094479 2.05325790 0.65363479 1.12904446
#> [493] 0.14725395 0.24921821 0.15962915 0.30562341 0.25809399 1.82848608
#> [499] 0.08763675 1.73952136

r2 = computemixdist(data, pi0 = 0.5, method = "nptll") # equivalent to normal
posteriormean(data, r2, fun = function(x) x^2)

#>   [1] 1.01239564 3.35412804 3.34663628 4.41204324 0.09980929 4.35227084
#>   [7] 2.33749426 2.53053949 2.96042958 4.45220735 1.07805261 4.21100741
#>  [13] 0.06397535 4.55252821 0.13169720 3.08813323 0.90464970 3.73273069
#>  [19] 3.16344330 2.34533466 0.14547238 3.87968985 0.98065311 6.79047835
#>  [25] 0.04045669 4.27866898 1.55371038 4.27472453 0.04242614 3.95566354
#>  [31] 0.15679764 0.58623471 0.07698771 1.68773657 3.08509446 2.40321010
#>  [37] 1.57039457 4.09694336 0.09255930 4.06249531 0.96984437 1.47126764
#>  [43] 1.24486956 3.61240372 0.24858649 2.02185001 0.27959059 1.60475588
#>  [49] 0.36561902 3.44790648 0.24597851 4.27781284 0.12588894 3.44850410
#>  [55] 1.50308177 4.33561100 0.31687126 3.69275067 1.98854093 3.92382388
#>  [61] 0.67634734 3.25713989 0.45484080 3.62737907 0.99947284 3.43024746
#>  [67] 0.22003719 3.52577770 0.16741146 4.40678435 0.05919080 4.30529832
#>  [73] 0.37451718 0.36785508 0.14363738 4.28961915 0.04089848 3.76229735
#>  [79] 0.04521920 3.70967190 1.82840461 3.34250868 1.68974828 4.15137262
#>  [85] 2.22578914 4.33896051 1.77957467 3.97386212 0.42586838 3.76733047
#>  [91] 0.71256728 1.98937441 0.13154899 2.83301159 0.14904041 3.22233877
#>  [97] 0.34568147 2.21665543 0.01248987 4.66100589 0.08914700 4.25148981
#> [103] 0.12621480 4.04131265 0.29889733 4.09473769 0.01904897 3.85846011
#> [109] 0.29229201 4.35491833 0.21068093 3.03920503 0.01928150 3.79286661
#> [115] 1.06639922 3.29738311 0.13414043 3.95316883 0.04752123 3.91998982
#> [121] 1.80461578 3.40293655 0.19015295 3.10810520 0.86883218 1.24827631
#> [127] 0.08526713 0.89787939 0.23658347 3.57304349 0.44404243 3.39785370
#> [133] 0.01951025 1.24105742 0.22454784 3.88980131 0.53003719 1.67804673
#> [139] 1.82410840 3.67091150 0.20399735 3.88406579 1.48156408 4.09506833
#> [145] 0.18755067 1.41663102 1.50129768 1.68535421 0.11179085 4.28667956
#> [151] 0.14824900 3.26394028 0.85129975 4.33967416 0.06707549 4.15708084
#> [157] 0.36260660 4.36510857 1.69713778 3.56229400 0.09920689 1.20401551
#> [163] 0.06425873 2.93961505 0.04668872 3.43289545 0.14599617 4.62983239
#> [169] 1.98985074 4.04613796 0.14019480 4.66045595 0.23163004 3.57420805
#> [175] 0.01730078 4.27241569 0.34613190 1.74642050 0.37632667 4.26749938
#> [181] 0.44060493 3.91138847 1.90464555 1.95420944 0.24421734 4.13529242
#> [187] 0.03507287 4.30105429 0.05746921 3.88411600 0.52038702 2.64051634
#> [193] 0.55651746 3.36204170 1.56237447 1.12463678 2.41475159 3.26864833
#> [199] 0.10147188 1.64902467 1.94747413 0.83563717 0.48611743 2.63667741
#> [205] 0.09685967 3.40821773 0.42324923 3.07318923 0.10043491 4.15141090
#> [211] 1.13518984 4.44515014 0.06656297 1.76179446 0.08934199 1.48126497
#> [217] 3.59989988 0.15298863 0.03422278 2.29290759 0.24849514 3.42652828
#> [223] 1.80969638 4.32425629 0.06099206 3.84772155 0.02332417 4.21080732
#> [229] 0.08073369 4.00884605 0.06587764 3.63861144 0.37652969 1.62911816
#> [235] 0.73193241 8.90341646 0.39551695 2.98797229 0.15145337 3.08823373
#> [241] 0.51280724 2.05098335 2.87105600 4.09522833 1.62140253 4.35382525
#> [247] 0.05640543 1.91162261 0.67264276 2.86811047 0.90406822 3.53254618
#> [253] 1.21133461 0.74806996 1.20581571 0.50023941 0.36931125 4.02436671
#> [259] 0.09253685 4.39050753 0.55820217 0.74479052 2.94747192 3.91686775
#> [265] 0.23267890 4.14976615 0.04368005 3.62083153 0.62337068 0.07454003
#> [271] 0.18284634 2.43955129 0.70346391 4.27767027 0.49436644 2.20079041
#> [277] 0.33280330 1.75742054 0.06441289 0.15467620 0.23913392 4.08806576
#> [283] 0.61119348 1.67831948 0.16466528 4.28257435 1.27008345 4.06320383
#> [289] 0.21456920 3.55436607 0.05417519 3.66370960 0.76241965 3.68135909
#> [295] 0.55488956 0.18464262 0.62687228 3.14584162 0.04292194 3.18944999
#> [301] 0.53185217 1.87236545 0.39619361 1.42087152 0.05711381 4.52370579
#> [307] 1.33290908 0.78261796 0.24938454 3.76095717 0.21992722 4.27259766
#> [313] 3.87603653 2.67170824 1.46598213 4.27910200 0.76437917 4.02427721
#> [319] 2.26179122 3.75078667 0.01637427 4.01032682 0.04122255 2.60376025
#> [325] 0.24854697 0.30689419 0.27484310 4.20934071 0.12506727 0.90740038
#> [331] 0.22213005 2.68298314 0.65310452 2.58434523 0.10892159 3.95130984
#> [337] 0.88255824 4.28656664 1.91654239 2.51343704 1.58560291 1.15887588
#> [343] 0.50509336 0.39022244 0.76257979 1.62778353 0.33457452 3.88838332
#> [349] 0.03013999 4.21288376 1.98287651 3.89104202 0.04471373 0.90998184
#> [355] 0.02998428 2.93197788 0.31545559 3.94807735 1.37296643 2.79296402
#> [361] 1.99273041 3.07187328 0.03010214 4.14504109 0.97074893 3.74779137
#> [367] 0.87611028 4.36788180 0.21114039 4.24791641 0.14211439 4.35416943
#> [373] 0.28387203 3.30017379 1.06852648 4.09759938 2.71235602 4.02405997
#> [379] 1.79506784 4.13258877 0.16129991 4.32047885 0.06888267 4.25128429
#> [385] 0.07791450 4.23856114 0.31250317 3.65940469 3.53409558 2.29101189
#> [391] 1.87847414 2.46379855 0.39872958 3.46093155 3.54687713 4.32184939
#> [397] 0.43577436 3.76402994 0.44064681 1.07083875 0.02344446 3.36961972
#> [403] 0.26433776 2.36324669 0.07829271 3.55441735 3.32579437 3.10486310
#> [409] 2.04902196 3.27885737 0.65162353 4.81269994 0.63588475 2.52846534
#> [415] 0.02276870 2.80357409 1.77427091 2.89585130 1.40986243 3.28828619
#> [421] 0.03469643 0.45810005 0.06422811 3.62758357 0.15652427 2.23287418
#> [427] 0.44681009 1.58880349 1.71348185 2.63318117 0.27954676 1.98264400
#> [433] 0.95675274 3.96321442 0.17175570 0.40277696 0.50501927 3.99694086
#> [439] 0.87577392 4.25950659 0.08863793 2.33781537 0.04177125 4.19933270
#> [445] 0.16730309 2.21031655 0.35148585 3.36392580 0.12232664 4.62111772
#> [451] 0.10169734 4.28556249 1.03657575 4.52492775 0.46290128 3.62946988
#> [457] 0.12654138 1.37759902 0.22224557 2.64290716 0.15123531 3.96186847
#> [463] 0.17106879 0.03412336 0.26257553 3.81508613 0.33645639 3.82659689
#> [469] 0.17009720 4.23036707 0.20781105 2.90851828 1.62408784 4.29981665
#> [475] 0.17997209 3.45640466 0.10723645 3.75968330 3.26278587 4.20569862
#> [481] 1.10963337 4.02206045 0.26134617 0.67882187 1.28744786 3.77438528
#> [487] 0.15048347 4.33285342 1.46768111 4.30456685 1.28456783 2.28689983
#> [493] 0.24173640 0.44674260 0.26627913 0.56195423 0.46480989 3.78349422
#> [499] 0.12607195 3.59095019

data = runif(500, min = -0.5, max = 0.5)
r3 = computemixdist(data, method = "npnormcll", beta = 100)
posteriormean(data, r3)

#>   [1] -0.361434595 -0.360495121  0.177725565 -0.203291204  0.368325176
#>   [6]  0.303317599  0.241969890 -0.075889603  0.367086144  0.369217132
#>  [11] -0.071804019 -0.072970888 -0.077635952  0.296228071  0.348610214
#>  [16]  0.358246051  0.218169588 -0.361780668  0.293560028  0.325263599
#>  [21] -0.071182608 -0.355558168 -0.063384662 -0.337505969  0.096166843
#>  [26]  0.066246633 -0.042603907  0.340314091  0.291180663 -0.361564833
#>  [31] -0.087653718  0.350585930 -0.346130972 -0.360725699  0.363375637
#>  [36] -0.358387868 -0.297486721  0.344594309  0.125594476 -0.173850149
#>  [41] -0.361635404  0.290157202 -0.238464286  0.330482963  0.369837850
#>  [46]  0.346601501 -0.022065469 -0.082179371  0.297306812  0.167564862
#>  [51]  0.227518828 -0.053824916 -0.360988055  0.292637010  0.307183893
#>  [56] -0.071906710 -0.354276810  0.368508839  0.270798573  0.365966541
#>  [61] -0.359422819  0.350089286 -0.060991206  0.354371688  0.224361867
#>  [66]  0.369639183 -0.081942492  0.156217924 -0.071693429 -0.166849292
#>  [71]  0.234809752 -0.360734418  0.272740306  0.259870968  0.364164431
#>  [76] -0.072880149  0.370152356 -0.002500766 -0.069784241  0.172432782
#>  [81]  0.358945935 -0.345706230  0.265438453  0.212019068 -0.032312037
#>  [86] -0.073712080 -0.064075475  0.334525040 -0.352836327  0.102939754
#>  [91]  0.331293631 -0.080472291  0.337506270 -0.035733132 -0.134791880
#>  [96]  0.352718784 -0.359363383 -0.125080126 -0.046919435  0.241704604
#> [101]  0.346019112 -0.236011813  0.367163217  0.369797161 -0.170980580
#> [106]  0.349903764 -0.070278790 -0.077330648  0.366901340  0.338895309
#> [111] -0.094929565 -0.337252107 -0.353657002  0.253744178  0.163185575
#> [116] -0.361697798  0.369777503 -0.355567851 -0.304108830  0.209409426
#> [121] -0.119701298  0.368527589 -0.069147575  0.338863985 -0.036839215
#> [126]  0.125466615 -0.091666598  0.308568564  0.211128433 -0.345345581
#> [131] -0.247601303  0.239003498 -0.071148323 -0.361780451  0.261027575
#> [136]  0.220343229  0.001956744  0.001062029  0.264684207 -0.060738676
#> [141] -0.360874100 -0.102346185 -0.330415376 -0.077546209 -0.072198639
#> [146]  0.222739306 -0.361370229 -0.361618221  0.046445526 -0.077879278
#> [151]  0.319776991  0.286680343 -0.070954352 -0.359859522 -0.361462933
#> [156] -0.327330550 -0.076275170  0.088616996 -0.360366591 -0.335095187
#> [161] -0.077194287 -0.113506319 -0.360916015 -0.069366769 -0.361700423
#> [166]  0.023538216  0.305504702 -0.358711790  0.103901544 -0.056292359
#> [171] -0.358779960 -0.053782717 -0.360959838 -0.075079629 -0.359705229
#> [176]  0.362982466 -0.156920749 -0.071381793  0.352489812 -0.068910482
#> [181]  0.177882488  0.315653433 -0.349915881  0.207472161  0.329303979
#> [186]  0.368448994 -0.360876819 -0.071969737  0.366522967 -0.360527249
#> [191] -0.076402792 -0.361136781  0.362240352  0.165919925 -0.354017120
#> [196]  0.366875231 -0.060807049 -0.356880716 -0.139769977 -0.075720269
#> [201]  0.158934414 -0.074571515 -0.350983519  0.353452810 -0.359664284
#> [206] -0.355096472 -0.349224413  0.368615331  0.367384560 -0.356129174
#> [211]  0.088011077 -0.222058052  0.336441675 -0.097661943 -0.308414043
#> [216]  0.327456369 -0.069688731 -0.076877032 -0.068204870 -0.361611553
#> [221] -0.077808510  0.351034003 -0.071372095 -0.050111774  0.130900486
#> [226]  0.242865301 -0.319571383  0.361658032  0.285822739 -0.077796871
#> [231]  0.367620993 -0.360126289 -0.065017794 -0.361652940  0.221798562
#> [236] -0.079260849  0.063102276 -0.358655579 -0.061823632  0.167074858
#> [241]  0.202488946 -0.357107671 -0.064642717 -0.360866448 -0.338692784
#> [246]  0.347129166 -0.067931311 -0.067788990  0.353461519  0.364427061
#> [251]  0.234068896 -0.070443284  0.184449406 -0.322758120 -0.069859909
#> [256] -0.034875948  0.229904249  0.229200211 -0.361696982  0.037081030
#> [261] -0.065955032  0.118814860 -0.361218028 -0.351594638 -0.095557805
#> [266] -0.349779954 -0.360914002 -0.068398915  0.356229614 -0.334800108
#> [271] -0.269882295  0.208467298 -0.127463647 -0.068671377  0.354531405
#> [276] -0.072493058  0.296538630 -0.069908515 -0.083876578 -0.325303329
#> [281] -0.361694335  0.337148430  0.302150583 -0.036298934 -0.088604498
#> [286] -0.078720537 -0.280403043  0.332862119 -0.158572143  0.044371235
#> [291] -0.357838489  0.363121296 -0.360985093 -0.361694899 -0.353768277
#> [296] -0.358002182  0.342733374  0.210554905 -0.355394993 -0.157161450
#> [301] -0.081708427  0.254402933  0.222925483  0.183043697  0.201809689
#> [306] -0.268866352 -0.361672145  0.367395333  0.362968901  0.299550488
#> [311]  0.036161690 -0.303168715  0.332545030 -0.294067805 -0.348014615
#> [316]  0.148625442 -0.109100821 -0.361663157 -0.077401485 -0.085990372
#> [321]  0.254791288 -0.097009795 -0.359620052 -0.077777675 -0.122634529
#> [326]  0.369992052  0.247848118 -0.091211736  0.343854129  0.313299655
#> [331] -0.344258626 -0.361417538  0.361688795 -0.361377446 -0.071032187
#> [336]  0.353276577 -0.073859798 -0.055479850  0.260315721 -0.079056031
#> [341]  0.310797643 -0.236933768 -0.306780376 -0.093231793  0.308322476
#> [346] -0.328090957 -0.349155964 -0.114216513  0.302509526  0.211304688
#> [351] -0.152539023 -0.323952593 -0.317715152  0.354106248  0.340114965
#> [356] -0.125539376  0.208925077 -0.361624737 -0.359581986 -0.070284788
#> [361]  0.309538702  0.216111809 -0.077313859  0.306192840  0.233700223
#> [366] -0.361507172  0.205214263 -0.090531812 -0.350954135  0.352821754
#> [371] -0.069926449  0.308320087 -0.059617335  0.267723180 -0.104024889
#> [376] -0.128442326 -0.361686965 -0.299739922 -0.138133930 -0.018645650
#> [381]  0.214285754  0.222045571  0.327469686 -0.085325245  0.348422265
#> [386] -0.071012353  0.271031497  0.321846409 -0.074241274  0.278113926
#> [391] -0.084293348 -0.073544972 -0.189406866 -0.279260084  0.322884235
#> [396]  0.293478200 -0.207255360 -0.054938725  0.057248103 -0.075349186
#> [401]  0.321005522  0.250177966  0.168414390 -0.068037128  0.043324929
#> [406]  0.367568554 -0.099508168 -0.360119027 -0.358811437 -0.360754616
#> [411] -0.077875168 -0.360491139 -0.336198570  0.367946576  0.007090511
#> [416] -0.133626353 -0.046539771  0.153596096 -0.083124891 -0.360565356
#> [421]  0.236526919 -0.258886103  0.362799698  0.100216410 -0.038235868
#> [426]  0.292445794  0.227001982  0.364224393 -0.354501631 -0.099243353
#> [431] -0.068849835 -0.007503610 -0.046709710  0.311902361  0.205579813
#> [436] -0.360055174  0.180045326 -0.072354461  0.032971069 -0.063684465
#> [441]  0.182894511  0.369764938 -0.341551464 -0.073932555  0.046761852
#> [446]  0.099488658 -0.361447052  0.287134668  0.317609861 -0.320101427
#> [451]  0.263161216  0.185763169 -0.063639838 -0.041410648  0.366638722
#> [456] -0.357575397 -0.164930724 -0.076226621  0.356466696  0.289183393
#> [461]  0.267684745  0.217933992  0.369415967  0.309054593  0.356304628
#> [466] -0.087145204  0.283552280 -0.358385773  0.237222440  0.277604789
#> [471]  0.225276030 -0.361618075  0.308090446 -0.346212408  0.284032527
#> [476] -0.247791006  0.352423280  0.368452025 -0.071464291  0.343787983
#> [481]  0.367419684 -0.074540620 -0.060002339  0.369180156 -0.360407340
#> [486] -0.360610762  0.356603925 -0.084373436 -0.360998114 -0.361765832
#> [491] -0.044369288 -0.349310839 -0.334576457 -0.049686977  0.358060679
#> [496] -0.113888578  0.258207523 -0.030415471 -0.265999759 -0.357701946