Find the posterior mean

Find the posterior mean given the observations and the mixing distribution based on the family in the result

posteriormean(x, result, fun = function(x) x)

posteriormean.npnorm(x, result, fun = function(x) x)

posteriormean.npt(x, result, fun = function(x) x)

posteriormean.npnormc(x, result, fun = function(x) x)

Arguments

x
a vector of observations

result
an object of class nspmix

fun
the function to transform the mean. It finds the posterior mean of fun(x). The function fun must be vectorised.

Examples

data = rnorm(500, c(0, 2))
r1 = computemixdist(data, pi0 = 0.5)
posteriormean(data, r1)

#>   [1] 0.52309011 1.62948136 1.62598709 2.08209915 0.07370999 2.06782611
#>   [7] 1.15249777 1.24344423 1.44537649 2.09041181 0.55456222 2.01984561
#>  [13] 0.05374445 2.10891039 0.09075926 1.50519198 0.47134029 1.80537596
#>  [19] 1.54042541 1.15619466 0.09799791 1.87305318 0.50785825 2.45830920
#>  [25] 0.03965607 2.04537134 0.78148088 2.04399688 0.04088983 1.90776910
#>  [31] 0.10390580 0.31740001 0.06113564 0.84514546 1.50376965 1.18347568
#>  [37] 0.78941183 1.97127389 0.06975789 1.95598425 0.50266905 0.74226610
#>  [43] 0.63434225 1.74965109 0.15082356 1.00343917 0.16641070 0.80574052
#>  [49] 0.20924139 1.67318444 0.14950775 2.04507441 0.08768721 1.67346272
#>  [55] 0.75740385 2.06327411 0.18503739 1.78688513 0.98768293 1.89325093
#>  [61] 0.36115595 1.58421425 0.25320626 1.75659756 0.51689036 1.66496032
#>  [67] 0.13637431 1.70941724 0.10941193 2.08094859 0.05096966 2.05420399
#>  [73] 0.21364376 0.21034810 0.09703717 2.04909931 0.03993400 1.81903215
#>  [79] 0.04261795 1.79471451 0.91185877 1.62406174 0.84610029 1.99497543
#>  [85] 1.09979750 2.06421297 0.88871243 1.91604298 0.23897002 1.82135513
#>  [91] 0.37869479 0.98807726 0.09068102 1.38561334 0.09986311 1.56795635
#>  [97] 0.19936053 1.09548596 0.01903943 2.12725427 0.06788527 2.03558500
#> [103] 0.08785991 1.94650507 0.17607121 1.97030040 0.02481304 1.86331222
#> [109] 0.17276971 2.06852313 0.13161514 1.48228444 0.02499895 1.83313274
#> [115] 0.54897946 1.60300472 0.09204779 1.90663343 0.04402520 1.89149944
#> [121] 0.90058394 1.65223605 0.12112491 1.51453887 0.45410579 0.63596905
#> [127] 0.06574529 0.46808387 0.14476109 1.73138044 0.24790430 1.64986726
#> [133] 0.02518091 0.63252175 0.13866423 1.87768727 0.29000981 0.84054605
#> [139] 0.90982277 1.77677358 0.12820734 1.87505912 0.74716604 1.97044638
#> [145] 0.11978985 0.71625370 0.75655512 0.84401469 0.08017279 2.04811142
#> [151] 0.09944972 1.58739025 0.44566336 2.06441143 0.05552438 1.99741661
#> [157] 0.20774997 2.07114262 0.84960738 1.72638753 0.07338290 0.61482614
#> [163] 0.05390771 1.43561921 0.04351795 1.66619370 0.09827195 2.12208155
#> [169] 0.98830261 1.94866882 0.09523196 2.12716353 0.14225411 1.73192127
#> [175] 0.02338255 2.04318482 0.19958402 0.87298966 0.21453847 2.04143752
#> [181] 0.24621550 1.88756782 0.94797433 0.97143755 0.14861876 1.98804696
#> [187] 0.03620701 2.05284993 0.04996211 1.87508214 0.28529723 1.29518324
#> [193] 0.30292702 1.63317196 0.78559963 0.57686611 1.18891430 1.58958889
#> [199] 0.07461160 0.82676722 0.96824980 0.43811768 0.26853781 1.29337809
#> [205] 0.07210624 1.65469704 0.23768129 1.49819676 0.07404945 1.99499184
#> [211] 0.58191595 2.08900402 0.05523102 0.88028120 0.06799250 0.74702371
#> [217] 1.74384891 0.10192279 0.03565056 1.13146923 0.15077748 1.66322789
#> [223] 0.90299215 2.05999911 0.05201851 1.85837955 0.02809929 2.01976502
#> [229] 0.06322899 1.93188801 0.05483820 1.76180592 0.21463884 0.81731345
#> [235] 0.38806186 2.78396528 0.22401650 1.45828443 0.10112241 1.50523902
#> [241] 0.28159368 1.01721590 1.40346557 1.97051702 0.81364863 2.06823620
#> [247] 0.04933691 0.95127795 0.35936062 1.40208364 0.47106063 1.71256381
#> [253] 0.61832350 0.39586263 0.61568639 0.27544884 0.21106866 1.93888767
#> [259] 0.06974560 2.07727455 0.30374815 0.39427774 1.43930254 1.89007273
#> [265] 0.14278520 1.99428659 0.04166863 1.75356077 0.33545454 0.05976026
#> [271] 0.11737311 1.20059855 0.37428902 2.04502489 0.27257555 1.08799605
#> [277] 0.19296504 0.87820687 0.05399647 0.10280184 0.14605070 1.96735074
#> [283] 0.32953797 0.84067552 0.10798993 2.04671588 0.64638031 1.95630039
#> [289] 0.13359451 1.72270436 0.04801928 1.77343754 0.40279540 1.78161173
#> [295] 0.30213350 0.11829641 0.33715523 1.53219344 0.04119837 1.55258492
#> [301] 0.29089583 0.93268669 0.22435035 0.71827330 0.04975347 2.10381481
#> [307] 0.67635441 0.41254803 0.15122604 1.81841352 0.13631844 2.04324902
#> [313] 1.87137798 1.30984786 0.73975053 2.04552123 0.40374184 1.93884737
#> [319] 1.11678857 1.81371762 0.02259800 1.93255669 0.04013744 1.27789712
#> [325] 0.15080362 0.18006348 0.16403012 2.01917364 0.08725158 0.47266316
#> [331] 0.13743714 1.31514760 0.34988723 1.26876397 0.07863215 1.90578694
#> [337] 0.46071243 2.04807328 0.95360731 1.23539356 0.79663977 0.59324652
#> [343] 0.27782271 0.22140345 0.40287275 0.81667954 0.19384530 1.87703762
#> [349] 0.03292285 2.02060013 0.98500296 1.87825563 0.04230697 0.47390456
#> [355] 0.03281679 1.43203857 0.18433209 1.90431460 0.69545063 1.36681379
#> [361] 0.98966497 1.49758071 0.03289708 1.99225606 0.50310337 1.81233425
#> [367] 0.45760921 2.07183873 0.13184915 2.03424674 0.09623901 2.06832667
#> [373] 0.16855582 1.60430735 0.54999866 1.97156326 1.32895159 1.93874954
#> [379] 0.89605783 1.98687520 0.10624484 2.05887736 0.05655606 2.03550833
#> [385] 0.06165482 2.03069223 0.18286068 1.77144308 1.71328404 1.13057496
#> [391] 0.93558012 1.21201993 0.22560137 1.67924874 1.71922452 2.05928623
#> [397] 0.24384149 1.81983186 0.24623608 0.55110646 0.02818826 1.63670559
#> [403] 0.15875463 1.16463950 0.06186644 1.72272818 1.61626399 1.51302171
#> [409] 1.01628851 1.59435596 0.34916882 2.15180826 0.34153122 1.24246793
#> [415] 0.02768628 1.37179521 0.88619760 1.41509699 0.71302980 1.59875810
#> [421] 0.03596105 0.25480564 0.05389007 1.75669240 0.10376361 1.10314172
#> [427] 0.24926366 0.79816071 0.85736331 1.29173403 0.16638873 0.98489295
#> [433] 0.49638205 1.91120440 0.11165786 0.22759732 0.27778648 1.92650567
#> [439] 0.45744731 2.03854579 0.06760516 1.15264918 0.04048105 2.01510567
#> [445] 0.10935584 1.09249349 0.20223960 1.63405056 0.08579657 2.12062380
#> [451] 0.07473374 2.04773350 0.53468580 2.10403360 0.25716095 1.75756718
#> [457] 0.08803294 0.69765835 0.13749578 1.29630741 0.10100867 1.91059229
#> [463] 0.11130302 0.03558525 0.15786862 1.84336854 0.19478032 1.84866633
#> [469] 0.11080094 2.02752214 0.13015271 1.42103790 0.81492416 2.05245117
#> [475] 0.11589439 1.67714125 0.07772535 1.81782546 1.58685113 2.01769966
#> [481] 0.56968491 1.93784889 0.15725033 0.36235502 0.65466785 1.82461031
#> [487] 0.10061642 2.06249196 0.74055916 2.05397209 0.65329346 1.12863511
#> [493] 0.14736581 0.24923051 0.15973035 0.30557662 0.25809699 1.82881173
#> [499] 0.08778421 1.73969482

r2 = computemixdist(data, pi0 = 0.5, method = "nptll") # equivalent to normal
posteriormean(data, r2, fun = function(x) x^2)

#>   [1] 1.01180084 3.35512598 3.34762361 4.41359277 0.09976381 4.35413982
#>   [7] 2.33734513 2.53055918 2.96091028 4.45350607 1.07744806 4.21318124
#>  [13] 0.06396719 4.55314864 0.13161844 3.08877232 0.90407881 3.73429558
#>  [19] 3.16418013 2.34519191 0.14537946 3.88148342 0.98006432 6.77885932
#>  [25] 0.04047120 4.28077961 1.55313905 4.27684258 0.04243890 3.95757261
#>  [31] 0.15669319 0.58579476 0.07696613 1.68720613 3.08572967 2.40311569
#>  [37] 1.56982760 4.09904348 0.09252142 4.06455403 0.96925783 1.47067788
#>  [43] 1.24425622 3.61378291 0.24839359 2.02147867 0.27936992 1.60419855
#>  [49] 0.36532685 3.44903951 0.24578800 4.27992512 0.12581619 3.44963801
#>  [55] 1.50249851 4.33755160 0.31661860 3.69425349 1.98815015 3.92568505
#>  [61] 0.67586087 3.25800265 0.45448287 3.62878116 0.99888038 3.43135475
#>  [67] 0.21987082 3.52702592 0.16729633 4.40836504 0.05918746 4.30734430
#>  [73] 0.37421807 0.36756116 0.14354634 4.29170632 0.04091261 3.76390830
#>  [79] 0.04522941 3.71120096 1.82793136 3.34349017 1.68921857 4.15352452
#>  [85] 2.22555361 4.34088746 1.77907994 3.97579811 0.42553089 3.76894926
#>  [91] 0.71206426 1.98898411 0.13147038 2.83334382 0.14894384 3.22315421
#>  [97] 0.34540515 2.21641320 0.01251763 4.66087206 0.08911271 4.25364119
#> [103] 0.12614172 4.04334373 0.29865990 4.09683532 0.01907746 3.86022083
#> [109] 0.29206027 4.35677511 0.21052345 3.03978233 0.01930993 3.79452523
#> [115] 1.06579614 3.29830138 0.13405914 3.95507419 0.04752929 3.92184516
#> [121] 1.80413185 3.40400429 0.19001531 3.10876990 0.86827146 1.24766302
#> [127] 0.08523691 0.89731034 0.23640162 3.57436284 0.44369202 3.39891412
#> [133] 0.01953862 1.24044402 0.22437722 3.89161046 0.52963022 1.67751289
#> [139] 1.82363320 3.67238054 0.20384627 3.88586610 1.48097634 4.09716634
#> [145] 0.18741558 1.41603191 1.50071403 1.68482293 0.11173280 4.28877345
#> [151] 0.14815324 3.26481236 0.85074442 4.34159815 0.06706415 4.15923678
#> [157] 0.36231680 4.36691644 1.69661072 3.56359710 0.09916203 1.20340223
#> [163] 0.06425027 2.94007084 0.04669756 3.43400659 0.14590271 4.62991525
#> [169] 1.98946071 4.04817546 0.14010729 4.66032594 0.23145279 3.57552916
#> [175] 0.01732954 4.27453784 0.34585521 1.74591217 0.37602617 4.26962965
#> [181] 0.44025693 3.91323071 1.90420929 1.95379944 0.24402845 4.13743130
#> [187] 0.03509194 4.30311231 0.05746760 3.88591639 0.51998602 2.64064302
#> [193] 0.55609457 3.36305087 1.56180538 1.12402748 2.41466708 3.26952687
#> [199] 0.10142465 1.64848105 1.94706047 0.83508688 0.48573841 2.63680022
#> [205] 0.09681728 3.40929308 0.42291363 3.07380930 0.10038876 4.15356283
#> [211] 1.13457972 4.44649429 0.06655216 1.76129233 0.08930749 1.48067716
#> [217] 3.60126000 0.15288804 0.03424252 2.29272296 0.24830233 3.42763016
#> [223] 1.80921469 4.32624028 0.06098692 3.84946561 0.02335111 4.21298117
#> [229] 0.08070820 4.01083246 0.06586753 3.64003076 0.37622903 1.62856819
#> [235] 0.73142104 8.88611184 0.39520181 2.98848636 0.15135435 3.08887295
#> [241] 0.51241101 2.05062965 2.87143153 4.09732652 1.62085018 4.35568710
#> [247] 0.05640487 1.91118994 0.67215804 2.86848262 0.90349749 3.53380454
#> [253] 1.21072122 0.74755190 1.20520240 0.49985119 0.36901619 4.02637478
#> [259] 0.09249900 4.39218150 0.55777828 0.74427380 2.94793708 3.91871834
#> [265] 0.23250067 4.15191686 0.04369168 3.62222360 0.62291062 0.07452099
#> [271] 0.18271587 2.43948837 0.70296493 4.27978283 0.49398202 2.20053669
#> [277] 0.33253745 1.75691663 0.06440427 0.15457390 0.23894970 4.09015571
#> [283] 0.61073987 1.67778574 0.16455290 4.28467710 1.26947074 4.06526344
#> [289] 0.21440801 3.55565721 0.05417684 3.66516753 0.76189584 3.68284428
#> [295] 0.55446763 0.18451038 0.62641039 3.14655533 0.04293425 3.19022130
#> [301] 0.53144409 1.87191302 0.39587796 1.42027305 0.05711255 4.52452506
#> [307] 1.33230013 0.78208636 0.24919092 3.76256603 0.21976096 4.27471948
#> [313] 3.87782446 2.67186666 1.46539136 4.28121178 0.76385458 4.02628515
#> [319] 2.26158262 3.75237967 0.01640306 4.01231532 0.04123639 2.60385032
#> [325] 0.24835411 0.30664994 0.27462661 4.21151468 0.12499538 0.90682876
#> [331] 0.22196170 2.68315318 0.65262931 2.58441630 0.10886654 3.95321242
#> [337] 0.88199349 4.28866078 1.91611227 2.51344080 1.58504009 1.15826423
#> [343] 0.50470203 0.38991130 0.76205591 1.62723314 0.33430722 3.89019029
#> [349] 0.03016281 4.21505737 1.98248250 3.89285308 0.04472441 0.90940954
#> [355] 0.03000721 2.93242459 0.31520412 3.94997510 1.37236151 2.79325159
#> [361] 1.99234203 3.07249168 0.03012498 4.14718814 0.97016219 3.74937971
#> [367] 0.87554741 4.36967584 0.21098247 4.25007145 0.14202490 4.35602969
#> [373] 0.28364761 3.30109594 1.06792312 4.09970024 2.71255672 4.02606762
#> [379] 1.79457974 4.13472525 0.16119091 4.32247625 0.06886948 4.25343589
#> [385] 0.07789196 4.24072409 0.31225419 3.66085599 3.53535626 2.29082578
#> [391] 1.87802471 2.46375712 0.39841203 3.46208367 3.54815700 4.32384198
#> [397] 0.43542979 3.76564358 0.44029879 1.07023510 0.02347135 3.37063969
#> [403] 0.26413062 2.36311866 0.07826977 3.55570857 3.32675233 3.10552363
#> [409] 2.04866705 3.27974998 0.65114905 4.81152575 0.63541821 2.52848309
#> [415] 0.02279589 2.80387340 1.77377394 2.89625554 1.40926232 3.28919183
#> [421] 0.03471580 0.45773987 0.06421969 3.62898598 0.15642010 2.23264388
#> [427] 0.44645775 1.58824157 1.71296080 2.63330047 0.27932612 1.98224987
#> [433] 0.95616905 3.96513471 0.17163623 0.40245639 0.50462798 3.99891034
#> [439] 0.87521114 4.26164839 0.08860417 2.33766650 0.04178460 4.20150641
#> [445] 0.16718806 2.21006971 0.35120487 3.36493766 0.12225760 4.62126122
#> [451] 0.10164987 4.28765885 1.03597693 4.52573863 0.46253783 3.63087518
#> [457] 0.12646796 1.37699464 0.22207711 2.64303625 0.15113651 3.96378677
#> [463] 0.17095000 0.03414319 0.26236996 3.81677940 0.33618755 3.82830809
#> [469] 0.16997939 4.23253509 0.20765632 2.90893732 1.62353632 4.30187807
#> [475] 0.17984446 3.45755014 0.10718317 3.76129017 3.26365636 4.20787271
#> [481] 1.10902541 4.02406533 0.26114171 0.67833423 1.28683588 3.77601508
#> [487] 0.15038543 4.33480498 1.46709067 4.30661495 1.28395571 2.28671052
#> [493] 0.24154979 0.44639030 0.26607025 0.56152814 0.46444513 3.78513823
#> [499] 0.12599902 3.59229670

data = runif(500, min = -0.5, max = 0.5)
r3 = computemixdist(data, method = "npnormcll", beta = 100)
posteriormean(data, r3)

#>   [1] -0.3613745134 -0.3604370893  0.1778048184 -0.2033984089  0.3683738473
#>   [6]  0.3032582014  0.2420271244 -0.0758255162  0.3671292287  0.3692699846
#>  [11] -0.0717478105 -0.0729107745 -0.0775734183  0.2961739870  0.3485910092
#>  [16]  0.3582550409  0.2182746632 -0.3617198334  0.2935086511  0.3252065290
#>  [21] -0.0711287551 -0.3555108836 -0.0633624321 -0.3374957678  0.0960869343
#>  [26]  0.0661298518 -0.0426487033  0.3402768302  0.2911320160 -0.3615044677
#>  [31] -0.0876123602  0.3505718415 -0.3461035689 -0.3606671637  0.3634032919
#>  [36] -0.3583344392 -0.2975420219  0.3445656896  0.1255666503 -0.1739424810
#>  [41] -0.3615748857  0.2901098194 -0.2385695321  0.3304309723  0.3698937003
#>  [46]  0.3465774257 -0.0221550815 -0.0821255841  0.2972517417  0.1676245231
#>  [51]  0.2276095257 -0.0538367115 -0.3609289475  0.2925866572  0.3071227919
#>  [56] -0.0718501265 -0.3542322858  0.3685583598  0.2707843896  0.3660047755
#>  [61] -0.3593671314  0.3500738807 -0.0609780675  0.3543683385  0.2244582656
#>  [66]  0.3696940655 -0.0818881858  0.1562539572 -0.0716376295 -0.1669358295
#>  [71]  0.2348847226 -0.3606758642  0.2727220657  0.2598820085  0.3641952172
#>  [76] -0.0728203062  0.3702097558 -0.0026174527 -0.0697360211  0.1725020977
#>  [81]  0.3589573087 -0.3456796968  0.2654361748  0.2121290405 -0.0323815872
#>  [86] -0.0736500465 -0.0640505469  0.3344782681 -0.3527948857  0.1028706260
#>  [91]  0.3312425958 -0.0804148528  0.3374640913 -0.0357949914 -0.1348405626
#>  [96]  0.3527105907 -0.3593078249 -0.1251136109 -0.0469523372  0.2417625196
#> [101]  0.3459936849 -0.2361178805  0.3672066422  0.3698528133 -0.1710706424
#> [106]  0.3498878723 -0.0702285448 -0.0772677120  0.3669436115  0.3388554904
#> [111] -0.0949046347 -0.3372423969 -0.3536138074  0.2537707194  0.1632362549
#> [116] -0.3616371438  0.3698330592 -0.3555205463 -0.3041549962  0.2095203094
#> [121] -0.1197256064  0.3685771967 -0.0691019882  0.3388241116 -0.0368984761
#> [126]  0.1254385364 -0.0916343786  0.3085070637  0.2112387942 -0.3453197858
#> [131] -0.2477025929  0.2390682649 -0.0710946043 -0.3617196162  0.2610357791
#> [136]  0.2204456655  0.0018357813  0.0009418711  0.2646836813 -0.0607264736
#> [141] -0.3608152406 -0.1023371782 -0.3304185375 -0.0774835529 -0.0721410185
#> [146]  0.2228383165 -0.3613102873 -0.3615577396  0.0463154666 -0.0778170951
#> [151]  0.3197165288  0.2866376472 -0.0709013950 -0.3598028793 -0.3614027893
#> [156] -0.3273392968 -0.0762112435  0.0885260589 -0.3603088408 -0.3350896106
#> [161] -0.0771311854 -0.1135193845 -0.3608570640 -0.0693202736 -0.3616397631
#> [166]  0.0234055814  0.3054442342 -0.3586576551  0.1038340121 -0.0562959396
#> [171] -0.3587256763 -0.0537946501 -0.3609007912 -0.0750156697 -0.3596489237
#> [176]  0.3630085874 -0.1569975594 -0.0713271708  0.3524809671 -0.0688658795
#> [181]  0.1779620242  0.3155916884 -0.3498806208  0.2075832797  0.3292506791
#> [186]  0.3684982373 -0.3608179537 -0.0719129265  0.3665635893 -0.3604691472
#> [191] -0.0763389460 -0.3610773485  0.3622636283  0.1659762405 -0.3539731529
#> [196]  0.3669173874 -0.0607945936 -0.3568305676 -0.1398257707 -0.0756561537
#> [201]  0.1589761949 -0.0745080103 -0.3509460107  0.3534467377 -0.3596080682
#> [206] -0.3550501837 -0.3491906013  0.3686653473  0.3674289675 -0.3560806555
#> [211]  0.0879193025 -0.2221668504  0.3363977818 -0.0976429894 -0.3084539001
#> [216]  0.3274012091 -0.0696409045 -0.0768135850 -0.0681631975 -0.3615510863
#> [221] -0.0777462228  0.3510211200 -0.0713175107 -0.0501352499  0.1308832835
#> [226]  0.2429202269 -0.3195935505  0.3616791193  0.2857812910 -0.0777345668
#> [231]  0.3676664589 -0.3600690640 -0.0649891335 -0.3615923831  0.2218989803
#> [236] -0.0792010160  0.0629827518 -0.3586015661 -0.0618073765  0.1671335266
#> [241]  0.2025990256 -0.3570570296 -0.0646155494 -0.3608076052 -0.3386802769
#> [246]  0.3471063385 -0.0678907728 -0.0677490414  0.3534554721  0.3644589072
#> [251]  0.2341455971 -0.0703923716  0.1845400907 -0.3227748848 -0.0698113777
#> [256] -0.0349397830  0.2299901527  0.2292875702 -0.3616363295  0.0369481716
#> [261] -0.0659226015  0.1187739241 -0.3611584188 -0.3515558368 -0.0955342596
#> [266] -0.3497449785 -0.3608550561 -0.0683564370  0.3562320051 -0.3347950911
#> [271] -0.2699680890  0.2085783425 -0.1275010249 -0.0686277670  0.3545285361
#> [276] -0.0724344389  0.2964842553 -0.0698597843 -0.0838265857 -0.3253156698
#> [281] -0.3616336885  0.3371056660  0.3020918679 -0.0363594708 -0.0885653175
#> [286] -0.0786597234 -0.2804786726  0.3328130615 -0.1586506948  0.0442403550
#> [291] -0.3577862578  0.3631479567 -0.3609259912 -0.3616342506 -0.3537248439
#> [296] -0.3579495948  0.3427008252  0.2106654741 -0.3553480616 -0.1572385170
#> [301] -0.0816536114  0.2544277759  0.2230242049  0.1831321258  0.2019194535
#> [306] -0.2689530317 -0.3616115462  0.3674397891  0.3629949696  0.2994935675
#> [311]  0.0360286836 -0.3032162208  0.3324955587 -0.2941275492 -0.3479833238
#> [316]  0.1486453105 -0.1091054518 -0.3616025775 -0.0773386382 -0.0859452010
#> [321]  0.2548151331 -0.0969894262 -0.3595639326 -0.0777153426 -0.1226639087
#> [326]  0.3700486598  0.2478900411 -0.0911784887  0.3438239149  0.3132376541
#> [331] -0.3442350419 -0.3613574934  0.3617099971 -0.3613174891 -0.0709789227
#> [336]  0.3532699907 -0.0737974500 -0.0554861769  0.2603256671 -0.0789958192
#> [341]  0.3107357395 -0.2370395387 -0.3068226618 -0.0932030892  0.3082610381
#> [346] -0.3280983405 -0.3491222946 -0.1142309045  0.3024505921  0.2114149788
#> [351] -0.1526109741 -0.3239672936 -0.3177403951  0.3541021038  0.3400773368
#> [356] -0.1255736187  0.2090360540 -0.3615642411 -0.3595259506 -0.0702345183
#> [361]  0.3094769902  0.2162189510 -0.0772509021  0.3061320930  0.2337777756
#> [366] -0.3614469325  0.2053251958 -0.0904970235 -0.3509166888  0.3528138551
#> [371] -0.0698776450  0.3082586504 -0.0596092369  0.2677157064 -0.1040193505
#> [376] -0.1284812714 -0.3616263341 -0.2997921927 -0.1381874371 -0.0187410190
#> [381]  0.2143943709  0.2221456278  0.3274145381 -0.0852785514  0.3484025903
#> [386] -0.0709591670  0.2710168194  0.3217869967 -0.0741782306  0.2780852454
#> [391] -0.0842443013 -0.0734833233 -0.1895089564 -0.2793369047  0.3228254534
#> [396]  0.2934269126 -0.2073634085 -0.0549468584  0.0571240961 -0.0752851082
#> [401]  0.3209456498  0.2502137851  0.1684757641 -0.0679961511  0.0431936782
#> [406]  0.3676137848 -0.0994931814 -0.3600618169 -0.3587570842 -0.3606960180
#> [411] -0.0778129781 -0.3604331163 -0.3361908867  0.3679935136  0.0069654405
#> [416] -0.1336733043 -0.0465737577  0.1536265565 -0.0830732055 -0.3605071715
#> [421]  0.2365977892 -0.2589805760  0.3628251130  0.1001428503 -0.0382917677
#> [426]  0.2923956597  0.2270936658  0.3642554206 -0.3544566242 -0.0992278003
#> [431] -0.0688054841 -0.0076146887 -0.0467432128  0.3118403677  0.2056908088
#> [436] -0.3599981039  0.1801286963 -0.0722963053  0.0328377316 -0.0636610683
#> [441]  0.1829826956  0.3698204324 -0.3415333201 -0.0738700607  0.0466319273
#> [446]  0.0994139361 -0.3613869437  0.2870913253  0.3175485910 -0.3201227072
#> [451]  0.2631642804  0.1858558884 -0.0636166155 -0.0414585685  0.3666798468
#> [456] -0.3575237384 -0.1650155199 -0.0761626679  0.3564698427  0.2891372625
#> [461]  0.2676773569  0.2180393255  0.3694697714  0.3089929798  0.3563072568
#> [466] -0.0871026804  0.2835143059 -0.3583323493  0.2372916167  0.2775770459
#> [471]  0.2253708589 -0.3615575944  0.3080290706 -0.3461848370  0.2839937976
#> [476] -0.2478921993  0.3524142462  0.3685012821 -0.0714093541  0.3437576279
#> [481]  0.3674642487 -0.0744771519 -0.0599928411  0.3692328322 -0.3603495002
#> [486] -0.3605524779  0.3566075114 -0.0843245711 -0.3609389838 -0.3617050298
#> [491] -0.0444093328 -0.3492768465 -0.3345718646 -0.0497117389  0.3580690454
#> [496] -0.1139023585  0.2582226995 -0.0304890641 -0.2660888452 -0.3576500120