Find the posterior mean

Find the posterior mean given the observations and the mixing distribution based on the family in the result

posteriormean(x, result, fun = function(x) x)

posteriormean.npnorm(x, result, fun = function(x) x)

posteriormean.npt(x, result, fun = function(x) x)

posteriormean.npnormc(x, result, fun = function(x) x)

Arguments

x
a vector of observations

result
an object of class nspmix

fun
the function to transform the mean. It finds the posterior mean of fun(x). The function fun must be vectorised.

Examples

data = rnorm(500, c(0, 2))
r1 = computemixdist(data, pi0 = 0.5)
posteriormean(data, r1)

#>   [1] 0.074718558 1.848281699 0.416716658 1.755668787 0.029776400 1.673455907
#>   [7] 0.714225125 1.853681665 0.988877947 1.890156694 0.027852384 1.917126221
#>  [13] 0.201652443 1.469148419 0.187738003 1.092716292 0.908703043 1.718454353
#>  [19] 0.026822681 1.882995457 0.072110375 1.604866477 0.584233242 1.832161484
#>  [25] 0.600590921 1.295277008 0.288795401 0.062502152 0.051475120 1.192309480
#>  [31] 0.081533209 1.843459892 0.108544816 1.775235995 0.537502178 1.855902182
#>  [37] 0.652507220 1.328447537 0.374332904 1.889261854 1.606675499 1.654818211
#>  [43] 1.317077050 1.320920454 0.181892716 1.275601484 0.121508389 1.840661285
#>  [49] 0.661082293 1.265953701 0.065392317 1.039718995 1.596158717 1.794020674
#>  [55] 0.397019463 1.171783698 0.312246016 1.525784607 1.510859715 1.307837124
#>  [61] 0.630536767 1.624374720 0.011360417 1.902141378 0.001900745 0.966891007
#>  [67] 0.075279646 1.261708629 0.591289998 1.536035949 0.058936747 0.678555909
#>  [73] 0.061402736 1.315406439 0.828195297 1.783282254 1.603469817 1.093416315
#>  [79] 0.243235020 1.436888451 0.104786876 1.649840358 0.002533070 1.746551907
#>  [85] 0.455065966 1.902225330 0.783327282 1.891386395 0.208673790 1.405990237
#>  [91] 1.430577286 1.528838196 0.240298387 1.370186805 0.070382162 1.643611704
#>  [97] 0.098484415 1.752795875 0.768204787 0.301268019 0.383299070 1.843825826
#> [103] 0.501494730 1.912870648 0.108733972 0.895645548 0.048085587 1.819751765
#> [109] 0.064005411 1.765783055 0.028732632 0.119569670 0.017324429 1.463404920
#> [115] 0.048642051 1.637952805 1.292033768 1.123653763 0.696726894 0.623607380
#> [121] 1.150680908 0.637268070 1.089565250 1.852798119 0.056285939 0.912614214
#> [127] 0.247082123 1.865197405 0.110106090 1.877378907 0.195730632 0.899066858
#> [133] 0.048330925 1.779934218 0.883636003 0.860073548 0.033722693 0.610985235
#> [139] 0.111148599 1.868183121 0.806218527 1.615446584 0.651586442 1.202001727
#> [145] 0.894360849 1.694410717 0.377007889 1.237162967 0.028730679 1.806350201
#> [151] 0.069463910 1.618730419 0.456077648 1.551816303 0.311743485 1.855456642
#> [157] 0.089861718 0.898445225 0.588873603 1.857294627 0.028729891 1.882707073
#> [163] 0.535787621 1.863207456 0.211068983 1.060549948 0.140157365 1.742597393
#> [169] 0.498111341 1.916493835 0.004634440 1.809172268 0.141993741 1.305473308
#> [175] 0.241725614 1.808390407 0.086387071 1.890473395 0.176025322 0.459342655
#> [181] 0.016865397 0.241546072 0.069183649 1.632065885 1.002058283 1.914658006
#> [187] 1.787978047 0.977964628 0.284767166 1.201372976 0.065199169 1.239953222
#> [193] 0.709839393 1.584355114 0.016635394 1.354877103 0.006258743 1.805678376
#> [199] 0.600853159 1.700102334 0.898517726 1.499403979 0.645953172 1.687443490
#> [205] 0.105913280 1.311735081 0.085865178 1.727537292 1.156733853 1.814310841
#> [211] 1.775895205 1.737053089 0.006551174 1.471119782 0.136815344 1.894007457
#> [217] 0.015699682 1.539563547 0.254165105 0.133968057 0.895493635 0.575168527
#> [223] 0.521521005 1.900093411 0.254164326 1.643532140 0.400387787 1.783536257
#> [229] 0.192668840 1.833646206 0.191618730 1.166241694 0.012988348 1.825201780
#> [235] 1.575891454 1.838611495 1.012093403 1.732445572 0.037724031 0.864936619
#> [241] 0.104227363 1.277602178 0.186422583 1.908493148 0.040998799 1.621699924
#> [247] 0.243246258 1.039257570 0.500312350 1.729981431 0.291867004 1.070507343
#> [253] 0.906010758 1.178968712 0.698160756 1.743008780 0.323967649 0.537099649
#> [259] 0.363830174 1.247562643 0.462569527 1.901100084 0.470022115 0.513259615
#> [265] 0.241096877 1.826247606 0.600761288 1.870672149 0.711970588 1.618456708
#> [271] 0.009873024 0.962576287 0.841041894 1.840784350 0.213998367 1.876205925
#> [277] 0.046967681 1.771295076 0.101287073 1.875984847 0.102723428 1.449755381
#> [283] 0.122695563 1.329737534 0.518527358 1.820782906 0.594638063 0.159778037
#> [289] 0.428873108 1.901269289 1.324509213 1.885257241 0.667740795 1.886770370
#> [295] 1.308128938 1.796995535 1.747288788 1.738020303 1.323886394 1.422718711
#> [301] 1.441095915 1.473946414 0.105961786 1.869098552 0.444497869 1.151204093
#> [307] 0.181240829 0.528226685 0.381615100 1.701108492 0.459995250 1.906899975
#> [313] 1.365196899 1.044089661 1.011168656 1.703563594 0.010658779 1.024208387
#> [319] 0.178013489 1.019835227 1.852108966 1.900877233 1.311014084 1.506365629
#> [325] 0.055807034 1.411983881 1.189784334 1.825037128 0.005452920 1.702220117
#> [331] 0.861560333 1.858106301 0.606519816 0.770473664 0.184677405 1.809784050
#> [337] 0.108653084 1.333411975 1.807331502 1.617347932 0.122450069 1.827571691
#> [343] 0.007644189 0.270435432 0.578274800 1.861879734 0.062027261 1.874383813
#> [349] 0.949598917 0.684355534 0.457523938 1.828942417 0.154194582 1.576705151
#> [355] 0.159470977 0.879065201 0.034671995 1.913674045 0.051739264 1.816372914
#> [361] 0.243157500 1.277863077 0.017338645 1.489284624 0.010548659 1.673927288
#> [367] 0.063774127 1.073604053 0.296333764 1.894808705 0.741429525 1.830441472
#> [373] 0.038937216 1.387151750 0.074213091 1.610479803 0.343147217 1.239275432
#> [379] 0.102688630 1.342625729 1.161289187 1.753147834 0.390919596 1.884378457
#> [385] 0.025319415 1.799742386 0.099187856 1.783193084 0.375224910 1.031448546
#> [391] 0.286267713 1.873485121 0.132848507 1.877850842 0.093689366 0.579993445
#> [397] 0.021410367 0.686364921 0.071334187 1.437627248 1.327253815 1.908306297
#> [403] 1.156727141 1.625093470 0.194527470 1.864500989 0.127976076 1.893616600
#> [409] 1.211897248 1.851277815 0.251347872 1.784911846 0.391466491 0.677266573
#> [415] 0.640616006 1.692669190 0.304820155 1.297282023 0.113658281 1.696131800
#> [421] 0.056792791 1.899661242 1.190091215 1.902847681 0.227593656 1.328683110
#> [427] 0.181378077 0.703088088 0.003269173 0.632983957 0.394183496 1.781599995
#> [433] 0.288498076 1.646487056 1.670045632 1.871628942 0.024502575 1.722105314
#> [439] 0.906949211 1.832036234 0.055713335 1.673703050 0.106364364 1.077295957
#> [445] 0.335294482 1.812969982 1.278466211 1.769054579 0.943224342 1.520632720
#> [451] 0.163940310 1.801558276 1.377995874 1.909413824 0.529338773 1.853950769
#> [457] 1.712930022 1.698444207 0.389131113 1.775667867 0.044303023 1.678435610
#> [463] 1.762653459 1.743097457 0.055788415 0.836841209 0.010349363 1.711008892
#> [469] 1.145782941 1.791457563 0.039431782 1.848097935 0.060740115 1.812331297
#> [475] 0.469810483 0.456905841 0.007568165 1.800960822 1.323397110 1.866751939
#> [481] 0.127902840 1.879945771 0.386892739 1.053589874 1.420817012 0.936160920
#> [487] 0.341690013 1.772696124 0.017359342 1.901633964 0.610788894 1.845614397
#> [493] 0.053605169 1.640204839 0.048694771 1.888580817 0.539477913 1.870020316
#> [499] 0.233633486 1.332118029

r2 = computemixdist(data, pi0 = 0.5, method = "nptll") # equivalent to normal
posteriormean(data, r2, fun = function(x) x^2)

#>   [1] 0.143698369 3.540391724 0.799852748 3.363009945 0.057314908 3.205622472
#>   [7] 1.369819020 3.550738910 1.895615545 3.620654940 0.053614672 3.672408057
#>  [13] 0.387409353 2.814634100 0.360706985 2.094341458 1.742156517 3.291760874
#>  [19] 0.051634264 3.606923956 0.138687133 3.074345902 1.120846260 3.509506776
#>  [25] 1.152181072 2.481945479 0.554585630 0.120224599 0.099031999 2.284921863
#>  [31] 0.156790804 3.531152966 0.208674570 3.400477564 1.031319337 3.554993984
#>  [37] 1.251622145 2.545414177 0.718605066 3.618939022 3.077808017 3.169948716
#>  [43] 2.523657858 2.531011841 0.349488839 2.444297968 0.233569223 3.525790974
#>  [49] 1.268045546 2.425837631 0.125778455 1.992918477 3.057681111 3.436450800
#>  [55] 0.762095961 2.245645295 0.599559421 2.923008641 2.894448982 2.505978132
#>  [61] 1.209541485 3.111681543 0.021885359 3.643641609 0.003667175 1.853533311
#>  [67] 0.144776389 2.417714971 1.134364390 2.942625463 0.113372798 1.301510701
#>  [73] 0.118111847 2.520461305 1.588038776 3.415885879 3.071672974 2.095681078
#>  [79] 0.467193008 2.752906069 0.201457340 3.160421084 0.004885981 3.345554077
#>  [85] 0.873354444 3.643802671 1.502136210 3.623013055 0.400882607 2.693784620
#>  [91] 2.740830059 2.928851920 0.461559244 2.625278082 0.135366521 3.148499582
#>  [97] 0.189352591 3.357509166 1.473181383 0.578506409 0.735794019 3.531854091
#> [103] 0.962327245 3.664233133 0.209037841 1.717161835 0.092516824 3.485734289
#> [109] 0.123113355 3.382376423 0.055307579 0.229846410 0.033362689 2.803644075
#> [115] 0.093586461 3.137668734 2.475739820 2.153544930 1.336309816 1.196268883
#> [121] 2.205263947 1.222434424 2.088311361 3.549045847 0.108278335 1.749643176
#> [127] 0.474573306 3.572807477 0.211672951 3.596156465 0.376045485 1.723710970
#> [133] 0.092988415 3.409474412 1.694172625 1.649066832 0.064903651 1.172091651
#> [139] 0.213675032 3.578529897 1.545963924 3.094594341 1.249858604 2.303468007
#> [145] 1.714702630 3.245733662 0.723733326 2.370748189 0.055303823 3.460064679
#> [151] 0.133602143 3.100879104 0.875293322 2.972823011 0.598595720 3.554140206
#> [157] 0.172790026 1.722521032 1.129735505 3.557662330 0.055302308 3.606371061
#> [163] 1.028034349 3.568993700 0.405478593 2.032784242 0.269376181 3.337982700
#> [169] 0.955844061 3.671192635 0.008934787 3.465469891 0.272901775 2.501455188
#> [175] 0.464297311 3.463972351 0.166115379 3.621262248 0.338227772 0.881550627
#> [181] 0.032479460 0.463952869 0.133063628 3.126401619 1.920841595 3.667665660
#> [187] 3.424878516 1.874728007 0.546859686 2.302264894 0.125407302 2.376087216
#> [193] 1.361420459 3.035091955 0.032036898 2.595984499 0.012063270 3.458777926
#> [199] 1.152683404 3.256628884 1.722659814 2.872528053 1.239069252 3.232396873
#> [205] 0.203620664 2.513436508 0.165112818 3.309149606 2.216846661 3.475312301
#> [211] 3.401739902 3.327367716 0.012626426 2.818406257 0.262959801 3.628039599
#> [217] 0.030236388 2.949375881 0.488160881 0.257493108 1.716871041 1.103481198
#> [223] 1.000699642 3.639712803 0.488159386 3.148347300 0.768552791 3.416372298
#> [229] 0.370169764 3.512351150 0.368154524 2.235040433 0.025018640 3.496174188
#> [235] 3.018894899 3.521863783 1.940047651 3.318546475 0.072597364 1.658376412
#> [241] 0.200382754 2.448126144 0.358182504 3.655829716 0.078893437 3.106562324
#> [247] 0.467214567 1.992035401 0.960061600 3.313828870 0.560476692 2.051840102
#> [253] 1.737002988 2.259394057 1.339055724 3.338770343 0.622037243 1.030548116
#> [259] 0.698469512 2.390647465 0.887734769 3.641643950 0.902017005 0.984870292
#> [265] 0.463091112 3.498177603 1.152507420 3.583300590 1.365501652 3.100355262
#> [271] 0.019022228 1.845274923 1.612632899 3.526026756 0.411099502 3.593907905
#> [277] 0.090367946 3.392931104 0.194735594 3.593484114 0.197494300 2.777526232
#> [283] 0.235848851 2.547882459 0.994963674 3.487709476 1.140777940 0.307042274
#> [289] 0.823153453 3.641968552 2.537878582 3.611260428 1.280797974 3.614161658
#> [295] 2.506536491 3.442148154 3.346964936 3.329219501 2.536686879 2.725793227
#> [301] 2.760956800 2.823814960 0.203713821 3.580284470 0.853100341 2.206265098
#> [307] 0.348237723 1.013547827 0.732565747 3.258554952 0.882801298 3.652772169
#> [313] 2.615730393 2.001283037 1.938277802 3.263254733 0.020534797 1.963233970
#> [319] 0.342043642 1.954864420 3.547725295 3.641216437 2.512056948 2.885849371
#> [325] 0.107357924 2.705252961 2.280089969 3.495858777 0.010511317 3.260682921
#> [331] 1.651913056 3.559217767 1.163538100 1.477525633 0.354833215 3.466641675
#> [337] 0.208882498 2.554913136 3.461944189 3.098233229 0.235377452 3.500714086
#> [343] 0.014731143 0.519370786 1.109431866 3.566449160 0.119312005 3.590415099
#> [349] 1.820435857 1.312617697 0.878065110 3.503339952 0.296324281 3.020452074
#> [355] 0.306452854 1.685422858 0.066729032 3.665775888 0.099539704 3.479262107
#> [361] 0.467044293 2.448625354 0.033390043 2.853164526 0.020322822 3.206524750
#> [367] 0.122668911 2.057766354 0.569043370 3.629576282 1.421912719 3.506211694
#> [373] 0.074929896 2.657738865 0.142727205 3.085088754 0.658813879 2.374790297
#> [379] 0.197427466 2.572542730 2.225563560 3.358183060 0.750402731 3.609575524
#> [385] 0.048742933 3.447408935 0.190703692 3.415715118 0.720315153 1.977090405
#> [391] 0.549737677 3.588692441 0.255343573 3.597061160 0.180142428 1.112724241
#> [397] 0.041223624 1.316465881 0.137195765 2.754319712 2.543130107 3.655471100
#> [403] 2.216833818 3.113057144 0.373736578 3.571472769 0.245988233 3.627290003
#> [409] 2.322402991 3.546132665 0.482756536 3.419006585 0.751451121 1.299041441
#> [415] 1.228846870 3.242399973 0.585318635 2.485781901 0.218494643 3.249028236
#> [421] 0.109252448 3.638883783 2.280677190 3.644996669 0.437184753 2.545864923
#> [427] 0.348501133 1.348491706 0.006304502 1.214228789 0.756659550 3.412664348
#> [433] 0.554015382 3.154002918 3.199094847 3.585134522 0.047171784 3.298750340
#> [439] 1.738799361 3.509266828 0.107177841 3.206095532 0.204486988 2.064831614
#> [445] 0.643756748 3.472743983 2.449779404 3.388640853 1.808234548 2.913150149
#> [451] 0.315031845 3.450886791 2.640219972 3.657596816 1.015678576 3.551254575
#> [457] 3.281185181 3.253454779 0.746974214 3.401304567 0.085245632 3.215154265
#> [463] 3.376383862 3.338940124 0.107322139 1.604590984 0.019939186 3.277507452
#> [469] 2.195891315 3.431542108 0.075880752 3.540039614 0.116838465 3.471520640
#> [475] 0.901611435 0.876880540 0.014584759 3.449742521 2.535750680 3.575786859
#> [481] 0.245847612 3.601077232 0.742683219 2.019464338 2.722154469 1.794714595
#> [487] 0.656019823 3.395613951 0.033429865 3.642668153 1.171715560 3.535281010
#> [493] 0.103126033 3.141978997 0.093687798 3.617633115 1.035104702 3.582051208
#> [499] 0.448772629 2.552437300

data = runif(500, min = -0.5, max = 0.5)
r3 = computemixdist(data, method = "npnormcll", beta = 100)
posteriormean(data, r3)

#>   [1] -0.126300080  0.108811101 -0.090274130 -0.287552909  0.211272560
#>   [6] -0.049100098 -0.381798077  0.270574487 -0.382526893 -0.391909405
#>  [11] -0.137121326 -0.156164363  0.043309102  0.079209987 -0.245442899
#>  [16] -0.387925835  0.369395007 -0.388714154 -0.116793456 -0.185798880
#>  [21] -0.385325482 -0.073011869  0.324507474 -0.287610245  0.236551485
#>  [26]  0.334830025  0.364323962 -0.392368055  0.370408804 -0.346545969
#>  [31] -0.079259358  0.010901049 -0.385398499  0.149714601  0.365008273
#>  [36] -0.045776558 -0.391639352  0.023585381 -0.134992224 -0.378485333
#>  [41] -0.392394333 -0.077212080  0.122046953  0.066975684 -0.389310608
#>  [46]  0.371795388 -0.083814962  0.065319245  0.332317993 -0.058505159
#>  [51]  0.355847479  0.250475970  0.368937496 -0.385837939 -0.390184198
#>  [56] -0.142865759 -0.392967749  0.151337917  0.103423675 -0.375305945
#>  [61]  0.350439837  0.368122315 -0.226099501 -0.203998490 -0.139651924
#>  [66]  0.256038396 -0.098772291 -0.297674627  0.089914721  0.098830787
#>  [71] -0.387840855  0.308615422  0.263478973  0.342603516 -0.392743205
#>  [76]  0.100120894 -0.034261691 -0.109221824 -0.161812941 -0.388899738
#>  [81] -0.110308210  0.295088613  0.190039520 -0.386050017 -0.086937245
#>  [86] -0.387584268  0.094560610  0.141019748  0.046412154  0.288505612
#>  [91]  0.106176840 -0.384316991  0.370003476  0.366812709  0.187052008
#>  [96]  0.127135770 -0.031790486 -0.391834952 -0.067447658 -0.126265412
#> [101]  0.103934706  0.348796338  0.171812942  0.130259337 -0.195275858
#> [106]  0.232056620  0.337948440  0.174791852 -0.246695151 -0.037215808
#> [111] -0.381382399  0.211302258  0.267646761  0.370709219 -0.148592054
#> [116] -0.199119786 -0.102752442 -0.246677392 -0.135112728 -0.167330516
#> [121]  0.363674595 -0.285819279 -0.155251949  0.139160031  0.360116908
#> [126]  0.145935250  0.079347913 -0.222249281 -0.385001842  0.348381307
#> [131] -0.382439709 -0.049359757 -0.392409566  0.118182111 -0.046989094
#> [136] -0.143017516  0.248361064 -0.386519069 -0.163151922 -0.327816022
#> [141]  0.370878445  0.307473686  0.371426026 -0.383001369  0.142523659
#> [146]  0.355024372 -0.190856621  0.051103618 -0.302202872  0.370005964
#> [151]  0.135699671  0.370080745 -0.387714561 -0.387447194 -0.142066960
#> [156]  0.343695816 -0.060436766  0.358582264 -0.392080617  0.371788560
#> [161] -0.350693066 -0.391085336  0.140739362 -0.393048191 -0.033178559
#> [166]  0.114734070 -0.040431247 -0.076257362 -0.019836994  0.229432824
#> [171] -0.252367246 -0.009221980 -0.369341030  0.141146252 -0.279827950
#> [176]  0.225627605 -0.391128361  0.133876676 -0.150604112 -0.274664431
#> [181] -0.392168578  0.317010109  0.371813853 -0.130838951  0.369540167
#> [186]  0.137607975  0.010748395  0.076495290  0.271251298 -0.391641426
#> [191] -0.381153626 -0.173391734 -0.072484450 -0.070494355 -0.206749292
#> [196]  0.363778672  0.365048454  0.014631561  0.099318760  0.135506715
#> [201] -0.390635628  0.366135888 -0.310445360  0.319909827 -0.065730383
#> [206]  0.208867755 -0.122902092 -0.387868658 -0.075662446  0.057626745
#> [211] -0.379242263  0.366248266 -0.391747750 -0.382105524 -0.046303744
#> [216] -0.346618233  0.010072778  0.370039993  0.151041201 -0.084620224
#> [221]  0.367974491  0.074716625 -0.147609122  0.338041696  0.241430667
#> [226] -0.254308563 -0.147055189 -0.142341694 -0.175638262  0.116896130
#> [231] -0.236787954 -0.382725299 -0.321097436 -0.391314478  0.017944514
#> [236] -0.331369435  0.077500926  0.119813860 -0.205885092  0.082022369
#> [241] -0.257839972 -0.372173955  0.371808011 -0.033878821 -0.384087120
#> [246] -0.249090923 -0.220613783 -0.391919232 -0.335401730  0.114238867
#> [251] -0.274899368 -0.388372824  0.129717301 -0.167052017  0.126099218
#> [256]  0.371298522 -0.348155029  0.036157279  0.087938633 -0.162316472
#> [261] -0.040411896  0.126335263 -0.386569171  0.371734503  0.358136413
#> [266]  0.312306536  0.355834030  0.311159445 -0.153891531 -0.077487495
#> [271] -0.278907050 -0.126606355 -0.391363701  0.371190060 -0.370161638
#> [276] -0.054624691  0.354228387 -0.363843181 -0.387795201  0.072959600
#> [281] -0.388999480  0.201897147  0.287852752  0.040144903  0.354957781
#> [286] -0.128622651 -0.079486874  0.129135683  0.369889113 -0.388820065
#> [291] -0.089426479 -0.392883517  0.335635537  0.147414858  0.138569786
#> [296] -0.157715318 -0.381598715 -0.032181822 -0.280744501 -0.198867652
#> [301] -0.392629930  0.099416964 -0.058803920  0.055075731  0.368884129
#> [306]  0.086180318 -0.387048352 -0.380768863  0.369095292 -0.227917996
#> [311] -0.263348371 -0.154890289 -0.117193153 -0.199007713  0.122049734
#> [316] -0.173012894 -0.205450431 -0.121392887 -0.114702529  0.337535215
#> [321] -0.061879912  0.178297559  0.302757389  0.325597733 -0.108054675
#> [326] -0.194378292  0.371569384 -0.223204059  0.356913544  0.195050762
#> [331]  0.272477679 -0.389512918 -0.377423821  0.202921693  0.119756646
#> [336]  0.282750131 -0.180438609  0.296418400  0.365966748 -0.045452690
#> [341]  0.008365339 -0.335363689 -0.142129205  0.113573707 -0.241571585
#> [346]  0.370094531 -0.388044900  0.332048140  0.368838657 -0.163609687
#> [351] -0.129119840 -0.388727775 -0.023375304 -0.390904219  0.364882248
#> [356]  0.078797365 -0.384332536  0.363550687 -0.193378524 -0.086371489
#> [361] -0.246515095 -0.389082243 -0.391555061  0.129726721 -0.100163339
#> [366]  0.150573583 -0.125982364 -0.313138668 -0.100908017  0.237321925
#> [371]  0.150568951 -0.200214109 -0.159197961 -0.062015596  0.022516719
#> [376] -0.026043949 -0.263028600  0.125323667 -0.392458724  0.370291036
#> [381] -0.095752389 -0.286056121  0.322254278 -0.094991572 -0.345796877
#> [386] -0.387065265 -0.173980602 -0.177409759 -0.332867969  0.365480697
#> [391] -0.388087211 -0.251101663 -0.073324791  0.061562945  0.370051191
#> [396]  0.341280081  0.349235138 -0.232533594 -0.372422944  0.359253213
#> [401] -0.328827195  0.371747315  0.371847991  0.370802493 -0.363410400
#> [406] -0.388692054 -0.243901929  0.368495853 -0.084528322 -0.142438772
#> [411]  0.272588617  0.159198309 -0.027711438  0.260487912  0.366677001
#> [416] -0.391193435  0.035179915  0.363450052 -0.214082775 -0.389270587
#> [421]  0.228198286  0.357806942  0.003883190  0.370762523  0.249256734
#> [426] -0.054808878  0.371403182 -0.037542143 -0.030659957  0.113521408
#> [431]  0.096297175  0.061993947 -0.015169426  0.369466632 -0.301006792
#> [436]  0.029904658  0.368293736  0.137224594 -0.254572513  0.055814371
#> [441]  0.367052379  0.295532043 -0.293097809  0.127056683  0.370974136
#> [446]  0.259062977  0.244339274  0.050474471  0.036734693 -0.255542859
#> [451]  0.118984189 -0.110079385  0.028335106 -0.193460982  0.093978083
#> [456]  0.169035259  0.371370955 -0.379328225 -0.340510010 -0.264300814
#> [461] -0.188248016  0.370072451  0.100949031 -0.131518784 -0.084404867
#> [466] -0.388408917  0.142805012  0.369556500  0.028078437  0.313020319
#> [471] -0.388781882  0.122361661  0.149109307  0.162198321  0.322112462
#> [476]  0.323110798 -0.382423623 -0.300563519 -0.364717101  0.143362065
#> [481] -0.371281996 -0.380103875  0.005394627  0.242629179 -0.014975469
#> [486] -0.288047067 -0.109792902  0.136377195 -0.016214956 -0.017834388
#> [491]  0.306473812 -0.382937087 -0.266449134 -0.247020283  0.174644611
#> [496] -0.387572390 -0.389286553 -0.042115317 -0.348683627  0.022158078