Find the posterior mean

Find the posterior mean given the observations and the mixing distribution based on the family in the result

posteriormean(x, result, fun = function(x) x)

posteriormean.npnorm(x, result, fun = function(x) x)

posteriormean.npt(x, result, fun = function(x) x)

posteriormean.npnormc(x, result, fun = function(x) x)

Arguments

x
a vector of observations

result
an object of class nspmix

fun
the function to transform the mean. It finds the posterior mean of fun(x). The function fun must be vectorised.

Examples

data = rnorm(500, c(0, 2))
r1 = computemixdist(data, pi0 = 0.5)
posteriormean(data, r1)

#>   [1]  0.1160961682  2.0020390322 -0.0467075902  1.9228958610  0.2702781576
#>   [6]  2.0282630884 -0.0710469390  1.9512021678  0.3351651444  1.4305666503
#>  [11]  0.2231489775  1.5255941229  1.5589726636  1.8859354301  0.1227637968
#>  [16]  1.7317197514  0.0150117689  1.8743918537 -0.0285503558  1.8463445326
#>  [21] -0.0888228178  1.5877098738  1.0318954647  1.4981854806  0.5962082609
#>  [26]  2.2270573108  0.9492950060  1.9714683489  0.6988159539  0.5653571907
#>  [31] -0.0238443250  2.2201921482  0.0275756860  2.0554969339  0.4441221289
#>  [36]  0.6258888326  0.6536967101  1.4021707223  0.0424061197  1.3710576361
#>  [41]  0.9355518254  1.9948946686  0.4671981518  1.9843908758 -0.0890820428
#>  [46]  1.7417233148  0.0187287158  1.1664464967 -0.0139692458  1.5112172999
#>  [51] -0.0127063731  0.8704904824 -0.0675164823  2.1816599674  0.4339921561
#>  [56]  1.8105825909  1.3186790611  0.8999076702  0.1821311027  2.0334046301
#>  [61]  0.0423283901  0.6332792417  0.0187272062  1.1929015583  0.4235109276
#>  [66]  2.0166341317  0.1772381182  1.5276325695 -0.1112649984  1.6230494289
#>  [71]  0.6034481232  0.9632580801  0.2939404332  1.8575444362 -0.0482628086
#>  [76]  2.1076337428  0.4810657994  2.0083277928  0.0346657944  2.1098034797
#>  [81]  0.0188833186  1.8482454722 -0.0544551594  1.1435946072  0.1232616199
#>  [86]  1.9896965510  0.0340451533  1.2956325057  0.1613209538  0.2689318923
#>  [91] -0.0378201076  0.4028628314  1.0536386204  1.6949991390  0.1909796057
#>  [96]  1.7008697780 -0.0604035102  1.1501016480  0.6507823042  1.8404071879
#> [101]  0.7731330724  1.8954647966  0.1297099035  1.4640015741  1.6919077729
#> [106]  1.5684974408  0.0896339334  1.8949546122  0.2692106211  0.5088363349
#> [111]  0.1023821365  1.9588676888 -0.0504924839  2.0621798476 -0.0363216328
#> [116]  0.2204400467 -0.0569699491  2.0654576576  0.9315140670  2.2294584671
#> [121]  0.1266846136  1.4637556491  0.9340107863  2.0310556744 -0.0826869240
#> [126]  1.6714805529  0.1257114687  0.3041828566  0.0838827810  1.0126425982
#> [131]  0.2803692045  2.0065044317 -0.0018813739  1.8644032727  0.0774905280
#> [136]  1.6036829505  1.5616847993  1.4611792879  0.2620622612  1.4688860243
#> [141]  0.3522233802  2.3319643933 -0.0182961182  0.8972522611  0.0630773688
#> [146]  1.0787712455  0.0908548651  0.4743789735  0.1991768001  0.1908738684
#> [151] -0.0133994892  1.6230433875 -0.0443484505  1.8819161218 -0.0371359132
#> [156]  1.7128002671  0.3012139656  1.0855533881 -0.0126045315  2.0346393887
#> [161]  0.0265962111  2.0417683634 -0.0296332380  0.5448871268  0.5511514484
#> [166]  1.7908396260  0.2030395684  0.6176458725  1.3204566495  1.7746875703
#> [171]  0.1831656306  1.1105035100 -0.1253546520  1.1454716342  0.0267623814
#> [176]  1.9000274115  0.0227641404  2.0625461505  0.1443343824  1.9129822629
#> [181]  0.4662842142  1.8749351998  0.2742116905  2.0439550303  0.3154336706
#> [186]  1.4459834060  1.8810147041  2.0387311940  0.0964733820  1.9190416599
#> [191]  1.1964392626  1.6130561885  1.0456724392  1.9870391376 -0.0067415813
#> [196]  0.3253244785  0.9955616040  2.4472702879  1.4277703246  1.8318260953
#> [201]  0.3884012622  2.5013946485  1.6632745510  1.9309983932  0.2542140692
#> [206]  2.0486783152  1.2042943997  1.8096842495 -0.0491633513  1.6066732045
#> [211] -0.1550264886  1.4722910783  0.0130168845  1.7715429283 -0.0909857636
#> [216]  0.3319297281  0.8643242413  1.5216248561  0.5264265042  2.0049882532
#> [221]  1.7363302649  1.3086883262  0.2236076947  2.5340114673  0.4279350714
#> [226]  2.1307410108  0.9798865222  1.1279836467  0.4929598676  0.4353632661
#> [231]  1.5966684845  1.8618653479  0.0119298368  1.4577234544  1.6896673018
#> [236]  1.8924984147  1.8271106596  1.1508863799  0.9426164791  0.8323481637
#> [241]  0.0758841340  2.0883229477  0.6092074786  1.9167776767  1.3380882149
#> [246]  2.1421967947  0.3348195672  1.8032486364  0.0871085906  1.8091920271
#> [251]  0.0566038562  0.5393801752  0.0653839471  1.2511198464  0.1523444317
#> [256]  1.2424739336 -0.0333604934  1.4924478898  0.3753036179  1.4956457887
#> [261] -0.0629374431  1.9913604012 -0.0291036597  0.0810610460  0.1222531117
#> [266]  1.6243895023  0.1664564165 -0.0404905852 -0.0094588608 -0.0126165114
#> [271] -0.0273337615  1.9344081407  1.0955975226  1.8704786582 -0.0111468132
#> [276]  1.8996099995  0.2586297297  1.8824737111 -0.0566068838  1.1262149495
#> [281]  0.0895242486  1.9183778521  0.7936640397  0.0664745489  0.3699825111
#> [286]  1.4817179323  0.0182753038  1.4189561283  0.1487427826  1.6778654555
#> [291]  0.1831522496  2.1051271530  0.1214039095  1.8440679507  0.0040521254
#> [296]  2.0401579081 -0.0054094717  1.7496012903  0.0031831338  2.0150803059
#> [301]  0.2955139404  2.2719570124  0.0125995535  2.0540642658  0.0128163956
#> [306]  1.0904200359  0.0037029728  1.0562219844  0.0235662945  1.9173759912
#> [311] -0.0333077034  1.9512249478 -0.0117120760  2.0552583522  0.0725054605
#> [316]  1.6355630757  1.5739122109  2.0523760695  0.7353056995  0.6312971913
#> [321]  1.2737912125  1.9942518277 -0.0505891384  0.6790957797  1.4074459787
#> [326]  2.0907205630  0.3668954976  1.9042530569  1.5110808021  0.4402253321
#> [331]  0.1707935009  1.6764721363  0.4161729498  1.4644642667  0.4896215538
#> [336]  1.7554349897  0.0455737579  2.1557316938 -0.0759056591  1.7510194928
#> [341] -0.0269554609  1.4010354915  1.0749762606  2.1172199409 -0.0496328298
#> [346]  1.9576638156 -0.1057066416 -0.0109864131 -0.0280048570  1.8393789944
#> [351] -0.0418157591  1.6504915165 -0.0850213016  2.1066294598 -0.0877525744
#> [356]  1.2216469523  0.0425564621  0.9611248066  0.0422387195  1.3128385243
#> [361]  0.0755303918  0.3488571881  0.4888593800  0.5164146217  0.0182473875
#> [366]  2.0373274681  0.0530360687  1.7207160044  0.0994164192  1.6605212168
#> [371]  1.3549550610  2.1956071052 -0.0383721701  1.8986938876  0.0193783151
#> [376]  1.4469509551  0.0498797227  0.6560735125  0.1601233964  1.9382757998
#> [381]  0.4696863751  1.0253879525  0.6616795046  0.3868139556 -0.0669516349
#> [386]  2.0305972872  0.6216987191  1.1906980325  0.0369430741  1.6813387035
#> [391] -0.0467207256  1.3499648205  0.0373441302  1.0219642278  0.0233186013
#> [396]  0.9782990018  0.0441179883  1.2059057801  0.8808947864  2.1489288016
#> [401]  0.0004350702  2.1285816133  0.1096504128  0.3838744099  0.0086178911
#> [406]  1.9614376711  0.0612345749  2.0164707469  0.1127393745  0.1671331035
#> [411] -0.0785044664  2.1794177708  0.9486114950  1.8744384033 -0.0171553431
#> [416]  1.9650190258 -0.0750834189  2.0376582353  0.3752645054  1.3402642833
#> [421]  0.0523263603  0.9912389610  0.4128759444  1.0239296031 -0.0589119262
#> [426]  1.9682803558  0.1405803060  1.8974217329 -0.0607801990  1.9247082676
#> [431]  0.6284174863  1.8542386371 -0.0496372147  2.0553074452  0.0543696467
#> [436]  2.0503783194  1.4952668100  1.8591946517  0.2128664193  2.1173270383
#> [441] -0.0144105956  2.0001783977  0.3673043431  1.5834866620 -0.0753215543
#> [446]  2.0388274628  0.6742436474  1.6639708943  0.4027067777  0.9626173743
#> [451]  0.3155276916  1.5147474036  0.1727020354  1.7421135324 -0.0182500729
#> [456]  1.9001608463  0.0400127109  2.0115011191  1.7520526163  1.6439418095
#> [461]  0.0689354902  1.5766527521  0.6333827250  1.6079140362  0.3151330736
#> [466]  0.9122318432  0.4622188533  1.7082315321 -0.0281682976  1.2644847724
#> [471]  0.4792250203  0.7653527540 -0.0383221922  1.9499466661  0.0443360952
#> [476]  0.2300000727  0.8218055738  1.4835218386 -0.0316884439  2.1175930836
#> [481] -0.0961235309  0.9556593919  0.5723263493  0.0836738568  0.9258196646
#> [486]  1.6936132735  0.0668621901  1.7324949146 -0.0555506421  1.2741658570
#> [491] -0.0459163212  1.7009927146  0.0293078290  1.6937944346  0.1382041951
#> [496]  1.0469101522 -0.1381097324  1.9959861243  1.9160619750  1.9306253943

r2 = computemixdist(data, pi0 = 0.5, method = "nptll") # equivalent to normal
posteriormean(data, r2, fun = function(x) x^2)

#>   [1] 0.32621838 4.18170465 0.07687016 3.96506005 0.61520839 4.26932562
#>   [7] 0.06466138 4.03681431 0.73995902 2.91673049 0.52545665 3.11049232
#>  [13] 3.17892385 3.87681625 0.33834279 3.53822136 0.15328269 3.85014268
#>  [19] 0.09419292 3.78660710 0.06416124 3.23803357 2.11415600 3.05445804
#>  [25] 1.24963318 5.21893463 1.94930103 4.09150936 1.45203594 1.18895717
#>  [31] 0.09958042 5.18253471 0.17300291 4.37294586 0.95152009 1.30808995
#>  [37] 1.36292743 2.85907990 0.19720548 2.79602291 1.92191317 4.15962010
#>  [43] 0.99657468 4.12832739 0.06419789 3.55943117 0.15903003 2.38362810
#>  [49] 0.11183821 3.08108374 0.11348849 1.79241568 0.06556768 4.97978777
#>  [55] 0.93176580 3.70757244 2.69010036 1.85093476 0.44816720 4.28785598
#>  [61] 0.19707644 1.32265758 0.15902768 2.43675654 0.91134272 4.22910710
#>  [67] 0.43901200 3.11466516 0.07056290 3.31101188 1.26388504 1.97713903
#>  [73] 0.66056451 3.81178532 0.07568368 4.60566375 1.02368488 4.20173470
#>  [79] 0.18446363 4.61610533 0.15927076 3.79086423 0.07148795 2.33777588
#>  [85] 0.33925010 4.14396800 0.18345194 2.64357885 0.40934272 0.61263320
#>  [91] 0.08459766 0.87115989 2.15762208 3.46086031 0.46476160 3.47318100
#>  [97] 0.06828896 2.35082851 1.35717716 3.77335082 1.59913918 3.89911585
#> [103] 0.35102722 2.98474735 3.45437905 3.19849420 0.27868286 3.89791498
#> [109] 0.61316631 1.07804645 0.30145701 4.05711694 0.07407417 4.40042857
#> [115] 0.08604933 0.52032568 0.07003393 4.41419341 1.91386881 5.23168116
#> [121] 0.34549621 2.98424649 1.91884286 4.27933178 0.06363045 3.41165882
#> [127] 0.34371914 0.68024874 0.26849659 2.07569301 0.63452978 4.19586778
#> [133] 0.12828746 3.82732617 0.25724647 3.27097649 3.18449436 2.97899989
#> [139] 0.59950224 2.99469710 0.77293134 5.78049092 0.10632025 1.84565018
#> [145] 0.23219411 2.20790148 0.28085273 1.01060953 0.48017567 0.46456303
#> [151] 0.11258061 3.31099938 0.07876736 3.86748906 0.08525545 3.49827239
#> [157] 0.67454002 2.22147656 0.11362233 4.29237662 0.17143710 4.31901875
#> [163] 0.09299963 1.14874958 1.16104950 3.66464666 0.48745220 1.29184737
#> [169] 2.69369063 3.62982411 0.45010486 2.27144245 0.07678485 2.34154075
#> [175] 0.17170243 3.90989279 0.16535493 4.40195764 0.37789479 3.94088986
#> [181] 0.99478889 3.85139053 0.62273604 4.32737709 0.70190376 2.94807371
#> [187] 3.86540324 4.30755493 0.29086966 3.95561269 2.44386501 3.29033989
#> [193] 2.14169366 4.13609331 0.12150598 0.72096793 2.04158830 6.40277254
#> [199] 2.91104875 3.75429362 0.84305918 6.69556029 3.39454606 3.98514878
#> [205] 0.58452107 4.34573105 2.45965204 3.70560962 0.07502050 3.27715115
#> [211] 0.09225258 3.00163539 0.15023119 3.62307198 0.06449981 0.73371260
#> [217] 1.78015618 3.10236919 1.11252655 4.19102660 3.54798876 2.66992717
#> [223] 0.52632583 6.87211891 0.91996129 4.71897151 2.01030664 2.30647336
#> [229] 1.04695636 0.93443869 3.25650160 3.82156428 0.14857856 2.97196380
#> [235] 3.44968451 3.89214495 3.74386922 2.35240282 1.93599038 1.71662120
#> [241] 0.25443192 4.51500762 1.27522587 3.95009372 2.72931834 4.77658081
#> [247] 0.73929172 3.69157609 0.27420285 3.70453456 0.22110430 1.13794014
#> [253] 0.23617131 2.55385892 0.39269427 2.53645140 0.08903429 3.04274450
#> [259] 0.81764120 3.04927248 0.06718113 4.14894157 0.09358095 0.26352077
#> [265] 0.33741232 3.31378622 0.41889530 0.08211318 0.11780949 0.11360658
#> [271] 0.09555477 3.99370194 2.24158633 3.84117655 0.11554999 3.90890400
#> [277] 0.59294727 3.86878038 0.07023472 2.30292788 0.27848805 3.95399222
#> [283] 1.63984602 0.23805630 0.80732402 3.02085318 0.15832480 2.89314615
#> [289] 0.38603295 3.42499261 0.45007979 4.59365982 0.33586575 3.78151707
#> [295] 0.13683407 4.31291903 0.12334366 3.57618076 0.13556528 4.22390395
#> [301] 0.66358661 5.45832033 0.14959586 4.36715778 0.14992584 2.23121945
#> [307] 0.13632361 2.16278840 0.16662216 3.95155011 0.08908887 4.03687402
#> [313] 0.11479985 4.37197941 0.24852965 3.33694043 3.20962897 4.36038493
#> [319] 1.52421442 1.31875020 2.59953427 4.15766605 0.07400688 1.41306990
#> [325] 2.86978241 4.52601869 0.80134108 3.91993632 3.08080471 0.94391923
#> [331] 0.42697779 3.42208146 0.89705453 2.98568972 1.04042297 3.58861207
#> [337] 0.20248208 4.84556498 0.06387929 3.57920062 0.09598274 2.85677716
#> [343] 2.20030677 4.65212934 0.07468179 4.05389945 0.06847520 0.11576345
#> [349] 0.09480080 3.77106119 0.08093072 3.36793940 0.06375351 4.60084660
#> [355] 0.06402158 2.49454324 0.19745514 1.97288517 0.19692759 2.67830607
#> [361] 0.25381284 0.76641968 1.03893152 1.09289704 0.15828141 4.30231356
#> [367] 0.21504099 3.51496183 0.29613647 3.38881013 2.76342967 5.05281643
#> [373] 0.08407314 3.90673592 0.16004237 2.95004187 0.20970786 1.36761748
#> [379] 0.40711798 4.00348040 1.00143712 2.10115282 1.37868139 0.83997709
#> [385] 0.06573979 4.27767983 1.29983268 2.43232935 0.18818887 3.43225307
#> [391] 0.07685992 2.75333444 0.18884702 2.09431267 0.16623052 2.00713938
#> [397] 0.20005285 2.46289112 1.81310658 4.81078085 0.13159088 4.70820588
#> [403] 0.31454916 0.83427055 0.14358974 4.06402348 0.22902611 4.22855816
#> [409] 0.32013474 0.42015547 0.06367257 4.96809433 1.94793863 3.85024955
#> [415] 0.10775388 4.07373735 0.06397481 4.30354537 0.81756535 2.73371758
#> [421] 0.21383923 2.03296039 0.89063755 2.09823914 0.06901266 4.08267802
#> [427] 0.37097833 3.90372982 0.06811457 3.96952603 1.31307379 3.80432899
#> [433] 0.07467865 4.37217819 0.21730311 4.35243705 3.04849876 3.81551554
#> [439] 0.50599890 4.65265312 0.11126558 4.17588653 0.80213335 3.22933452
#> [445] 0.06394563 4.30791571 1.40348694 3.39599719 0.87085646 1.97586140
#> [451] 0.70208487 3.08830130 0.43053870 3.56025983 0.10637781 3.91020900
#> [457] 0.19324176 4.21206494 3.58140141 3.35432957 0.24232004 3.21526715
#> [463] 1.32286160 3.27971410 0.70132476 1.87546705 0.98684671 3.48865510
#> [469] 0.09461820 2.58077965 1.02008487 1.58372058 0.08412039 4.03352918
#> [475] 0.20041635 0.53844806 1.69568796 3.02453223 0.09078535 4.65395469
#> [481] 0.06558229 1.96198813 1.20265569 0.26812766 1.90252566 3.45795420
#> [487] 0.23872697 3.53986262 0.07083644 2.60028942 0.07749355 3.47343919
#> [493] 0.17578268 3.45833403 0.36660741 2.14416821 0.08316985 4.16295002
#> [499] 3.94835358 3.98421683

data = runif(500, min = -0.5, max = 0.5)
r3 = computemixdist(data, method = "npnormcll", beta = 100)
posteriormean(data, r3)

#>   [1]  0.0321789599  0.3656832298  0.3726729131  0.3300651055  0.1136038041
#>   [6] -0.1857469612 -0.3631191237 -0.2478250691  0.2835413653 -0.3819341831
#>  [11]  0.2107105218 -0.0598221135 -0.3676120474  0.3727269644 -0.3815397861
#>  [16]  0.3397178472  0.0326295073  0.0864083018 -0.2577004605 -0.3079829921
#>  [21]  0.1008878470 -0.3815271200  0.0430934430  0.3736779182  0.3738836978
#>  [26] -0.1562218492  0.3732162703 -0.3825432921  0.1914620908 -0.0714461158
#>  [31] -0.2447749274  0.0645267331 -0.3755136085 -0.2712690996 -0.2083360643
#>  [36] -0.3814671374 -0.3653666403  0.0686390982  0.3281282368  0.3297755427
#>  [41] -0.2420658932  0.2857131905  0.0710028751 -0.1173978165  0.2288545281
#>  [46]  0.3745471822  0.3593676552 -0.3112008928  0.3538686571  0.0256987040
#>  [51] -0.0203674541  0.3498456606  0.0542246865 -0.1553437597 -0.1985167126
#>  [56]  0.0611031599  0.0488710341  0.3711096457 -0.3792431938  0.0219069478
#>  [61] -0.3760056198 -0.1982379089  0.1553299677 -0.2224756050  0.0421361153
#>  [66] -0.0927591950  0.0598021924 -0.1831870570 -0.3604981750  0.1430876467
#>  [71]  0.0750709968 -0.2303959765  0.3753433284 -0.3014300709  0.3689709222
#>  [76]  0.0097430365 -0.2084121355  0.2408627514 -0.3254052092  0.0451938237
#>  [81]  0.3719382178 -0.0202015100  0.0647787437  0.0204737195 -0.2123645769
#>  [86]  0.1051546282 -0.3798212899 -0.3670664942 -0.1919859860 -0.1809623344
#>  [91] -0.3499489584  0.0536707702 -0.3789040833  0.2387245017  0.2531574791
#>  [96] -0.3341789948  0.2967589057 -0.3248711202  0.1712657731 -0.2620214571
#> [101]  0.0659270568 -0.3584778639 -0.3125315271 -0.2055886684  0.1098751818
#> [106]  0.3603248663  0.3061317496 -0.3610228880  0.0605630296  0.3231231808
#> [111]  0.3750461804  0.3257855348 -0.2516385725  0.3003878588  0.2509011666
#> [116]  0.0423145497 -0.3805329072 -0.3662147879  0.3755187200  0.0625709104
#> [121] -0.2907663837  0.0643255214 -0.3689187144  0.0580097439 -0.3260814545
#> [126]  0.3743626247  0.0677425177 -0.1815311052 -0.3443825090  0.0662308910
#> [131] -0.2084995084  0.0345500548  0.0667662627  0.3708387030 -0.0941714003
#> [136] -0.3818305385  0.0418289546 -0.2141776747 -0.3239108841 -0.1799495198
#> [141] -0.2080097529  0.3650816103 -0.1972211264  0.0351558971  0.3732601688
#> [146] -0.3702562126 -0.1505698386  0.0689787685 -0.2391833619  0.0848812842
#> [151]  0.2698193271  0.0062316086  0.3263831830 -0.1819137713  0.0374184286
#> [156] -0.3774492677  0.3482047638  0.0617671458 -0.2594516004 -0.3704178762
#> [161]  0.3698702473  0.2965680930 -0.2930186647  0.2055904905  0.3055664200
#> [166] -0.1677245934 -0.2581929848  0.3726055574  0.3689663474  0.3637395924
#> [171]  0.0242514637 -0.3778244057 -0.3471547825 -0.2636296736 -0.1844344616
#> [176]  0.2112807046  0.0732902885 -0.3822867382  0.0356850152  0.2390918994
#> [181] -0.3798498745  0.3745099195  0.3687770774  0.0560677462 -0.0063161769
#> [186]  0.3548072358 -0.3805230029  0.3705180574  0.2433008547 -0.2726343103
#> [191] -0.1948937687 -0.3010326916  0.2634478874  0.2192015109  0.2490825733
#> [196] -0.3633414410 -0.3802207209 -0.3707391017  0.2194245968 -0.3764166933
#> [201] -0.2096783372  0.0984875508 -0.1712809273 -0.1008770070  0.0329600025
#> [206]  0.3744505879  0.0665213794 -0.0270707499 -0.0132506251  0.0548759231
#> [211]  0.2930008144 -0.3552975199 -0.3734139966 -0.1633392337 -0.2123299378
#> [216]  0.3724122240  0.0709304436  0.0536891037  0.0400880988 -0.3767081534
#> [221]  0.3713263646 -0.1220993987  0.0361673660 -0.3784655692  0.3648138868
#> [226] -0.3823321152 -0.3797779672 -0.3796498453  0.0132197001 -0.3762032079
#> [231]  0.0441404240  0.3331690491 -0.3482065366 -0.3678314691  0.3728191399
#> [236]  0.0300731595 -0.2299249623  0.3682794773  0.1655728750 -0.1092472418
#> [241]  0.3537007303 -0.0508186454  0.3744585305 -0.1627305609  0.1571727925
#> [246] -0.1644575544 -0.1787047939 -0.2468213214 -0.3832098988  0.3736190893
#> [251]  0.0441294911  0.3704314470 -0.1622112545 -0.0906019550  0.3307559374
#> [256] -0.3607654750  0.3721573705  0.3699804839 -0.3791793059  0.2970020655
#> [261]  0.0532807759 -0.0926082244 -0.3340765990  0.3751111444  0.3608890662
#> [266] -0.3597824138  0.0310614097  0.3451437860  0.1513503937 -0.3803367441
#> [271] -0.1471037558 -0.0512815878  0.3712039256  0.0630169810 -0.3740190266
#> [276]  0.2789144425  0.3614174835  0.0607862721  0.3749811018 -0.1883564798
#> [281]  0.3063677884 -0.2213449971 -0.0529109464  0.0666697243 -0.1645344754
#> [286]  0.2420752176 -0.0421389574  0.1402256908  0.0451583458 -0.0520511598
#> [291]  0.0999814373 -0.1718374213  0.3728363129 -0.3712978174  0.1120457470
#> [296]  0.0705223125  0.0772227544  0.2040516022  0.1242249902 -0.3726539977
#> [301]  0.3671835333  0.2809279916  0.3630183057  0.3751240838 -0.2173520082
#> [306]  0.3412281857  0.0502145303 -0.3764271980 -0.1157265890  0.3677468278
#> [311] -0.2784568534  0.0394294684 -0.1886561416  0.0764805198  0.0425052411
#> [316]  0.0273802910 -0.3806854464 -0.3568187438  0.0498395219 -0.3767714159
#> [321] -0.3805533372  0.3708782495  0.0282397686 -0.1622438053 -0.2612783690
#> [326] -0.2716293837 -0.2322984296 -0.2451284901 -0.2764675628  0.3725597953
#> [331] -0.3056589726 -0.2660917439  0.3477520423 -0.3819817816 -0.1910933711
#> [336]  0.0932210403  0.0405867662 -0.0145078183  0.3518989399 -0.2543152782
#> [341]  0.0566751784 -0.3608091997  0.0475337789 -0.3181032736 -0.3484806830
#> [346] -0.3482084156  0.3044680751 -0.3358402612  0.3702349570  0.1982632114
#> [351]  0.1602219904  0.2931790207 -0.1728559380  0.0396723845  0.2800981929
#> [356] -0.2860551116 -0.3832382516  0.3732248701 -0.2394046443  0.0131651052
#> [361]  0.0450607232 -0.3456095517  0.3749432349  0.3686109272 -0.3602027571
#> [366]  0.3567576388 -0.2646798330  0.2934667132 -0.2022665264 -0.2193788856
#> [371] -0.1997490894  0.2407745169  0.0642447935 -0.3044711727  0.1131167511
#> [376] -0.0196146550 -0.2046740106 -0.1747572291  0.0479203409  0.3419218172
#> [381]  0.0499770012  0.0371571173 -0.3353127977 -0.3335464752 -0.3754441489
#> [386] -0.2276743965  0.0642915189 -0.2116640357 -0.0107031250  0.3743534640
#> [391] -0.0199630447  0.0446026091  0.3730202291 -0.0240504361  0.3549311553
#> [396] -0.1104927180  0.1514942640 -0.0130188965 -0.3744718315 -0.0016997424
#> [401] -0.3299518167 -0.3818357293  0.3715197842 -0.0253340867  0.3676816977
#> [406]  0.1698551817 -0.3772253067 -0.2768625961  0.3236275389 -0.1209160266
#> [411] -0.2900011693 -0.3544660798  0.3723290188  0.3345617087 -0.0341053521
#> [416] -0.2789617453  0.1690887501  0.3494119752  0.3698065300 -0.2014490399
#> [421]  0.3632503698 -0.2569931398  0.0009927975  0.3740354945  0.3646758323
#> [426]  0.0550777630 -0.2354228261  0.3706596154  0.0806239774  0.3709906059
#> [431] -0.2109615757  0.2226502550  0.1098753048 -0.3335215576 -0.3711195125
#> [436] -0.2588081086  0.2190932744  0.3683603566  0.3669572441 -0.3570413440
#> [441]  0.3681301188 -0.3758283680 -0.0245894707 -0.3794559285  0.3753284933
#> [446]  0.0557229152 -0.2166345690  0.3631109372  0.0889978820 -0.3221572975
#> [451] -0.3713792275  0.3438267204  0.0952882419 -0.3377372796  0.0685606932
#> [456]  0.2832098360 -0.3763995556 -0.2207637193  0.0382573326 -0.2657351532
#> [461]  0.0825885881 -0.3795108008  0.1534315094  0.0894375142  0.0604559767
#> [466]  0.3724131053 -0.2643477743 -0.1784203240 -0.3733515302 -0.3581457714
#> [471]  0.3717899428 -0.3742344401 -0.1374130679  0.0720354759  0.3670089152
#> [476] -0.0217534688 -0.3820984992 -0.2346742985  0.3748415412  0.0356069716
#> [481] -0.3514131944 -0.3805563694 -0.1149637006  0.3674337773  0.3163026462
#> [486] -0.0856280362  0.0688983425  0.0278832971 -0.2616515872  0.3740745672
#> [491] -0.3830278309  0.3659967928  0.0386032124 -0.0469817315 -0.3125368173
#> [496]  0.3542896597 -0.3834462079 -0.3664910346 -0.1749985649  0.2877539477