Find the posterior mean

Find the posterior mean given the observations and the mixing distribution based on the family in the result

posteriormean(x, result, fun = function(x) x)

posteriormean.npnorm(x, result, fun = function(x) x)

posteriormean.npt(x, result, fun = function(x) x)

posteriormean.npnormc(x, result, fun = function(x) x)

Arguments

x
a vector of observations

result
an object of class nspmix

fun
the function to transform the mean. It finds the posterior mean of fun(x). The function fun must be vectorised.

Examples

data = rnorm(500, c(0, 2))
r1 = computemixdist(data, pi0 = 0.5)
posteriormean(data, r1)

#>   [1] 0.52334114 1.62947189 1.62598239 2.08145705 0.07356137 2.06713115
#>   [7] 1.15289892 1.24380358 1.44558215 2.08981688 0.55483761 2.01915716
#>  [13] 0.05360459 2.10845214 0.09061258 1.50533571 0.47154803 1.80509202
#>  [19] 1.54052908 1.15659444 0.09785373 1.87264563 0.50809695 2.46058101
#>  [25] 0.03953543 2.04465673 0.78188255 2.04328280 0.04076686 1.90729464
#>  [31] 0.10376423 0.31745909 0.06099066 0.84556592 1.50391493 1.18386448
#>  [37] 0.78981625 1.97067361 0.06960984 1.95541427 0.50290346 0.74265246
#>  [43] 0.63467220 1.74946149 0.15071460 1.00387285 0.16631552 0.80615025
#>  [49] 0.20918777 1.67311314 0.14939769 2.04435990 0.08753974 1.67339101
#>  [55] 0.75779646 2.06256976 0.18495976 1.78663325 0.98811747 1.89280471
#>  [61] 0.36125975 1.58426443 0.25319813 1.75639658 0.51713641 1.66490097
#>  [67] 0.13625361 1.70929149 0.10927319 2.08030087 0.05083247 2.05348991
#>  [73] 0.21359461 0.21029560 0.09689262 2.04838404 0.03981283 1.81872405
#>  [79] 0.04249195 1.79444915 0.91229069 1.62405965 0.84652097 1.99432955
#>  [85] 1.10021522 2.06351028 0.88914134 1.91555229 0.23894701 1.82104288
#>  [91] 0.37881604 0.98851178 0.09053432 1.38587334 0.09971970 1.56802688
#>  [97] 0.19929697 1.09590479 0.01899066 2.12695462 0.06773766 2.03487683
#> [103] 0.08771247 1.94595397 0.17598500 1.96970204 0.02473635 1.86292305
#> [109] 0.17268040 2.06782991 0.13149084 1.48245278 0.02492147 1.83279927
#> [115] 0.54925064 1.60303068 0.09190149 1.90616120 0.04389693 1.89105660
#> [121] 0.90101452 1.65219491 0.12099326 1.51467222 0.45429827 0.63630000
#> [127] 0.06559833 0.46828875 0.14464707 1.73122028 0.24789061 1.64982948
#> [133] 0.02510268 0.63285057 0.13854532 1.87727091 0.29004032 0.84096541
#> [139] 0.91025447 1.77653892 0.12808056 1.87464777 0.74755447 1.96984774
#> [145] 0.11965732 0.71662836 0.75694739 0.84443488 0.08002414 2.04739621
#> [151] 0.09930614 1.58743640 0.44584821 2.06370911 0.05538304 1.99676622
#> [157] 0.20769484 2.07045668 0.85002887 1.72623529 0.07323430 0.61514372
#> [163] 0.05376770 1.43583425 0.04339047 1.66613257 0.09812789 2.12173591
#> [169] 0.98873713 1.94811340 0.09508673 2.12686306 0.14213805 1.73176025
#> [175] 0.02331212 2.04247111 0.19952069 0.87341593 0.21449023 2.04072477
#> [181] 0.24620005 1.88713256 0.94840890 0.97187247 0.14850794 1.98741413
#> [187] 0.03609380 2.05213528 0.04982600 1.87467074 0.28532280 1.29551127
#> [193] 0.30297104 1.63315743 0.78600275 0.57715780 1.18930073 1.58963222
#> [199] 0.07446291 0.82718304 0.96868474 0.43829564 0.26854578 1.29370731
#> [205] 0.07195779 1.65465241 0.23765693 1.49834812 0.07390080 1.99434592
#> [211] 0.58221123 2.08840034 0.05508991 0.88070876 0.06784487 0.74741208
#> [217] 1.74366875 0.10178030 0.03553867 1.13187767 0.15066849 1.66317105
#> [223] 0.90342304 2.05928998 0.05188025 1.85799961 0.02800980 2.01907670
#> [229] 0.06308301 1.93136599 0.05469741 1.76159633 0.21459071 0.81772664
#> [235] 0.38819232 2.78763084 0.22397798 1.45847737 0.10097955 1.50538269
#> [241] 0.28161536 1.01764842 1.40371012 1.96991823 0.81406075 2.06754225
#> [247] 0.04920151 0.95171263 0.35946261 1.40232941 0.47126812 1.71243320
#> [253] 0.61864335 0.39600069 0.61600453 0.27546406 0.21101689 1.93835173
#> [259] 0.06959755 2.07661047 0.30379303 0.39441426 1.43951407 1.88963264
#> [265] 0.14266958 1.99364199 0.04154425 1.75336478 0.33553224 0.05961604
#> [271] 0.11723904 1.20097969 0.37440592 2.04431040 0.27258775 1.08841672
#> [277] 0.19289515 0.87863407 0.05385639 0.10265975 0.14593774 1.96675821
#> [283] 0.32960960 0.84109491 0.10785043 2.04600090 0.64671757 1.95572979
#> [289] 0.13347170 1.72255793 0.04788545 1.77320851 0.40294015 1.78136886
#> [295] 0.30217670 0.11816293 0.33723468 1.53230671 0.04107483 1.55267415
#> [301] 0.29092727 0.93312044 0.22431218 0.71864891 0.04961759 2.10331564
#> [307] 0.67670877 0.41270209 0.15111743 1.81810652 0.13619770 2.04253528
#> [313] 1.87097360 1.31016601 0.74013582 2.04480657 0.40388749 1.93831151
#> [319] 1.11720156 1.81341897 0.02253117 1.93203334 0.04001587 1.27823622
#> [325] 0.15069465 0.17998107 0.16393277 2.01848621 0.08710400 0.47287205
#> [331] 0.13731726 1.31546207 0.34997965 1.26910868 0.07848341 1.90531636
#> [337] 0.46091080 2.04735807 0.95404205 1.23575730 0.79704661 0.59354968
#> [343] 0.27784043 0.22136225 0.40301757 0.81709254 0.19377628 1.87662251
#> [349] 0.03281806 2.01991056 0.98543760 1.87783820 0.04218148 0.47411454
#> [355] 0.03271230 1.43225701 0.18425377 1.90384689 0.69581506 1.36708933
#> [361] 0.99009943 1.49773274 0.03279236 1.99161526 0.50333815 1.81203805
#> [367] 0.45780482 2.07115493 0.13172503 2.03353988 0.09609415 2.06763295
#> [373] 0.16846260 1.60433161 0.55027061 1.97096242 1.32925620 1.93821388
#> [379] 0.89648780 1.98624460 0.10610443 2.05816693 0.05641394 2.03480024
#> [385] 0.06150957 2.02998922 0.18278094 1.77121740 1.71315231 1.13098369
#> [391] 0.93601406 1.21239563 0.22556448 1.67916853 1.71908354 2.05857626
#> [397] 0.24382355 1.81952234 0.24622065 0.55137924 0.02809845 1.63668619
#> [403] 0.15865256 1.16503604 0.06172109 1.72258171 1.61627240 1.51315677
#> [409] 1.01672112 1.59439317 0.34926051 2.15173223 0.34161514 1.24282782
#> [415] 0.02759829 1.37206670 0.88662612 1.41533112 0.71340293 1.59878961
#> [421] 0.03584842 0.25479918 0.05375008 1.75649126 0.10362198 1.10355855
#> [427] 0.24925139 0.79856804 0.85778650 1.29206432 0.16629353 0.98532760
#> [433] 0.49661122 1.91072321 0.11152038 0.22756250 0.27780416 1.92599432
#> [439] 0.45764278 2.03783506 0.06745763 1.15305028 0.04035883 2.01442457
#> [445] 0.10921708 1.09291307 0.20217893 1.63403482 0.08564869 2.12026534
#> [451] 0.07458504 2.04701834 0.53494598 2.10353614 0.25715696 1.75736460
#> [457] 0.08788554 0.69802390 0.13737596 1.29663470 0.10086576 1.91011230
#> [463] 0.11116533 0.03547352 0.15776576 1.84301640 0.19471222 1.84830443
#> [469] 0.11066297 2.02682300 0.13002734 1.42126660 0.81533665 2.05173639
#> [475] 0.11575941 1.67706414 0.07757658 1.81751951 1.58689796 2.01701449
#> [481] 0.56997144 1.93731502 0.15714693 0.36246001 0.65500999 1.82429223
#> [487] 0.10047334 2.06178632 0.74094479 2.05325790 0.65363479 1.12904446
#> [493] 0.14725395 0.24921821 0.15962915 0.30562341 0.25809399 1.82848608
#> [499] 0.08763675 1.73952136

r2 = computemixdist(data, pi0 = 0.5, method = "nptll") # equivalent to normal
posteriormean(data, r2, fun = function(x) x^2)

#>   [1] 1.01265269 3.35413831 3.34665156 4.41131054 0.09973356 4.35145670
#>   [7] 2.33789921 2.53090930 2.96065738 4.45154800 1.07833030 4.21020078
#>  [13] 0.06389839 4.55208330 0.13162723 3.08830033 0.90487057 3.73243730
#>  [19] 3.16357029 2.34573851 0.14540570 3.87924943 0.98089978 6.79441330
#>  [25] 0.04038537 4.27782484 1.55410098 4.27388115 0.04235400 3.95514042
#>  [31] 0.15673393 0.58633518 0.07691027 1.68814593 3.08526311 2.40360496
#>  [37] 1.57078784 4.09625768 0.09248274 4.06184973 0.97008746 1.47164368
#>  [43] 1.24519452 3.61221807 0.24855256 2.02227788 0.27956812 1.60515436
#>  [49] 0.36563010 3.44785090 0.24594364 4.27696885 0.12581770 3.44844808
#>  [55] 1.50346369 4.33478254 0.31686308 3.69249431 1.98896885 3.92333595
#>  [61] 0.67648354 3.25721232 0.45488797 3.62718061 0.99972569 3.43020473
#>  [67] 0.21999320 3.52566293 0.16735073 4.40604293 0.05911439 4.30445511
#>  [73] 0.37453182 0.36786706 0.14357024 4.28877409 0.04082697 3.76197577
#>  [79] 0.04514603 3.70940002 1.82882683 3.34252671 1.69015788 4.15062526
#>  [85] 2.22620695 4.33813459 1.77999324 3.97331858 0.42590377 3.76700402
#>  [91] 0.71271753 1.98980234 0.13147897 2.83329119 0.14897464 3.22243208
#>  [97] 0.34568461 2.21707405 0.01244955 4.66081179 0.08907014 4.25065499
#> [103] 0.12614363 4.04069172 0.29888220 4.09405457 0.01899571 3.85804200
#> [109] 0.29227436 4.35410685 0.21063377 3.03939646 0.01922789 3.79251515
#> [115] 1.06667333 3.29743047 0.13407100 3.95264850 0.04744732 3.91950608
#> [121] 1.80503634 3.40291327 0.19009910 3.10826198 0.86904044 1.24860215
#> [127] 0.08519000 0.89809790 0.23654525 3.57289072 0.44408520 3.39783399
#> [133] 0.01945631 1.24138139 0.22450540 3.88935014 0.53011496 1.67845495
#> [139] 1.82453034 3.67067484 0.20394796 3.88362073 1.48194206 4.09438483
#> [145] 0.18749599 1.41699620 1.50167928 1.68576329 0.11171691 4.28583458
#> [151] 0.14818303 3.26400854 0.85150174 4.33884880 0.06699829 4.15632728
#> [157] 0.36261647 4.36430824 1.69754822 3.56215002 0.09913107 1.20432962
#> [163] 0.06418174 2.93985191 0.04661507 3.43285080 0.14592961 4.62956543
#> [169] 1.99027867 4.04551142 0.14012681 4.66026056 0.23159007 3.57405432
#> [175] 0.01725034 4.27157286 0.34613522 1.74683610 0.37634204 4.26665794
#> [181] 0.44064630 3.91091408 1.90507171 1.95463696 0.24418185 4.13456288
#> [187] 0.03500430 4.30021019 0.05739306 3.88367089 0.52046087 2.64085849
#> [193] 0.55660596 3.36204663 1.56276647 1.12492844 2.41514448 3.26871369
#> [199] 0.10139636 1.64942925 1.94790153 0.83583351 0.48617734 2.63702061
#> [205] 0.09678358 3.40819073 0.42328355 3.07336390 0.10035925 4.15066350
#> [211] 1.13548459 4.44447714 0.06648580 1.76221149 0.08926515 1.48164289
#> [217] 3.59972480 0.15292389 0.03415472 2.29331832 0.24846118 3.42648823
#> [223] 1.81011730 4.32342057 0.06091542 3.84731459 0.02326531 4.21000086
#> [229] 0.08065633 4.00826265 0.06580052 3.63840326 0.37654514 1.62952008
#> [235] 0.73209009 8.90956992 0.39554005 2.98818776 0.15138823 3.08840078
#> [241] 0.51287802 2.05141081 2.87132102 4.09454464 1.62180338 4.35301266
#> [247] 0.05632945 1.91204902 0.67277750 2.86837665 0.90428889 3.53242607
#> [253] 1.21165070 0.74823379 1.20613031 0.50030508 0.36932380 4.02376542
#> [259] 0.09246030 4.38974082 0.55829135 0.74495310 2.94770538 3.91638741
#> [265] 0.23263930 4.14902055 0.04360743 3.62063869 0.62348598 0.07446258
#> [271] 0.18279021 2.43993975 0.70361064 4.27682631 0.49442971 2.20121037
#> [277] 0.33280136 1.75783717 0.06433589 0.15461191 0.23909660 4.08739039
#> [283] 0.61130394 1.67872772 0.16460376 4.28172968 1.27041488 4.06255741
#> [289] 0.21452335 3.55422851 0.05409963 3.66347936 0.76258890 3.68111306
#> [295] 0.55497740 0.18458704 0.62698897 3.14597828 0.04284960 3.18956237
#> [301] 0.53193068 1.87279021 0.39621698 1.42123757 0.05703771 4.52319648
#> [307] 1.33325578 0.78279478 0.24935090 3.76063688 0.21988319 4.27175478
#> [313] 3.87559998 2.67204149 1.46635716 4.27825779 0.76454915 4.02367602
#> [319] 2.26220547 3.75047614 0.01632549 4.00974172 0.04115091 2.60411228
#> [325] 0.24851303 0.30688214 0.27481885 4.20853551 0.12499586 0.90762221
#> [331] 0.22208678 2.68331306 0.65323160 2.58470222 0.10884717 3.95079158
#> [337] 0.88277137 4.28572166 1.91696896 2.51381065 1.58599853 1.15917743
#> [343] 0.50516100 0.39024340 0.76274910 1.62818526 0.33457327 3.88793367
#> [349] 0.03007474 4.21207554 1.98330440 3.89058953 0.04464073 0.91020457
#> [355] 0.02991915 2.93221800 0.31544686 3.94756269 1.37332224 2.79325816
#> [361] 1.99315835 3.07204861 0.03003691 4.14430069 0.97099232 3.74748371
#> [367] 0.87632113 4.36708475 0.21109338 4.24708347 0.14204687 4.35335719
#> [373] 0.28385118 3.30021937 1.06880124 4.09691294 2.71267697 4.02345904
#> [379] 1.79548767 4.13186227 0.16123744 4.31964117 0.06880537 4.25044958
#> [385] 0.07783707 4.23773375 0.31249330 3.65917827 3.53397424 2.29142284
#> [391] 1.87889919 2.46418241 0.39875397 3.46086636 3.54674560 4.32101238
#> [397] 0.43581378 3.76370668 0.44068821 1.07111423 0.02338546 3.36961951
#> [403] 0.26430959 2.36364792 0.07821529 3.55427974 3.32582342 3.10502158
#> [409] 2.04944945 3.27891639 0.65175002 4.81286148 0.63600502 2.52883562
#> [415] 0.02271047 2.80386446 1.77468903 2.89610642 1.41022619 3.28833930
#> [421] 0.03462809 0.45814855 0.06415113 3.62738494 0.15646049 2.23329133
#> [427] 0.44685400 1.58919959 1.71389410 2.63352534 0.27952427 1.98307189
#> [433] 0.95699145 3.96268286 0.17169623 0.40280299 0.50508688 3.99637110
#> [439] 0.87598465 4.25866807 0.08856103 2.33822028 0.04169937 4.19853645
#> [445] 0.16724232 2.21073572 0.35149129 3.36392947 0.12225466 4.62083045
#> [451] 0.10162185 4.28471757 1.03684052 4.52442113 0.46295174 3.62926962
#> [457] 0.12647028 1.37795585 0.22220233 2.64324864 0.15117011 3.96133842
#> [463] 0.17100911 0.03405537 0.26254671 3.81471248 0.33645589 3.82621159
#> [469] 0.17003725 4.22954522 0.20776293 2.90876822 1.62448906 4.29897235
#> [475] 0.17991508 3.45634282 0.10716177 3.75936424 3.26285484 4.20489661
#> [481] 1.10992060 4.02146182 0.26131690 0.67895904 1.28778363 3.77405197
#> [487] 0.15041807 4.33202301 1.46805647 4.30372346 1.28490288 2.28731127
#> [493] 0.24170001 0.44678648 0.26625168 0.56204492 0.46486113 3.78315200
#> [499] 0.12600075 3.59078260

data = runif(500, min = -0.5, max = 0.5)
r3 = computemixdist(data, method = "npnormcll", beta = 100)
posteriormean(data, r3)

#>   [1] -0.3613611484 -0.3604237289  0.1777899557 -0.2034152222  0.3683695359
#>   [6]  0.3032663681  0.2420179138 -0.0758284619  0.3671256549  0.3692651183
#>  [11] -0.0717481640 -0.0729118800 -0.0775773242  0.2961812026  0.3485953588
#>  [16]  0.3582558763  0.2182582498 -0.3617064944  0.2935154163  0.3252151605
#>  [21] -0.0711287182 -0.3554980206 -0.0633581892 -0.3374863707  0.0960933115
#>  [26]  0.0661425431 -0.0426370076  0.3402833330  0.2911383385 -0.3614911093
#>  [31] -0.0876203983  0.3505755646 -0.3460923501 -0.3606537964  0.3634017328
#>  [36] -0.3583212314 -0.2975420540  0.3445711751  0.1255653762 -0.1739611299
#>  [41] -0.3615615322  0.2901159399 -0.2385816885  0.3304391090  0.3698884344
#>  [46]  0.3465823661 -0.0221391925 -0.0821315709  0.2972591253  0.1676119336
#>  [51]  0.2275953995 -0.0538285306 -0.3609155762  0.2925932552  0.3071313206
#>  [56] -0.0718505453 -0.3542196105  0.3685539359  0.2707854884  0.3660018383
#>  [61] -0.3593538325  0.3500777656 -0.0609727374  0.3543707292  0.2244432622
#>  [66]  0.3696889288 -0.0818940746  0.1562442554 -0.0716379129 -0.1669545916
#>  [71]  0.2348729364 -0.3606624967  0.2727237617  0.2598794399  0.3641932517
#>  [76] -0.0728213527  0.3702042829 -0.0025996758 -0.0697351378  0.1724883716
#>  [81]  0.3589578410 -0.3456685627  0.2654355347  0.2121116938 -0.0323674712
#>  [86] -0.0736516351 -0.0640466327  0.3344858474 -0.3527824386  0.1028753401
#>  [91]  0.3312506332 -0.0804201115  0.3374711581 -0.0357816040 -0.1348579863
#>  [96]  0.3527135850 -0.3592945306 -0.1251299448 -0.0469418739  0.2417532107
#> [101]  0.3459987881 -0.2361304366  0.3672030235  0.3698475739 -0.1710893538
#> [106]  0.3498918169 -0.0702279557 -0.0772714584  0.3669401446  0.3388622882
#> [111] -0.0949149400 -0.3372330561 -0.3536012283  0.2537659200  0.1632247467
#> [116] -0.3616237955  0.3698278327 -0.3555076819 -0.3041534621  0.2095027152
#> [121] -0.1197411784  0.3685727613 -0.0691007347  0.3388309157 -0.0368853400
#> [126]  0.1254372974 -0.0916437204  0.3085156942  0.2112213531 -0.3453087240
#> [131] -0.2477131727  0.2390579652 -0.0710945461 -0.3617062771  0.2610336200
#> [136]  0.2204296963  0.0018537438  0.0009598031  0.2646827866 -0.0607210332
#> [141] -0.3608018704 -0.1023494061 -0.3304107494 -0.0774874123 -0.0721416240
#> [146]  0.2228229015 -0.3612969200 -0.3615443848  0.0463311694 -0.0778211260
#> [151]  0.3197254184  0.2866430303 -0.0709012168 -0.3597895504 -0.3613894255
#> [156] -0.3273322276 -0.0762144116  0.0885341995 -0.3602954854 -0.3350807523
#> [161] -0.0771348595 -0.1135339270 -0.3608436933 -0.0693191465 -0.3616264151
#> [166]  0.0234233177  0.3054526196 -0.3586444162  0.1038384843 -0.0562886788
#> [171] -0.3587124311 -0.0537864539 -0.3608874201 -0.0750181287 -0.3596356047
#> [176]  0.3630072268 -0.1570162462 -0.0713272581  0.3524840423 -0.0688644904
#> [181]  0.1779471297  0.3156006023 -0.3498686802  0.2075655589  0.3292589494
#> [186]  0.3684938502 -0.3608045836 -0.0719133855  0.3665603390 -0.3604557856
#> [191] -0.0763421863 -0.3610639772  0.3622626352  0.1659640512 -0.3539605175
#> [196]  0.3669139356 -0.0607891830 -0.3568175295 -0.1398436200 -0.0756589998
#> [201]  0.1589657782 -0.0745101516 -0.3509338790  0.3534494683 -0.3595947520
#> [206] -0.3550373865 -0.3491787877  0.3686608577  0.3674252193 -0.3560677143
#> [211]  0.0879275800 -0.2221814768  0.3364050417 -0.0976540435 -0.3084513389
#> [216]  0.3274096630 -0.0696399652 -0.0768170881 -0.0681614123 -0.3615377310
#> [221] -0.0777502176  0.3510246946 -0.0713175919 -0.0501257913  0.1308805448
#> [226]  0.2429113492 -0.3195883190  0.3616784081  0.2857864804 -0.0777385556
#> [231]  0.3676625711 -0.3600557199 -0.0649856794 -0.3615790309  0.2218833398
#> [236] -0.0792057224  0.0629959905 -0.3585883324 -0.0618024158  0.1671210555
#> [241]  0.2025811933 -0.3570439636 -0.0646119104 -0.3607942352 -0.3386706184
#> [246]  0.3471111282 -0.0678888367 -0.0677470275  0.3534581995  0.3644568039
#> [251]  0.2341335570 -0.0703918815  0.1845239774 -0.3227688947 -0.0698105391
#> [256] -0.0349262062  0.2299767491  0.2292739485 -0.3616229812  0.0369649089
#> [261] -0.0659196192  0.1187744961 -0.3611450482 -0.3515435987 -0.0955447415
#> [266] -0.3497330627 -0.3608416854 -0.0683547597  0.3562336752 -0.3347862994
#> [271] -0.2699743064  0.2085606808 -0.1275176594 -0.0686262425  0.3545308666
#> [276] -0.0724352342  0.2964915200 -0.0698589744 -0.0838332492 -0.3253090774
#> [281] -0.3616203399  0.3371127989  0.3020999032 -0.0363462110 -0.0885736776
#> [286] -0.0786641721 -0.2804826152  0.3328208889 -0.1586694140  0.0442563098
#> [291] -0.3577731075  0.3631465263 -0.3609126199 -0.3616209021 -0.3537122473
#> [296] -0.3579364268  0.3427067769  0.2106479776 -0.3553352216 -0.1572572091
#> [301] -0.0816594025  0.2544232204  0.2230088357  0.1831162532  0.2019016289
#> [306] -0.2689594621 -0.3615981956  0.3674360346  0.3629936157  0.2995012741
#> [311]  0.0360455078 -0.3032149102  0.3325034305 -0.2941283811 -0.3479717379
#> [316]  0.1486376588 -0.1091191554 -0.3615892261 -0.0773424219 -0.0859526537
#> [321]  0.2548107210 -0.0970003061 -0.3595506195 -0.0777193216 -0.1226799109
#> [326]  0.3700432928  0.2478830333 -0.0911876898  0.3438295903  0.3132465187
#> [331] -0.3442242009 -0.3613441278  0.3617092712 -0.3613041221 -0.0709787926
#> [336]  0.3532727851 -0.0737991346 -0.0554786064  0.2603232565 -0.0790004288
#> [341]  0.3107445024 -0.2370519457 -0.3068204909 -0.0932129031  0.3082696516
#> [346] -0.3280910932 -0.3491104937 -0.1142455737  0.3024586689  0.2113975556
#> [351] -0.1526295353 -0.3239610203 -0.3177356070  0.3541045938  0.3400838823
#> [356] -0.1255900123  0.2090184237 -0.3615508867 -0.3595126401 -0.0702339327
#> [361]  0.3094856831  0.2162021711 -0.0772546396  0.3061405397  0.2337656103
#> [366] -0.3614335707  0.2053073871 -0.0905060109 -0.3509045623  0.3528168128
#> [371] -0.0698768457  0.3082672637 -0.0596033161  0.2677158230 -0.1040319679
#> [376] -0.1284980229 -0.3616129849 -0.2997916942 -0.1382051561 -0.0187246653
#> [381]  0.2143773112  0.2221300456  0.3274229908 -0.0852857616  0.3484069971
#> [386] -0.0709590246  0.2710179908  0.3217958197 -0.0741801614  0.2780884931
#> [391] -0.0842511248 -0.0734848031 -0.1895268958 -0.2793411001  0.3228342291
#> [396]  0.2934336632 -0.2073798209 -0.0549390854  0.0571382867 -0.0752877321
#> [401]  0.3209545044  0.2502076536  0.1684629708 -0.0679942732  0.0432097553
#> [406]  0.3676099281 -0.0995047139 -0.3600484732 -0.3587438362 -0.3606826500
#> [411] -0.0778170069 -0.3604197561 -0.3361817806  0.3679894315  0.0069835138
#> [416] -0.1336906157 -0.0465631803  0.1536175550 -0.0830795745 -0.3604938087
#> [421]  0.2365866024 -0.2589890314  0.3628238437  0.1001482416 -0.0382789601
#> [426]  0.2924022223  0.2270793895  0.3642534238 -0.3544439149 -0.0992392655
#> [431] -0.0688040606 -0.0075972232 -0.0467326864  0.3118491817  0.2056730100
#> [436] -0.3599847636  0.1801133749 -0.0722970111  0.0328548364 -0.0636569671
#> [441]  0.1829668492  0.3698152140 -0.3415230444 -0.0738717924  0.0466475905
#> [446]  0.0994195065 -0.3613735792  0.2870968092  0.3175575112 -0.3201173493
#> [451]  0.2631628650  0.1858395614 -0.0636124931 -0.0414465560  0.3666765306
#> [456] -0.3575106174 -0.1650342895 -0.0761658083  0.3564714176  0.2891431844
#> [461]  0.2676774610  0.2180228674  0.3694647783  0.3090016426  0.3563088970
#> [466] -0.0871105426  0.2835189603 -0.3583191418  0.2372806766  0.2775801532
#> [471]  0.2253560995 -0.3615442395  0.3080376677 -0.3461736021  0.2839985678
#> [476] -0.2479027450  0.3524173448  0.3684968931 -0.0714094931  0.3437633200
#> [481]  0.3674604799 -0.0744792736 -0.0599870835  0.3692279893 -0.3603361431
#> [486] -0.3605391136  0.3566090308 -0.0843314251 -0.3609256124 -0.3616916888
#> [491] -0.0443981245 -0.3492650169 -0.3345631234 -0.0497021418  0.3580699599
#> [496] -0.1139169694  0.2582195348 -0.0304745748 -0.2660958702 -0.3576368768