Find the posterior mean

Find the posterior mean given the observations and the mixing distribution based on the family in the result

posteriormean(x, result, fun = function(x) x)

posteriormean.npnorm(x, result, fun = function(x) x)

posteriormean.npt(x, result, fun = function(x) x)

posteriormean.npnormc(x, result, fun = function(x) x)

Arguments

x
a vector of observations

result
an object of class nspmix

fun
the function to transform the mean. It finds the posterior mean of fun(x). The function fun must be vectorised.

Examples

data = rnorm(500, c(0, 2))
r1 = computemixdist(data, pi0 = 0.5)
posteriormean(data, r1)

#>   [1] 0.52334114 1.62947189 1.62598239 2.08145705 0.07356137 2.06713115
#>   [7] 1.15289892 1.24380358 1.44558215 2.08981688 0.55483761 2.01915716
#>  [13] 0.05360459 2.10845214 0.09061258 1.50533571 0.47154803 1.80509202
#>  [19] 1.54052908 1.15659444 0.09785373 1.87264563 0.50809695 2.46058101
#>  [25] 0.03953543 2.04465673 0.78188255 2.04328280 0.04076686 1.90729464
#>  [31] 0.10376423 0.31745909 0.06099066 0.84556592 1.50391493 1.18386448
#>  [37] 0.78981625 1.97067361 0.06960984 1.95541427 0.50290346 0.74265246
#>  [43] 0.63467220 1.74946149 0.15071460 1.00387285 0.16631552 0.80615025
#>  [49] 0.20918777 1.67311314 0.14939769 2.04435990 0.08753974 1.67339101
#>  [55] 0.75779646 2.06256976 0.18495976 1.78663325 0.98811747 1.89280471
#>  [61] 0.36125975 1.58426443 0.25319813 1.75639658 0.51713641 1.66490097
#>  [67] 0.13625361 1.70929149 0.10927319 2.08030087 0.05083247 2.05348991
#>  [73] 0.21359461 0.21029560 0.09689262 2.04838404 0.03981283 1.81872405
#>  [79] 0.04249195 1.79444915 0.91229069 1.62405965 0.84652097 1.99432955
#>  [85] 1.10021522 2.06351028 0.88914134 1.91555229 0.23894701 1.82104288
#>  [91] 0.37881604 0.98851178 0.09053432 1.38587334 0.09971970 1.56802688
#>  [97] 0.19929697 1.09590479 0.01899066 2.12695462 0.06773766 2.03487683
#> [103] 0.08771247 1.94595397 0.17598500 1.96970204 0.02473635 1.86292305
#> [109] 0.17268040 2.06782991 0.13149084 1.48245278 0.02492147 1.83279927
#> [115] 0.54925064 1.60303068 0.09190149 1.90616120 0.04389693 1.89105660
#> [121] 0.90101452 1.65219491 0.12099326 1.51467222 0.45429827 0.63630000
#> [127] 0.06559833 0.46828875 0.14464707 1.73122028 0.24789061 1.64982948
#> [133] 0.02510268 0.63285057 0.13854532 1.87727091 0.29004032 0.84096541
#> [139] 0.91025447 1.77653892 0.12808056 1.87464777 0.74755447 1.96984774
#> [145] 0.11965732 0.71662836 0.75694739 0.84443488 0.08002414 2.04739621
#> [151] 0.09930614 1.58743640 0.44584821 2.06370911 0.05538304 1.99676622
#> [157] 0.20769484 2.07045668 0.85002887 1.72623529 0.07323430 0.61514372
#> [163] 0.05376770 1.43583425 0.04339047 1.66613257 0.09812789 2.12173591
#> [169] 0.98873713 1.94811340 0.09508673 2.12686306 0.14213805 1.73176025
#> [175] 0.02331212 2.04247111 0.19952069 0.87341593 0.21449023 2.04072477
#> [181] 0.24620005 1.88713256 0.94840890 0.97187247 0.14850794 1.98741413
#> [187] 0.03609380 2.05213528 0.04982600 1.87467074 0.28532280 1.29551127
#> [193] 0.30297104 1.63315743 0.78600275 0.57715780 1.18930073 1.58963222
#> [199] 0.07446291 0.82718304 0.96868474 0.43829564 0.26854578 1.29370731
#> [205] 0.07195779 1.65465241 0.23765693 1.49834812 0.07390080 1.99434592
#> [211] 0.58221123 2.08840034 0.05508991 0.88070876 0.06784487 0.74741208
#> [217] 1.74366875 0.10178030 0.03553867 1.13187767 0.15066849 1.66317105
#> [223] 0.90342304 2.05928998 0.05188025 1.85799961 0.02800980 2.01907670
#> [229] 0.06308301 1.93136599 0.05469741 1.76159633 0.21459071 0.81772664
#> [235] 0.38819232 2.78763084 0.22397798 1.45847737 0.10097955 1.50538269
#> [241] 0.28161536 1.01764842 1.40371012 1.96991823 0.81406075 2.06754225
#> [247] 0.04920151 0.95171263 0.35946261 1.40232941 0.47126812 1.71243320
#> [253] 0.61864335 0.39600069 0.61600453 0.27546406 0.21101689 1.93835173
#> [259] 0.06959755 2.07661047 0.30379303 0.39441426 1.43951407 1.88963264
#> [265] 0.14266958 1.99364199 0.04154425 1.75336478 0.33553224 0.05961604
#> [271] 0.11723904 1.20097969 0.37440592 2.04431040 0.27258775 1.08841672
#> [277] 0.19289515 0.87863407 0.05385639 0.10265975 0.14593774 1.96675821
#> [283] 0.32960960 0.84109491 0.10785043 2.04600090 0.64671757 1.95572979
#> [289] 0.13347170 1.72255793 0.04788545 1.77320851 0.40294015 1.78136886
#> [295] 0.30217670 0.11816293 0.33723468 1.53230671 0.04107483 1.55267415
#> [301] 0.29092727 0.93312044 0.22431218 0.71864891 0.04961759 2.10331564
#> [307] 0.67670877 0.41270209 0.15111743 1.81810652 0.13619770 2.04253528
#> [313] 1.87097360 1.31016601 0.74013582 2.04480657 0.40388749 1.93831151
#> [319] 1.11720156 1.81341897 0.02253117 1.93203334 0.04001587 1.27823622
#> [325] 0.15069465 0.17998107 0.16393277 2.01848621 0.08710400 0.47287205
#> [331] 0.13731726 1.31546207 0.34997965 1.26910868 0.07848341 1.90531636
#> [337] 0.46091080 2.04735807 0.95404205 1.23575730 0.79704661 0.59354968
#> [343] 0.27784043 0.22136225 0.40301757 0.81709254 0.19377628 1.87662251
#> [349] 0.03281806 2.01991056 0.98543760 1.87783820 0.04218148 0.47411454
#> [355] 0.03271230 1.43225701 0.18425377 1.90384689 0.69581506 1.36708933
#> [361] 0.99009943 1.49773274 0.03279236 1.99161526 0.50333815 1.81203805
#> [367] 0.45780482 2.07115493 0.13172503 2.03353988 0.09609415 2.06763295
#> [373] 0.16846260 1.60433161 0.55027061 1.97096242 1.32925620 1.93821388
#> [379] 0.89648780 1.98624460 0.10610443 2.05816693 0.05641394 2.03480024
#> [385] 0.06150957 2.02998922 0.18278094 1.77121740 1.71315231 1.13098369
#> [391] 0.93601406 1.21239563 0.22556448 1.67916853 1.71908354 2.05857626
#> [397] 0.24382355 1.81952234 0.24622065 0.55137924 0.02809845 1.63668619
#> [403] 0.15865256 1.16503604 0.06172109 1.72258171 1.61627240 1.51315677
#> [409] 1.01672112 1.59439317 0.34926051 2.15173223 0.34161514 1.24282782
#> [415] 0.02759829 1.37206670 0.88662612 1.41533112 0.71340293 1.59878961
#> [421] 0.03584842 0.25479918 0.05375008 1.75649126 0.10362198 1.10355855
#> [427] 0.24925139 0.79856804 0.85778650 1.29206432 0.16629353 0.98532760
#> [433] 0.49661122 1.91072321 0.11152038 0.22756250 0.27780416 1.92599432
#> [439] 0.45764278 2.03783506 0.06745763 1.15305028 0.04035883 2.01442457
#> [445] 0.10921708 1.09291307 0.20217893 1.63403482 0.08564869 2.12026534
#> [451] 0.07458504 2.04701834 0.53494598 2.10353614 0.25715696 1.75736460
#> [457] 0.08788554 0.69802390 0.13737596 1.29663470 0.10086576 1.91011230
#> [463] 0.11116533 0.03547352 0.15776576 1.84301640 0.19471222 1.84830443
#> [469] 0.11066297 2.02682300 0.13002734 1.42126660 0.81533665 2.05173639
#> [475] 0.11575941 1.67706414 0.07757658 1.81751951 1.58689796 2.01701449
#> [481] 0.56997144 1.93731502 0.15714693 0.36246001 0.65500999 1.82429223
#> [487] 0.10047334 2.06178632 0.74094479 2.05325790 0.65363479 1.12904446
#> [493] 0.14725395 0.24921821 0.15962915 0.30562341 0.25809399 1.82848608
#> [499] 0.08763675 1.73952136

r2 = computemixdist(data, pi0 = 0.5, method = "nptll") # equivalent to normal
posteriormean(data, r2, fun = function(x) x^2)

#>   [1] 1.01265268 3.35413832 3.34665158 4.41131057 0.09973356 4.35145673
#>   [7] 2.33789920 2.53090930 2.96065738 4.45154803 1.07833029 4.21020082
#>  [13] 0.06389839 4.55208332 0.13162723 3.08830034 0.90487056 3.73243733
#>  [19] 3.16357030 2.34573850 0.14540570 3.87924946 0.98089977 6.79441310
#>  [25] 0.04038537 4.27782488 1.55410096 4.27388119 0.04235400 3.95514045
#>  [31] 0.15673393 0.58633517 0.07691027 1.68814591 3.08526312 2.40360495
#>  [37] 1.57078783 4.09625772 0.09248274 4.06184976 0.97008745 1.47164367
#>  [43] 1.24519451 3.61221809 0.24855256 2.02227786 0.27956812 1.60515435
#>  [49] 0.36563010 3.44785091 0.24594364 4.27696889 0.12581770 3.44844810
#>  [55] 1.50346368 4.33478257 0.31686307 3.69249434 1.98896884 3.92333598
#>  [61] 0.67648353 3.25721233 0.45488797 3.62718063 0.99972568 3.43020474
#>  [67] 0.21999320 3.52566294 0.16735073 4.40604296 0.05911439 4.30445514
#>  [73] 0.37453182 0.36786705 0.14357024 4.28877412 0.04082697 3.76197579
#>  [79] 0.04514603 3.70940005 1.82882682 3.34252672 1.69015787 4.15062530
#>  [85] 2.22620694 4.33813463 1.77999323 3.97331861 0.42590376 3.76700404
#>  [91] 0.71271752 1.98980233 0.13147897 2.83329119 0.14897464 3.22243209
#>  [97] 0.34568461 2.21707404 0.01244955 4.66081179 0.08907015 4.25065503
#> [103] 0.12614363 4.04069176 0.29888219 4.09405461 0.01899572 3.85804203
#> [109] 0.29227435 4.35410688 0.21063377 3.03939646 0.01922790 3.79251517
#> [115] 1.06667332 3.29743048 0.13407100 3.95264853 0.04744733 3.91950611
#> [121] 1.80503633 3.40291329 0.19009910 3.10826199 0.86904043 1.24860214
#> [127] 0.08519000 0.89809789 0.23654525 3.57289074 0.44408520 3.39783401
#> [133] 0.01945631 1.24138138 0.22450540 3.88935017 0.53011495 1.67845493
#> [139] 1.82453033 3.67067486 0.20394796 3.88362076 1.48194205 4.09438486
#> [145] 0.18749599 1.41699619 1.50167927 1.68576328 0.11171691 4.28583461
#> [151] 0.14818303 3.26400855 0.85150174 4.33884884 0.06699829 4.15632731
#> [157] 0.36261647 4.36430827 1.69754821 3.56215003 0.09913108 1.20432961
#> [163] 0.06418174 2.93985191 0.04661507 3.43285082 0.14592961 4.62956543
#> [169] 1.99027866 4.04551146 0.14012681 4.66026056 0.23159007 3.57405434
#> [175] 0.01725035 4.27157290 0.34613521 1.74683609 0.37634204 4.26665798
#> [181] 0.44064630 3.91091411 1.90507170 1.95463695 0.24418184 4.13456292
#> [187] 0.03500430 4.30021023 0.05739306 3.88367092 0.52046086 2.64085849
#> [193] 0.55660595 3.36204665 1.56276646 1.12492843 2.41514447 3.26871370
#> [199] 0.10139636 1.64942923 1.94790152 0.83583350 0.48617733 2.63702061
#> [205] 0.09678358 3.40819074 0.42328355 3.07336391 0.10035925 4.15066354
#> [211] 1.13548458 4.44447717 0.06648580 1.76221147 0.08926515 1.48164288
#> [217] 3.59972482 0.15292389 0.03415472 2.29331831 0.24846118 3.42648824
#> [223] 1.81011729 4.32342060 0.06091542 3.84731461 0.02326531 4.21000090
#> [229] 0.08065633 4.00826268 0.06580052 3.63840328 0.37654514 1.62952006
#> [235] 0.73209008 8.90956960 0.39554004 2.98818777 0.15138823 3.08840079
#> [241] 0.51287801 2.05141080 2.87132102 4.09454468 1.62180337 4.35301269
#> [247] 0.05632945 1.91204900 0.67277750 2.86837665 0.90428888 3.53242609
#> [253] 1.21165069 0.74823378 1.20613030 0.50030507 0.36932380 4.02376545
#> [259] 0.09246030 4.38974085 0.55829134 0.74495309 2.94770538 3.91638744
#> [265] 0.23263930 4.14902059 0.04360743 3.62063871 0.62348598 0.07446258
#> [271] 0.18279020 2.43993975 0.70361064 4.27682635 0.49442971 2.20121036
#> [277] 0.33280136 1.75783715 0.06433589 0.15461191 0.23909660 4.08739043
#> [283] 0.61130393 1.67872771 0.16460376 4.28172972 1.27041487 4.06255745
#> [289] 0.21452334 3.55422853 0.05409963 3.66347938 0.76258889 3.68111308
#> [295] 0.55497739 0.18458704 0.62698897 3.14597829 0.04284960 3.18956237
#> [301] 0.53193067 1.87279020 0.39621697 1.42123756 0.05703772 4.52319649
#> [307] 1.33325577 0.78279477 0.24935090 3.76063690 0.21988319 4.27175482
#> [313] 3.87560001 2.67204149 1.46635715 4.27825783 0.76454915 4.02367606
#> [319] 2.26220546 3.75047617 0.01632549 4.00974175 0.04115091 2.60411228
#> [325] 0.24851303 0.30688213 0.27481884 4.20853555 0.12499586 0.90762220
#> [331] 0.22208677 2.68331306 0.65323159 2.58470221 0.10884717 3.95079161
#> [337] 0.88277136 4.28572169 1.91696894 2.51381065 1.58599852 1.15917742
#> [343] 0.50516099 0.39024339 0.76274909 1.62818525 0.33457327 3.88793370
#> [349] 0.03007474 4.21207558 1.98330439 3.89058956 0.04464073 0.91020456
#> [355] 0.02991915 2.93221800 0.31544686 3.94756272 1.37332223 2.79325815
#> [361] 1.99315834 3.07204862 0.03003692 4.14430073 0.97099231 3.74748373
#> [367] 0.87632112 4.36708479 0.21109337 4.24708351 0.14204687 4.35335722
#> [373] 0.28385118 3.30021938 1.06880123 4.09691298 2.71267697 4.02345907
#> [379] 1.79548765 4.13186231 0.16123744 4.31964120 0.06880537 4.25044962
#> [385] 0.07783708 4.23773379 0.31249329 3.65917829 3.53397426 2.29142283
#> [391] 1.87889918 2.46418241 0.39875397 3.46086637 3.54674561 4.32101242
#> [397] 0.43581377 3.76370671 0.44068820 1.07111422 0.02338546 3.36961952
#> [403] 0.26430959 2.36364791 0.07821529 3.55427976 3.32582343 3.10502158
#> [409] 2.04944944 3.27891640 0.65175001 4.81286147 0.63600502 2.52883562
#> [415] 0.02271047 2.80386446 1.77468902 2.89610643 1.41022618 3.28833931
#> [421] 0.03462809 0.45814854 0.06415113 3.62738496 0.15646049 2.23329132
#> [427] 0.44685399 1.58919958 1.71389409 2.63352533 0.27952427 1.98307188
#> [433] 0.95699144 3.96268289 0.17169623 0.40280299 0.50508687 3.99637113
#> [439] 0.87598464 4.25866811 0.08856103 2.33822027 0.04169937 4.19853648
#> [445] 0.16724232 2.21073571 0.35149129 3.36392948 0.12225466 4.62083046
#> [451] 0.10162185 4.28471760 1.03684051 4.52442115 0.46295173 3.62926964
#> [457] 0.12647028 1.37795584 0.22220233 2.64324864 0.15117011 3.96133845
#> [463] 0.17100911 0.03405537 0.26254671 3.81471251 0.33645588 3.82621162
#> [469] 0.17003725 4.22954526 0.20776293 2.90876822 1.62448905 4.29897239
#> [475] 0.17991508 3.45634283 0.10716177 3.75936426 3.26285485 4.20489665
#> [481] 1.10992058 4.02146186 0.26131690 0.67895903 1.28778361 3.77405200
#> [487] 0.15041807 4.33202304 1.46805646 4.30372350 1.28490287 2.28731126
#> [493] 0.24170001 0.44678647 0.26625168 0.56204491 0.46486112 3.78315203
#> [499] 0.12600075 3.59078262

data = runif(500, min = -0.5, max = 0.5)
r3 = computemixdist(data, method = "npnormcll", beta = 100)
posteriormean(data, r3)

#>   [1] -0.3613611492 -0.3604237297  0.1777899561 -0.2034152207  0.3683695365
#>   [6]  0.3032663677  0.2420179132 -0.0758284627  0.3671256555  0.3692651189
#>  [11] -0.0717481648 -0.0729118808 -0.0775773249  0.2961812022  0.3485953592
#>  [16]  0.3582558769  0.2182582494 -0.3617064952  0.2935154159  0.3252151605
#>  [21] -0.0711287189 -0.3554980213 -0.0633581898 -0.3374863709  0.0960933126
#>  [26]  0.0661425441 -0.0426370078  0.3402833333  0.2911383381 -0.3614911101
#>  [31] -0.0876203987  0.3505755650 -0.3460923505 -0.3606537972  0.3634017334
#>  [36] -0.3583212321 -0.2975420533  0.3445711754  0.1255653772 -0.1739611285
#>  [41] -0.3615615330  0.2901159395 -0.2385816871  0.3304391091  0.3698884351
#>  [46]  0.3465823665 -0.0221391924 -0.0821315714  0.2972591249  0.1676119341
#>  [51]  0.2275953990 -0.0538285310 -0.3609155770  0.2925932548  0.3071313203
#>  [56] -0.0718505461 -0.3542196111  0.3685539366  0.2707854878  0.3660018389
#>  [61] -0.3593538333  0.3500777661 -0.0609727380  0.3543707297  0.2244432617
#>  [66]  0.3696889294 -0.0818940751  0.1562442561 -0.0716379137 -0.1669545903
#>  [71]  0.2348729358 -0.3606624975  0.2727237611  0.2598794392  0.3641932523
#>  [76] -0.0728213534  0.3702042836 -0.0025996754 -0.0697351386  0.1724883721
#>  [81]  0.3589578416 -0.3456685631  0.2654355340  0.2121116936 -0.0323674712
#>  [86] -0.0736516358 -0.0640466333  0.3344858475 -0.3527824392  0.1028753412
#>  [91]  0.3312506333 -0.0804201121  0.3374711583 -0.0357816042 -0.1348579855
#>  [96]  0.3527135855 -0.3592945313 -0.1251299442 -0.0469418743  0.2417532101
#> [101]  0.3459987884 -0.2361304351  0.3672030242  0.3698475746 -0.1710893525
#> [106]  0.3498918173 -0.0702279564 -0.0772714591  0.3669401453  0.3388622884
#> [111] -0.0949149402 -0.3372330562 -0.3536012289  0.2537659193  0.1632247473
#> [116] -0.3616237964  0.3698278334 -0.3555076826 -0.3041534615  0.2095027151
#> [121] -0.1197411779  0.3685727619 -0.0691007354  0.3388309160 -0.0368853401
#> [126]  0.1254372984 -0.0916437207  0.3085156939  0.2112213528 -0.3453087244
#> [131] -0.2477131713  0.2390579646 -0.0710945468 -0.3617062779  0.2610336194
#> [136]  0.2204296959  0.0018537443  0.0009598036  0.2646827860 -0.0607210338
#> [141] -0.3608018712 -0.1023494060 -0.3304107494 -0.0774874130 -0.0721416248
#> [146]  0.2228229011 -0.3612969208 -0.3615443856  0.0463311703 -0.0778211266
#> [151]  0.3197254183  0.2866430298 -0.0709012176 -0.3597895512 -0.3613894263
#> [156] -0.3273322276 -0.0762144123  0.0885342005 -0.3602954862 -0.3350807524
#> [161] -0.0771348602 -0.1135339266 -0.3608436941 -0.0693191473 -0.3616264159
#> [166]  0.0234233184  0.3054526193 -0.3586444169  0.1038384854 -0.0562886793
#> [171] -0.3587124319 -0.0537864544 -0.3608874210 -0.0750181295 -0.3596356055
#> [176]  0.3630072274 -0.1570162451 -0.0713272588  0.3524840428 -0.0688644912
#> [181]  0.1779471301  0.3156006022 -0.3498686807  0.2075655587  0.3292589495
#> [186]  0.3684938508 -0.3608045844 -0.0719133862  0.3665603397 -0.3604557864
#> [191] -0.0763421871 -0.3610639780  0.3622626358  0.1659640518 -0.3539605182
#> [196]  0.3669139362 -0.0607891836 -0.3568175302 -0.1398436191 -0.0756590006
#> [201]  0.1589657789 -0.0745101524 -0.3509338795  0.3534494688 -0.3595947528
#> [206] -0.3550373872 -0.3491787882  0.3686608584  0.3674252200 -0.3560677150
#> [211]  0.0879275811 -0.2221814753  0.3364050419 -0.0976540435 -0.3084513384
#> [216]  0.3274096630 -0.0696399659 -0.0768170888 -0.0681614130 -0.3615377318
#> [221] -0.0777502183  0.3510246950 -0.0713175926 -0.0501257917  0.1308805458
#> [226]  0.2429113486 -0.3195883188  0.3616784087  0.2857864799 -0.0777385563
#> [231]  0.3676625718 -0.3600557207 -0.0649856800 -0.3615790317  0.2218833394
#> [236] -0.0792057231  0.0629959916 -0.3585883332 -0.0618024164  0.1671210560
#> [241]  0.2025811932 -0.3570439643 -0.0646119111 -0.3607942360 -0.3386706187
#> [246]  0.3471111285 -0.0678888374 -0.0677470282  0.3534582000  0.3644568045
#> [251]  0.2341335564 -0.0703918822  0.1845239777 -0.3227688945 -0.0698105398
#> [256] -0.0349262063  0.2299767486  0.2292739480 -0.3616229821  0.0369649098
#> [261] -0.0659196198  0.1187744971 -0.3611450490 -0.3515435992 -0.0955447417
#> [266] -0.3497330632 -0.3608416862 -0.0683547604  0.3562336757 -0.3347862995
#> [271] -0.2699743053  0.2085606806 -0.1275176587 -0.0686262432  0.3545308671
#> [276] -0.0724352350  0.2964915196 -0.0698589752 -0.0838332497 -0.3253090773
#> [281] -0.3616203408  0.3371127991  0.3020999029 -0.0363462111 -0.0885736779
#> [286] -0.0786641727 -0.2804826142  0.3328208891 -0.1586694128  0.0442563107
#> [291] -0.3577731082  0.3631465269 -0.3609126207 -0.3616209029 -0.3537122479
#> [296] -0.3579364275  0.3427067772  0.2106479774 -0.3553352223 -0.1572572079
#> [301] -0.0816594030  0.2544232198  0.2230088353  0.1831162535  0.2019016288
#> [306] -0.2689594610 -0.3615981965  0.3674360352  0.3629936163  0.2995012737
#> [311]  0.0360455086 -0.3032149096  0.3325034307 -0.2941283803 -0.3479717384
#> [316]  0.1486376596 -0.1091191552 -0.3615892270 -0.0773424226 -0.0859526541
#> [321]  0.2548107204 -0.0970003061 -0.3595506202 -0.0777193223 -0.1226799103
#> [326]  0.3700432935  0.2478830326 -0.0911876901  0.3438295906  0.3132465185
#> [331] -0.3442242013 -0.3613441286  0.3617092718 -0.3613041230 -0.0709787933
#> [336]  0.3532727856 -0.0737991353 -0.0554786069  0.2603232558 -0.0790004295
#> [341]  0.3107445022 -0.2370519443 -0.3068204903 -0.0932129033  0.3082696514
#> [346] -0.3280910932 -0.3491104942 -0.1142455733  0.3024586685  0.2113975553
#> [351] -0.1526295342 -0.3239610202 -0.3177356067  0.3541045943  0.3400838826
#> [356] -0.1255900117  0.2090184235 -0.3615508875 -0.3595126408 -0.0702339335
#> [361]  0.3094856829  0.2162021708 -0.0772546403  0.3061405394  0.2337656098
#> [366] -0.3614335716  0.2053073870 -0.0905060112 -0.3509045628  0.3528168133
#> [371] -0.0698768464  0.3082672635 -0.0596033166  0.2677158223 -0.1040319678
#> [376] -0.1284980222 -0.3616129857 -0.2997916936 -0.1382051552 -0.0187246652
#> [381]  0.2143773109  0.2221300452  0.3274229908 -0.0852857620  0.3484069975
#> [386] -0.0709590254  0.2710179902  0.3217958197 -0.0741801621  0.2780884925
#> [391] -0.0842511253 -0.0734848039 -0.1895268944 -0.2793410991  0.3228342291
#> [396]  0.2934336627 -0.2073798194 -0.0549390859  0.0571382877 -0.0752877329
#> [401]  0.3209545043  0.2502076529  0.1684629713 -0.0679942739  0.0432097562
#> [406]  0.3676099288 -0.0995047139 -0.3600484740 -0.3587438369 -0.3606826508
#> [411] -0.0778170076 -0.3604197569 -0.3361817807  0.3679894322  0.0069835143
#> [416] -0.1336906149 -0.0465631806  0.1536175558 -0.0830795750 -0.3604938095
#> [421]  0.2365866018 -0.2589890301  0.3628238443  0.1001482427 -0.0382789603
#> [426]  0.2924022219  0.2270793891  0.3642534245 -0.3544439155 -0.0992392655
#> [431] -0.0688040613 -0.0075972228 -0.0467326867  0.3118491815  0.2056730099
#> [436] -0.3599847644  0.1801133753 -0.0722970119  0.0328548372 -0.0636569678
#> [441]  0.1829668495  0.3698152147 -0.3415230447 -0.0738717932  0.0466475914
#> [446]  0.0994195076 -0.3613735801  0.2870968087  0.3175575111 -0.3201173491
#> [451]  0.2631628644  0.1858395616 -0.0636124938 -0.0414465562  0.3666765312
#> [456] -0.3575106182 -0.1650342883 -0.0761658090  0.3564714181  0.2891431839
#> [461]  0.2676774603  0.2180228671  0.3694647790  0.3090016423  0.3563088975
#> [466] -0.0871105430  0.2835189597 -0.3583191425  0.2372806760  0.2775801526
#> [471]  0.2253560991 -0.3615442403  0.3080376674 -0.3461736025  0.2839985673
#> [476] -0.2479027436  0.3524173453  0.3684968937 -0.0714094939  0.3437633203
#> [481]  0.3674604805 -0.0744792744 -0.0599870841  0.3692279900 -0.3603361439
#> [486] -0.3605391144  0.3566090313 -0.0843314256 -0.3609256132 -0.3616916896
#> [491] -0.0443981247 -0.3492650174 -0.3345631235 -0.0497021422  0.3580699604
#> [496] -0.1139169691  0.2582195341 -0.0304745748 -0.2660958690 -0.3576368776