Find the posterior mean

Find the posterior mean given the observations and the mixing distribution based on the family in the result

posteriormean(x, result, fun = function(x) x)

posteriormean.npnorm(x, result, fun = function(x) x)

posteriormean.npt(x, result, fun = function(x) x)

posteriormean.npnormc(x, result, fun = function(x) x)

Arguments

x
a vector of observations

result
an object of class nspmix

fun
the function to transform the mean. It finds the posterior mean of fun(x). The function fun must be vectorised.

Examples

data = rnorm(500, c(0, 2))
r1 = computemixdist(data, pi0 = 0.5)
posteriormean(data, r1)

#>   [1] 0.074718558 1.848281699 0.416716658 1.755668787 0.029776400 1.673455907
#>   [7] 0.714225125 1.853681665 0.988877947 1.890156694 0.027852384 1.917126221
#>  [13] 0.201652443 1.469148419 0.187738003 1.092716292 0.908703043 1.718454353
#>  [19] 0.026822681 1.882995457 0.072110375 1.604866477 0.584233242 1.832161484
#>  [25] 0.600590921 1.295277008 0.288795401 0.062502152 0.051475120 1.192309480
#>  [31] 0.081533209 1.843459892 0.108544816 1.775235995 0.537502178 1.855902182
#>  [37] 0.652507220 1.328447537 0.374332904 1.889261854 1.606675499 1.654818211
#>  [43] 1.317077050 1.320920454 0.181892716 1.275601484 0.121508389 1.840661285
#>  [49] 0.661082293 1.265953701 0.065392317 1.039718995 1.596158717 1.794020674
#>  [55] 0.397019463 1.171783698 0.312246016 1.525784607 1.510859715 1.307837124
#>  [61] 0.630536767 1.624374720 0.011360417 1.902141378 0.001900745 0.966891007
#>  [67] 0.075279646 1.261708629 0.591289998 1.536035949 0.058936747 0.678555909
#>  [73] 0.061402736 1.315406439 0.828195297 1.783282254 1.603469817 1.093416315
#>  [79] 0.243235020 1.436888451 0.104786876 1.649840358 0.002533070 1.746551907
#>  [85] 0.455065966 1.902225330 0.783327282 1.891386395 0.208673790 1.405990237
#>  [91] 1.430577286 1.528838196 0.240298387 1.370186805 0.070382162 1.643611704
#>  [97] 0.098484415 1.752795875 0.768204787 0.301268019 0.383299070 1.843825826
#> [103] 0.501494730 1.912870648 0.108733972 0.895645548 0.048085587 1.819751765
#> [109] 0.064005411 1.765783055 0.028732632 0.119569670 0.017324429 1.463404920
#> [115] 0.048642051 1.637952805 1.292033768 1.123653763 0.696726894 0.623607380
#> [121] 1.150680908 0.637268070 1.089565250 1.852798119 0.056285939 0.912614214
#> [127] 0.247082123 1.865197405 0.110106090 1.877378907 0.195730632 0.899066858
#> [133] 0.048330925 1.779934218 0.883636003 0.860073548 0.033722693 0.610985235
#> [139] 0.111148599 1.868183121 0.806218527 1.615446584 0.651586442 1.202001727
#> [145] 0.894360849 1.694410717 0.377007889 1.237162967 0.028730679 1.806350201
#> [151] 0.069463910 1.618730419 0.456077648 1.551816303 0.311743485 1.855456642
#> [157] 0.089861718 0.898445225 0.588873603 1.857294627 0.028729891 1.882707073
#> [163] 0.535787621 1.863207456 0.211068983 1.060549948 0.140157365 1.742597393
#> [169] 0.498111341 1.916493835 0.004634440 1.809172268 0.141993741 1.305473308
#> [175] 0.241725614 1.808390407 0.086387071 1.890473395 0.176025322 0.459342655
#> [181] 0.016865397 0.241546072 0.069183649 1.632065885 1.002058283 1.914658006
#> [187] 1.787978047 0.977964628 0.284767166 1.201372976 0.065199169 1.239953222
#> [193] 0.709839393 1.584355114 0.016635394 1.354877103 0.006258743 1.805678376
#> [199] 0.600853159 1.700102334 0.898517726 1.499403979 0.645953172 1.687443490
#> [205] 0.105913280 1.311735081 0.085865178 1.727537292 1.156733853 1.814310841
#> [211] 1.775895205 1.737053089 0.006551174 1.471119782 0.136815344 1.894007457
#> [217] 0.015699682 1.539563547 0.254165105 0.133968057 0.895493635 0.575168527
#> [223] 0.521521005 1.900093411 0.254164326 1.643532140 0.400387787 1.783536257
#> [229] 0.192668840 1.833646206 0.191618730 1.166241694 0.012988348 1.825201780
#> [235] 1.575891454 1.838611495 1.012093403 1.732445572 0.037724031 0.864936619
#> [241] 0.104227363 1.277602178 0.186422583 1.908493148 0.040998799 1.621699924
#> [247] 0.243246258 1.039257570 0.500312350 1.729981431 0.291867004 1.070507343
#> [253] 0.906010758 1.178968712 0.698160756 1.743008780 0.323967649 0.537099649
#> [259] 0.363830174 1.247562643 0.462569527 1.901100084 0.470022115 0.513259615
#> [265] 0.241096877 1.826247606 0.600761288 1.870672149 0.711970588 1.618456708
#> [271] 0.009873024 0.962576287 0.841041894 1.840784350 0.213998367 1.876205925
#> [277] 0.046967681 1.771295076 0.101287073 1.875984847 0.102723428 1.449755381
#> [283] 0.122695563 1.329737534 0.518527358 1.820782906 0.594638063 0.159778037
#> [289] 0.428873108 1.901269289 1.324509213 1.885257241 0.667740795 1.886770370
#> [295] 1.308128938 1.796995535 1.747288788 1.738020303 1.323886394 1.422718711
#> [301] 1.441095915 1.473946414 0.105961786 1.869098552 0.444497869 1.151204093
#> [307] 0.181240829 0.528226685 0.381615100 1.701108492 0.459995250 1.906899975
#> [313] 1.365196899 1.044089661 1.011168656 1.703563594 0.010658779 1.024208387
#> [319] 0.178013489 1.019835227 1.852108966 1.900877233 1.311014084 1.506365629
#> [325] 0.055807034 1.411983881 1.189784334 1.825037128 0.005452920 1.702220117
#> [331] 0.861560333 1.858106301 0.606519816 0.770473664 0.184677405 1.809784050
#> [337] 0.108653084 1.333411975 1.807331502 1.617347932 0.122450069 1.827571691
#> [343] 0.007644189 0.270435432 0.578274800 1.861879734 0.062027261 1.874383813
#> [349] 0.949598917 0.684355534 0.457523938 1.828942417 0.154194582 1.576705151
#> [355] 0.159470977 0.879065201 0.034671995 1.913674045 0.051739264 1.816372914
#> [361] 0.243157500 1.277863077 0.017338645 1.489284624 0.010548659 1.673927288
#> [367] 0.063774127 1.073604053 0.296333764 1.894808705 0.741429525 1.830441472
#> [373] 0.038937216 1.387151750 0.074213091 1.610479803 0.343147217 1.239275432
#> [379] 0.102688630 1.342625729 1.161289187 1.753147834 0.390919596 1.884378457
#> [385] 0.025319415 1.799742386 0.099187856 1.783193084 0.375224910 1.031448546
#> [391] 0.286267713 1.873485121 0.132848507 1.877850842 0.093689366 0.579993445
#> [397] 0.021410367 0.686364921 0.071334187 1.437627248 1.327253815 1.908306297
#> [403] 1.156727141 1.625093470 0.194527470 1.864500989 0.127976076 1.893616600
#> [409] 1.211897248 1.851277815 0.251347872 1.784911846 0.391466491 0.677266573
#> [415] 0.640616006 1.692669190 0.304820155 1.297282023 0.113658281 1.696131800
#> [421] 0.056792791 1.899661242 1.190091215 1.902847681 0.227593656 1.328683110
#> [427] 0.181378077 0.703088088 0.003269173 0.632983957 0.394183496 1.781599995
#> [433] 0.288498076 1.646487056 1.670045632 1.871628942 0.024502575 1.722105314
#> [439] 0.906949211 1.832036234 0.055713335 1.673703050 0.106364364 1.077295957
#> [445] 0.335294482 1.812969982 1.278466211 1.769054579 0.943224342 1.520632720
#> [451] 0.163940310 1.801558276 1.377995874 1.909413824 0.529338773 1.853950769
#> [457] 1.712930022 1.698444207 0.389131113 1.775667867 0.044303023 1.678435610
#> [463] 1.762653459 1.743097457 0.055788415 0.836841209 0.010349363 1.711008892
#> [469] 1.145782941 1.791457563 0.039431782 1.848097935 0.060740115 1.812331297
#> [475] 0.469810483 0.456905841 0.007568165 1.800960822 1.323397110 1.866751939
#> [481] 0.127902840 1.879945771 0.386892739 1.053589874 1.420817012 0.936160920
#> [487] 0.341690013 1.772696124 0.017359342 1.901633964 0.610788894 1.845614397
#> [493] 0.053605169 1.640204839 0.048694771 1.888580817 0.539477913 1.870020316
#> [499] 0.233633486 1.332118029

r2 = computemixdist(data, pi0 = 0.5, method = "nptll") # equivalent to normal
posteriormean(data, r2, fun = function(x) x^2)

#>   [1] 0.162703628 3.212495008 0.810552566 3.052083196 0.068132540 2.913577608
#>   [7] 1.329351865 3.222091589 1.789669393 3.288219778 0.063949844 3.340366965
#>  [13] 0.413258887 2.576369902 0.386567650 1.960838486 1.656602406 2.989071609
#>  [19] 0.061704831 3.275007278 0.157339441 2.799578984 1.105956614 3.184056129
#>  [25] 1.134315819 2.292333213 0.577328934 0.137468161 0.114424855 2.124098050
#>  [31] 0.176663259 3.203957232 0.231292279 3.085478839 1.024500032 3.226049258
#>  [37] 1.223831730 2.346477756 0.734132214 3.286560525 2.802572421 2.882488817
#>  [43] 2.327918304 2.334191719 0.375308549 2.260210695 0.257162830 3.199014570
#>  [49] 1.238548699 2.244456731 0.143464442 1.873621783 2.785178580 3.117746880
#>  [55] 0.775133498 2.090506498 0.620691316 2.669242020 2.644737729 2.312836098
#>  [61] 1.186038019 2.831893982 0.027282828 3.310733979 0.004974343 1.753265233
#>  [67] 0.163856018 2.237524177 1.118200314 2.686087829 0.130047076 1.268480586
#>  [73] 0.135182687 2.325191425 1.521990150 3.099273698 2.797268317 1.961988704
#>  [79] 0.492170053 2.523586184 0.223753354 2.874200896 0.006539378 3.036589768
#>  [85] 0.879060747 3.310894025 1.446467232 3.290504180 0.426670618 2.473085428
#>  [91] 2.513267282 2.674258610 0.486636482 2.414612786 0.153778257 2.863839074
#>  [97] 0.211067530 3.047196600 1.420923693 0.600429494 0.750363860 3.204604190
#> [103] 0.961262531 3.331622282 0.231671266 1.634842209 0.107284078 3.162346864
#> [109] 0.140589113 3.069319845 0.065865372 0.253306791 0.040745544 2.566967252
#> [115] 0.108458398 2.854433124 2.287038702 2.011632079 1.299527865 1.174091490
#> [121] 2.055943577 1.197630735 1.955660420 3.220518731 0.124511782 1.663115213
#> [127] 0.499410699 3.242696052 0.234419036 3.264738686 0.401918084 1.640546364
#> [133] 0.107801920 3.093528979 1.614804646 1.575422793 0.076665508 1.152296846
#> [139] 0.236505140 3.248072976 1.485046432 2.817095199 1.222250328 2.139951832
#> [145] 1.632699827 2.948650811 0.738978570 2.197428477 0.065861125 3.139054346
#> [151] 0.151883837 2.822536383 0.880860318 2.712045056 0.619765218 3.225254607
#> [157] 0.193621022 1.639510088 1.114009416 3.228534618 0.065859412 3.274478166
#> [163] 1.021498503 3.239120806 0.431237309 1.907936225 0.294038747 3.029881479
#> [169] 0.955298123 3.339029378 0.011624171 3.143947178 0.297649749 2.308977575
#> [175] 0.489326552 3.142590988 0.186559478 3.288807635 0.363978172 0.886665461
#> [181] 0.039719182 0.488988221 0.151305321 2.844656014 1.811463010 3.335235254
#> [187] 3.107342660 1.771607825 0.569849532 2.138923520 0.143064238 2.201987413
#> [193] 1.321883487 2.765680107 0.039204361 2.389618978 0.015473025 3.137890456
#> [199] 1.134769836 2.958201110 1.639630952 2.625945099 1.212570626 2.936974413
#> [205] 0.226015007 2.319198744 0.185497220 3.004396777 2.065860314 3.152872503
#> [211] 3.086607579 3.020488117 0.016160313 2.579597776 0.287458088 3.295390651
#> [217] 0.037106038 2.691887616 0.512716640 0.281842281 1.634588891 1.090207667
#> [223] 0.996486281 3.306841586 0.512715178 2.863706775 0.781201940 3.099709807
#> [229] 0.396043712 3.186663504 0.394027251 2.081432394 0.030987509 3.171863145
#> [235] 2.751713393 3.195400066 1.828041909 3.012692016 0.085260008 1.583558453
#> [241] 0.222629313 2.263477385 0.384036359 3.322973781 0.092254836 2.827458647
#> [247] 0.492191218 1.872861167 0.959178620 3.008526256 0.583025924 1.924323222
#> [253] 1.652117904 2.102267950 1.301974430 3.030579010 0.642255092 1.023795442
#> [259] 0.715072571 2.214418885 0.892398879 3.308752104 0.905625686 0.981971846
#> [265] 0.488141659 3.173692386 1.134610781 3.252567603 1.325513209 2.822082772
#> [271] 0.023874260 1.746113846 1.543544595 3.199231727 0.436816740 3.262603288
#> [277] 0.104922483 3.078736134 0.216715491 3.262201153 0.219605893 2.544630957
#> [283] 0.259521904 2.348583347 0.991229362 3.164145291 1.124004282 0.332446667
#> [289] 0.832332125 3.309073749 2.340049455 3.279164622 1.249963519 3.281953654
#> [295] 2.313312430 3.122878027 3.037840577 3.022125835 2.339032863 2.500421221
#> [301] 2.530466710 2.584226574 0.226112361 3.249724718 0.860239038 2.056800826
#> [307] 0.374051141 1.008250922 0.747318071 2.959890528 0.887825287 3.319875759
#> [313] 2.406466088 1.880825279 1.826514644 2.964014285 0.025677944 1.848041203
#> [319] 0.367820740 1.840823983 3.219292661 3.308328714 2.318021867 2.637363812
#> [325] 0.123510074 2.482878192 2.119966804 3.171575247 0.013570665 2.961757433
#> [331] 1.577910533 3.229984671 1.144575871 1.424759045 0.380675887 3.145008689
#> [337] 0.231509208 2.354580949 3.140755004 2.820245390 0.259034209 3.176009812
#> [343] 0.018716258 0.543162831 1.095607204 3.236738948 0.136481263 3.259291983
#> [349] 1.724589439 1.278398575 0.883432298 3.178410645 0.321555051 2.753055649
#> [355] 0.331848447 1.607172284 0.078709526 3.333238560 0.114980124 3.156460257
#> [361] 0.492024054 2.263903369 0.040777308 2.609355233 0.025425583 2.914365142
#> [367] 0.140109240 1.929417531 0.591301099 3.296889435 1.375580871 3.181038350
#> [373] 0.087855273 2.442314537 0.161664998 2.808869564 0.677387565 2.200880014
#> [379] 0.219535901 2.369620330 2.073321844 3.047795038 0.764131440 3.277547730
#> [385] 0.058418774 3.127621372 0.212486136 3.099120608 0.735748649 1.859985214
#> [391] 0.572636747 3.257661326 0.279631772 3.265598688 0.201378913 1.098593446
#> [397] 0.049822234 1.281832943 0.155740695 2.524794276 2.344529312 3.322609290
#> [403] 2.065849319 2.833086069 0.399610575 3.241444088 0.269995385 3.294660405
#> [409] 2.156133228 3.217814812 0.507428085 3.102072319 0.765118506 1.266274566
#> [415] 1.203392112 2.945730727 0.606992705 2.295606281 0.241521527 2.951537588
#> [421] 0.125571378 3.306022965 2.120468894 3.312081727 0.462630051 2.346862269
#> [427] 0.374315905 1.310378137 0.008337788 1.190253989 0.770020390 3.096386346
#> [433] 0.576777194 2.868621261 2.907881965 3.254298438 0.056628771 2.995227803
#> [439] 1.653681199 3.183836275 0.123314026 2.913990500 0.226920249 1.935489769
#> [445] 0.663025450 3.150541466 2.264888133 3.074906645 1.714007975 2.660780667
#> [451] 0.340547086 3.130760061 2.427363197 3.324774590 1.010200581 3.222570852
#> [457] 2.979766166 2.955417730 0.760902573 3.086218289 0.099279768 2.921900411
#> [463] 3.063980769 3.030729377 0.123471119 1.536499818 0.024968518 2.976532769
#> [469] 2.047917226 3.113330756 0.088911912 3.212169097 0.133804000 3.149431654
#> [475] 0.905250363 0.882333214 0.018539100 3.129727124 2.338234231 3.245493649
#> [481] 0.269850342 3.269422342 0.756859518 1.896476042 2.497313053 1.702276353
#> [487] 0.674724734 3.081132246 0.040823548 3.309767488 1.151957478 3.207768618
#> [493] 0.118897708 2.858175448 0.108569612 3.285299395 1.027957618 3.251389432
#> [499] 0.474056351 2.352468904

data = runif(500, min = -0.5, max = 0.5)
r3 = computemixdist(data, method = "npnormcll", beta = 100)
posteriormean(data, r3)

#>   [1] -0.126256286  0.108743980 -0.090244864 -0.287960887  0.211351763
#>   [6] -0.049018721 -0.381503333  0.270547821 -0.382220176 -0.391434035
#>  [11] -0.137049496 -0.156047968  0.043220350  0.079079514 -0.245732629
#>  [16] -0.387525836  0.369402542 -0.388299768 -0.116766622 -0.185730940
#>  [21] -0.384971475 -0.072957270  0.324438675 -0.288018232  0.236583555
#>  [26]  0.334765469  0.364311456 -0.391883480  0.370421378 -0.346695851
#>  [31] -0.079214503  0.010898309 -0.385043229  0.149823737  0.364998113
#>  [36] -0.045693951 -0.391169337  0.023547620 -0.134926327 -0.378243362
#>  [41] -0.391909225 -0.077164031  0.122041586  0.066848240 -0.388885182
#>  [46]  0.371815995 -0.083777009  0.065193256  0.332251905 -0.058431070
#>  [51]  0.355811750  0.250482448  0.368942910 -0.385475033 -0.389742345
#>  [56] -0.142777908 -0.392470893  0.151450521  0.103336714 -0.375112218
#>  [61]  0.350393480  0.368124142 -0.226276778 -0.204031715 -0.139572980
#>  [66]  0.256035139 -0.098750860 -0.298076828  0.089794539  0.098729867
#>  [71] -0.387442393  0.308549354  0.263463319  0.342545998 -0.392250983
#>  [76]  0.100023615 -0.034180797 -0.109201489 -0.161692078 -0.388481932
#>  [81] -0.110287337  0.295032031  0.190152081 -0.385683407 -0.086903705
#>  [86] -0.387190434  0.094449190  0.141103257  0.046316499  0.288455625
#>  [91]  0.106099562 -0.383980273  0.370013972  0.366809203  0.187168208
#>  [96]  0.127156539 -0.031711183 -0.391361063 -0.067384759 -0.126221696
#> [101]  0.103849469  0.348747257  0.171940731  0.130295835 -0.195256412
#> [106]  0.232097167  0.337886273  0.174918491 -0.246991202 -0.037133577
#> [111] -0.381094425  0.211381408  0.267624534  0.370723397 -0.148489637
#> [116] -0.199122841 -0.102732698 -0.246973354 -0.135046498 -0.167211363
#> [121]  0.363659942 -0.286226786 -0.155136804  0.139236506  0.360091686
#> [126]  0.146034729  0.079217491 -0.222401764 -0.384653415  0.348331573
#> [131] -0.382134432 -0.049278507 -0.391924150  0.118157290 -0.046906849
#> [136] -0.142929253  0.248371331 -0.386144215 -0.163030897 -0.328108867
#> [141]  0.370893554  0.307408186  0.371444309 -0.382686783  0.142612463
#> [146]  0.354986846 -0.190813145  0.050998400 -0.302597637  0.370016473
#> [151]  0.135761725  0.370091630 -0.387318380 -0.387055824 -0.141981295
#> [156]  0.343639590 -0.060364795  0.358553020 -0.391601828  0.371809122
#> [161] -0.350803008 -0.390626173  0.140821844 -0.392549663 -0.033098309
#> [166]  0.114692704 -0.040348266 -0.076207818 -0.019771216  0.229478334
#> [171] -0.252690110 -0.009175412 -0.369231899  0.141230220 -0.280230645
#> [176]  0.225680308 -0.390668361  0.133930508 -0.150497256 -0.275059132
#> [181] -0.391688025  0.316941334  0.371834581 -0.130784180  0.369548392
#> [186]  0.137678187  0.010746069  0.076364229  0.271223629 -0.391171371
#> [191] -0.380869358 -0.173281878 -0.072429038 -0.070435918 -0.206800106
#> [196]  0.363764358  0.365038435  0.014618617  0.099219191  0.135567918
#> [201] -0.390185152  0.366129801 -0.310818616  0.319840758 -0.065665088
#> [206]  0.208951201 -0.122865403 -0.387469693 -0.075611971  0.057510293
#> [211] -0.378988466  0.366242600 -0.391275591 -0.381805746 -0.046221283
#> [216] -0.346767445  0.010072279  0.370050673  0.151153201 -0.084583440
#> [221]  0.367975691  0.074585588 -0.147509024  0.337979609  0.241453620
#> [226] -0.254639883 -0.146956438 -0.142255274 -0.175533813  0.116865029
#> [231] -0.237030462 -0.382415298 -0.321427159 -0.390850847  0.017922403
#> [236] -0.331639755  0.077369994  0.119797163 -0.205930344  0.081893333
#> [241] -0.258185671 -0.372025582  0.371828700 -0.033798145 -0.383754303
#> [246] -0.249398671 -0.220755600 -0.391443667 -0.335643917  0.114195207
#> [251] -0.275294508 -0.387964695  0.129751109 -0.166932625  0.126114679
#> [256]  0.371316042 -0.348289781  0.036085869  0.087815540 -0.162195506
#> [261] -0.040328918  0.126351935 -0.386193433  0.371754714  0.358106054
#> [266]  0.312238951  0.355798270  0.311092278 -0.153778493 -0.077439877
#> [271] -0.279308577 -0.126561870 -0.390899107  0.371206942 -0.370041310
#> [276] -0.054546954  0.354189189 -0.363805574 -0.387397564  0.072828891
#> [281] -0.388579830  0.201992357  0.287803482  0.040063602  0.354920114
#> [286] -0.128573428 -0.079442372  0.129166583  0.369899039 -0.388403729
#> [291] -0.089396179 -0.392388405  0.335571545  0.147518387  0.138643921
#> [296] -0.157597119 -0.381307223 -0.032102239 -0.281148247 -0.198869193
#> [301] -0.392140029  0.099317672 -0.058730147  0.054963360  0.368889301
#> [306]  0.086055031 -0.386664113 -0.380490800  0.369101428 -0.228106790
#> [311] -0.263713786 -0.154775676 -0.117165798 -0.199010094  0.122044381
#> [316] -0.172902230 -0.205492899 -0.121358998 -0.114678134  0.337472699
#> [321] -0.061809641  0.178422072  0.302694708  0.325529136 -0.108034774
#> [326] -0.194353791  0.371588548 -0.223362735  0.356880245  0.195156493
#> [331]  0.272448213 -0.389083716 -0.377198212  0.203015232  0.119739662
#> [336]  0.282706835 -0.180349074  0.296360633  0.365960032 -0.045370004
#> [341]  0.008369422 -0.335606155 -0.142043368  0.113527005 -0.241840924
#> [346]  0.370105485 -0.387642742  0.331981907  0.368843623 -0.163488689
#> [351] -0.129069401 -0.388313138 -0.023304556 -0.390448567  0.364871649
#> [356]  0.078666754 -0.383995554  0.363535633 -0.193348493 -0.086337175
#> [361] -0.246810245 -0.388661060 -0.391086709  0.129760576 -0.100142660
#> [366]  0.150684604 -0.125939278 -0.313502616 -0.100887671  0.237352548
#> [371]  0.150679962 -0.200223793 -0.159078434 -0.061945490  0.022481916
#> [376] -0.025969977 -0.263392965  0.125335142 -0.391972310  0.370302997
#> [381] -0.095728767 -0.286463717  0.322185230 -0.094967286 -0.345953645
#> [386] -0.386680725 -0.173872063 -0.177310293 -0.333128155  0.365472210
#> [391] -0.387684285 -0.251418811 -0.073270675  0.061441064  0.370061926
#> [396]  0.341221102  0.349186763 -0.232750907 -0.372271042  0.359225691
#> [401] -0.329113845  0.371767609  0.371868944  0.370817181 -0.363378170
#> [406] -0.388278074 -0.244183693  0.368499295 -0.084491405 -0.142352086
#> [411]  0.272558990  0.159322418 -0.027635688  0.260477146  0.366672969
#> [416] -0.390732168  0.035111003  0.363434673 -0.214181594 -0.388845907
#> [421]  0.228246132  0.357775774  0.003899014  0.370776992  0.249265391
#> [426] -0.054731294  0.371421328 -0.037459801 -0.030581525  0.113474469
#> [431]  0.096189739  0.061871540 -0.015111346  0.369474506 -0.301403819
#> [436]  0.029849633  0.368296299  0.137293208 -0.254904952  0.055700775
#> [441]  0.367049814  0.295475060 -0.293504451  0.127077048  0.370989781
#> [446]  0.259054586  0.244356859  0.050370468  0.036661819 -0.255879354
#> [451]  0.118963341 -0.110058636  0.028284321 -0.193431403  0.093865414
#> [456]  0.169163509  0.371388906 -0.379073077 -0.340712568 -0.264669288
#> [461] -0.188191473  0.370083295  0.100854205 -0.131462249 -0.084367772
#> [466] -0.388000128  0.142894767  0.369564804  0.028028348  0.312952498
#> [471] -0.388366249  0.122357905  0.149217047  0.162324761  0.322043406
#> [476]  0.323041818 -0.382118611 -0.300961329 -0.364668525  0.143453665
#> [481] -0.371146157 -0.379836451  0.005406544  0.242649913 -0.014917736
#> [486] -0.288455103 -0.109772301  0.136442201 -0.016155044 -0.017771760
#> [491]  0.306408847 -0.382623571 -0.266824118 -0.247317952  0.174771322
#> [496] -0.387178770 -0.388861575 -0.042032237 -0.348813307  0.022124269