Find the posterior mean

Find the posterior mean given the observations and the mixing distribution based on the family in the result

posteriormean(x, result, fun = function(x) x)

posteriormean.npnorm(x, result, fun = function(x) x)

posteriormean.npt(x, result, fun = function(x) x)

posteriormean.npnormc(x, result, fun = function(x) x)

Arguments

x
a vector of observations

result
an object of class nspmix

fun
the function to transform the mean. It finds the posterior mean of fun(x). The function fun must be vectorised.

Examples

data = rnorm(500, c(0, 2))
r1 = computemixdist(data, pi0 = 0.5)
posteriormean(data, r1)

#>   [1] 0.52334114 1.62947189 1.62598239 2.08145705 0.07356137 2.06713115
#>   [7] 1.15289892 1.24380358 1.44558215 2.08981688 0.55483761 2.01915716
#>  [13] 0.05360459 2.10845214 0.09061258 1.50533571 0.47154803 1.80509202
#>  [19] 1.54052908 1.15659444 0.09785373 1.87264563 0.50809695 2.46058101
#>  [25] 0.03953543 2.04465673 0.78188255 2.04328280 0.04076686 1.90729464
#>  [31] 0.10376423 0.31745909 0.06099066 0.84556592 1.50391493 1.18386448
#>  [37] 0.78981625 1.97067361 0.06960984 1.95541427 0.50290346 0.74265246
#>  [43] 0.63467220 1.74946149 0.15071460 1.00387285 0.16631552 0.80615025
#>  [49] 0.20918777 1.67311314 0.14939769 2.04435990 0.08753974 1.67339101
#>  [55] 0.75779646 2.06256976 0.18495976 1.78663325 0.98811747 1.89280471
#>  [61] 0.36125975 1.58426443 0.25319813 1.75639658 0.51713641 1.66490097
#>  [67] 0.13625361 1.70929149 0.10927319 2.08030087 0.05083247 2.05348991
#>  [73] 0.21359461 0.21029560 0.09689262 2.04838404 0.03981283 1.81872405
#>  [79] 0.04249195 1.79444915 0.91229069 1.62405965 0.84652097 1.99432955
#>  [85] 1.10021522 2.06351028 0.88914134 1.91555229 0.23894701 1.82104288
#>  [91] 0.37881604 0.98851178 0.09053432 1.38587334 0.09971970 1.56802688
#>  [97] 0.19929697 1.09590479 0.01899066 2.12695462 0.06773766 2.03487683
#> [103] 0.08771247 1.94595397 0.17598500 1.96970204 0.02473635 1.86292305
#> [109] 0.17268040 2.06782991 0.13149084 1.48245278 0.02492147 1.83279927
#> [115] 0.54925064 1.60303068 0.09190149 1.90616120 0.04389693 1.89105660
#> [121] 0.90101452 1.65219491 0.12099326 1.51467222 0.45429827 0.63630000
#> [127] 0.06559833 0.46828875 0.14464707 1.73122028 0.24789061 1.64982948
#> [133] 0.02510268 0.63285057 0.13854532 1.87727091 0.29004032 0.84096541
#> [139] 0.91025447 1.77653892 0.12808056 1.87464777 0.74755447 1.96984774
#> [145] 0.11965732 0.71662836 0.75694739 0.84443488 0.08002414 2.04739621
#> [151] 0.09930614 1.58743640 0.44584821 2.06370911 0.05538304 1.99676622
#> [157] 0.20769484 2.07045668 0.85002887 1.72623529 0.07323430 0.61514372
#> [163] 0.05376770 1.43583425 0.04339047 1.66613257 0.09812789 2.12173591
#> [169] 0.98873713 1.94811340 0.09508673 2.12686306 0.14213805 1.73176025
#> [175] 0.02331212 2.04247111 0.19952069 0.87341593 0.21449023 2.04072477
#> [181] 0.24620005 1.88713256 0.94840890 0.97187247 0.14850794 1.98741413
#> [187] 0.03609380 2.05213528 0.04982600 1.87467074 0.28532280 1.29551127
#> [193] 0.30297104 1.63315743 0.78600275 0.57715780 1.18930073 1.58963222
#> [199] 0.07446291 0.82718304 0.96868474 0.43829564 0.26854578 1.29370731
#> [205] 0.07195779 1.65465241 0.23765693 1.49834812 0.07390080 1.99434592
#> [211] 0.58221123 2.08840034 0.05508991 0.88070876 0.06784487 0.74741208
#> [217] 1.74366875 0.10178030 0.03553867 1.13187767 0.15066849 1.66317105
#> [223] 0.90342304 2.05928998 0.05188025 1.85799961 0.02800980 2.01907670
#> [229] 0.06308301 1.93136599 0.05469741 1.76159633 0.21459071 0.81772664
#> [235] 0.38819232 2.78763084 0.22397798 1.45847737 0.10097955 1.50538269
#> [241] 0.28161536 1.01764842 1.40371012 1.96991823 0.81406075 2.06754225
#> [247] 0.04920151 0.95171263 0.35946261 1.40232941 0.47126812 1.71243320
#> [253] 0.61864335 0.39600069 0.61600453 0.27546406 0.21101689 1.93835173
#> [259] 0.06959755 2.07661047 0.30379303 0.39441426 1.43951407 1.88963264
#> [265] 0.14266958 1.99364199 0.04154425 1.75336478 0.33553224 0.05961604
#> [271] 0.11723904 1.20097969 0.37440592 2.04431040 0.27258775 1.08841672
#> [277] 0.19289515 0.87863407 0.05385639 0.10265975 0.14593774 1.96675821
#> [283] 0.32960960 0.84109491 0.10785043 2.04600090 0.64671757 1.95572979
#> [289] 0.13347170 1.72255793 0.04788545 1.77320851 0.40294015 1.78136886
#> [295] 0.30217670 0.11816293 0.33723468 1.53230671 0.04107483 1.55267415
#> [301] 0.29092727 0.93312044 0.22431218 0.71864891 0.04961759 2.10331564
#> [307] 0.67670877 0.41270209 0.15111743 1.81810652 0.13619770 2.04253528
#> [313] 1.87097360 1.31016601 0.74013582 2.04480657 0.40388749 1.93831151
#> [319] 1.11720156 1.81341897 0.02253117 1.93203334 0.04001587 1.27823622
#> [325] 0.15069465 0.17998107 0.16393277 2.01848621 0.08710400 0.47287205
#> [331] 0.13731726 1.31546207 0.34997965 1.26910868 0.07848341 1.90531636
#> [337] 0.46091080 2.04735807 0.95404205 1.23575730 0.79704661 0.59354968
#> [343] 0.27784043 0.22136225 0.40301757 0.81709254 0.19377628 1.87662251
#> [349] 0.03281806 2.01991056 0.98543760 1.87783820 0.04218148 0.47411454
#> [355] 0.03271230 1.43225701 0.18425377 1.90384689 0.69581506 1.36708933
#> [361] 0.99009943 1.49773274 0.03279236 1.99161526 0.50333815 1.81203805
#> [367] 0.45780482 2.07115493 0.13172503 2.03353988 0.09609415 2.06763295
#> [373] 0.16846260 1.60433161 0.55027061 1.97096242 1.32925620 1.93821388
#> [379] 0.89648780 1.98624460 0.10610443 2.05816693 0.05641394 2.03480024
#> [385] 0.06150957 2.02998922 0.18278094 1.77121740 1.71315231 1.13098369
#> [391] 0.93601406 1.21239563 0.22556448 1.67916853 1.71908354 2.05857626
#> [397] 0.24382355 1.81952234 0.24622065 0.55137924 0.02809845 1.63668619
#> [403] 0.15865256 1.16503604 0.06172109 1.72258171 1.61627240 1.51315677
#> [409] 1.01672112 1.59439317 0.34926051 2.15173223 0.34161514 1.24282782
#> [415] 0.02759829 1.37206670 0.88662612 1.41533112 0.71340293 1.59878961
#> [421] 0.03584842 0.25479918 0.05375008 1.75649126 0.10362198 1.10355855
#> [427] 0.24925139 0.79856804 0.85778650 1.29206432 0.16629353 0.98532760
#> [433] 0.49661122 1.91072321 0.11152038 0.22756250 0.27780416 1.92599432
#> [439] 0.45764278 2.03783506 0.06745763 1.15305028 0.04035883 2.01442457
#> [445] 0.10921708 1.09291307 0.20217893 1.63403482 0.08564869 2.12026534
#> [451] 0.07458504 2.04701834 0.53494598 2.10353614 0.25715696 1.75736460
#> [457] 0.08788554 0.69802390 0.13737596 1.29663470 0.10086576 1.91011230
#> [463] 0.11116533 0.03547352 0.15776576 1.84301640 0.19471222 1.84830443
#> [469] 0.11066297 2.02682300 0.13002734 1.42126660 0.81533665 2.05173639
#> [475] 0.11575941 1.67706414 0.07757658 1.81751951 1.58689796 2.01701449
#> [481] 0.56997144 1.93731502 0.15714693 0.36246001 0.65500999 1.82429223
#> [487] 0.10047334 2.06178632 0.74094479 2.05325790 0.65363479 1.12904446
#> [493] 0.14725395 0.24921821 0.15962915 0.30562341 0.25809399 1.82848608
#> [499] 0.08763675 1.73952136

r2 = computemixdist(data, pi0 = 0.5, method = "nptll") # equivalent to normal
posteriormean(data, r2, fun = function(x) x^2)

#>   [1] 1.01239564 3.35412804 3.34663628 4.41204325 0.09980929 4.35227084
#>   [7] 2.33749426 2.53053949 2.96042958 4.45220735 1.07805261 4.21100741
#>  [13] 0.06397535 4.55252821 0.13169720 3.08813323 0.90464970 3.73273069
#>  [19] 3.16344330 2.34533466 0.14547238 3.87968985 0.98065311 6.79047835
#>  [25] 0.04045669 4.27866898 1.55371038 4.27472453 0.04242614 3.95566354
#>  [31] 0.15679764 0.58623471 0.07698771 1.68773657 3.08509446 2.40321010
#>  [37] 1.57039457 4.09694336 0.09255930 4.06249531 0.96984437 1.47126764
#>  [43] 1.24486956 3.61240372 0.24858649 2.02185001 0.27959059 1.60475588
#>  [49] 0.36561902 3.44790648 0.24597851 4.27781284 0.12588894 3.44850410
#>  [55] 1.50308177 4.33561100 0.31687126 3.69275067 1.98854093 3.92382388
#>  [61] 0.67634734 3.25713989 0.45484080 3.62737907 0.99947284 3.43024746
#>  [67] 0.22003719 3.52577770 0.16741146 4.40678435 0.05919080 4.30529832
#>  [73] 0.37451718 0.36785508 0.14363738 4.28961915 0.04089848 3.76229735
#>  [79] 0.04521920 3.70967190 1.82840461 3.34250868 1.68974828 4.15137262
#>  [85] 2.22578914 4.33896051 1.77957467 3.97386212 0.42586838 3.76733047
#>  [91] 0.71256728 1.98937441 0.13154899 2.83301159 0.14904041 3.22233877
#>  [97] 0.34568147 2.21665543 0.01248987 4.66100589 0.08914700 4.25148981
#> [103] 0.12621480 4.04131265 0.29889733 4.09473769 0.01904897 3.85846011
#> [109] 0.29229201 4.35491833 0.21068093 3.03920503 0.01928150 3.79286661
#> [115] 1.06639922 3.29738311 0.13414043 3.95316883 0.04752123 3.91998982
#> [121] 1.80461578 3.40293655 0.19015295 3.10810520 0.86883218 1.24827631
#> [127] 0.08526713 0.89787939 0.23658347 3.57304349 0.44404243 3.39785370
#> [133] 0.01951025 1.24105742 0.22454784 3.88980131 0.53003719 1.67804673
#> [139] 1.82410840 3.67091150 0.20399735 3.88406579 1.48156408 4.09506833
#> [145] 0.18755067 1.41663102 1.50129768 1.68535421 0.11179085 4.28667956
#> [151] 0.14824900 3.26394028 0.85129975 4.33967416 0.06707549 4.15708084
#> [157] 0.36260660 4.36510857 1.69713778 3.56229400 0.09920689 1.20401551
#> [163] 0.06425873 2.93961505 0.04668872 3.43289545 0.14599617 4.62983239
#> [169] 1.98985074 4.04613796 0.14019480 4.66045595 0.23163004 3.57420805
#> [175] 0.01730078 4.27241569 0.34613190 1.74642050 0.37632667 4.26749938
#> [181] 0.44060493 3.91138847 1.90464555 1.95420944 0.24421734 4.13529242
#> [187] 0.03507287 4.30105429 0.05746921 3.88411600 0.52038702 2.64051634
#> [193] 0.55651746 3.36204170 1.56237447 1.12463678 2.41475159 3.26864833
#> [199] 0.10147188 1.64902467 1.94747413 0.83563717 0.48611743 2.63667741
#> [205] 0.09685967 3.40821773 0.42324923 3.07318923 0.10043491 4.15141090
#> [211] 1.13518984 4.44515014 0.06656297 1.76179446 0.08934199 1.48126497
#> [217] 3.59989988 0.15298863 0.03422278 2.29290759 0.24849514 3.42652828
#> [223] 1.80969638 4.32425629 0.06099206 3.84772155 0.02332417 4.21080732
#> [229] 0.08073369 4.00884605 0.06587764 3.63861144 0.37652969 1.62911816
#> [235] 0.73193241 8.90341644 0.39551695 2.98797229 0.15145337 3.08823373
#> [241] 0.51280724 2.05098335 2.87105599 4.09522833 1.62140253 4.35382526
#> [247] 0.05640543 1.91162261 0.67264276 2.86811047 0.90406822 3.53254618
#> [253] 1.21133461 0.74806996 1.20581571 0.50023941 0.36931125 4.02436671
#> [259] 0.09253685 4.39050753 0.55820217 0.74479052 2.94747192 3.91686775
#> [265] 0.23267890 4.14976615 0.04368005 3.62083153 0.62337068 0.07454003
#> [271] 0.18284634 2.43955129 0.70346391 4.27767027 0.49436644 2.20079041
#> [277] 0.33280330 1.75742054 0.06441289 0.15467620 0.23913392 4.08806576
#> [283] 0.61119348 1.67831948 0.16466528 4.28257435 1.27008345 4.06320383
#> [289] 0.21456920 3.55436607 0.05417519 3.66370960 0.76241965 3.68135909
#> [295] 0.55488956 0.18464262 0.62687228 3.14584162 0.04292194 3.18944999
#> [301] 0.53185217 1.87236545 0.39619361 1.42087152 0.05711381 4.52370579
#> [307] 1.33290908 0.78261796 0.24938454 3.76095717 0.21992722 4.27259766
#> [313] 3.87603653 2.67170824 1.46598213 4.27910200 0.76437917 4.02427721
#> [319] 2.26179122 3.75078667 0.01637427 4.01032682 0.04122255 2.60376025
#> [325] 0.24854697 0.30689419 0.27484310 4.20934071 0.12506727 0.90740038
#> [331] 0.22213005 2.68298314 0.65310452 2.58434523 0.10892159 3.95130984
#> [337] 0.88255824 4.28656664 1.91654239 2.51343704 1.58560291 1.15887588
#> [343] 0.50509336 0.39022244 0.76257979 1.62778353 0.33457452 3.88838332
#> [349] 0.03013999 4.21288376 1.98287651 3.89104202 0.04471373 0.90998184
#> [355] 0.02998428 2.93197788 0.31545559 3.94807735 1.37296643 2.79296402
#> [361] 1.99273041 3.07187328 0.03010214 4.14504109 0.97074893 3.74779137
#> [367] 0.87611028 4.36788180 0.21114039 4.24791641 0.14211439 4.35416943
#> [373] 0.28387203 3.30017379 1.06852648 4.09759938 2.71235602 4.02405997
#> [379] 1.79506784 4.13258877 0.16129991 4.32047885 0.06888267 4.25128429
#> [385] 0.07791450 4.23856114 0.31250317 3.65940469 3.53409558 2.29101189
#> [391] 1.87847414 2.46379855 0.39872958 3.46093155 3.54687713 4.32184939
#> [397] 0.43577436 3.76402994 0.44064681 1.07083875 0.02344446 3.36961972
#> [403] 0.26433776 2.36324669 0.07829271 3.55441735 3.32579437 3.10486310
#> [409] 2.04902196 3.27885737 0.65162353 4.81269995 0.63588475 2.52846534
#> [415] 0.02276870 2.80357409 1.77427091 2.89585130 1.40986243 3.28828619
#> [421] 0.03469643 0.45810005 0.06422811 3.62758357 0.15652427 2.23287418
#> [427] 0.44681009 1.58880349 1.71348185 2.63318116 0.27954676 1.98264400
#> [433] 0.95675274 3.96321442 0.17175570 0.40277696 0.50501927 3.99694086
#> [439] 0.87577392 4.25950659 0.08863793 2.33781537 0.04177125 4.19933270
#> [445] 0.16730309 2.21031655 0.35148585 3.36392580 0.12232664 4.62111772
#> [451] 0.10169734 4.28556249 1.03657575 4.52492775 0.46290128 3.62946988
#> [457] 0.12654138 1.37759902 0.22224557 2.64290716 0.15123531 3.96186847
#> [463] 0.17106879 0.03412336 0.26257553 3.81508613 0.33645639 3.82659689
#> [469] 0.17009720 4.23036707 0.20781105 2.90851828 1.62408784 4.29981665
#> [475] 0.17997209 3.45640466 0.10723645 3.75968330 3.26278587 4.20569862
#> [481] 1.10963337 4.02206045 0.26134617 0.67882187 1.28744786 3.77438528
#> [487] 0.15048347 4.33285342 1.46768111 4.30456685 1.28456783 2.28689983
#> [493] 0.24173640 0.44674260 0.26627913 0.56195423 0.46480989 3.78349422
#> [499] 0.12607195 3.59095019

data = runif(500, min = -0.5, max = 0.5)
r3 = computemixdist(data, method = "npnormcll", beta = 100)
posteriormean(data, r3)

#>   [1] -0.361437517 -0.360497631  0.177723491 -0.203278993  0.368318741
#>   [6]  0.303322621  0.241964621 -0.075904321  0.367080265  0.369210283
#>  [11] -0.071817597 -0.072984984 -0.077650578  0.296232674  0.348610641
#>  [16]  0.358243596  0.218160940 -0.361783759  0.293564404  0.325268071
#>  [21] -0.071195877 -0.355558813 -0.063393717 -0.337501266  0.096183957
#>  [26]  0.066266392 -0.042603173  0.340316402  0.291184806 -0.361567816
#>  [31] -0.087666086  0.350585830 -0.346128617 -0.360728307  0.363371301
#>  [36] -0.358389542 -0.297474629  0.344595713  0.125606261 -0.173842050
#>  [41] -0.361638422  0.290161236 -0.238449608  0.330486861  0.369830704
#>  [46]  0.346602433 -0.022057280 -0.082193130  0.297311497  0.167565643
#>  [51]  0.227510941 -0.053829161 -0.360990777  0.292641299  0.307189026
#>  [56] -0.071920337 -0.354277013  0.368502320  0.270799645  0.365961145
#>  [61] -0.359424894  0.350089322 -0.060999006  0.354370488  0.224353619
#>  [66]  0.369632133 -0.081956307  0.156221917 -0.071706953 -0.166842425
#>  [71]  0.234803027 -0.360737030  0.272741744  0.259869762  0.364159780
#>  [76] -0.072894209  0.370145057 -0.002487281 -0.069796773  0.172432184
#>  [81]  0.358943239 -0.345703751  0.265438449  0.212010467 -0.032307313
#>  [86] -0.073726440 -0.064084899  0.334528366 -0.352836048  0.102955865
#>  [91]  0.331297423 -0.080486432  0.337509104 -0.035729676 -0.134791787
#>  [96]  0.352718079 -0.359365434 -0.125082416 -0.046920535  0.241699277
#> [101]  0.346020184 -0.235997215  0.367157304  0.369790035 -0.170972974
#> [106]  0.349903850 -0.070291587 -0.077345307  0.366895542  0.338897891
#> [111] -0.094939946 -0.337247340 -0.353656997  0.253741585  0.163187603
#> [116] -0.361700847  0.369770387 -0.355568500 -0.304097612  0.209400979
#> [121] -0.119704965  0.368521061 -0.069159762  0.338866573 -0.036836181
#> [126]  0.125478428 -0.091677878  0.308573715  0.211119876 -0.345342999
#> [131] -0.247586449  0.238997599 -0.071161575 -0.361783541  0.261026625
#> [136]  0.220334672  0.001971204  0.001076300  0.264684045 -0.060746345
#> [141] -0.360876772 -0.102354514 -0.330408970 -0.077560845 -0.072212403
#> [146]  0.222730912 -0.361373121 -0.361621230  0.046465326 -0.077893875
#> [151]  0.319781877  0.286683968 -0.070967504 -0.359861771 -0.361465868
#> [156] -0.327323461 -0.076289890  0.088635057 -0.360369048 -0.335089884
#> [161] -0.077208958 -0.113511616 -0.360918705 -0.069379075 -0.361703473
#> [166]  0.023556241  0.305509796 -0.358713588  0.103917502 -0.056297793
#> [171] -0.358781784 -0.053786943 -0.360962547 -0.075094288 -0.359707416
#> [176]  0.362978283 -0.156915787 -0.071395163  0.352489174 -0.068922540
#> [181]  0.177880370  0.315658504 -0.349914667  0.207463880  0.329308022
#> [186]  0.368442503 -0.360879492 -0.071983394  0.366517333 -0.360529773
#> [191] -0.076417510 -0.361139568  0.362236455  0.165921175 -0.354017235
#> [196]  0.366869444 -0.060814754 -0.356881832 -0.139768718 -0.075734981
#> [201]  0.158937646 -0.074586094 -0.350982641  0.353451888 -0.359666455
#> [206] -0.355096956 -0.349222985  0.368608763  0.367378549 -0.356130021
#> [211]  0.088029206 -0.222044178  0.336444691 -0.097671567 -0.308403462
#> [216]  0.327460621 -0.069701211 -0.076891728 -0.068216541 -0.361614559
#> [221] -0.077823116  0.351033779 -0.071385460 -0.050114295  0.130911059
#> [226]  0.242860228 -0.319562721  0.361654356  0.285826254 -0.077811478
#> [231]  0.367614877 -0.360128647 -0.065027724 -0.361655966  0.221790096
#> [236] -0.079275228  0.063122137 -0.358657355 -0.061831864  0.167075779
#> [241]  0.202481271 -0.357108869 -0.064652445 -0.360869116 -0.338688386
#> [246]  0.347129969 -0.067942831 -0.067800432  0.353460595  0.364422304
#> [251]  0.234062036 -0.070456169  0.184445535 -0.322750079 -0.069872482
#> [256] -0.034872169  0.229896696  0.229192555 -0.361700030  0.037100347
#> [261] -0.065965471  0.118828088 -0.361220852 -0.351593956 -0.095568012
#> [266] -0.349778698 -0.360916692 -0.068410692  0.356227830 -0.334794732
#> [271] -0.269867874  0.208458927 -0.127465340 -0.068683303  0.354530156
#> [276] -0.072506955  0.296543257 -0.069921114 -0.083889926 -0.325295810
#> [281] -0.361697382  0.337151326  0.302155555 -0.036295693 -0.088616612
#> [286] -0.078735010 -0.280389267  0.332865694 -0.158566852  0.044390957
#> [291] -0.357839957  0.363117059 -0.360987813 -0.361697946 -0.353768309
#> [296] -0.358003711  0.342735189  0.210546381 -0.355395582 -0.157156439
#> [301] -0.081722296  0.254400491  0.222917104  0.183040191  0.201802115
#> [306] -0.268851884 -0.361675181  0.367389318  0.362964723  0.299555321
#> [311]  0.036180942 -0.303157365  0.332548650 -0.294055311 -0.348012818
#> [316]  0.148631524 -0.109107302 -0.361666188 -0.077416136 -0.086003180
#> [321]  0.254788935 -0.097019600 -0.359622205 -0.077792284 -0.122637440
#> [326]  0.369984831  0.247844167 -0.091223141  0.343855700  0.313304787
#> [331] -0.344255733 -0.361420452  0.361685107 -0.361380342 -0.071045379
#> [336]  0.353275707 -0.073874202 -0.055484889  0.260314614 -0.079070447
#> [341]  0.310802801 -0.236919138 -0.306769547 -0.093242643  0.308327624
#> [346] -0.328084034 -0.349154514 -0.114221621  0.302514513  0.211296121
#> [351] -0.152534958 -0.323944794 -0.317706144  0.354105129  0.340117316
#> [356] -0.125541550  0.208916668 -0.361627749 -0.359584124 -0.070297588
#> [361]  0.309543860  0.216103126 -0.077328519  0.306197952  0.233693296
#> [366] -0.361510128  0.205206225 -0.090543402 -0.350953248  0.352821019
#> [371] -0.069939058  0.308325236 -0.059624430  0.267723646 -0.104032756
#> [376] -0.128443776 -0.361690008 -0.299728113 -0.138133050 -0.018636413
#> [381]  0.214277083  0.222037123  0.327473936 -0.085338225  0.348422741
#> [386] -0.071025535  0.271032614  0.321851161 -0.074255781  0.278116304
#> [391] -0.084306592 -0.073559278 -0.189396373 -0.279246223  0.322888910
#> [396]  0.293482569 -0.207242734 -0.054943503  0.057268057 -0.075363874
#> [401]  0.321010332  0.250174551  0.168414929 -0.068048706  0.043344605
#> [406]  0.367562461 -0.099517281 -0.360121381 -0.358813273 -0.360757237
#> [411] -0.077889764 -0.360493648 -0.336193538  0.367940313  0.007105992
#> [416] -0.133626538 -0.046540706  0.153600819 -0.083138424 -0.360567897
#> [421]  0.236520522 -0.258871314  0.362795587  0.100232941 -0.038233375
#> [426]  0.292450065  0.226994029  0.364219718 -0.354501912 -0.099252539
#> [431] -0.068861859 -0.007491320 -0.046710719  0.311907512  0.205571732
#> [436] -0.360057503  0.180042619 -0.072368296  0.032990069 -0.063693680
#> [441]  0.182891044  0.369757827 -0.341547822 -0.073946980  0.046781663
#> [446]  0.099505298 -0.361449980  0.287138349  0.317614857 -0.320092867
#> [451]  0.263160730  0.185758962 -0.063649029 -0.041409424  0.366633038
#> [456] -0.357576767 -0.164924211 -0.076241342  0.356464835  0.289187319
#> [461]  0.267685202  0.217925338  0.369409023  0.309059748  0.356302819
#> [466] -0.087157708  0.283555487 -0.358387447  0.237216180  0.277607083
#> [471]  0.225267876 -0.361621085  0.308095592 -0.346210076  0.284035801
#> [476] -0.247776151  0.352422661  0.368445532 -0.071477702  0.343789568
#> [481]  0.367413658 -0.074555192 -0.060009630  0.369173324 -0.360409814
#> [486] -0.360613321  0.356602019 -0.084386659 -0.361000840 -0.361768915
#> [491] -0.044369292 -0.349309438 -0.334571027 -0.049689305  0.358058287
#> [496] -0.113893773  0.258205947 -0.030410068 -0.265985172 -0.357703363