Find the posterior mean

Find the posterior mean given the observations and the mixing distribution based on the family in the result

posteriormean(x, result, fun = function(x) x)

posteriormean.npnorm(x, result, fun = function(x) x)

posteriormean.npt(x, result, fun = function(x) x)

posteriormean.npnormc(x, result, fun = function(x) x)

Arguments

x
a vector of observations

result
an object of class nspmix

fun
the function to transform the mean. It finds the posterior mean of fun(x). The function fun must be vectorised.

Examples

data = rnorm(500, c(0, 2))
r1 = computemixdist(data, pi0 = 0.5)
posteriormean(data, r1)

#>   [1]  0.092424444  1.776406463  1.620919795  1.791570585  0.065285554
#>   [6]  1.918812236  0.053188710  1.632550917  0.806120490  1.111488631
#>  [11]  1.333542436  2.005027411  0.203199307  1.952965694  1.257791358
#>  [16]  0.614007219  0.627030590  0.628742794  0.117577864  2.057776991
#>  [21]  0.683850082  2.089440957  0.618132249  1.987873527 -0.304686648
#>  [26]  0.616303670  0.385488856  0.169400951  0.138976412  1.299378706
#>  [31]  0.500848845  1.733618450  0.905448260  1.268577241  0.287556139
#>  [36]  0.726047215  0.635529470  0.764594452  0.709547452  2.225816329
#>  [41]  0.233454286  1.918381614  1.139179072  1.897874870  1.233796163
#>  [46]  0.999294339 -0.075520331  1.911206428  0.198820127  1.448682401
#>  [51]  0.332085261  0.892719422  0.014932163  2.009135532  1.087091182
#>  [56]  1.608080315  1.209915199  1.942901527  0.371842348  0.358213145
#>  [61]  1.086799491  2.074405712  0.343611956  1.670438276  0.310451278
#>  [66]  1.765015034 -0.012666253  1.146515631  1.159704679  1.677857595
#>  [71]  0.150001439  1.937991976  1.184621352  1.123098690  0.276059059
#>  [76]  1.320440280  0.383688401  2.058557646  1.172556128  1.826735411
#>  [81]  0.204629003  1.355526672 -0.102214958  0.922391157  0.339745429
#>  [86]  0.572155201  0.499250506  1.896887745  0.055976811  1.820496058
#>  [91]  0.619235996  0.816233968 -0.096967791  1.940824284  1.147005334
#>  [96]  2.119577303  0.498145064  0.513235184  0.019420544  1.429052029
#> [101]  0.788324088  1.848084842  0.183063480  1.017689618  0.708853651
#> [106]  1.977326158  0.937141871  0.647051532  0.848768468  0.713444460
#> [111]  0.474969006  2.019179173  0.075713262  1.979703348  0.123109541
#> [116]  1.845514283  1.826512347  1.681428693  0.058693595  1.469947603
#> [121]  0.123594066  1.957062885  0.009410326  1.293934247  0.782949583
#> [126]  1.777977694  1.892347042  1.822582984  0.072128317  0.980644252
#> [131]  1.199086992  1.159493211  0.318047084  1.287336944  1.127540385
#> [136]  1.941659424  0.292277018  2.054893648 -0.032201055  1.652265660
#> [141] -0.003845999  1.817974516  0.840610296  1.860049667  0.046905455
#> [146]  1.524244292  0.093258245  1.445113772  0.176252260  1.950269762
#> [151]  0.027378470  1.436995316  1.336426576  1.965465164  0.800270990
#> [156]  2.036999308  0.385034348  1.985762202  0.096551466  0.828624220
#> [161]  0.035551047  1.157226347 -0.034203930  2.001246389  0.070444732
#> [166]  0.359667922  0.020756648  0.804803741 -0.090796401  1.880562455
#> [171]  0.359337958  0.355767753  1.484948537  1.248780519  0.104368014
#> [176]  1.423777196 -0.335157485  1.846059272  0.177545510  1.523920699
#> [181]  0.689729612  1.845700542  0.077158516  1.536325163 -0.037253585
#> [186]  2.266905393  0.023086429  1.222841769  1.505156768  1.852804020
#> [191] -0.005715016  1.897475638  0.078964872  1.583726910  0.870781088
#> [196]  0.948409295  0.043158918  1.746988075 -0.001578258  1.707348238
#> [201] -0.032172993  1.169182857  1.699075143  1.915625275  0.193232984
#> [206]  0.809963981  0.769106647  0.895459849  0.100927963  0.895417048
#> [211]  0.077742906  1.966296114  0.080060885  2.056780763  0.360759149
#> [216]  1.716585708 -0.003939404  1.852427928  0.213689167  1.933347949
#> [221]  0.033446800  2.078142157  0.433633232  1.012491963  0.274567577
#> [226]  1.765161802  0.413246472  0.594626274  0.818829900  1.011842677
#> [231]  0.099865807  1.905372997  0.119769572  1.627423857  0.048103866
#> [236]  1.112388751  0.609653982  1.991820744  0.012985262  1.220810835
#> [241]  0.076600126  0.398362574  0.173175985  1.820317098  0.231179425
#> [246]  1.532888489  0.638771002  2.241922172 -0.212304345  0.366460193
#> [251]  0.378902456  2.055790836 -0.032492002  1.666999170  1.174723538
#> [256]  1.977766993  0.297250604  1.738420796  0.616747354  1.602403395
#> [261]  1.516323137  1.355254178  0.072559728  1.987689675  0.468122157
#> [266]  2.035173972 -0.142121803  1.646203680  0.123751539  1.497655005
#> [271]  0.985571637  0.603065171  0.305957896  2.014802278  0.689308834
#> [276]  0.331149828  0.100876197  0.976965945  0.038588738  1.080530935
#> [281]  0.700747433  0.915643075  0.773532736  1.733666521  0.008579770
#> [286]  1.823076185  1.085998624  1.859771841  0.925019246  1.865552955
#> [291]  0.306242270  0.195897186  0.963847758  0.490038254 -0.010503936
#> [296]  1.895079977  0.473991064  1.850432992  1.536485413  1.988847231
#> [301]  0.084898270  0.871890173  1.425387797  1.568888813  0.240838274
#> [306]  2.024874714  0.126004927  1.460614760  0.058703879  1.954735304
#> [311]  0.047063963  0.193130610  0.022247181  1.659368481  0.036630333
#> [316]  1.834621806  0.406331383  0.454270000  0.161060988  1.980716961
#> [321]  0.452204490  1.768300112 -0.066639530  1.781096119  1.311571743
#> [326]  2.029061469  0.110220884  1.981872246  0.848347398  2.047176271
#> [331]  0.789524540  0.595634550  0.738156032  0.228863271  1.069795366
#> [336]  2.032763522  0.367160126  0.273616331  1.020035047  1.983641214
#> [341]  1.229369266  1.633744613  0.229255057  2.040374145  0.594346454
#> [346]  0.878634441  0.631075508  1.878317601  0.681473591  1.905148639
#> [351]  0.545327218  1.952870309  0.151569285  2.050784855  0.041974859
#> [356]  1.904609750 -0.113668397  1.755230039  0.231704729  1.389789689
#> [361] -0.033964660  0.272185237  0.197007085  0.346561539  0.207745327
#> [366]  0.947069219  0.969537897  1.956491031  0.091579492  1.987986916
#> [371]  1.737761039  1.949339955  0.499735709  1.274575816  0.047757481
#> [376]  1.667580969  0.739682588  1.033073873  0.051608918  1.401282783
#> [381]  0.070658817  1.650039853  0.010862435  0.523108943  1.363415705
#> [386]  1.920065731  0.428857761  0.447912836 -0.485251280  1.988310091
#> [391]  0.677378420  1.911599699  0.709111774  1.978480676  0.270837434
#> [396]  0.469534412  0.194445398  2.068275711  0.087964436  0.886379005
#> [401] -0.086492905  1.848601467  0.702333000  1.978910068  0.606407484
#> [406]  1.996203434  1.247161281  1.429540833  0.054937452  1.960383049
#> [411]  1.615971265  1.763802079  0.857950203  1.994150784  0.111182830
#> [416]  1.784465987  0.486221968  1.944278004  0.117030951  0.757157402
#> [421]  0.662934550  0.525317444  0.109089747  1.997327060  0.106920188
#> [426]  1.805312793  0.047349889  1.836473491  0.716025466  2.015617001
#> [431]  0.596636144  1.656545089  0.044316230  1.567409642  0.703719721
#> [436]  2.084450435  0.374958774  1.385287873  0.013633874  1.523429369
#> [441]  1.462533859  0.560000577  0.029276048  1.845439312 -0.747819030
#> [446]  2.044724893  0.159650668  1.590411348  0.580430875  2.245235029
#> [451]  0.029900366  1.259488711  0.576085933  0.477537173  0.233546435
#> [456]  1.814996370  0.368712803  1.745536373  0.687692689  1.494360175
#> [461]  0.361338672  1.987120916  0.594134282  1.811384792  0.293053248
#> [466]  2.177069189  0.159920641  1.587076450  0.364077935  1.991301339
#> [471]  0.245879221  1.739258770  0.058823571  2.045565491  0.123603802
#> [476]  1.845592586  1.142123630  0.430025744  1.280844899  1.913467345
#> [481]  0.162610262  1.931790947  0.089143538  1.715792338  0.052272717
#> [486]  1.894285698  0.159431278  1.593268970  1.322931410  2.055430519
#> [491]  0.068987532  1.973733390  0.523307941  1.952248420  0.629153482
#> [496]  1.324165764  0.250156700  1.271589387  0.148114108  1.398813121

r2 = computemixdist(data, pi0 = 0.5, method = "nptll") # equivalent to normal
posteriormean(data, r2, fun = function(x) x^2)

#>   [1] 0.1916858 3.5684794 3.2228259 3.6033657 0.1517477 3.9123149 0.1363689
#>   [8] 3.2481083 1.5476567 2.1600550 2.6151143 4.1582844 0.3843637 4.0038091
#>  [15] 2.4587846 1.1692798 1.1947589 1.1981106 0.2327005 4.3470814 1.3062194
#>  [22] 4.4813975 1.1773472 4.1047457 0.6309460 1.1737707 0.7269692 0.3231145
#>  [29] 0.2693319 2.5444552 0.9490429 3.4714081 1.7453490 2.4809690 0.5411205
#>  [36] 1.3893017 1.2114001 1.4654204 1.3567846 5.1696900 0.4400813 3.9111954
#>  [43] 2.2163008 3.8586784 2.4095155 1.9334357 0.1930080 3.8926480 0.3763606
#>  [50] 2.8553736 0.6252428 1.7199365 0.1067314 4.1715892 2.1106041 3.1949994
#>  [57] 2.3605928 3.9762365 0.7008990 0.6749116 2.1100135 4.4156704 0.6471295
#>  [64] 3.3310007 0.5842825 3.5424501 0.1126748 2.2312251 2.2580781 3.3473377
#>  [71] 0.2886315 3.9629857 2.3088930 2.1836220 0.5195273 2.5879846 0.7235269
#>  [78] 4.3502006 2.2842734 3.6854641 0.3869803 2.6607256 0.2391091 1.7792109
#>  [85] 0.6397832 1.0875789 0.9459478 3.8561866 0.1397353 3.6707615 1.1795063
#>  [92] 1.5677217 0.2297881 3.9706149 2.2322216 4.6220857 0.9438074 0.9730442
#>  [99] 0.1081654 2.8141574 1.5123837 3.7362847 0.3477205 1.9704571 1.3554181
#> [106] 4.0732461 1.8087253 1.2339783 1.6323679 1.3644611 0.8989859 4.2050023
#> [113] 0.1663684 4.0802610 0.2420471 3.7301208 3.6849374 3.3552142 0.1431241
#> [120] 2.9001537 0.2428703 4.0152047 0.1058781 2.5332187 1.5017400 3.5720813
#> [127] 3.8447639 3.6756722 0.1612218 1.8959537 2.3384458 2.2576474 0.5986455
#> [134] 2.5196114 2.1926441 3.9728723 0.5500033 4.3356477 0.1305309 3.2911342
#> [141] 0.1080449 3.6648374 1.6161435 3.7651526 0.1292405 3.0151670 0.1929971
#> [148] 2.8478724 0.3354245 3.9963664 0.1122432 2.8308216 2.6210916 4.0389139
#> [155] 1.5360579 4.2677230 0.7261002 4.0983607 0.1982137 1.5923236 0.1182734
#> [162] 2.2530299 0.1328469 4.1462127 0.1588463 0.6776831 0.1087163 1.5450453
#> [169] 0.2189693 3.8154005 0.6770544 0.6702544 2.9318289 2.4402693 0.2108152
#> [176] 2.8031013 0.6918376 3.7314266 0.3377549 3.0144785 1.3177802 3.7305670
#> [183] 0.1684754 3.0409009 0.1365162 5.3910134 0.1098086 2.3870607 2.9746184
#> [190] 3.7476352 0.1088220 3.8576703 0.1711332 3.1424423 1.6761919 1.8312914
#> [197] 0.1253267 3.5015419 0.1072552 3.4126547 0.1304990 2.2773948 3.3942670
#> [204] 3.9040471 0.3661761 1.5552804 1.4743444 1.7254057 0.2052334 1.7253202
#> [211] 0.1693323 4.0412849 0.1727587 4.3431154 0.6797623 3.4332491 0.1080811
#> [218] 3.7467290 0.4036021 3.9505628 0.1165570 4.4316883 0.8193057 1.9599914
#> [225] 0.5167304 3.5427846 0.7801403 1.1314112 1.5728744 1.9586843 0.2035211
#> [232] 3.8777091 0.2363922 3.2369547 0.1305477 2.1618813 1.1607689 4.1167966
#> [239] 0.1063140 2.3829001 0.1676592 0.7516064 0.3298896 3.6703408 0.4358689
#> [246] 3.0335747 1.2177500 5.2560949 0.4480217 0.6906306 0.7143806 4.3391908
#> [253] 0.1308625 3.3234401 2.2886941 4.0745434 0.5593715 3.4822130 1.1746385
#> [260] 3.1827225 2.9983236 2.6601595 0.1618349 4.1041881 0.8857640 4.2610811
#> [267] 0.3126397 3.2778805 0.2431380 2.9587176 1.9058514 1.1478929 0.5757956
#> [274] 4.1902814 1.3169526 0.6234685 0.2051498 1.8885673 0.1209364 2.0973238
#> [281] 1.3394579 1.7657193 1.4831008 3.4715162 0.1058387 3.6768338 2.1083919
#> [288] 3.7644788 1.7844671 3.7785347 0.5763325 0.3710288 1.8622413 0.9281177
#> [295] 0.1113182 3.8516313 0.8970969 3.7419268 3.0412426 4.1077045 0.1800411
#> [302] 1.6784017 2.8064763 3.1105544 0.4537773 4.2245480 0.2469763 2.8804834
#> [309] 0.1431372 4.0087183 0.1294119 0.3659898 0.1093960 3.3066917 0.1191950
#> [316] 3.7041412 0.7668769 0.8590424 0.3082174 4.0832666 0.8550612 3.5499417
#> [323] 0.1785944 3.5792385 2.5696434 4.2392095 0.2204294 4.0867030 1.6315302
#> [330] 4.3057176 1.5147616 1.1333799 1.4131900 0.4315835 2.0756062 4.2523887
#> [337] 0.6919655 0.5149471 1.9751811 4.0919876 2.4004383 3.2507072 0.4323081
#> [344] 4.2801399 1.1308649 1.6918434 1.2026776 3.8098505 1.3015479 3.8771369
#> [351] 1.0353553 4.0035450 0.2913952 4.3195912 0.1241472 3.8757633 0.2597874
#> [358] 3.5202040 0.4368413 2.7320484 0.1325662 0.5122651 0.3730524 0.6527367
#> [365] 0.3926898 1.8286065 1.8736573 4.0136080 0.1903611 4.1050898 3.4807274
#> [372] 3.9938093 0.9468873 2.4933172 0.1301672 3.3247186 1.4162031 2.0014585
#> [379] 0.1345146 2.7560404 0.1591468 3.2862653 0.1060032 0.9921965 2.6771215
#> [386] 3.9155778 0.8101232 0.8467920 0.9936894 4.1060712 1.2935001 3.8936596
#> [393] 1.3559265 4.0766471 0.5097400 0.8884905 0.3683836 4.3898660 0.1847418
#> [400] 1.7072868 0.2115282 3.7375251 1.3425789 4.0779148 1.1544237 4.1303583
#> [407] 2.4369438 2.8151823 0.1384668 4.0245171 3.2120911 3.5396870 1.6506390
#> [414] 4.1239823 0.2220226 3.5869860 0.9207360 3.9799743 0.2317817 1.4507181
#> [421] 1.2651345 0.9964827 0.2185605 4.1338672 0.2149901 3.6352360 0.1297222
#> [428] 3.7085423 1.3695465 4.1930024 1.1353356 3.3005039 0.1265068 3.1073809
#> [435] 1.3453089 4.4592120 0.7068484 2.7226602 0.1064391 3.0134331 2.8845260
#> [442] 1.0639039 0.1134864 3.7299413 1.5250803 4.2964140 0.3057085 3.1568400
#> [449] 1.1037122 5.2739279 0.1139169 2.4622741 1.0952404 0.9039476 0.4402520
#> [456] 3.6578533 0.6949273 3.4982618 1.3137744 2.9517400 0.6808667 4.1024649
#> [463] 1.1304507 3.6494018 0.5514648 4.9121230 0.3061885 3.1496543 0.6860880
#> [470] 4.1152022 0.4631465 3.4841006 0.1432888 4.2995933 0.2428869 3.7303084
#> [477] 2.2222896 0.8123686 2.5062303 3.8984712 0.3109770 3.9464209 0.1865658
#> [484] 3.4314776 0.1352889 3.8496331 0.3053185 3.1630014 2.5931398 4.3377663
#> [491] 0.1568129 4.0627321 0.9925826 4.0018247 1.1989146 2.5956947 0.4711083
#> [498] 2.4871686 0.2853106 2.7508819

data = runif(500, min = -0.5, max = 0.5)
r3 = computemixdist(data, method = "npnormcll", beta = 100)
posteriormean(data, r3)

#>   [1] -0.0893439532 -0.1178647919  0.3045453013  0.0300234701 -0.3709208744
#>   [6] -0.1104023003  0.1423684399 -0.0777766844  0.0980890322  0.1741712390
#>  [11]  0.3492045253 -0.0364502811 -0.3811490006  0.3317627277  0.1763520648
#>  [16]  0.3742411123  0.1230640196  0.1383979169 -0.1496241026 -0.0835448456
#>  [21]  0.0625309552 -0.1505052821 -0.1339956167  0.0186263489  0.3762769604
#>  [26]  0.1255740512  0.1187067294  0.0186857193 -0.0810847601  0.0397575946
#>  [31] -0.0473368631 -0.2657498597 -0.1808332578  0.3098315281  0.1114693155
#>  [36]  0.1930866554 -0.3414317471  0.2053646836  0.0898342760 -0.3829721889
#>  [41] -0.3828071375  0.1190940416 -0.1234981471 -0.2823419321  0.3649892519
#>  [46] -0.3471112800 -0.1380058094  0.3482458822  0.2383121093 -0.0910448028
#>  [51]  0.3770710617  0.1694757649  0.3581564852  0.1380698922  0.2210781315
#>  [56] -0.3598736837  0.2955265273  0.3056948660 -0.3781960713  0.3718145804
#>  [61] -0.3278392306  0.1629479483  0.3770443878 -0.3795014028 -0.1574394293
#>  [66] -0.0110239124  0.3671935298  0.3749176800  0.1560641753 -0.3832715384
#>  [71] -0.3762410759 -0.0890512350  0.2724749718  0.1343163713 -0.1561097746
#>  [76] -0.1351645030 -0.1428519150  0.3684579276 -0.1647898163  0.3768715040
#>  [81]  0.3566454468 -0.1503829418  0.2865023039 -0.1375098409  0.1655456244
#>  [86] -0.1231564819  0.2571766568  0.1407961260 -0.3828199273 -0.3832683375
#>  [91] -0.0364158582 -0.3218555799 -0.1724099663  0.1076052273 -0.2601433946
#>  [96] -0.3664192708 -0.2908243113  0.3466623048 -0.1360933566  0.3631630900
#> [101] -0.2854549404  0.0787577067 -0.3569306774 -0.0262315851 -0.3805114547
#> [106] -0.1137097711  0.3768209490 -0.3810573494  0.1522998127  0.3725297649
#> [111]  0.0513294241 -0.1524213210  0.3741805145 -0.1549686232 -0.3522903109
#> [116]  0.3770203424  0.0101804164  0.3758794580  0.3542350407 -0.1216551081
#> [121] -0.1429075260  0.3735087039 -0.3759575712 -0.3421602676 -0.1724130938
#> [126] -0.3677328350 -0.2461130418  0.3487844983  0.1486393680  0.1499387840
#> [131]  0.3336799822 -0.3832815009  0.0026596570 -0.2209416127 -0.1336959376
#> [136] -0.1496918179 -0.3834230244 -0.1212522140 -0.1146072479  0.2577474480
#> [141]  0.2427978948 -0.0404824328 -0.3393188945  0.2645555646  0.1091708710
#> [146] -0.1941559238  0.0993149017  0.2636035361 -0.3722072490 -0.1534981601
#> [151]  0.3752001083  0.3636880213  0.1718767790 -0.3807495534  0.3184639531
#> [156] -0.1663521735 -0.1492865906 -0.0357644154  0.3753089457 -0.3832521735
#> [161] -0.1720899590  0.3177636556  0.0915305180 -0.1533372234 -0.3761565475
#> [166] -0.1193122527 -0.2718772165 -0.1053273046 -0.0804394129 -0.2409580727
#> [171]  0.2722536127 -0.3720801528 -0.1498982854 -0.3818481219 -0.1936049923
#> [176] -0.0180600017 -0.0227960318  0.3405803889 -0.3825468570 -0.0670797044
#> [181] -0.3603940299  0.1823788500 -0.3812810819 -0.1022350653  0.3662936939
#> [186] -0.0197924692 -0.0797865102  0.1329093040  0.0232055091  0.1399887099
#> [191] -0.0753987708  0.1173049177 -0.3786347348 -0.1195854609 -0.2685847750
#> [196]  0.3769577982  0.2708521921  0.3065737957  0.0843003436  0.0969530912
#> [201]  0.3243329565  0.3744348259  0.1243437729  0.2902914820 -0.0717412042
#> [206] -0.2697323927 -0.1433759656 -0.3781937198  0.3768078570  0.1397758372
#> [211] -0.1271942944  0.1835209958 -0.1148244529 -0.0264750581  0.1240443395
#> [216]  0.3126447631  0.0975073048  0.3760259686 -0.3795525954 -0.3828998668
#> [221] -0.1039864299 -0.2491571099  0.1910213861  0.3709356914 -0.3738280118
#> [226] -0.0169637680  0.3693054562  0.3745409666 -0.3579619480 -0.1276005249
#> [231] -0.3828498447  0.1526508501 -0.3773319429 -0.3684757037  0.3701495958
#> [236]  0.1373188625  0.3561763845  0.1613136125  0.3770797408  0.0701864280
#> [241] -0.0758987639  0.0611338960 -0.1434175988  0.3422353520  0.3564678154
#> [246]  0.3494829062  0.3520591536 -0.3800434695 -0.3824346758 -0.3757205179
#> [251] -0.0946398785  0.3766573620 -0.0772077285 -0.1326591474 -0.3822828442
#> [256] -0.0899104181 -0.1189355637 -0.1318757222 -0.3257184354 -0.3805500471
#> [261]  0.1032104647  0.3712651931  0.3146363221  0.3771058906 -0.2382319186
#> [266]  0.1324618385  0.1045407981 -0.1103178200 -0.1230316096  0.1593213724
#> [271] -0.1198884513  0.3770896050  0.3762906956  0.1543109137 -0.1438570355
#> [276] -0.2724680371  0.0933935222 -0.1493458687 -0.1425599665  0.0230396812
#> [281]  0.1278477230 -0.1414032171  0.3772766639  0.2497254406 -0.3082941407
#> [286] -0.1352805892 -0.3527413226  0.2993924032 -0.1310064842 -0.1204239829
#> [291] -0.3017696047  0.3764624694 -0.1138935329  0.3677609210 -0.3802126226
#> [296]  0.2739649303  0.0344862636 -0.1292305022 -0.3830056450  0.0353626218
#> [301] -0.2884139440 -0.3602486503 -0.1146090829 -0.1328551305 -0.3686922987
#> [306] -0.0708176864  0.1162587311  0.3768118919 -0.3593107892 -0.3594939729
#> [311] -0.0655622062 -0.1370884716  0.1120085851  0.3416049789 -0.2989246204
#> [316] -0.3795922675 -0.2850178817  0.3770984626  0.3553015746 -0.1462251679
#> [321] -0.1587981544  0.0297074137  0.3605345133  0.0934420366 -0.3441626281
#> [326]  0.3728584256 -0.3787319703 -0.1600208411  0.2103738230  0.3770253482
#> [331]  0.0259650848 -0.3833101122  0.0395129762  0.3751010960 -0.3765446102
#> [336]  0.0181538569  0.3742055370  0.3350881427  0.0045471162  0.0494412122
#> [341]  0.0117438451 -0.0196620865 -0.2878613808 -0.1334609359 -0.3835574802
#> [346]  0.1637985887 -0.2709319078  0.3708511222 -0.1585630682  0.0923898879
#> [351]  0.1270319714 -0.3464855032 -0.1647833044  0.1526414373  0.3694054328
#> [356]  0.3571074254 -0.0500296308  0.3732804042  0.1351071711  0.1458210489
#> [361]  0.1651452217 -0.1674593985  0.1118260431  0.1941728711  0.1001641886
#> [366] -0.1295516042  0.3618607723 -0.0938941366  0.1170000455 -0.3789122675
#> [371] -0.1465837964 -0.3823524108  0.3772256831 -0.1342101870  0.3584861514
#> [376]  0.3753246483 -0.1093252839  0.0388677528  0.1393801021 -0.1979972313
#> [381]  0.3718259454  0.3740644879 -0.1174741342 -0.1505894946  0.3658348885
#> [386] -0.0389432424  0.2647412582 -0.1571272009 -0.2593016960 -0.0417785935
#> [391] -0.3821037630 -0.3001077882  0.3623980204  0.1569425239  0.3773446001
#> [396]  0.3333855381  0.3753973401  0.2413365668  0.0756694699  0.3681282508
#> [401]  0.1377947940 -0.1079807860  0.3773592881  0.2325535695  0.3767147516
#> [406] -0.3568509208  0.1495945974  0.3604220404  0.0883992953  0.3627294765
#> [411]  0.3643515208 -0.3757121224  0.3720240847  0.0132602466  0.0502199900
#> [416] -0.1512073927 -0.3728919453 -0.0415805431 -0.1191446383 -0.3669043499
#> [421] -0.3836570963  0.3537468538 -0.1420027159 -0.1067857249 -0.1859639577
#> [426] -0.1058912582  0.0004115184 -0.3823219606 -0.3829078378  0.2288585969
#> [431] -0.3820442927 -0.2765248538  0.3463441285 -0.1098796634 -0.1345767002
#> [436]  0.3759564080  0.1957414476 -0.3759479710  0.3113223031 -0.1309843718
#> [441]  0.3769341972 -0.2468189456 -0.1878396579 -0.2153248468  0.0629946347
#> [446] -0.1241891511  0.3773655525  0.3770311645  0.1501566489 -0.3777235612
#> [451]  0.3760612491 -0.2261564934  0.3610930963  0.2602597076 -0.2795874528
#> [456]  0.1030225146  0.1776847131  0.1196822754 -0.3808396268 -0.3401570297
#> [461]  0.3764555287  0.1214552605  0.1026552180  0.0587016258 -0.1889480389
#> [466] -0.1430787129 -0.1444562756 -0.1028262169  0.3768009215 -0.3821963967
#> [471]  0.1226021070 -0.1341427934  0.3771358176  0.3770397269 -0.2681848682
#> [476] -0.1694936358  0.3574382967 -0.1300547949 -0.3619076156  0.3715114347
#> [481] -0.1239937509 -0.1929332699  0.3737908686  0.1165194144  0.2463430592
#> [486] -0.3800366812  0.1377298998  0.1333603894  0.3744428930 -0.1676309661
#> [491]  0.3763200427 -0.1926619376  0.1229638654  0.2125353049  0.0421242470
#> [496] -0.3834180795 -0.1166837850  0.3197686762  0.0496098013  0.0097457941