Find the posterior mean

Find the posterior mean given the observations and the mixing distribution based on the family in the result

posteriormean(x, result, fun = function(x) x)

posteriormean.npnorm(x, result, fun = function(x) x)

posteriormean.npt(x, result, fun = function(x) x)

posteriormean.npnormc(x, result, fun = function(x) x)

Arguments

x
a vector of observations

result
an object of class nspmix

fun
the function to transform the mean. It finds the posterior mean of fun(x). The function fun must be vectorised.

Examples

data = rnorm(500, c(0, 2))
r1 = computemixdist(data, pi0 = 0.5)
posteriormean(data, r1)

#>   [1]  0.1216704298  1.6969860052  0.1155136470  1.8207810457  0.3699749965
#>   [6]  1.2946382584  0.0109829667  1.2741076208  1.6406418192  1.3328125864
#>  [11]  0.0687466808  1.7742662066  0.5373622540  1.7859865600  0.1809826946
#>  [16]  1.8157457606  0.0373453177  1.6062603110  0.0868994174  1.7777720490
#>  [21]  0.3464424053  1.8150022517  1.6448302946  1.6705760959  0.1169189129
#>  [26]  1.1353617671  0.1855418942  1.5981316889  1.6215590667  0.8786010317
#>  [31]  0.0431928902  1.4307543966  0.6457711225  1.7931220495  0.0604273929
#>  [36]  1.6665214854  0.2653155335  1.8196087292  1.0636604385  1.4849691625
#>  [41]  0.3048554356  1.7742371748  0.0150059304  1.4860484226  0.6359641596
#>  [46]  1.3581343225  0.2582477164  1.6168340913  0.0839529567  0.9932562379
#>  [51]  0.2059581673  1.4659232202  0.8473564316  1.8321412287  0.2785843743
#>  [56]  1.7333438273  0.2539689325  1.7673076861  0.3948170719  1.6968988158
#>  [61]  1.0833585232  0.9992390913  1.3387036713  1.7231784505  0.3736430489
#>  [66]  1.3331692760  0.2184481303  1.7283199537  0.0061228712  1.3044182801
#>  [71]  0.6445553496  1.4944427602  0.0732398868  1.2799210253  0.4048014747
#>  [76]  0.4853959400  0.8788892976  0.2166543625  0.1234199736  1.1644844255
#>  [81]  0.0324591031  0.9222353315  0.4132300897  1.5774921771  0.1741948600
#>  [86]  0.1684733736  0.0281957806  1.5725516263  0.0637615081  1.1568476396
#>  [91]  1.0304516622  1.6806072474  0.0472591874  1.3934768433  0.1789025478
#>  [96]  1.2470441146  0.0546134666  1.6051260249  0.7694035004  1.6943981145
#> [101]  0.2058635611  1.7311868768  0.8134422867  1.5685541260  0.5476331415
#> [106]  0.8646042164  0.1897119153  1.4233642244  0.3818824596  1.7249182831
#> [111]  0.2331907504  1.1810642075  0.0581041561  1.7072378349  1.2744856349
#> [116]  1.1645216887  0.1512013458  1.7191868946  0.1129292883  1.6648705834
#> [121]  0.2633471566  1.1731492103  0.8451088618  1.7955041159  1.0090237139
#> [126]  1.8230111349  0.1227774725  1.3961348649  0.4198693383  1.7712567414
#> [131]  0.1632094517  1.4215243433  0.1010629733  1.8307843119  0.0007916430
#> [136]  1.5565256776  0.3012138058  0.9378400601  0.4989050159  0.2987531212
#> [141]  0.9502674688  0.4667756765  0.0041612538  1.1831474163  0.3634296070
#> [146]  1.2683353452  0.1120169809  1.7197405329  0.0459819664  1.5915981381
#> [151]  0.5780800071  1.8347251550  0.5646094454  1.8364181748  0.5457525834
#> [156]  1.6233649961  0.0034904765  0.7978565529  0.1689839152  1.7815481757
#> [161]  0.1679787150  1.7587707876  0.0455725911  1.5050417818  0.3939636184
#> [166]  1.4589732070  0.3375999349  1.2541650574  0.4914867051  1.5935410078
#> [171]  0.2635753668  0.3073178355  0.0916966922  0.5872323920  0.0164870597
#> [176]  1.3773709982  0.2195466229  1.7084451530  1.0653850769  1.7420267400
#> [181]  0.9988082926  1.3560878141  0.1721674643  1.8135449788  0.2843720449
#> [186]  0.4559805566  0.0674609901  1.8084966989  0.1024040922  1.7197928681
#> [191]  0.2902060685  1.3757685075  0.1601651747  1.7319664831  0.3897278528
#> [196]  1.8215084443  0.2059742164  1.0533105553  0.3654955939  1.7807344387
#> [201]  1.7080580079  1.3966375632  0.1668977374  1.7987470575  1.2179500325
#> [206]  1.2631006572  0.7067149570  1.3297950566  0.0547988139  1.7703498808
#> [211]  0.3343834534  1.7314894107  0.1079551130  1.8378696826  0.7072933291
#> [216]  1.4504032778  0.7377270374  0.8063589064  0.2279212802  1.8282663331
#> [221]  0.2097175796  1.8011189865  0.1986160098  0.1750209214  0.0635129017
#> [226]  1.6933303015  1.3378108718  1.3101474606  0.4287814700  1.0183506669
#> [231]  0.0496698812  1.1723070073  1.2833676807  1.8035203607  0.6505157733
#> [236]  1.8153348571  0.1413344564  1.7868468824  0.3897565113  0.9467989643
#> [241]  0.0693559642  1.6420037291  0.3874894700  1.4800638702  0.0515007128
#> [246]  0.2457936994  0.5261309258  1.4940263651  0.6163088541  0.5464262340
#> [251]  0.0321755750  1.5810617144  0.0238791267  0.2914844120  0.8047730491
#> [256]  1.8003169846  0.0484666602  1.5061071466  0.0400899481  1.7160188461
#> [261]  0.2709117092  1.0978973869  0.0767513577  0.9357325135  0.5339068204
#> [266]  1.7629212253  0.4929070127  1.1196457275  0.0114683183  1.7567672050
#> [271]  0.2137993591  1.6099043405  1.0157737780  1.0780326182  1.1089541727
#> [276]  1.4276344395  0.0855212930  1.5388481650  0.5020010973  1.5381506701
#> [281]  0.0612397588  1.5011676501  0.0004666221  1.3678454848  0.5477477372
#> [286]  1.5557644684  0.7259571063  1.7493777275  0.6915280317  1.7746649395
#> [291]  0.0126478164  1.6260380515  1.4529684296  1.3859849127  0.5454986448
#> [296]  1.6713815285  0.6222787647  1.5061210870  0.0958799015  1.6492632006
#> [301]  0.3887267516  1.5484075757  0.2980728166  1.7358165892  0.0764834893
#> [306]  1.8166976498  1.6875139689  1.8306351985  0.1342615926  0.3716047760
#> [311]  0.3087333195  1.7985539536  0.8001721662  0.7935999102  0.0416478723
#> [316]  1.6685285714  0.6759657468  1.7979099862  0.8323817778  1.0156864047
#> [321]  0.0288190181  0.2783233501  0.5832273986  1.7937943301  0.4662676471
#> [326]  0.9382537407  1.4160939286  1.4528459167  1.5315649856  1.7599663115
#> [331]  1.1465009341  1.7191696793  0.2563106360  0.9779161564  0.1121675881
#> [336]  1.6069435489  0.4324996933  1.6122507876  0.2134805503  1.6828026085
#> [341]  0.1190124680  1.3496519813  0.0220151721  1.7955007525  0.3052970994
#> [346]  1.5925014202  0.0814248823  1.5183992705  0.0437032518  0.9466315400
#> [351]  0.0086227937  1.8066512004  0.3443451641  0.8891003468  1.1250757592
#> [356]  1.7705167457  0.7703883431  1.5292348556  0.1417098071  1.5865967063
#> [361]  0.5119689964  1.7910355624  1.4292252861  1.3674791906  0.0497502897
#> [366]  1.6584921738  1.8094946438  1.4183278217  1.1103172582  1.5207257804
#> [371]  0.1288099375  1.2242911764  0.2021414282  1.8348996212  0.4447152013
#> [376]  1.7496074840  0.8512889245  1.8415174211  0.7980887698  0.6344367041
#> [381]  0.0481552825  0.5159804229 -0.0023737721  1.8239858734  0.1254848995
#> [386]  1.7326123856  1.1395059273  1.3656655834  0.2289506418  1.7234222918
#> [391]  0.3990678723  1.1746792468  0.7608116479  1.7590786345  0.1778462324
#> [396]  0.7997637283  0.6817171071  1.8290765301  1.2916198572  1.7978837517
#> [401]  0.1277957042  1.3527274252  0.1066117025  1.4285280312  0.7430009932
#> [406]  1.7018660904  0.0098886269  0.3513597576  0.2738483023  1.6732934041
#> [411]  0.0028920541  1.1410652436  0.2962854136  1.8301664195  0.0069401025
#> [416]  0.7721917151  0.0442227050  1.6805534693  0.4630482538  1.4804271763
#> [421]  0.0162748309  1.2836465123  0.1117836895  1.5403460781  1.4488407571
#> [426]  1.2496782238  0.3193588850  1.2281116992  0.0329661448  1.7720400666
#> [431]  0.0466795610  1.8453947327  0.3273911682  1.6454124285  0.2555394750
#> [436]  1.8386520116  0.7061965531  1.5167911318  0.0377885807  0.9556380403
#> [441]  0.0752948033  1.8205758855  0.2096687037  1.1881387180  0.8118796894
#> [446]  1.8261638862  0.0883859997  1.4083269877  0.4723678525  1.6227822438
#> [451]  0.5131153308  1.8364071883  0.2580555302  1.5016460776  0.5689362823
#> [456]  1.6060793650  0.6772991874  1.5301083415  0.5389328447  0.5338262135
#> [461]  0.2900619085  1.3627785913  0.0361329570  1.4277076510  0.4899223021
#> [466]  1.6558129550  0.7301577399  1.8035665480  0.8425885058  1.4783605654
#> [471]  0.0279945062  1.5018365246  1.1298370096  1.5767675967  0.0154947222
#> [476]  1.7502960689  0.4323438343  1.4351233575  0.3695519191  1.8270656365
#> [481]  1.2315989677  1.4012672805  0.2886017221  0.6138460173  0.0833459632
#> [486]  1.8408360351  1.3721073831  1.5275381292  0.0887088027  0.6817513539
#> [491]  1.6662244615  1.6627544392  0.1054526287  1.8131112314  0.0426933788
#> [496]  1.7831949029  1.3339308033  1.5367445002  1.3410255198  1.7727809787

r2 = computemixdist(data, pi0 = 0.5, method = "nptll") # equivalent to normal
posteriormean(data, r2, fun = function(x) x^2)

#>   [1] 0.225109053 3.134352843 0.213756495 3.363003172 0.683464563 2.391218525
#>   [7] 0.023145338 2.353298786 3.030284937 2.461725936 0.127618235 3.277089855
#>  [13] 0.992583274 3.298737442 0.334536819 3.353702952 0.070062292 2.966782240
#>  [19] 0.161025502 3.283565174 0.640009595 3.352329684 3.038021061 3.085573633
#>  [25] 0.216347521 2.097038419 0.342951239 2.951768670 2.995039058 1.622812793
#>  [31] 0.080737477 2.642623410 1.192797423 3.311916752 0.112329242 3.078084754
#>  [37] 0.490213863 3.360837892 1.964608427 2.742757736 0.563219041 3.277036233
#>  [43] 0.029981418 2.744751124 1.174685257 2.508494856 0.477164813 2.986312022
#>  [49] 0.155599797 1.834574755 0.380634516 2.707579985 1.565106102 3.383985545
#>  [55] 0.514712290 3.201505936 0.469265181 3.264237415 0.729338791 3.134191804
#>  [61] 2.000990086 1.845624815 2.472606690 3.182730425 0.690238043 2.462384737
#>  [67] 0.403689980 3.192226811 0.015528185 2.409282069 1.190552049 2.760255426
#>  [73] 0.135882292 2.364036043 0.747776628 0.896612377 1.623345202 0.400378743
#>  [79] 0.228335368 2.150827156 0.061172189 1.703402735 0.763341568 2.913647493
#>  [85] 0.322009811 0.311451328 0.053448272 2.904522299 0.118454451 2.136722218
#>  [91] 1.903273102 3.104101207 0.088177216 2.573772102 0.330697816 2.303312997
#>  [97] 0.101656806 2.964687212 1.421133161 3.129572997 0.380459886 3.197522035
#> [103] 1.502469167 2.897138919 1.011551695 1.596961586 0.350647639 2.628973822
#> [109] 0.705453154 3.185943906 0.430905367 2.181449585 0.108063137 3.153288022
#> [115] 2.353996972 2.150895980 0.279581114 3.175357999 0.208991722 3.075035532
#> [121] 0.486579696 2.166830787 1.560954998 3.316316449 1.863696582 3.367122169
#> [127] 0.227150524 2.578681441 0.775602196 3.271531350 0.301737747 2.625575577
#> [133] 0.187118838 3.381479306 0.009713975 2.874922389 0.556495002 1.732223813
#> [139] 0.921560596 0.551951530 1.755176583 0.862225228 0.012891386 2.185297220
#> [145] 0.671377816 2.342637497 0.207309789 3.176380572 0.085839200 2.939701200
#> [151] 1.067781800 3.388758084 1.042903927 3.391885109 1.008078648 2.998374613
#> [157] 0.012092938 1.473683548 0.312393464 3.290539710 0.310538507 3.248469700
#> [163] 0.085090040 2.779831774 0.727762756 2.694743353 0.623681534 2.316465240
#> [169] 0.907860627 2.943289681 0.487001034 0.567765732 0.169861388 1.084684783
#> [175] 0.032558236 2.544024774 0.405717776 3.155517945 1.967793770 3.217543330
#> [181] 1.844829149 2.504714980 0.318268374 3.349638087 0.525398383 0.842289400
#> [187] 0.125254312 3.340313866 0.189590429 3.176477235 0.536170199 2.541064991
#> [193] 0.296120344 3.198961973 0.719940775 3.364346685 0.380664140 1.945492579
#> [199] 0.675192868 3.289036731 3.154802886 2.579609920 0.308543734 3.322306190
#> [205] 2.249576835 2.332969124 1.305353506 2.456152600 0.101996850 3.269856371
#> [211] 0.617742189 3.198080817 0.199821884 3.394566060 1.306421698 2.678914750
#> [217] 1.362629662 1.489386716 0.421177550 3.376828573 0.387573935 3.326687164
#> [223] 0.367082201 0.323534277 0.117997621 3.127600742 2.470957702 2.419863777
#> [229] 0.792060229 1.880923065 0.092592742 2.165275260 2.370401956 3.331122524
#> [235] 1.201560183 3.352944009 0.261377868 3.300326465 0.719993698 1.748770436
#> [241] 0.128738613 3.032800388 0.715807276 2.733697686 0.095948241 0.454172084
#> [247] 0.971841183 2.759486346 1.138384631 1.009322757 0.060657429 2.920240427
#> [253] 0.045674324 0.538530523 1.486457763 3.325205860 0.090388463 2.781799498
#> [259] 0.075068792 3.169506596 0.500546052 2.027842876 0.142343272 1.728331288
#> [265] 0.986201747 3.256135589 0.910483615 2.068011367 0.023953272 3.244769069
#> [271] 0.395108556 2.973512765 1.876163661 1.991153325 2.048264400 2.636860874
#> [277] 0.158487654 2.842272057 0.927278383 2.840983786 0.113821370 2.772676268
#> [283] 0.009544299 2.526431258 1.011763332 2.873516437 1.340891762 3.231120653
#> [289] 1.277304916 3.277826318 0.025938683 3.003311750 2.683652568 2.559934582
#> [295] 1.007609670 3.087061270 1.149410209 2.781825246 0.177568073 3.046208650
#> [301] 0.718092095 2.859928269 0.550695401 3.206073142 0.141850332 3.355461100
#> [307] 3.116857950 3.381203892 0.248330963 0.686474135 0.570379358 3.321949526
#> [313] 1.477960299 1.465821875 0.077913892 3.081791848 1.248563162 3.320760112
#> [319] 1.537448985 1.876002286 0.054574983 0.514230350 1.077288198 3.313158460
#> [325] 0.861287025 1.732987859 2.615545647 2.683426287 2.828820043 3.250677842
#> [331] 2.117612141 3.175326202 0.473588499 1.806242338 0.207587446 2.968044181
#> [337] 0.798926703 2.977846655 0.394520056 3.108156048 0.220207792 2.492828077
#> [343] 0.042337940 3.316310237 0.564034549 2.941369562 0.150945369 2.804503007
#> [349] 0.081670600 1.748461212 0.019315497 3.336905214 0.636136918 1.642204369
#> [355] 2.078040469 3.270164572 1.422952079 2.824516300 0.262070298 2.930463555
#> [361] 0.945686974 3.308062992 2.639799155 2.525754717 0.092740077 3.063254589
#> [367] 3.342157080 2.619671628 2.050781976 2.808800063 0.238275729 2.261288790
#> [373] 0.373589424 3.389080325 0.821485305 3.231545015 1.572369143 3.401303472
#> [379] 1.474112434 1.171864252 0.089818155 0.953095249 0.009389525 3.368922521
#> [385] 0.232143441 3.200154959 2.104692563 2.522405005 0.423077809 3.183180801
#> [391] 0.737188571 2.169656725 1.405264776 3.249038296 0.328748364 1.477205947
#> [397] 1.259185254 3.378325015 2.385643587 3.320711657 0.236405158 2.498508383
#> [403] 0.197345572 2.638511331 1.372370151 3.143366384 0.021346849 0.649089812
#> [409] 0.505967928 3.090592511 0.011438817 2.107572575 0.547395112 3.380338052
#> [415] 0.016725305 1.426282750 0.082620546 3.104001879 0.855341617 2.734368710
#> [421] 0.032187553 2.370916953 0.206879699 2.845038699 2.676028785 2.308178145
#> [427] 0.589999172 2.268345214 0.062093086 3.272978157 0.087116054 3.408464909
#> [433] 0.604830786 3.039096264 0.472164757 3.396011030 1.304396070 2.801532780
#> [439] 0.070870306 1.765095754 0.139663014 3.362624239 0.387483715 2.194516031
#> [445] 1.499583171 3.372945332 0.163763314 2.601200176 0.872552597 2.997298268
#> [451] 0.947804015 3.391864817 0.476809990 2.773559922 1.050894859 2.966448032
#> [457] 1.251025870 2.826129625 0.995483859 0.986052881 0.535904024 2.517072765
#> [463] 0.067853488 2.636996095 0.904971529 3.058306064 1.348649913 3.331207833
#> [469] 1.556300078 2.730551691 0.053084603 2.773911677 2.086834356 2.912309193
#> [475] 0.030829074 3.232816836 0.798638876 2.650692847 0.682683304 3.374610874
#> [481] 2.274786125 2.588160963 0.533207955 1.133836125 0.154482190 3.400044945
#> [487] 2.534302935 2.821382451 0.164357846 1.259248504 3.077536149 3.071127007
#> [493] 0.195209140 3.348836951 0.079824388 3.293581229 2.463791267 2.838386593
#> [499] 2.476895114 3.274346628

data = runif(500, min = -0.5, max = 0.5)
r3 = computemixdist(data, method = "npnormcll", beta = 100)
posteriormean(data, r3)

#>   [1] -0.226845409  0.354460560  0.222767928  0.124201137 -0.233511816
#>   [6] -0.374318930 -0.197518864  0.043225508  0.309942679 -0.210916730
#>  [11] -0.187210090  0.353610592  0.002289812  0.354203150  0.343375224
#>  [16]  0.354317591 -0.361041520 -0.188852654  0.353519586  0.113341247
#>  [21]  0.354313601  0.354185791  0.245371327 -0.127200744  0.138332587
#>  [26] -0.220420148 -0.190970640  0.341964730  0.174141581  0.304573792
#>  [31] -0.191952380  0.113359424  0.088837600 -0.246588001 -0.228818197
#>  [36]  0.349031879  0.313502508 -0.388249897  0.282149529  0.351056637
#>  [41]  0.350657589 -0.357594264  0.354261462  0.351946140 -0.219183596
#>  [46]  0.350271894  0.339757521 -0.061728411 -0.377826106  0.347397535
#>  [51] -0.183765927 -0.241187495  0.337981736  0.354487867 -0.243059739
#>  [56] -0.373532274  0.107166072  0.345704786 -0.295070877  0.016460617
#>  [61] -0.163103482  0.336760810 -0.004504166 -0.241358852 -0.211929634
#>  [66]  0.336175535 -0.388228807 -0.109242447  0.076586093  0.034377918
#>  [71]  0.165240296  0.018446521  0.050871573  0.342177947 -0.189601576
#>  [76] -0.386882037  0.333196367  0.308055950  0.346239778 -0.380998761
#>  [81] -0.389981546  0.073734263  0.347927787 -0.051716741 -0.386639824
#>  [86] -0.103307874 -0.155041453 -0.012459438 -0.347161955 -0.009196521
#>  [91] -0.215612484  0.001964196  0.351682814  0.012521812  0.199368808
#>  [96] -0.382878230  0.175798822  0.323813650 -0.389017067 -0.386753982
#> [101]  0.065041227  0.317689147  0.252962958 -0.212928719  0.327780078
#> [106] -0.082550325 -0.316742559  0.082044625  0.337239447  0.310212777
#> [111] -0.376255493 -0.379630441  0.353911440 -0.372938770 -0.261783127
#> [116] -0.061846140  0.307398393  0.344645588  0.097864976 -0.131688655
#> [121] -0.338501760  0.349498707  0.348511080 -0.175024337  0.022938470
#> [126]  0.005924439  0.320187799  0.352432624  0.070337286  0.336367212
#> [131]  0.328928656 -0.205366407  0.019144997 -0.198246059  0.347627461
#> [136]  0.216309500 -0.157540173  0.354554442  0.195114518 -0.236221197
#> [141]  0.119537865 -0.171544468  0.324970146 -0.183019096 -0.213710484
#> [146] -0.379906016  0.060608714 -0.390367730 -0.205733651 -0.206181975
#> [151]  0.334194711 -0.358958801 -0.217028315 -0.061804002  0.353897675
#> [156] -0.186974447  0.313978975 -0.348870439  0.006500257  0.031893024
#> [161]  0.349985271 -0.158109090 -0.024861898 -0.060586547  0.216317187
#> [166] -0.384287303  0.195647226  0.138350446  0.138851591  0.011531611
#> [171] -0.208865752 -0.147590332 -0.365858179 -0.222050833 -0.385199142
#> [176] -0.020749064 -0.183533548 -0.264330186  0.278566465 -0.230694128
#> [181] -0.109761513  0.199182419 -0.360779432 -0.346286102  0.354391045
#> [186]  0.354269959  0.286404392 -0.387697996  0.354225595  0.288392861
#> [191]  0.350120136 -0.251290480 -0.211740604  0.084319047 -0.174019030
#> [196]  0.284262468 -0.378424334  0.350432448  0.300046768  0.349614769
#> [201]  0.353614180 -0.183579020 -0.301337089 -0.332756870 -0.360642884
#> [206] -0.018718450  0.330943450  0.347490451 -0.381260356  0.247659722
#> [211]  0.354197933 -0.190306472  0.221502440  0.076035013 -0.385744474
#> [216] -0.047718924  0.110497138  0.325000526  0.043107763  0.349515481
#> [221] -0.390154080 -0.050964423 -0.004959990  0.320068251 -0.207056830
#> [226]  0.073004908 -0.290865973  0.328843201  0.075465841 -0.210619253
#> [231]  0.326162869  0.082648358  0.353886381 -0.301396528 -0.390362386
#> [236] -0.148978979  0.301043105  0.062185785 -0.310853899 -0.373300884
#> [241]  0.348838539 -0.336804392  0.107244068 -0.051793321  0.352796750
#> [246]  0.087717547 -0.380103698  0.025670658  0.176213901 -0.003246320
#> [251]  0.090881107 -0.009567582  0.344911878 -0.213476635 -0.105634910
#> [256]  0.324337684  0.348504295  0.041184779 -0.181264997  0.065327003
#> [261] -0.177947174 -0.323995762  0.020363758 -0.386804710 -0.339657525
#> [266] -0.309175059 -0.183587506 -0.164640531  0.347452519 -0.162411150
#> [271]  0.353698722 -0.005918160  0.347595242 -0.186975271 -0.176984446
#> [276] -0.320901686  0.098046517 -0.295705935  0.041400498 -0.280444942
#> [281] -0.366633804 -0.180070671  0.080323395  0.177563793 -0.270180682
#> [286] -0.096716208 -0.337883532 -0.267099044  0.295056250 -0.006787993
#> [291] -0.234436636  0.322238869 -0.056779561 -0.311433873 -0.316871044
#> [296]  0.354121562  0.029794189  0.029959637  0.174922199  0.316180527
#> [301]  0.180724940  0.349236014 -0.340859822 -0.215008842  0.350897896
#> [306] -0.164382254  0.353077114 -0.181855546  0.117299539  0.029085727
#> [311]  0.055825250  0.010532755  0.035814540 -0.137229811  0.080359981
#> [316] -0.216354079  0.118682937  0.141702013  0.050897814 -0.191766474
#> [321]  0.265460173  0.343652740 -0.152041120 -0.058910733  0.348075082
#> [326] -0.105286002 -0.003177323  0.014931481 -0.386332218 -0.122999355
#> [331] -0.032588298  0.052424376  0.150563427  0.164794864  0.350435144
#> [336]  0.141677608  0.031028517 -0.190351445 -0.388186510 -0.350025487
#> [341]  0.013355219  0.021017287 -0.360286251 -0.221203573 -0.317524444
#> [346] -0.206564749 -0.330111520 -0.076784066 -0.261975270 -0.377425711
#> [351]  0.348307163 -0.259643088 -0.272568158 -0.187530558 -0.180045475
#> [356]  0.298698964  0.062145501 -0.320129785  0.331964314  0.352615426
#> [361] -0.223886036  0.276940186 -0.270817218  0.338494463 -0.194854861
#> [366]  0.295820968 -0.356782451  0.343065849  0.335461753  0.345171199
#> [371] -0.389452448 -0.386586553 -0.196428077 -0.182098315 -0.378444234
#> [376]  0.207206944 -0.388306729  0.336593059  0.021429199  0.295931806
#> [381]  0.353295098  0.043156747  0.353648186 -0.187328104 -0.202574473
#> [386]  0.354116244 -0.231498145  0.324412096 -0.198255646  0.188027079
#> [391]  0.289502209 -0.209797950 -0.390284249 -0.086639071  0.077775125
#> [396] -0.057586057 -0.172488746  0.025546090 -0.244700475 -0.201206809
#> [401] -0.388627555  0.034118427 -0.091847821  0.354235969 -0.384600165
#> [406]  0.110327990  0.005883598 -0.390324845 -0.293306136  0.106285585
#> [411]  0.026940197  0.350530898  0.157791437  0.023299336 -0.192261029
#> [416] -0.120366529 -0.329347445 -0.140827772 -0.211813588  0.353538760
#> [421] -0.224051332  0.253708312 -0.380481022 -0.149892063 -0.256000740
#> [426]  0.349181970 -0.084879724 -0.185178839 -0.100095557 -0.326185922
#> [431]  0.101652402  0.298878606  0.351376618  0.086422101  0.103035665
#> [436]  0.136553344  0.352983192  0.354496634 -0.219517537 -0.200474144
#> [441]  0.352042531 -0.147473139  0.233716377  0.341831237 -0.359369915
#> [446]  0.343039240  0.129398997 -0.389940047 -0.364553295 -0.079312270
#> [451]  0.338142023 -0.177106414 -0.198520016  0.351731223 -0.108236162
#> [456]  0.058884151 -0.227027291 -0.346685650 -0.279389750  0.035327565
#> [461]  0.353829760 -0.089698563  0.341658645 -0.376862466  0.275285360
#> [466] -0.214480101  0.034796374 -0.385420400  0.298706161  0.351976076
#> [471]  0.269180062 -0.369853232  0.353548070 -0.301492135  0.326021354
#> [476]  0.148747561 -0.092049539  0.279741809 -0.096856988  0.075265143
#> [481] -0.185799093  0.118450945 -0.161140135  0.348249408 -0.108536065
#> [486] -0.038519507  0.350305977  0.124602063 -0.389881185  0.128616717
#> [491] -0.232795837  0.354058545 -0.050514761  0.145137696 -0.368777303
#> [496] -0.268055132  0.353599649  0.059111777 -0.107102143  0.066251601