Find the posterior mean

Find the posterior mean given the observations and the mixing distribution based on the family in the result

posteriormean(x, result, fun = function(x) x)

posteriormean.npnorm(x, result, fun = function(x) x)

posteriormean.npt(x, result, fun = function(x) x)

posteriormean.npnormc(x, result, fun = function(x) x)

Arguments

x
a vector of observations

result
an object of class nspmix

fun
the function to transform the mean. It finds the posterior mean of fun(x). The function fun must be vectorised.

Examples

data = rnorm(500, c(0, 2))
r1 = computemixdist(data, pi0 = 0.5)
posteriormean(data, r1)

#>   [1]  0.092424444  1.776406463  1.620919795  1.791570585  0.065285554
#>   [6]  1.918812236  0.053188710  1.632550917  0.806120490  1.111488631
#>  [11]  1.333542436  2.005027411  0.203199307  1.952965694  1.257791358
#>  [16]  0.614007219  0.627030590  0.628742794  0.117577864  2.057776991
#>  [21]  0.683850082  2.089440957  0.618132249  1.987873527 -0.304686648
#>  [26]  0.616303670  0.385488856  0.169400951  0.138976412  1.299378706
#>  [31]  0.500848845  1.733618450  0.905448260  1.268577241  0.287556139
#>  [36]  0.726047215  0.635529470  0.764594452  0.709547452  2.225816329
#>  [41]  0.233454286  1.918381614  1.139179072  1.897874870  1.233796163
#>  [46]  0.999294339 -0.075520331  1.911206428  0.198820127  1.448682401
#>  [51]  0.332085261  0.892719422  0.014932163  2.009135532  1.087091182
#>  [56]  1.608080315  1.209915199  1.942901527  0.371842348  0.358213145
#>  [61]  1.086799491  2.074405712  0.343611956  1.670438276  0.310451278
#>  [66]  1.765015034 -0.012666253  1.146515631  1.159704679  1.677857595
#>  [71]  0.150001439  1.937991976  1.184621352  1.123098690  0.276059059
#>  [76]  1.320440280  0.383688401  2.058557646  1.172556128  1.826735411
#>  [81]  0.204629003  1.355526672 -0.102214958  0.922391157  0.339745429
#>  [86]  0.572155201  0.499250506  1.896887745  0.055976811  1.820496058
#>  [91]  0.619235996  0.816233968 -0.096967791  1.940824284  1.147005334
#>  [96]  2.119577303  0.498145064  0.513235184  0.019420544  1.429052029
#> [101]  0.788324088  1.848084842  0.183063480  1.017689618  0.708853651
#> [106]  1.977326158  0.937141871  0.647051532  0.848768468  0.713444460
#> [111]  0.474969006  2.019179173  0.075713262  1.979703348  0.123109541
#> [116]  1.845514283  1.826512347  1.681428693  0.058693595  1.469947603
#> [121]  0.123594066  1.957062885  0.009410326  1.293934247  0.782949583
#> [126]  1.777977694  1.892347042  1.822582984  0.072128317  0.980644252
#> [131]  1.199086992  1.159493211  0.318047084  1.287336944  1.127540385
#> [136]  1.941659424  0.292277018  2.054893648 -0.032201055  1.652265660
#> [141] -0.003845999  1.817974516  0.840610296  1.860049667  0.046905455
#> [146]  1.524244292  0.093258245  1.445113772  0.176252260  1.950269762
#> [151]  0.027378470  1.436995316  1.336426576  1.965465164  0.800270990
#> [156]  2.036999308  0.385034348  1.985762202  0.096551466  0.828624220
#> [161]  0.035551047  1.157226347 -0.034203930  2.001246389  0.070444732
#> [166]  0.359667922  0.020756648  0.804803741 -0.090796401  1.880562455
#> [171]  0.359337958  0.355767753  1.484948537  1.248780519  0.104368014
#> [176]  1.423777196 -0.335157485  1.846059272  0.177545510  1.523920699
#> [181]  0.689729612  1.845700542  0.077158516  1.536325163 -0.037253585
#> [186]  2.266905393  0.023086429  1.222841769  1.505156768  1.852804020
#> [191] -0.005715016  1.897475638  0.078964872  1.583726910  0.870781088
#> [196]  0.948409295  0.043158918  1.746988075 -0.001578258  1.707348238
#> [201] -0.032172993  1.169182857  1.699075143  1.915625275  0.193232984
#> [206]  0.809963981  0.769106647  0.895459849  0.100927963  0.895417048
#> [211]  0.077742906  1.966296114  0.080060885  2.056780763  0.360759149
#> [216]  1.716585708 -0.003939404  1.852427928  0.213689167  1.933347949
#> [221]  0.033446800  2.078142157  0.433633232  1.012491963  0.274567577
#> [226]  1.765161802  0.413246472  0.594626274  0.818829900  1.011842677
#> [231]  0.099865807  1.905372997  0.119769572  1.627423857  0.048103866
#> [236]  1.112388751  0.609653982  1.991820744  0.012985262  1.220810835
#> [241]  0.076600126  0.398362574  0.173175985  1.820317098  0.231179425
#> [246]  1.532888489  0.638771002  2.241922172 -0.212304345  0.366460193
#> [251]  0.378902456  2.055790836 -0.032492002  1.666999170  1.174723538
#> [256]  1.977766993  0.297250604  1.738420796  0.616747354  1.602403395
#> [261]  1.516323137  1.355254178  0.072559728  1.987689675  0.468122157
#> [266]  2.035173972 -0.142121803  1.646203680  0.123751539  1.497655005
#> [271]  0.985571637  0.603065171  0.305957896  2.014802278  0.689308834
#> [276]  0.331149828  0.100876197  0.976965945  0.038588738  1.080530935
#> [281]  0.700747433  0.915643075  0.773532736  1.733666521  0.008579770
#> [286]  1.823076185  1.085998624  1.859771841  0.925019246  1.865552955
#> [291]  0.306242270  0.195897186  0.963847758  0.490038254 -0.010503936
#> [296]  1.895079977  0.473991064  1.850432992  1.536485413  1.988847231
#> [301]  0.084898270  0.871890173  1.425387797  1.568888813  0.240838274
#> [306]  2.024874714  0.126004927  1.460614760  0.058703879  1.954735304
#> [311]  0.047063963  0.193130610  0.022247181  1.659368481  0.036630333
#> [316]  1.834621806  0.406331383  0.454270000  0.161060988  1.980716961
#> [321]  0.452204490  1.768300112 -0.066639530  1.781096119  1.311571743
#> [326]  2.029061469  0.110220884  1.981872246  0.848347398  2.047176271
#> [331]  0.789524540  0.595634550  0.738156032  0.228863271  1.069795366
#> [336]  2.032763522  0.367160126  0.273616331  1.020035047  1.983641214
#> [341]  1.229369266  1.633744613  0.229255057  2.040374145  0.594346454
#> [346]  0.878634441  0.631075508  1.878317601  0.681473591  1.905148639
#> [351]  0.545327218  1.952870309  0.151569285  2.050784855  0.041974859
#> [356]  1.904609750 -0.113668397  1.755230039  0.231704729  1.389789689
#> [361] -0.033964660  0.272185237  0.197007085  0.346561539  0.207745327
#> [366]  0.947069219  0.969537897  1.956491031  0.091579492  1.987986916
#> [371]  1.737761039  1.949339955  0.499735709  1.274575816  0.047757481
#> [376]  1.667580969  0.739682588  1.033073873  0.051608918  1.401282783
#> [381]  0.070658817  1.650039853  0.010862435  0.523108943  1.363415705
#> [386]  1.920065731  0.428857761  0.447912836 -0.485251280  1.988310091
#> [391]  0.677378420  1.911599699  0.709111774  1.978480676  0.270837434
#> [396]  0.469534412  0.194445398  2.068275711  0.087964436  0.886379005
#> [401] -0.086492905  1.848601467  0.702333000  1.978910068  0.606407484
#> [406]  1.996203434  1.247161281  1.429540833  0.054937452  1.960383049
#> [411]  1.615971265  1.763802079  0.857950203  1.994150784  0.111182830
#> [416]  1.784465987  0.486221968  1.944278004  0.117030951  0.757157402
#> [421]  0.662934550  0.525317444  0.109089747  1.997327060  0.106920188
#> [426]  1.805312793  0.047349889  1.836473491  0.716025466  2.015617001
#> [431]  0.596636144  1.656545089  0.044316230  1.567409642  0.703719721
#> [436]  2.084450435  0.374958774  1.385287873  0.013633874  1.523429369
#> [441]  1.462533859  0.560000577  0.029276048  1.845439312 -0.747819030
#> [446]  2.044724893  0.159650668  1.590411348  0.580430875  2.245235029
#> [451]  0.029900366  1.259488711  0.576085933  0.477537173  0.233546435
#> [456]  1.814996370  0.368712803  1.745536373  0.687692689  1.494360175
#> [461]  0.361338672  1.987120916  0.594134282  1.811384792  0.293053248
#> [466]  2.177069189  0.159920641  1.587076450  0.364077935  1.991301339
#> [471]  0.245879221  1.739258770  0.058823571  2.045565491  0.123603802
#> [476]  1.845592586  1.142123630  0.430025744  1.280844899  1.913467345
#> [481]  0.162610262  1.931790947  0.089143538  1.715792338  0.052272717
#> [486]  1.894285698  0.159431278  1.593268970  1.322931410  2.055430519
#> [491]  0.068987532  1.973733390  0.523307941  1.952248420  0.629153482
#> [496]  1.324165764  0.250156700  1.271589387  0.148114108  1.398813121

r2 = computemixdist(data, pi0 = 0.5, method = "nptll") # equivalent to normal
posteriormean(data, r2, fun = function(x) x^2)

#>   [1] 0.1915808 3.5690056 3.2235192 3.6038522 0.1516781 3.9121603 0.1363142
#>   [8] 3.2488001 1.5474848 2.1602389 2.6155906 4.1572686 0.3841298 4.0033509
#>  [15] 2.4591640 1.1689707 1.1944562 1.1978088 0.2325627 4.3457520 1.3059500
#>  [22] 4.4804291 1.1770401 4.1039178 0.6305219 1.1734627 0.7266287 0.3229151
#>  [29] 0.2691676 2.5448887 0.9486979 3.4720188 1.7452786 2.4813625 0.5408209
#>  [36] 1.3890622 1.2111019 1.4652117 1.3565329 5.1762933 0.4398205 3.9110442
#>  [43] 2.2165217 3.8586778 2.4098632 1.9334754 0.1929363 3.8925521 0.3761309
#>  [50] 2.8559777 0.6249202 1.7198520 0.1067111 4.1705302 2.1107559 3.1956929
#>  [57] 2.3609086 3.9758746 0.7005621 0.6745789 2.1101648 4.4144567 0.6468021
#>  [64] 3.3316776 0.5839700 3.5430026 0.1126573 2.2314557 2.2583265 3.3480098
#>  [71] 0.2884540 3.9626687 2.3091748 2.1838214 0.5192350 2.5884447 0.7231868
#>  [78] 4.3488734 2.2845390 3.6858343 0.3867450 2.6612286 0.2390024 1.7791594
#>  [85] 0.6394573 1.0872523 0.9456025 3.8561928 0.1396773 3.6711551 1.1791997
#>  [92] 1.5675594 0.2296886 3.9702721 2.2324529 4.6220441 0.9434620 0.9727013
#>  [99] 0.1081427 2.8147421 1.5121956 3.7365651 0.3475066 1.9705197 1.3551659
#> [106] 4.0725342 1.8086906 1.2336864 1.6322373 1.3642123 0.8986381 4.2038443
#> [113] 0.1662854 4.0795231 0.2419024 3.7304129 3.6853085 3.3558837 0.1430628
#> [120] 2.9007771 0.2427249 4.0147058 0.1058601 2.5336451 1.5015472 3.5726036
#> [127] 3.8448001 3.6760581 0.1611434 1.8959705 2.3387470 2.2578954 0.5983293
#> [134] 2.5200294 2.1928494 3.9725219 0.5497009 4.3343129 0.1305033 3.2918203
#> [141] 0.1080289 3.6652400 1.6160048 3.7653752 0.1291931 3.0158312 0.1928910
#> [148] 2.8484730 0.3352177 3.9959345 0.1122148 2.8314143 2.6215715 4.0383287
#> [155] 1.5358806 4.2664302 0.7257598 4.0975560 0.1981033 1.5921731 0.1182378
#> [162] 2.2532749 0.1328178 4.1452375 0.1587701 0.6773499 0.1086927 1.5448722
#> [169] 0.2188780 3.8155103 0.6767213 0.6699225 2.9324649 2.4406368 0.2106945
#> [176] 2.8036807 0.6913635 3.7317162 0.3375467 3.0151424 1.3175147 3.7308582
#> [183] 0.1683904 3.0415721 0.1364846 5.4008896 0.1097835 2.3873937 2.9752698
#> [190] 3.7478934 0.1088059 3.8576724 0.1710460 3.1431324 1.6760840 1.8312698
#> [197] 0.1252834 3.5021302 0.1072392 3.4133003 0.1304714 2.2776558 3.3949212
#> [204] 3.9039175 0.3659519 1.5551121 1.4741395 1.7253242 0.2051172 1.7252387
#> [211] 0.1692466 4.0406910 0.1726700 4.3417837 0.6794288 3.4338836 0.1080651
#> [218] 3.7469890 0.4033584 3.9502870 0.1165234 4.4305224 0.8189579 1.9600474
#> [225] 0.5164390 3.5433367 0.7797946 1.1310934 1.5727145 1.9587395 0.2034063
#> [232] 3.8776561 0.2362517 3.2376473 0.1304990 2.1620665 1.1604577 4.1159249
#> [239] 0.1062947 2.3832305 0.1675750 0.7512631 0.3296862 3.6707349 0.4356099
#> [246] 3.0342439 1.2174536 5.2639585 0.4477478 0.6902952 0.7140417 4.3378572
#> [253] 0.1308347 3.3241191 2.2889627 4.0738267 0.5590662 3.4828160 1.1743307
#> [260] 3.1834157 2.9989827 2.6606622 0.1617560 4.1033621 0.8854158 4.2597988
#> [267] 0.3124753 3.2785688 0.2429924 2.9593635 1.9058742 1.1475787 0.5754856
#> [274] 4.1891652 1.3166869 0.6231462 0.2050337 1.8885797 0.1208979 2.0974670
#> [281] 1.3391999 1.7656602 1.4828998 3.4721267 0.1058210 3.6772178 2.1085422
#> [288] 3.7647029 1.7844185 3.7787288 0.5760223 0.3708019 1.8622379 0.9277713
#> [295] 0.1113012 3.8516496 0.8967490 3.7421963 3.0419139 4.1068658 0.1799460
#> [302] 1.6782950 2.8070573 3.1112402 0.4535105 4.2233399 0.2468279 2.8810985
#> [309] 0.1430758 4.0082426 0.1293643 0.3657657 0.1093715 3.3073746 0.1191585
#> [316] 3.7044801 0.7665323 0.8586939 0.3080272 4.0825175 0.8547127 3.5504868
#> [323] 0.1785334 3.5797530 2.5700923 4.2379684 0.2203011 4.0859413 1.6313993
#> [330] 4.3043861 1.5145746 1.1330624 1.4129598 0.4313265 2.0757354 4.2511215
#> [337] 0.6916299 0.5146564 1.9752466 4.0912064 2.4007800 3.2513987 0.4320508
#> [344] 4.2788305 1.1305469 1.6917438 1.2023771 3.8099735 1.3012770 3.8770855
#> [351] 1.0350199 4.0030878 0.2912159 4.3182550 0.1241052 3.8757158 0.2596647
#> [358] 3.5207767 0.4365819 2.7325911 0.1325373 0.5119753 0.3728245 0.6524081
#> [365] 0.3924515 1.8285834 1.8736608 4.0131147 0.1902572 4.1042606 3.4813314
#> [372] 3.9933864 0.9465421 2.4937186 0.1301189 3.3253973 1.4159741 2.0015404
#> [379] 0.1344618 2.7565958 0.1590704 3.2869522 0.1059847 0.9918555 2.6776339
#> [386] 3.9154133 0.8097758 0.8464436 0.9929720 4.1052384 1.2932265 3.8935607
#> [393] 1.3556745 4.0759226 0.5094511 0.8881423 0.3681582 4.3885919 0.1846427
#> [400] 1.7071955 0.2114426 3.7378031 1.3423221 4.0771856 1.1541110 4.1294383
#> [407] 2.4373092 2.8157676 0.1384100 4.0239845 3.2127847 3.5402420 1.6505179
#> [414] 4.1230848 0.2218931 3.5874917 0.9203892 3.9795996 0.2316447 1.4505033
#> [421] 1.2648519 0.9961422 0.2184336 4.1329348 0.2148661 3.6356813 0.1296743
#> [428] 3.7088735 1.3692996 4.1918782 1.1350186 3.3011881 0.1264623 3.1080662
#> [435] 1.3450530 4.4581464 0.7065105 2.7231979 0.1064195 3.0140967 2.8851428
#> [442] 1.0635731 0.1134565 3.7302336 1.5239714 4.2950885 0.3055200 3.1575315
#> [449] 1.1033886 5.2820548 0.1138865 2.4626558 1.0949152 0.9036000 0.4399911
#> [456] 3.6582664 0.6945913 3.4988526 1.3135076 2.9523834 0.6805329 4.1016453
#> [463] 1.1301326 3.6498272 0.5511619 4.9152073 0.3059997 3.1503452 0.6857534
#> [470] 4.1143363 0.4628757 3.4847021 0.1432273 4.2982655 0.2427415 3.7306001
#> [477] 2.2225144 0.8120210 2.5066399 3.8983582 0.3107851 3.9461587 0.1864651
#> [484] 3.4321132 0.1352353 3.8496566 0.3051302 3.1636934 2.5936030 4.3364321
#> [491] 0.1567386 4.0620591 0.9922418 4.0013735 1.1986130 2.5961594 0.4708342
#> [498] 2.4875661 0.2851353 2.7514347

data = runif(500, min = -0.5, max = 0.5)
r3 = computemixdist(data, method = "npnormcll", beta = 100)
posteriormean(data, r3)

#>   [1] -0.0893767502 -0.1178863349  0.3046499766  0.0300817996 -0.3709197364
#>   [6] -0.1104305172  0.1422837090 -0.0778056568  0.0981042380  0.1741407192
#>  [11]  0.3492156684 -0.0364470065 -0.3811433702  0.3318212696  0.1763266994
#>  [16]  0.3741525619  0.1230071016  0.1383133827 -0.1496095296 -0.0835761592
#>  [21]  0.0625908910 -0.1504904964 -0.1339937026  0.0186784143  0.3761774099
#>  [26]  0.1255100660  0.1186632158  0.0187378228 -0.0811151729  0.0398192889
#>  [31] -0.0473437739 -0.2657563713 -0.1808211151  0.3099297110  0.1114486258
#>  [36]  0.1931007623 -0.3414382380  0.2054058174  0.0898659392 -0.3829655264
#>  [41] -0.3828005745  0.1190493027 -0.1235125164 -0.2823502864  0.3649431401
#>  [46] -0.3471167388 -0.1379986620  0.3482600313  0.2384090807 -0.0910778482
#>  [51]  0.3769668369  0.1694343274  0.3581370667  0.1379854989  0.2211494026
#>  [56] -0.3598761164  0.2956396722  0.3057982268 -0.3781919098  0.3717380965
#>  [61] -0.3278475032  0.1628920973  0.3769403258 -0.3794966155 -0.1574239254
#>  [66] -0.0109960341  0.3671380752  0.3748255871  0.1559948961 -0.3832646913
#>  [71] -0.3762377964 -0.0890839803  0.2725955146  0.1342350481 -0.1560943010
#>  [76] -0.1351609269 -0.1428404051  0.3683969280 -0.1647747535  0.3767684859
#>  [81]  0.3566315169 -0.1503681837  0.2866207300 -0.1375032641  0.1654953659
#>  [86] -0.1231713366  0.2572914640  0.1407111509 -0.3828133567 -0.3832614924
#>  [91] -0.0364125505 -0.3218643536 -0.1723960718  0.1075960060 -0.2601490928
#>  [96] -0.3664197123 -0.2908332549  0.3466813122 -0.1360885349  0.3631244317
#> [101] -0.2854635400  0.0788051625 -0.3569339019 -0.0262183737 -0.3805061586
#> [106] -0.1137354197  0.3767182325 -0.3810517677  0.1522244136  0.3724498141
#> [111]  0.0513921481 -0.1524061821  0.3740922770 -0.1549532109 -0.3522946661
#> [116]  0.3769164268  0.0102265447  0.3757821451  0.3542295808 -0.1216720222
#> [121] -0.1428959782  0.3734238913 -0.3759544148 -0.3421666361 -0.1723992000
#> [126] -0.3677328348 -0.2461163257  0.3487969645  0.1485591889  0.1498601501
#> [131]  0.3337340375 -0.3832746475  0.0026997673 -0.2209394348 -0.1336944645
#> [136] -0.1496772273 -0.3834160815 -0.1212696596 -0.1146320943  0.2578626141
#> [141]  0.2429001305 -0.0404830021 -0.3393257243  0.2646741551  0.1091570958
#> [146] -0.1941470146  0.0993272908  0.2637217435 -0.3722056233 -0.1534828807
#> [151]  0.3751065072  0.3636472458  0.1718408856 -0.3807441337  0.3185490287
#> [156] -0.1663373046 -0.1492721083 -0.0357604808  0.3752147582 -0.3832453385
#> [161] -0.1720760057  0.3178499126  0.0915591469 -0.1533219625 -0.3761533049
#> [166] -0.1193321272 -0.2718845100 -0.1053583537 -0.0804695647 -0.2409603453
#> [171]  0.2723741313 -0.3720785760 -0.1498836426 -0.3818421110 -0.1935959438
#> [176] -0.0180388334 -0.0227794640  0.3406168517 -0.3825404479 -0.0671024762
#> [181] -0.3603963163  0.1823678278 -0.3812753809 -0.1022672406  0.3662421011
#> [186] -0.0197729784 -0.0798163879  0.1328298383  0.0232603579  0.1399037702
#> [191] -0.0754265574  0.1172658610 -0.3786303666 -0.1196050066 -0.2685916626
#> [196]  0.3768542614  0.2709725161  0.3066761152  0.0843407341  0.0969708194
#> [201]  0.3244072933  0.3743452707  0.1242831692  0.2904080726 -0.0717669423
#> [206] -0.2697394255 -0.1433641079 -0.3781895594  0.3767052184  0.1396909255
#> [211] -0.1272030798  0.1835126903 -0.1148490983 -0.0264620848  0.1239845841
#> [216]  0.3127390196  0.0975238128  0.3759278374 -0.3795477828 -0.3828932482
#> [221] -0.1040180201 -0.2491609610  0.1910307267  0.3708633804 -0.3738257455
#> [226] -0.0169415424  0.3692406572  0.3744508577 -0.3579649019 -0.1276086674
#> [231] -0.3828432561  0.1525759748 -0.3773281788 -0.3684754470  0.3700809421
#> [236]  0.1372348672  0.3561641300  0.1612543686  0.3769754629  0.0702418704
#> [241] -0.0759268097  0.0611944113 -0.1434057143  0.3422672504  0.3564545216
#> [246]  0.3494931671  0.3520610490 -0.3800384120 -0.3824283320 -0.3757174636
#> [251] -0.0946731413  0.3765556124 -0.0772364281 -0.1326592401 -0.3822765880
#> [256] -0.0899433078 -0.1189558843 -0.1318770319 -0.3257269045 -0.3805447311
#> [261]  0.1032132477  0.3711913284  0.3147275984  0.3770014524 -0.2382336318
#> [266]  0.1323830441  0.1045400745 -0.1103460948 -0.1230466402  0.1592581493
#> [271] -0.1199076272  0.3769852667  0.3761910668  0.1542386518 -0.1438448774
#> [276] -0.2724753999  0.0934186153 -0.1493313700 -0.1425486602  0.0230944351
#> [281]  0.1277779997 -0.1413927878  0.3771711656  0.2498345698 -0.3083034658
#> [286] -0.1352768536 -0.3527455739  0.2995023147 -0.1310091686 -0.1204424921
#> [291] -0.3017789243  0.3763618544 -0.1139190210  0.3677029957 -0.3802074795
#> [296]  0.2740855907  0.0345463887 -0.1292360360 -0.3829989622  0.0354230506
#> [301] -0.2884227459 -0.3602509777 -0.1146339277 -0.1328549230 -0.3686919663
#> [306] -0.0708428670  0.1162230094  0.3767092294 -0.3593133781 -0.3594965112
#> [311] -0.0655839347 -0.1370823978  0.1119862384  0.3416386321 -0.2989338856
#> [316] -0.3795874353 -0.2850264490  0.3769940700  0.3552924095 -0.1462117745
#> [321] -0.1587826603  0.0297656008  0.3605061649  0.0934670347 -0.3441686452
#> [326]  0.3727768583 -0.3787275559 -0.1600053874  0.2104251768  0.3769214021
#> [331]  0.0260214433 -0.3833032409  0.0395746118  0.3750080257 -0.3765411977
#> [336]  0.0182056165  0.3741171703  0.3351387930  0.0045887930  0.0495040173
#> [341]  0.0117911435 -0.0196424692 -0.2878701472 -0.1334598126 -0.3835504507
#> [346]  0.1637445441 -0.2709390881  0.3707792080 -0.1585475697  0.0924169125
#> [351]  0.1269642230 -0.3464910847 -0.1647682408  0.1525665479  0.3693401810
#> [356]  0.3570918324 -0.0500389446  0.3731967384  0.1350249740  0.1457381788
#> [361]  0.1650940870 -0.1674446832  0.1118042586  0.1941894637  0.1001745683
#> [366] -0.1295566214  0.3618272817 -0.0939273901  0.1169619606 -0.3789077667
#> [371] -0.1465702481 -0.3823461145  0.3771205038 -0.1342079605  0.3584655166
#> [376]  0.3752303761 -0.1093542103  0.0389292267  0.1392952639 -0.1979892988
#> [381]  0.3717494070  0.3739768476 -0.1174961063 -0.1505746904  0.3657852390
#> [386] -0.0389423524  0.2648599196 -0.1571117003 -0.2593072643 -0.0417803829
#> [391] -0.3820976087 -0.3001170798  0.3623624124  0.1568748162  0.3772386727
#> [396]  0.3334402936  0.3753026743  0.2414371555  0.0757201931  0.3680687109
#> [401]  0.1377105342 -0.1080105143  0.3772532674  0.2326428943  0.3766126644
#> [406] -0.3568541660  0.1495155370  0.3603941226  0.0884333910  0.3626925521
#> [411]  0.3643080403 -0.3757090717  0.3719465921  0.0133086507  0.0502827743
#> [416] -0.1511924611 -0.3728900526 -0.0415821466 -0.1191647123 -0.3669046301
#> [421] -0.3836500016  0.3537430678 -0.1419918177 -0.1068160902 -0.1859530171
#> [426] -0.1059220548  0.0004497189 -0.3823156819 -0.3829012143  0.2289425067
#> [431] -0.3820381720 -0.2765326624  0.3463640962 -0.1099082322 -0.1345739476
#> [436]  0.3758586662  0.1957616015 -0.3759448188  0.3114184465 -0.1309870914
#> [441]  0.3768308027 -0.2468223632 -0.1878291731 -0.2153213011  0.0630543625
#> [446] -0.1242025209  0.3772594919  0.3769271831  0.1500782912 -0.3777196185
#> [451]  0.3759629198 -0.2261555453  0.3610625966  0.2603763219 -0.2795955627
#> [456]  0.1030257839  0.1776625124  0.1196356838 -0.3808341600 -0.3401637279
#> [461]  0.3763549538  0.1214031697  0.1026594308  0.0587629852 -0.1889378265
#> [466] -0.1430670498 -0.1444437678 -0.1028582097  0.3766983242 -0.3821901898
#> [471]  0.1225465553 -0.1341406647  0.3770311953  0.3769356933 -0.2681917041
#> [476] -0.1694792332  0.3574215045 -0.1300590029 -0.3619094650  0.3714364006
#> [481] -0.1240074057 -0.1929240520  0.3737046270  0.1164828618  0.2464490131
#> [486] -0.3800316272  0.1376456736  0.1332802861  0.3743532958 -0.1676162756
#> [491]  0.3762202462 -0.1926526514  0.1229072422  0.2125909001  0.0421864352
#> [496] -0.3834111398 -0.1167065975  0.3198514943  0.0496726034  0.0097915917