Find the posterior mean

Find the posterior mean given the observations and the mixing distribution based on the family in the result

posteriormean(x, result, fun = function(x) x)

posteriormean.npnorm(x, result, fun = function(x) x)

posteriormean.npt(x, result, fun = function(x) x)

posteriormean.npnormc(x, result, fun = function(x) x)

Arguments

x
a vector of observations

result
an object of class nspmix

fun
the function to transform the mean. It finds the posterior mean of fun(x). The function fun must be vectorised.

Examples

data = rnorm(500, c(0, 2))
r1 = computemixdist(data, pi0 = 0.5)
posteriormean(data, r1)

#>   [1] 0.52334114 1.62947189 1.62598239 2.08145705 0.07356137 2.06713115
#>   [7] 1.15289892 1.24380358 1.44558215 2.08981688 0.55483761 2.01915716
#>  [13] 0.05360459 2.10845214 0.09061258 1.50533571 0.47154803 1.80509202
#>  [19] 1.54052908 1.15659444 0.09785373 1.87264563 0.50809695 2.46058101
#>  [25] 0.03953543 2.04465673 0.78188255 2.04328280 0.04076686 1.90729464
#>  [31] 0.10376423 0.31745909 0.06099066 0.84556592 1.50391493 1.18386448
#>  [37] 0.78981625 1.97067361 0.06960984 1.95541427 0.50290346 0.74265246
#>  [43] 0.63467220 1.74946149 0.15071460 1.00387285 0.16631552 0.80615025
#>  [49] 0.20918777 1.67311314 0.14939769 2.04435990 0.08753974 1.67339101
#>  [55] 0.75779646 2.06256976 0.18495976 1.78663325 0.98811747 1.89280471
#>  [61] 0.36125975 1.58426443 0.25319813 1.75639658 0.51713641 1.66490097
#>  [67] 0.13625361 1.70929149 0.10927319 2.08030087 0.05083247 2.05348991
#>  [73] 0.21359461 0.21029560 0.09689262 2.04838404 0.03981283 1.81872405
#>  [79] 0.04249195 1.79444915 0.91229069 1.62405965 0.84652097 1.99432955
#>  [85] 1.10021522 2.06351028 0.88914134 1.91555229 0.23894701 1.82104288
#>  [91] 0.37881604 0.98851178 0.09053432 1.38587334 0.09971970 1.56802688
#>  [97] 0.19929697 1.09590479 0.01899066 2.12695462 0.06773766 2.03487683
#> [103] 0.08771247 1.94595397 0.17598500 1.96970204 0.02473635 1.86292305
#> [109] 0.17268040 2.06782991 0.13149084 1.48245278 0.02492147 1.83279927
#> [115] 0.54925064 1.60303068 0.09190149 1.90616120 0.04389693 1.89105660
#> [121] 0.90101452 1.65219491 0.12099326 1.51467222 0.45429827 0.63630000
#> [127] 0.06559833 0.46828875 0.14464707 1.73122028 0.24789061 1.64982948
#> [133] 0.02510268 0.63285057 0.13854532 1.87727091 0.29004032 0.84096541
#> [139] 0.91025447 1.77653892 0.12808056 1.87464777 0.74755447 1.96984774
#> [145] 0.11965732 0.71662836 0.75694739 0.84443488 0.08002414 2.04739621
#> [151] 0.09930614 1.58743640 0.44584821 2.06370911 0.05538304 1.99676622
#> [157] 0.20769484 2.07045668 0.85002887 1.72623529 0.07323430 0.61514372
#> [163] 0.05376770 1.43583425 0.04339047 1.66613257 0.09812789 2.12173591
#> [169] 0.98873713 1.94811340 0.09508673 2.12686306 0.14213805 1.73176025
#> [175] 0.02331212 2.04247111 0.19952069 0.87341593 0.21449023 2.04072477
#> [181] 0.24620005 1.88713256 0.94840890 0.97187247 0.14850794 1.98741413
#> [187] 0.03609380 2.05213528 0.04982600 1.87467074 0.28532280 1.29551127
#> [193] 0.30297104 1.63315743 0.78600275 0.57715780 1.18930073 1.58963222
#> [199] 0.07446291 0.82718304 0.96868474 0.43829564 0.26854578 1.29370731
#> [205] 0.07195779 1.65465241 0.23765693 1.49834812 0.07390080 1.99434592
#> [211] 0.58221123 2.08840034 0.05508991 0.88070876 0.06784487 0.74741208
#> [217] 1.74366875 0.10178030 0.03553867 1.13187767 0.15066849 1.66317105
#> [223] 0.90342304 2.05928998 0.05188025 1.85799961 0.02800980 2.01907670
#> [229] 0.06308301 1.93136599 0.05469741 1.76159633 0.21459071 0.81772664
#> [235] 0.38819232 2.78763084 0.22397798 1.45847737 0.10097955 1.50538269
#> [241] 0.28161536 1.01764842 1.40371012 1.96991823 0.81406075 2.06754225
#> [247] 0.04920151 0.95171263 0.35946261 1.40232941 0.47126812 1.71243320
#> [253] 0.61864335 0.39600069 0.61600453 0.27546406 0.21101689 1.93835173
#> [259] 0.06959755 2.07661047 0.30379303 0.39441426 1.43951407 1.88963264
#> [265] 0.14266958 1.99364199 0.04154425 1.75336478 0.33553224 0.05961604
#> [271] 0.11723904 1.20097969 0.37440592 2.04431040 0.27258775 1.08841672
#> [277] 0.19289515 0.87863407 0.05385639 0.10265975 0.14593774 1.96675821
#> [283] 0.32960960 0.84109491 0.10785043 2.04600090 0.64671757 1.95572979
#> [289] 0.13347170 1.72255793 0.04788545 1.77320851 0.40294015 1.78136886
#> [295] 0.30217670 0.11816293 0.33723468 1.53230671 0.04107483 1.55267415
#> [301] 0.29092727 0.93312044 0.22431218 0.71864891 0.04961759 2.10331564
#> [307] 0.67670877 0.41270209 0.15111743 1.81810652 0.13619770 2.04253528
#> [313] 1.87097360 1.31016601 0.74013582 2.04480657 0.40388749 1.93831151
#> [319] 1.11720156 1.81341897 0.02253117 1.93203334 0.04001587 1.27823622
#> [325] 0.15069465 0.17998107 0.16393277 2.01848621 0.08710400 0.47287205
#> [331] 0.13731726 1.31546207 0.34997965 1.26910868 0.07848341 1.90531636
#> [337] 0.46091080 2.04735807 0.95404205 1.23575730 0.79704661 0.59354968
#> [343] 0.27784043 0.22136225 0.40301757 0.81709254 0.19377628 1.87662251
#> [349] 0.03281806 2.01991056 0.98543760 1.87783820 0.04218148 0.47411454
#> [355] 0.03271230 1.43225701 0.18425377 1.90384689 0.69581506 1.36708933
#> [361] 0.99009943 1.49773274 0.03279236 1.99161526 0.50333815 1.81203805
#> [367] 0.45780482 2.07115493 0.13172503 2.03353988 0.09609415 2.06763295
#> [373] 0.16846260 1.60433161 0.55027061 1.97096242 1.32925620 1.93821388
#> [379] 0.89648780 1.98624460 0.10610443 2.05816693 0.05641394 2.03480024
#> [385] 0.06150957 2.02998922 0.18278094 1.77121740 1.71315231 1.13098369
#> [391] 0.93601406 1.21239563 0.22556448 1.67916853 1.71908354 2.05857626
#> [397] 0.24382355 1.81952234 0.24622065 0.55137924 0.02809845 1.63668619
#> [403] 0.15865256 1.16503604 0.06172109 1.72258171 1.61627240 1.51315677
#> [409] 1.01672112 1.59439317 0.34926051 2.15173223 0.34161514 1.24282782
#> [415] 0.02759829 1.37206670 0.88662612 1.41533112 0.71340293 1.59878961
#> [421] 0.03584842 0.25479918 0.05375008 1.75649126 0.10362198 1.10355855
#> [427] 0.24925139 0.79856804 0.85778650 1.29206432 0.16629353 0.98532760
#> [433] 0.49661122 1.91072321 0.11152038 0.22756250 0.27780416 1.92599432
#> [439] 0.45764278 2.03783506 0.06745763 1.15305028 0.04035883 2.01442457
#> [445] 0.10921708 1.09291307 0.20217893 1.63403482 0.08564869 2.12026534
#> [451] 0.07458504 2.04701834 0.53494598 2.10353614 0.25715696 1.75736460
#> [457] 0.08788554 0.69802390 0.13737596 1.29663470 0.10086576 1.91011230
#> [463] 0.11116533 0.03547352 0.15776576 1.84301640 0.19471222 1.84830443
#> [469] 0.11066297 2.02682300 0.13002734 1.42126660 0.81533665 2.05173639
#> [475] 0.11575941 1.67706414 0.07757658 1.81751951 1.58689796 2.01701449
#> [481] 0.56997144 1.93731502 0.15714693 0.36246001 0.65500999 1.82429223
#> [487] 0.10047334 2.06178632 0.74094479 2.05325790 0.65363479 1.12904446
#> [493] 0.14725395 0.24921821 0.15962915 0.30562341 0.25809399 1.82848608
#> [499] 0.08763675 1.73952136

r2 = computemixdist(data, pi0 = 0.5, method = "nptll") # equivalent to normal
posteriormean(data, r2, fun = function(x) x^2)

#>   [1] 1.01265268 3.35413832 3.34665157 4.41131056 0.09973356 4.35145673
#>   [7] 2.33789920 2.53090929 2.96065738 4.45154802 1.07833029 4.21020082
#>  [13] 0.06389839 4.55208332 0.13162723 3.08830033 0.90487056 3.73243732
#>  [19] 3.16357030 2.34573850 0.14540570 3.87924946 0.98089977 6.79441313
#>  [25] 0.04038537 4.27782488 1.55410096 4.27388119 0.04235400 3.95514045
#>  [31] 0.15673393 0.58633517 0.07691027 1.68814591 3.08526312 2.40360495
#>  [37] 1.57078783 4.09625771 0.09248275 4.06184976 0.97008745 1.47164367
#>  [43] 1.24519451 3.61221808 0.24855256 2.02227786 0.27956812 1.60515435
#>  [49] 0.36563010 3.44785091 0.24594364 4.27696888 0.12581770 3.44844810
#>  [55] 1.50346368 4.33478257 0.31686307 3.69249433 1.98896884 3.92333598
#>  [61] 0.67648353 3.25721233 0.45488797 3.62718062 0.99972568 3.43020474
#>  [67] 0.21999320 3.52566294 0.16735073 4.40604296 0.05911439 4.30445514
#>  [73] 0.37453182 0.36786706 0.14357024 4.28877412 0.04082697 3.76197579
#>  [79] 0.04514603 3.70940004 1.82882682 3.34252672 1.69015787 4.15062529
#>  [85] 2.22620694 4.33813462 1.77999323 3.97331861 0.42590376 3.76700404
#>  [91] 0.71271752 1.98980233 0.13147898 2.83329119 0.14897464 3.22243209
#>  [97] 0.34568461 2.21707404 0.01244955 4.66081179 0.08907015 4.25065503
#> [103] 0.12614363 4.04069175 0.29888220 4.09405460 0.01899572 3.85804202
#> [109] 0.29227435 4.35410688 0.21063377 3.03939646 0.01922790 3.79251517
#> [115] 1.06667332 3.29743047 0.13407100 3.95264853 0.04744733 3.91950610
#> [121] 1.80503633 3.40291328 0.19009910 3.10826198 0.86904043 1.24860214
#> [127] 0.08519000 0.89809789 0.23654525 3.57289074 0.44408520 3.39783400
#> [133] 0.01945631 1.24138138 0.22450540 3.88935017 0.53011495 1.67845493
#> [139] 1.82453033 3.67067486 0.20394796 3.88362075 1.48194205 4.09438486
#> [145] 0.18749599 1.41699619 1.50167927 1.68576328 0.11171691 4.28583461
#> [151] 0.14818303 3.26400855 0.85150174 4.33884883 0.06699829 4.15632731
#> [157] 0.36261647 4.36430827 1.69754821 3.56215003 0.09913108 1.20432961
#> [163] 0.06418175 2.93985191 0.04661507 3.43285081 0.14592961 4.62956543
#> [169] 1.99027866 4.04551145 0.14012681 4.66026057 0.23159007 3.57405434
#> [175] 0.01725035 4.27157289 0.34613522 1.74683608 0.37634204 4.26665798
#> [181] 0.44064630 3.91091411 1.90507170 1.95463695 0.24418184 4.13456291
#> [187] 0.03500430 4.30021023 0.05739306 3.88367091 0.52046087 2.64085849
#> [193] 0.55660595 3.36204664 1.56276646 1.12492843 2.41514447 3.26871369
#> [199] 0.10139636 1.64942923 1.94790151 0.83583350 0.48617733 2.63702061
#> [205] 0.09678358 3.40819073 0.42328355 3.07336391 0.10035925 4.15066353
#> [211] 1.13548458 4.44447716 0.06648580 1.76221147 0.08926515 1.48164288
#> [217] 3.59972481 0.15292389 0.03415472 2.29331831 0.24846118 3.42648824
#> [223] 1.81011729 4.32342060 0.06091542 3.84731461 0.02326531 4.21000090
#> [229] 0.08065633 4.00826268 0.06580052 3.63840328 0.37654514 1.62952006
#> [235] 0.73209008 8.90956965 0.39554004 2.98818777 0.15138823 3.08840078
#> [241] 0.51287801 2.05141080 2.87132102 4.09454467 1.62180337 4.35301269
#> [247] 0.05632945 1.91204900 0.67277750 2.86837665 0.90428888 3.53242608
#> [253] 1.21165069 0.74823378 1.20613030 0.50030507 0.36932380 4.02376545
#> [259] 0.09246030 4.38974085 0.55829134 0.74495309 2.94770538 3.91638743
#> [265] 0.23263930 4.14902058 0.04360743 3.62063871 0.62348598 0.07446258
#> [271] 0.18279021 2.43993975 0.70361064 4.27682634 0.49442971 2.20121036
#> [277] 0.33280136 1.75783715 0.06433589 0.15461191 0.23909660 4.08739042
#> [283] 0.61130393 1.67872771 0.16460376 4.28172972 1.27041487 4.06255744
#> [289] 0.21452334 3.55422852 0.05409963 3.66347938 0.76258889 3.68111308
#> [295] 0.55497739 0.18458704 0.62698897 3.14597829 0.04284960 3.18956237
#> [301] 0.53193068 1.87279020 0.39621697 1.42123756 0.05703772 4.52319649
#> [307] 1.33325577 0.78279477 0.24935090 3.76063690 0.21988319 4.27175481
#> [313] 3.87560000 2.67204148 1.46635715 4.27825783 0.76454915 4.02367605
#> [319] 2.26220546 3.75047616 0.01632549 4.00974175 0.04115091 2.60411228
#> [325] 0.24851303 0.30688214 0.27481885 4.20853555 0.12499586 0.90762220
#> [331] 0.22208677 2.68331306 0.65323159 2.58470221 0.10884718 3.95079160
#> [337] 0.88277136 4.28572169 1.91696894 2.51381065 1.58599852 1.15917742
#> [343] 0.50516099 0.39024339 0.76274909 1.62818525 0.33457327 3.88793369
#> [349] 0.03007474 4.21207558 1.98330438 3.89058956 0.04464073 0.91020456
#> [355] 0.02991915 2.93221799 0.31544686 3.94756271 1.37332223 2.79325815
#> [361] 1.99315834 3.07204861 0.03003692 4.14430072 0.97099231 3.74748373
#> [367] 0.87632112 4.36708478 0.21109338 4.24708351 0.14204687 4.35335722
#> [373] 0.28385118 3.30021937 1.06880123 4.09691297 2.71267697 4.02345907
#> [379] 1.79548765 4.13186231 0.16123744 4.31964120 0.06880537 4.25044961
#> [385] 0.07783708 4.23773378 0.31249329 3.65917829 3.53397425 2.29142283
#> [391] 1.87889918 2.46418240 0.39875397 3.46086637 3.54674561 4.32101242
#> [397] 0.43581377 3.76370670 0.44068820 1.07111422 0.02338546 3.36961952
#> [403] 0.26430959 2.36364791 0.07821529 3.55427976 3.32582342 3.10502158
#> [409] 2.04944944 3.27891640 0.65175001 4.81286147 0.63600502 2.52883561
#> [415] 0.02271047 2.80386445 1.77468902 2.89610642 1.41022618 3.28833931
#> [421] 0.03462809 0.45814854 0.06415113 3.62738495 0.15646049 2.23329132
#> [427] 0.44685399 1.58919958 1.71389409 2.63352533 0.27952427 1.98307188
#> [433] 0.95699144 3.96268288 0.17169623 0.40280299 0.50508687 3.99637113
#> [439] 0.87598464 4.25866810 0.08856103 2.33822027 0.04169937 4.19853648
#> [445] 0.16724232 2.21073571 0.35149129 3.36392947 0.12225466 4.62083046
#> [451] 0.10162185 4.28471760 1.03684051 4.52442115 0.46295173 3.62926964
#> [457] 0.12647028 1.37795584 0.22220233 2.64324863 0.15117011 3.96133844
#> [463] 0.17100911 0.03405537 0.26254671 3.81471250 0.33645588 3.82621161
#> [469] 0.17003725 4.22954526 0.20776293 2.90876822 1.62448905 4.29897239
#> [475] 0.17991508 3.45634283 0.10716177 3.75936426 3.26285485 4.20489665
#> [481] 1.10992059 4.02146185 0.26131690 0.67895903 1.28778361 3.77405199
#> [487] 0.15041807 4.33202304 1.46805646 4.30372349 1.28490287 2.28731126
#> [493] 0.24170001 0.44678647 0.26625168 0.56204491 0.46486112 3.78315202
#> [499] 0.12600075 3.59078262

data = runif(500, min = -0.5, max = 0.5)
r3 = computemixdist(data, method = "npnormcll", beta = 100)
posteriormean(data, r3)

#>   [1] -0.3613611488 -0.3604237293  0.1777899544 -0.2034152214  0.3683695348
#>   [6]  0.3032663691  0.2420179133 -0.0758284626  0.3671256539  0.3692651172
#>  [11] -0.0717481645 -0.0729118806 -0.0775773249  0.2961812036  0.3485953590
#>  [16]  0.3582558759  0.2182582483 -0.3617064948  0.2935154173  0.3252151613
#>  [21] -0.0711287186 -0.3554980209 -0.0633581892 -0.3374863707  0.0960933122
#>  [26]  0.0661425444 -0.0426370068  0.3402833335  0.2911383395 -0.3614911097
#>  [31] -0.0876203988  0.3505755646 -0.3460923503 -0.3606537968  0.3634017321
#>  [36] -0.3583212318 -0.2975420535  0.3445711754  0.1255653762 -0.1739611292
#>  [41] -0.3615615326  0.2901159409 -0.2385816876  0.3304391098  0.3698884332
#>  [46]  0.3465823663 -0.0221391912 -0.0821315715  0.2972591263  0.1676119324
#>  [51]  0.2275953984 -0.0538285302 -0.3609155766  0.2925932562  0.3071313216
#>  [56] -0.0718505458 -0.3542196108  0.3685539349  0.2707854889  0.3660018374
#>  [61] -0.3593538329  0.3500777657 -0.0609727373  0.3543707291  0.2244432609
#>  [66]  0.3696889276 -0.0818940752  0.1562442545 -0.0716379134 -0.1669545910
#>  [71]  0.2348729355 -0.3606624971  0.2727237622  0.2598794400  0.3641932509
#>  [76] -0.0728213532  0.3702042817 -0.0025996742 -0.0697351382  0.1724883703
#>  [81]  0.3589578406 -0.3456685628  0.2654355350  0.2121116922 -0.0323674701
#>  [86] -0.0736516356 -0.0640466327  0.3344858480 -0.3527824388  0.1028753406
#>  [91]  0.3312506339 -0.0804201121  0.3374711587 -0.0357816030 -0.1348579861
#>  [96]  0.3527135849 -0.3592945309 -0.1251299447 -0.0469418733  0.2417532101
#> [101]  0.3459987883 -0.2361304357  0.3672030225  0.3698475728 -0.1710893532
#> [106]  0.3498918170 -0.0702279561 -0.0772714590  0.3669401437  0.3388622887
#> [111] -0.0949149404 -0.3372330561 -0.3536012285  0.2537659199  0.1632247456
#> [116] -0.3616237960  0.3698278315 -0.3555076822 -0.3041534616  0.2095027136
#> [121] -0.1197411784  0.3685727602 -0.0691007351  0.3388309162 -0.0368853390
#> [126]  0.1254372974 -0.0916437209  0.3085156952  0.2112213515 -0.3453087242
#> [131] -0.2477131719  0.2390579645 -0.0710945465 -0.3617062775  0.2610336202
#> [136]  0.2204296949  0.0018537455  0.0009598047  0.2646827869 -0.0607210331
#> [141] -0.3608018708 -0.1023494063 -0.3304107493 -0.0774874129 -0.0721416245
#> [146]  0.2228229002 -0.3612969204 -0.3615443852  0.0463311709 -0.0778211266
#> [151]  0.3197254194  0.2866430312 -0.0709012173 -0.3597895508 -0.3613894259
#> [156] -0.3273322275 -0.0762144122  0.0885342003 -0.3602954858 -0.3350807522
#> [161] -0.0771348601 -0.1135339271 -0.3608436937 -0.0693191469 -0.3616264155
#> [166]  0.0234233194  0.3054526206 -0.3586444166  0.1038384848 -0.0562886785
#> [171] -0.3587124315 -0.0537864535 -0.3608874206 -0.0750181293 -0.3596356051
#> [176]  0.3630072262 -0.1570162457 -0.0713272585  0.3524840422 -0.0688644908
#> [181]  0.1779471283  0.3156006033 -0.3498686804  0.2075655573  0.3292589502
#> [186]  0.3684938491 -0.3608045840 -0.0719133860  0.3665603381 -0.3604557860
#> [191] -0.0763421870 -0.3610639776  0.3622626346  0.1659640501 -0.3539605178
#> [196]  0.3669139346 -0.0607891829 -0.3568175298 -0.1398436197 -0.0756590004
#> [201]  0.1589657773 -0.0745101522 -0.3509338792  0.3534494682 -0.3595947524
#> [206] -0.3550373868 -0.3491787879  0.3686608567  0.3674252184 -0.3560677146
#> [211]  0.0879275809 -0.2221814759  0.3364050423 -0.0976540438 -0.3084513385
#> [216]  0.3274096638 -0.0696399655 -0.0768170887 -0.0681614126 -0.3615377314
#> [221] -0.0777502182  0.3510246946 -0.0713175923 -0.0501257907  0.1308805447
#> [226]  0.2429113486 -0.3195883188  0.3616784075  0.2857864812 -0.0777385562
#> [231]  0.3676625701 -0.3600557203 -0.0649856795 -0.3615790313  0.2218833385
#> [236] -0.0792057230  0.0629959919 -0.3585883328 -0.0618024157  0.1671210543
#> [241]  0.2025811916 -0.3570439640 -0.0646119105 -0.3607942356 -0.3386706185
#> [246]  0.3471111283 -0.0678888370 -0.0677470278  0.3534581994  0.3644568032
#> [251]  0.2341335561 -0.0703918819  0.1845239759 -0.3227688945 -0.0698105394
#> [256] -0.0349262052  0.2299767480  0.2292739474 -0.3616229817  0.0369649106
#> [261] -0.0659196193  0.1187744962 -0.3611450486 -0.3515435989 -0.0955447419
#> [266] -0.3497330629 -0.3608416858 -0.0683547600  0.3562336750 -0.3347862994
#> [271] -0.2699743057  0.2085606792 -0.1275176592 -0.0686262428  0.3545308665
#> [276] -0.0724352347  0.2964915211 -0.0698589748 -0.0838332498 -0.3253090772
#> [281] -0.3616203404  0.3371127995  0.3020999043 -0.0363462100 -0.0885736781
#> [286] -0.0786641727 -0.2804826146  0.3328208896 -0.1586694135  0.0442563113
#> [291] -0.3577731078  0.3631465256 -0.3609126203 -0.3616209025 -0.3537122476
#> [296] -0.3579364272  0.3427067772  0.2106479761 -0.3553352219 -0.1572572086
#> [301] -0.0816594031  0.2544232204  0.2230088344  0.1831162517  0.2019016272
#> [306] -0.2689594614 -0.3615981961  0.3674360336  0.3629936151  0.2995012751
#> [311]  0.0360455094 -0.3032149097  0.3325034312 -0.2941283806 -0.3479717381
#> [316]  0.1486376581 -0.1091191555 -0.3615892266 -0.0773424226 -0.0859526542
#> [321]  0.2548107210 -0.0970003064 -0.3595506199 -0.0777193222 -0.1226799108
#> [326]  0.3700432916  0.2478830329 -0.0911876902  0.3438295906  0.3132465197
#> [331] -0.3442242010 -0.3613441282  0.3617092706 -0.3613041226 -0.0709787930
#> [336]  0.3532727850 -0.0737991351 -0.0554786061  0.2603232566 -0.0790004294
#> [341]  0.3107445035 -0.2370519449 -0.3068204905 -0.0932129035  0.3082696527
#> [346] -0.3280910931 -0.3491104939 -0.1142455738  0.3024586699  0.2113975540
#> [351] -0.1526295348 -0.3239610202 -0.3177356068  0.3541045937  0.3400838828
#> [356] -0.1255900122  0.2090184221 -0.3615508871 -0.3595126404 -0.0702339331
#> [361]  0.3094856842  0.2162021696 -0.0772546402  0.3061405408  0.2337656094
#> [366] -0.3614335712  0.2053073854 -0.0905060113 -0.3509045625  0.3528168127
#> [371] -0.0698768461  0.3082672648 -0.0596033159  0.2677158234 -0.1040319681
#> [376] -0.1284980228 -0.3616129853 -0.2997916938 -0.1382051558 -0.0187246640
#> [381]  0.2143773096  0.2221300442  0.3274229916 -0.0852857621  0.3484069972
#> [386] -0.0709590251  0.2710179913  0.3217958206 -0.0741801620  0.2780884938
#> [391] -0.0842511253 -0.0734848037 -0.1895268951 -0.2793410994  0.3228342300
#> [396]  0.2934336642 -0.2073798201 -0.0549390850  0.0571382881 -0.0752877328
#> [401]  0.3209545053  0.2502076533  0.1684629696 -0.0679942734  0.0432097568
#> [406]  0.3676099271 -0.0995047142 -0.3600484736 -0.3587438366 -0.3606826504
#> [411] -0.0778170075 -0.3604197565 -0.3361817805  0.3679894305  0.0069835154
#> [416] -0.1336906155 -0.0465631797  0.1536175542 -0.0830795750 -0.3604938091
#> [421]  0.2365866015 -0.2589890306  0.3628238431  0.1001482422 -0.0382789592
#> [426]  0.2924022233  0.2270793884  0.3642534231 -0.3544439152 -0.0992392658
#> [431] -0.0688040609 -0.0075972217 -0.0467326857  0.3118491827  0.2056730084
#> [436] -0.3599847640  0.1801133735 -0.0722970116  0.0328548380 -0.0636569672
#> [441]  0.1829668477  0.3698152129 -0.3415230445 -0.0738717930  0.0466475920
#> [446]  0.0994195071 -0.3613735797  0.2870968101  0.3175575122 -0.3201173490
#> [451]  0.2631628653  0.1858395599 -0.0636124932 -0.0414465551  0.3666765297
#> [456] -0.3575106178 -0.1650342889 -0.0761658089  0.3564714173  0.2891431853
#> [461]  0.2676774614  0.2180228660  0.3694647772  0.3090016436  0.3563088967
#> [466] -0.0871105431  0.2835189611 -0.3583191422  0.2372806758  0.2775801539
#> [471]  0.2253560983 -0.3615442399  0.3080376687 -0.3461736022  0.2839985687
#> [476] -0.2479027442  0.3524173447  0.3684968920 -0.0714094936  0.3437633203
#> [481]  0.3674604789 -0.0744792742 -0.0599870834  0.3692279882 -0.3603361435
#> [486] -0.3605391140  0.3566090305 -0.0843314256 -0.3609256128 -0.3616916892
#> [491] -0.0443981237 -0.3492650171 -0.3345631234 -0.0497021412  0.3580699595
#> [496] -0.1139169695  0.2582195348 -0.0304745737 -0.2660958695 -0.3576368772